Struct gdnative::api::animation_node_state_machine::AnimationNodeStateMachine

pub struct AnimationNodeStateMachine { /* fields omitted */ }

core class AnimationNodeStateMachine inherits AnimationRootNode (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

AnimationNodeStateMachine inherits methods from:

AnimationRootNode

AnimationNode

Resource

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl AnimationNodeStateMachine

pub fn new() -> Ref<AnimationNodeStateMachine, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn add_node(
&self,
name: impl Into<GodotString>,
node: impl AsArg<AnimationNode>,
position: Vector2
)

Adds a new node to the graph. The position is used for display in the editor.

Default Arguments

position - Vector2( 0, 0 )

pub fn add_transition(
&self,
from: impl Into<GodotString>,
to: impl Into<GodotString>,
transition: impl AsArg<AnimationNodeStateMachineTransition>
)

Adds a transition between the given nodes.

pub fn get_end_node(&self) -> GodotString

Returns the graph’s end node.

pub fn get_graph_offset(&self) -> Vector2

Returns the draw offset of the graph. Used for display in the editor.

pub fn get_node(
&self,
name: impl Into<GodotString>
) -> Option<Ref<AnimationNode, Shared>>

Returns the animation node with the given name.

pub fn get_node_name(&self, node: impl AsArg<AnimationNode>) -> GodotString

Returns the given animation node’s name.

pub fn get_node_position(&self, name: impl Into<GodotString>) -> Vector2

Returns the given node’s coordinates. Used for display in the editor.

pub fn get_start_node(&self) -> GodotString

Returns the graph’s end node.

pub fn get_transition(
&self,
idx: i64
) -> Option<Ref<AnimationNodeStateMachineTransition, Shared>>

Returns the given transition.

pub fn get_transition_count(&self) -> i64

Returns the number of connections in the graph.

pub fn get_transition_from(&self, idx: i64) -> GodotString

Returns the given transition’s start node.

pub fn get_transition_to(&self, idx: i64) -> GodotString

Returns the given transition’s end node.

pub fn has_node(&self, name: impl Into<GodotString>) -> bool

Returns true if the graph contains the given node.

pub fn has_transition(
&self,
from: impl Into<GodotString>,
to: impl Into<GodotString>
) -> bool

Returns true if there is a transition between the given nodes.

pub fn remove_node(&self, name: impl Into<GodotString>)

Deletes the given node from the graph.

pub fn remove_transition(
&self,
from: impl Into<GodotString>,
to: impl Into<GodotString>
)

Deletes the transition between the two specified nodes.

pub fn remove_transition_by_index(&self, idx: i64)

Deletes the given transition by index.

pub fn rename_node(
&self,
name: impl Into<GodotString>,
new_name: impl Into<GodotString>
)

Renames the given node.

pub fn replace_node(
&self,
name: impl Into<GodotString>,
node: impl AsArg<AnimationNode>
)

Replaces the node and keeps its transitions unchanged.

pub fn set_end_node(&self, name: impl Into<GodotString>)

Sets the given node as the graph end point.

pub fn set_graph_offset(&self, offset: Vector2)

Sets the draw offset of the graph. Used for display in the editor.

pub fn set_node_position(&self, name: impl Into<GodotString>, position: Vector2)

Sets the node’s coordinates. Used for display in the editor.

pub fn set_start_node(&self, name: impl Into<GodotString>)

Sets the given node as the graph start point.