Struct gdnative::api::audio_effect_compressor::AudioEffectCompressor

pub struct AudioEffectCompressor { /* fields omitted */ }

core class AudioEffectCompressor inherits AudioEffect (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

AudioEffectCompressor inherits methods from:

AudioEffect

Resource

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl AudioEffectCompressor

pub fn new() -> Ref<AudioEffectCompressor, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn attack_us(&self) -> f64

Compressor’s reaction time when the signal exceeds the threshold, in microseconds. Value can range from 20 to 2000.

pub fn gain(&self) -> f64

Gain applied to the output signal.

pub fn mix(&self) -> f64

Balance between original signal and effect signal. Value can range from 0 (totally dry) to 1 (totally wet).

pub fn ratio(&self) -> f64

Amount of compression applied to the audio once it passes the threshold level. The higher the ratio, the more the loud parts of the audio will be compressed. Value can range from 1 to 48.

pub fn release_ms(&self) -> f64

Compressor’s delay time to stop reducing the signal after the signal level falls below the threshold, in milliseconds. Value can range from 20 to 2000.

pub fn sidechain(&self) -> GodotString

Reduce the sound level using another audio bus for threshold detection.

pub fn threshold(&self) -> f64

The level above which compression is applied to the audio. Value can range from -60 to 0.

pub fn set_attack_us(&self, attack_us: f64)

Compressor’s reaction time when the signal exceeds the threshold, in microseconds. Value can range from 20 to 2000.

pub fn set_gain(&self, gain: f64)

Gain applied to the output signal.

pub fn set_mix(&self, mix: f64)

Balance between original signal and effect signal. Value can range from 0 (totally dry) to 1 (totally wet).

pub fn set_ratio(&self, ratio: f64)

Amount of compression applied to the audio once it passes the threshold level. The higher the ratio, the more the loud parts of the audio will be compressed. Value can range from 1 to 48.

pub fn set_release_ms(&self, release_ms: f64)

Compressor’s delay time to stop reducing the signal after the signal level falls below the threshold, in milliseconds. Value can range from 20 to 2000.

pub fn set_sidechain(&self, sidechain: impl Into<GodotString>)

Reduce the sound level using another audio bus for threshold detection.

pub fn set_threshold(&self, threshold: f64)

The level above which compression is applied to the audio. Value can range from -60 to 0.