Struct gdnative::api::bitmap_font::BitmapFont

pub struct BitmapFont { /* fields omitted */ }

core class BitmapFont inherits Font (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

BitmapFont inherits methods from:

Font

Resource

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl BitmapFont

pub fn new() -> Ref<BitmapFont, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn add_char(
&self,
character: i64,
texture: i64,
rect: Rect2,
align: Vector2,
advance: f64
)

Adds a character to the font, where character is the Unicode value, texture is the texture index, rect is the region in the texture (in pixels!), align is the (optional) alignment for the character and advance is the (optional) advance.

Default Arguments

align - Vector2( 0, 0 )

advance - -1

pub fn add_kerning_pair(&self, char_a: i64, char_b: i64, kerning: i64)

Adds a kerning pair to the BitmapFont as a difference. Kerning pairs are special cases where a typeface advance is determined by the next character.

pub fn add_texture(&self, texture: impl AsArg<Texture>)

Adds a texture to the BitmapFont.

pub fn clear(&self)

Clears all the font data and settings.

pub fn create_from_fnt(
&self,
path: impl Into<GodotString>
) -> Result<(), GodotError>

Creates a BitmapFont from the *.fnt file at path.

pub fn fallback(&self) -> Option<Ref<BitmapFont, Shared>>

The fallback font.

pub fn get_kerning_pair(&self, char_a: i64, char_b: i64) -> i64

Returns a kerning pair as a difference.

pub fn get_texture(&self, idx: i64) -> Option<Ref<Texture, Shared>>

Returns the font atlas texture at index idx.

pub fn get_texture_count(&self) -> i64

Returns the number of textures in the BitmapFont atlas.

pub fn set_ascent(&self, px: f64)

Ascent (number of pixels above the baseline).

pub fn set_distance_field_hint(&self, enable: bool)

If true, distance field hint is enabled.

pub fn set_fallback(&self, fallback: impl AsArg<BitmapFont>)

The fallback font.

pub fn set_height(&self, px: f64)

Total font height (ascent plus descent) in pixels.