Struct gdnative::api::box_container::BoxContainer

pub struct BoxContainer { /* fields omitted */ }

core class BoxContainer inherits Container (unsafe).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Class hierarchy

BoxContainer inherits methods from:

Container

Control

CanvasItem

Node

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl BoxContainer

Constants

pub const ALIGN_BEGIN: i64

pub const ALIGN_CENTER: i64

pub const ALIGN_END: i64

impl BoxContainer

pub fn add_spacer(&self, begin: bool)

Adds a control to the box as a spacer. If true, begin will insert the spacer control in front of other children.

pub fn alignment(&self) -> AlignMode

The alignment of the container’s children (must be one of ALIGN_BEGIN, ALIGN_CENTER or ALIGN_END).

pub fn set_alignment(&self, alignment: i64)

The alignment of the container’s children (must be one of ALIGN_BEGIN, ALIGN_CENTER or ALIGN_END).