Struct gdnative::api::camera_2d::Camera2D

pub struct Camera2D { /* fields omitted */ }

core class Camera2D inherits Node2D (unsafe).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non reference counted objects such as the ones of this type are usually owned by the engine.

Camera2D is a reference-only type. Persistent references can only exist in the unsafe Ref<Camera2D> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

Camera2D inherits methods from:

Node2D

CanvasItem

Node

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Camera2D

Constants

pub const ANCHOR_MODE_FIXED_TOP_LEFT: i64

pub const CAMERA2D_PROCESS_PHYSICS: i64

pub const ANCHOR_MODE_DRAG_CENTER: i64

pub const CAMERA2D_PROCESS_IDLE: i64

impl Camera2D

pub fn new() -> Ref<Camera2D, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn align(&self)

Aligns the camera to the tracked node.

pub fn clear_current(&self)

Removes any Camera2D from the ancestor Viewport’s internal currently-assigned camera.

pub fn force_update_scroll(&self)

Forces the camera to update scroll immediately.

pub fn anchor_mode(&self) -> AnchorMode

The Camera2D’s anchor point. See AnchorMode constants.

pub fn get_camera_position(&self) -> Vector2

Returns the camera position.

pub fn get_camera_screen_center(&self) -> Vector2

Returns the location of the Camera2D’s screen-center, relative to the origin.

pub fn custom_viewport(&self) -> Option<Ref<Node, Shared>>

The custom Viewport node attached to the Camera2D. If null or not a Viewport, uses the default viewport instead.

pub fn drag_margin(&self, margin: i64) -> f64

Top margin needed to drag the camera. A value of 1 makes the camera move only when reaching the edge of the screen.

pub fn follow_smoothing(&self) -> f64

Speed in pixels per second of the camera’s smoothing effect when [smoothing_enabled][Self::smoothing_enabled] is true.

pub fn h_offset(&self) -> f64

The horizontal offset of the camera, relative to the drag margins. Note: Offset H is used only to force offset relative to margins. It’s not updated in any way if drag margins are enabled and can be used to set initial offset.

pub fn limit(&self, margin: i64) -> i64

Top scroll limit in pixels. The camera stops moving when reaching this value.

pub fn offset(&self) -> Vector2

The camera’s offset, useful for looking around or camera shake animations.

pub fn process_mode(&self) -> Camera2DProcessMode

The camera’s process callback. See Camera2DProcessMode.

pub fn v_offset(&self) -> f64

The vertical offset of the camera, relative to the drag margins. Note: Used the same as [offset_h][Self::offset_h].

pub fn zoom(&self) -> Vector2

The camera’s zoom relative to the viewport. Values larger than Vector2(1, 1) zoom out and smaller values zoom in. For an example, use Vector2(0.5, 0.5) for a 2× zoom-in, and Vector2(4, 4) for a 4× zoom-out.

pub fn is_current(&self) -> bool

If true, the camera is the active camera for the current scene. Only one camera can be current, so setting a different camera current will disable this one.

pub fn is_follow_smoothing_enabled(&self) -> bool

If true, the camera smoothly moves towards the target at [smoothing_speed][Self::smoothing_speed].

pub fn is_h_drag_enabled(&self) -> bool

If true, the camera only moves when reaching the horizontal drag margins. If false, the camera moves horizontally regardless of margins.

pub fn is_limit_drawing_enabled(&self) -> bool

If true, draws the camera’s limits rectangle in the editor.

pub fn is_limit_smoothing_enabled(&self) -> bool

If true, the camera smoothly stops when reaches its limits.

pub fn is_margin_drawing_enabled(&self) -> bool

If true, draws the camera’s drag margin rectangle in the editor.

pub fn is_rotating(&self) -> bool

If true, the camera rotates with the target.

pub fn is_screen_drawing_enabled(&self) -> bool

If true, draws the camera’s screen rectangle in the editor.

pub fn is_v_drag_enabled(&self) -> bool

If true, the camera only moves when reaching the vertical drag margins. If false, the camera moves vertically regardless of margins.

pub fn make_current(&self)

Make this the current 2D camera for the scene (viewport and layer), in case there are many cameras in the scene.

pub fn reset_smoothing(&self)

Sets the camera’s position immediately to its current smoothing destination. This has no effect if smoothing is disabled.

pub fn set_anchor_mode(&self, anchor_mode: i64)

The Camera2D’s anchor point. See AnchorMode constants.

pub fn set_custom_viewport(&self, viewport: impl AsArg<Node>)

The custom Viewport node attached to the Camera2D. If null or not a Viewport, uses the default viewport instead.

pub fn set_drag_margin(&self, margin: i64, drag_margin: f64)

Top margin needed to drag the camera. A value of 1 makes the camera move only when reaching the edge of the screen.

pub fn set_enable_follow_smoothing(&self, follow_smoothing: bool)

If true, the camera smoothly moves towards the target at [smoothing_speed][Self::smoothing_speed].

pub fn set_follow_smoothing(&self, follow_smoothing: f64)

Speed in pixels per second of the camera’s smoothing effect when [smoothing_enabled][Self::smoothing_enabled] is true.

pub fn set_h_drag_enabled(&self, enabled: bool)

If true, the camera only moves when reaching the horizontal drag margins. If false, the camera moves horizontally regardless of margins.

pub fn set_h_offset(&self, ofs: f64)

The horizontal offset of the camera, relative to the drag margins. Note: Offset H is used only to force offset relative to margins. It’s not updated in any way if drag margins are enabled and can be used to set initial offset.

pub fn set_limit(&self, margin: i64, limit: i64)

Top scroll limit in pixels. The camera stops moving when reaching this value.

pub fn set_limit_drawing_enabled(&self, limit_drawing_enabled: bool)

If true, draws the camera’s limits rectangle in the editor.

pub fn set_limit_smoothing_enabled(&self, limit_smoothing_enabled: bool)

If true, the camera smoothly stops when reaches its limits.

pub fn set_margin_drawing_enabled(&self, margin_drawing_enabled: bool)

If true, draws the camera’s drag margin rectangle in the editor.

pub fn set_offset(&self, offset: Vector2)

The camera’s offset, useful for looking around or camera shake animations.

pub fn set_process_mode(&self, mode: i64)

The camera’s process callback. See Camera2DProcessMode.

pub fn set_rotating(&self, rotating: bool)

If true, the camera rotates with the target.

pub fn set_screen_drawing_enabled(&self, screen_drawing_enabled: bool)

If true, draws the camera’s screen rectangle in the editor.

pub fn set_v_drag_enabled(&self, enabled: bool)

If true, the camera only moves when reaching the vertical drag margins. If false, the camera moves vertically regardless of margins.

pub fn set_v_offset(&self, ofs: f64)

The vertical offset of the camera, relative to the drag margins. Note: Used the same as [offset_h][Self::offset_h].

pub fn set_zoom(&self, zoom: Vector2)

The camera’s zoom relative to the viewport. Values larger than Vector2(1, 1) zoom out and smaller values zoom in. For an example, use Vector2(0.5, 0.5) for a 2× zoom-in, and Vector2(4, 4) for a 4× zoom-out.