Struct gdnative::api::class_db::ClassDB

pub struct ClassDB { /* fields omitted */ }

core singleton class ClassDB inherits Object (unsafe).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Class hierarchy

ClassDB inherits methods from:

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl ClassDB

pub fn godot_singleton() -> &'static ClassDB

Returns a reference to the singleton instance.

pub fn can_instance(&self, class: impl Into<GodotString>) -> bool

Returns true if you can instance objects from the specified class, false in other case.

pub fn class_exists(&self, class: impl Into<GodotString>) -> bool

Returns whether the specified class is available or not.

pub fn class_get_category(&self, class: impl Into<GodotString>) -> GodotString

Returns a category associated with the class for use in documentation and the Asset Library. Debug mode required.

pub fn class_get_integer_constant(
&self,
class: impl Into<GodotString>,
name: impl Into<GodotString>
) -> i64

Returns the value of the integer constant name of class or its ancestry. Always returns 0 when the constant could not be found.

pub fn class_get_integer_constant_list(
&self,
class: impl Into<GodotString>,
no_inheritance: bool
) -> TypedArray<GodotString>

Returns an array with the names all the integer constants of class or its ancestry.

Default Arguments

no_inheritance - false

pub fn class_get_method_list(
&self,
class: impl Into<GodotString>,
no_inheritance: bool
) -> VariantArray<Shared>

Returns an array with all the methods of class or its ancestry if no_inheritance is false. Every element of the array is a Dictionary with the following keys: args, default_args, flags, id, name, return: (class_name, hint, hint_string, name, type, usage). Note: In exported release builds the debug info is not available, so the returned dictionaries will contain only method names.

Default Arguments

no_inheritance - false

pub fn class_get_property(
&self,
object: impl AsArg<Object>,
property: impl Into<GodotString>
) -> Variant

Returns the value of property of class or its ancestry.

pub fn class_get_property_list(
&self,
class: impl Into<GodotString>,
no_inheritance: bool
) -> VariantArray<Shared>

Returns an array with all the properties of class or its ancestry if no_inheritance is false.

Default Arguments

no_inheritance - false

pub fn class_get_signal(
&self,
class: impl Into<GodotString>,
signal: impl Into<GodotString>
) -> Dictionary<Shared>

Returns the signal data of class or its ancestry. The returned value is a Dictionary with the following keys: args, default_args, flags, id, name, return: (class_name, hint, hint_string, name, type, usage).

pub fn class_get_signal_list(
&self,
class: impl Into<GodotString>,
no_inheritance: bool
) -> VariantArray<Shared>

Returns an array with all the signals of class or its ancestry if no_inheritance is false. Every element of the array is a Dictionary as described in class_get_signal.

Default Arguments

no_inheritance - false

pub fn class_has_integer_constant(
&self,
class: impl Into<GodotString>,
name: impl Into<GodotString>
) -> bool

Returns whether class or its ancestry has an integer constant called name or not.

pub fn class_has_method(
&self,
class: impl Into<GodotString>,
method: impl Into<GodotString>,
no_inheritance: bool
) -> bool

Returns whether class (or its ancestry if no_inheritance is false) has a method called method or not.

Default Arguments

no_inheritance - false

pub fn class_has_signal(
&self,
class: impl Into<GodotString>,
signal: impl Into<GodotString>
) -> bool

Returns whether class or its ancestry has a signal called signal or not.

pub fn class_set_property(
&self,
object: impl AsArg<Object>,
property: impl Into<GodotString>,
value: impl OwnedToVariant
) -> Result<(), GodotError>

Sets property value of class to value.

pub fn get_class_list(&self) -> TypedArray<GodotString>

Returns the names of all the classes available.

pub fn get_inheriters_from_class(
&self,
class: impl Into<GodotString>
) -> TypedArray<GodotString>

Returns the names of all the classes that directly or indirectly inherit from class.

pub fn get_parent_class(&self, class: impl Into<GodotString>) -> GodotString

Returns the parent class of class.

pub fn instance(&self, class: impl Into<GodotString>) -> Variant

Creates an instance of class.

pub fn is_class_enabled(&self, class: impl Into<GodotString>) -> bool

Returns whether this class is enabled or not.

pub fn is_parent_class(
&self,
class: impl Into<GodotString>,
inherits: impl Into<GodotString>
) -> bool

Returns whether inherits is an ancestor of class or not.