Struct gdnative::api::file_dialog::FileDialog

pub struct FileDialog { /* fields omitted */ }

core class FileDialog inherits ConfirmationDialog (unsafe).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non reference counted objects such as the ones of this type are usually owned by the engine.

FileDialog is a reference-only type. Persistent references can only exist in the unsafe Ref<FileDialog> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

FileDialog inherits methods from:

ConfirmationDialog

AcceptDialog

WindowDialog

Popup

Control

CanvasItem

Node

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl FileDialog

Constants

pub const ACCESS_RESOURCES: i64

pub const MODE_OPEN_FILE: i64

pub const ACCESS_USERDATA: i64

pub const MODE_OPEN_FILES: i64

pub const ACCESS_FILESYSTEM: i64

pub const MODE_OPEN_DIR: i64

pub const MODE_OPEN_ANY: i64

pub const MODE_SAVE_FILE: i64

impl FileDialog

pub fn new() -> Ref<FileDialog, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn add_filter(&self, filter: impl Into<GodotString>)

Adds filter as a custom filter; filter should be of the form "filename.extension ; Description". For example, "*.png ; PNG Images".

pub fn clear_filters(&self)

Clear all the added filters in the dialog.

pub fn deselect_items(&self)

Clear currently selected items in the dialog.

pub fn access(&self) -> Access

The file system access scope. See enum Access constants. Warning: Currently, in sandboxed environments such as HTML5 builds or sandboxed macOS apps, FileDialog cannot access the host file system. See godot-proposals#1123.

pub fn current_dir(&self) -> GodotString

The current working directory of the file dialog.

pub fn current_file(&self) -> GodotString

The currently selected file of the file dialog.

pub fn current_path(&self) -> GodotString

The currently selected file path of the file dialog.

pub fn filters(&self) -> TypedArray<GodotString>

The available file type filters. For example, this shows only .png and .gd files: set_filters(PoolStringArray(["*.png ; PNG Images","*.gd ; GDScript Files"])).

pub fn get_line_edit(&self) -> Option<Ref<LineEdit, Shared>>

Returns the LineEdit for the selected file.

pub fn mode(&self) -> Mode

The dialog’s open or save mode, which affects the selection behavior. See enum Mode constants.

pub fn get_vbox(&self) -> Option<Ref<VBoxContainer, Shared>>

Returns the vertical box container of the dialog, custom controls can be added to it.

pub fn invalidate(&self)

Invalidate and update the current dialog content list.

pub fn is_mode_overriding_title(&self) -> bool

If true, changing the Mode property will set the window title accordingly (e.g. setting mode to MODE_OPEN_FILE will change the window title to “Open a File”).

pub fn is_showing_hidden_files(&self) -> bool

If true, the dialog will show hidden files.

pub fn set_access(&self, access: i64)

The file system access scope. See enum Access constants. Warning: Currently, in sandboxed environments such as HTML5 builds or sandboxed macOS apps, FileDialog cannot access the host file system. See godot-proposals#1123.

pub fn set_current_dir(&self, dir: impl Into<GodotString>)

The current working directory of the file dialog.

pub fn set_current_file(&self, file: impl Into<GodotString>)

The currently selected file of the file dialog.

pub fn set_current_path(&self, path: impl Into<GodotString>)

The currently selected file path of the file dialog.

pub fn set_filters(&self, filters: TypedArray<GodotString>)

The available file type filters. For example, this shows only .png and .gd files: set_filters(PoolStringArray(["*.png ; PNG Images","*.gd ; GDScript Files"])).

pub fn set_mode(&self, mode: i64)

The dialog’s open or save mode, which affects the selection behavior. See enum Mode constants.

pub fn set_mode_overrides_title(&self, _override: bool)

If true, changing the Mode property will set the window title accordingly (e.g. setting mode to MODE_OPEN_FILE will change the window title to “Open a File”).

pub fn set_show_hidden_files(&self, show: bool)

If true, the dialog will show hidden files.