Struct gdnative::api::label::Label

pub struct Label { /* fields omitted */ }

core class Label inherits Control (unsafe).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non reference counted objects such as the ones of this type are usually owned by the engine.

Label is a reference-only type. Persistent references can only exist in the unsafe Ref<Label> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

Label inherits methods from:

Control

CanvasItem

Node

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Label

Constants

pub const ALIGN_LEFT: i64

pub const VALIGN_TOP: i64

pub const ALIGN_CENTER: i64

pub const VALIGN_CENTER: i64

pub const ALIGN_RIGHT: i64

pub const VALIGN_BOTTOM: i64

pub const ALIGN_FILL: i64

pub const VALIGN_FILL: i64

impl Label

pub fn new() -> Ref<Label, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn align(&self) -> Align

Controls the text’s horizontal align. Supports left, center, right, and fill, or justify. Set it to one of the Align constants.

pub fn get_line_count(&self) -> i64

Returns the amount of lines of text the Label has.

pub fn get_line_height(&self) -> i64

Returns the font size in pixels.

pub fn lines_skipped(&self) -> i64

The node ignores the first lines_skipped lines before it starts to display text.

pub fn max_lines_visible(&self) -> i64

Limits the lines of text the node shows on screen.

pub fn percent_visible(&self) -> f64

Limits the amount of visible characters. If you set percent_visible to 0.5, only up to half of the text’s characters will display on screen. Useful to animate the text in a dialog box.

pub fn text(&self) -> GodotString

The text to display on screen.

pub fn get_total_character_count(&self) -> i64

Returns the total number of printable characters in the text (excluding spaces and newlines).

pub fn valign(&self) -> VAlign

Controls the text’s vertical align. Supports top, center, bottom, and fill. Set it to one of the VAlign constants.

pub fn visible_characters(&self) -> i64

Restricts the number of characters to display. Set to -1 to disable.

pub fn get_visible_line_count(&self) -> i64

Returns the number of lines shown. Useful if the Label’s height cannot currently display all lines.

pub fn has_autowrap(&self) -> bool

If true, wraps the text inside the node’s bounding rectangle. If you resize the node, it will change its height automatically to show all the text.

pub fn is_clipping_text(&self) -> bool

If true, the Label only shows the text that fits inside its bounding rectangle. It also lets you scale the node down freely.

pub fn is_uppercase(&self) -> bool

If true, all the text displays as UPPERCASE.

pub fn set_align(&self, align: i64)

Controls the text’s horizontal align. Supports left, center, right, and fill, or justify. Set it to one of the Align constants.

pub fn set_autowrap(&self, enable: bool)

If true, wraps the text inside the node’s bounding rectangle. If you resize the node, it will change its height automatically to show all the text.

pub fn set_clip_text(&self, enable: bool)

If true, the Label only shows the text that fits inside its bounding rectangle. It also lets you scale the node down freely.

pub fn set_lines_skipped(&self, lines_skipped: i64)

The node ignores the first lines_skipped lines before it starts to display text.

pub fn set_max_lines_visible(&self, lines_visible: i64)

Limits the lines of text the node shows on screen.

pub fn set_percent_visible(&self, percent_visible: f64)

Limits the amount of visible characters. If you set percent_visible to 0.5, only up to half of the text’s characters will display on screen. Useful to animate the text in a dialog box.

pub fn set_text(&self, text: impl Into<GodotString>)

The text to display on screen.

pub fn set_uppercase(&self, enable: bool)

If true, all the text displays as UPPERCASE.

pub fn set_valign(&self, valign: i64)

Controls the text’s vertical align. Supports top, center, bottom, and fill. Set it to one of the VAlign constants.

pub fn set_visible_characters(&self, amount: i64)

Restricts the number of characters to display. Set to -1 to disable.