Struct gdnative::api::light_2d::Light2D

pub struct Light2D { /* fields omitted */ }

core class Light2D inherits Node2D (unsafe).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non reference counted objects such as the ones of this type are usually owned by the engine.

Light2D is a reference-only type. Persistent references can only exist in the unsafe Ref<Light2D> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

Light2D inherits methods from:

Node2D

CanvasItem

Node

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Light2D

Constants

pub const MODE_ADD: i64

pub const SHADOW_FILTER_NONE: i64

pub const MODE_SUB: i64

pub const SHADOW_FILTER_PCF3: i64

pub const MODE_MIX: i64

pub const SHADOW_FILTER_PCF5: i64

pub const MODE_MASK: i64

pub const SHADOW_FILTER_PCF7: i64

pub const SHADOW_FILTER_PCF9: i64

pub const SHADOW_FILTER_PCF13: i64

impl Light2D

pub fn new() -> Ref<Light2D, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn color(&self) -> Color

The Light2D’s Color.

pub fn energy(&self) -> f64

The Light2D’s energy value. The larger the value, the stronger the light.

pub fn height(&self) -> f64

The height of the Light2D. Used with 2D normal mapping.

pub fn item_cull_mask(&self) -> i64

The layer mask. Only objects with a matching mask will be affected by the Light2D.

pub fn item_shadow_cull_mask(&self) -> i64

The shadow mask. Used with LightOccluder2D to cast shadows. Only occluders with a matching light mask will cast shadows.

pub fn layer_range_max(&self) -> i64

Maximum layer value of objects that are affected by the Light2D.

pub fn layer_range_min(&self) -> i64

Minimum layer value of objects that are affected by the Light2D.

pub fn mode(&self) -> Mode

The Light2D’s mode. See Mode constants for values.

pub fn shadow_buffer_size(&self) -> i64

Shadow buffer size.

pub fn shadow_color(&self) -> Color

Color of shadows cast by the Light2D.

pub fn shadow_filter(&self) -> ShadowFilter

Shadow filter type. See ShadowFilter for possible values.

pub fn shadow_gradient_length(&self) -> f64

Smooth shadow gradient length.

pub fn shadow_smooth(&self) -> f64

Smoothing value for shadows.

pub fn texture(&self) -> Option<Ref<Texture, Shared>>

Texture used for the Light2D’s appearance.

pub fn texture_offset(&self) -> Vector2

The offset of the Light2D’s texture.

pub fn texture_scale(&self) -> f64

The texture’s scale factor.

pub fn z_range_max(&self) -> i64

Maximum z value of objects that are affected by the Light2D.

pub fn z_range_min(&self) -> i64

Minimum z value of objects that are affected by the Light2D.

pub fn is_editor_only(&self) -> bool

If true, Light2D will only appear when editing the scene.

pub fn is_enabled(&self) -> bool

If true, Light2D will emit light.

pub fn is_shadow_enabled(&self) -> bool

If true, the Light2D will cast shadows.

pub fn set_color(&self, color: Color)

The Light2D’s Color.

pub fn set_editor_only(&self, editor_only: bool)

If true, Light2D will only appear when editing the scene.

pub fn set_enabled(&self, enabled: bool)

If true, Light2D will emit light.

pub fn set_energy(&self, energy: f64)

The Light2D’s energy value. The larger the value, the stronger the light.

pub fn set_height(&self, height: f64)

The height of the Light2D. Used with 2D normal mapping.

pub fn set_item_cull_mask(&self, item_cull_mask: i64)

The layer mask. Only objects with a matching mask will be affected by the Light2D.

pub fn set_item_shadow_cull_mask(&self, item_shadow_cull_mask: i64)

The shadow mask. Used with LightOccluder2D to cast shadows. Only occluders with a matching light mask will cast shadows.

pub fn set_layer_range_max(&self, layer: i64)

Maximum layer value of objects that are affected by the Light2D.

pub fn set_layer_range_min(&self, layer: i64)

Minimum layer value of objects that are affected by the Light2D.

pub fn set_mode(&self, mode: i64)

The Light2D’s mode. See Mode constants for values.

pub fn set_shadow_buffer_size(&self, size: i64)

Shadow buffer size.

pub fn set_shadow_color(&self, shadow_color: Color)

Color of shadows cast by the Light2D.

pub fn set_shadow_enabled(&self, enabled: bool)

If true, the Light2D will cast shadows.

pub fn set_shadow_filter(&self, filter: i64)

Shadow filter type. See ShadowFilter for possible values.

pub fn set_shadow_gradient_length(&self, multiplier: f64)

Smooth shadow gradient length.

pub fn set_shadow_smooth(&self, smooth: f64)

Smoothing value for shadows.

pub fn set_texture(&self, texture: impl AsArg<Texture>)

Texture used for the Light2D’s appearance.

pub fn set_texture_offset(&self, texture_offset: Vector2)

The offset of the Light2D’s texture.

pub fn set_texture_scale(&self, texture_scale: f64)

The texture’s scale factor.

pub fn set_z_range_max(&self, z: i64)

Maximum z value of objects that are affected by the Light2D.

pub fn set_z_range_min(&self, z: i64)

Minimum z value of objects that are affected by the Light2D.