Struct gdnative::api::networked_multiplayer_peer::NetworkedMultiplayerPeer

pub struct NetworkedMultiplayerPeer { /* fields omitted */ }

core class NetworkedMultiplayerPeer inherits PacketPeer (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

NetworkedMultiplayerPeer inherits methods from:

PacketPeer

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl NetworkedMultiplayerPeer

Constants

pub const CONNECTION_DISCONNECTED: i64

pub const TARGET_PEER_BROADCAST: i64

pub const TRANSFER_MODE_UNRELIABLE: i64

pub const CONNECTION_CONNECTING: i64

pub const TARGET_PEER_SERVER: i64

pub const TRANSFER_MODE_UNRELIABLE_ORDERED: i64

pub const CONNECTION_CONNECTED: i64

pub const TRANSFER_MODE_RELIABLE: i64

impl NetworkedMultiplayerPeer

pub fn get_connection_status(&self) -> ConnectionStatus

Returns the current state of the connection. See ConnectionStatus.

pub fn get_packet_peer(&self) -> i64

Returns the ID of the NetworkedMultiplayerPeer who sent the most recent packet.

pub fn transfer_mode(&self) -> TransferMode

The manner in which to send packets to the target_peer. See TransferMode.

pub fn get_unique_id(&self) -> i64

Returns the ID of this NetworkedMultiplayerPeer.

pub fn is_refusing_new_connections(&self) -> bool

If true, this NetworkedMultiplayerPeer refuses new connections.

pub fn poll(&self)

Waits up to 1 second to receive a new network event.

pub fn set_refuse_new_connections(&self, enable: bool)

If true, this NetworkedMultiplayerPeer refuses new connections.

pub fn set_target_peer(&self, id: i64)

Sets the peer to which packets will be sent. The id can be one of: TARGET_PEER_BROADCAST to send to all connected peers, TARGET_PEER_SERVER to send to the peer acting as server, a valid peer ID to send to that specific peer, a negative peer ID to send to all peers except that one. By default, the target peer is TARGET_PEER_BROADCAST.

pub fn set_transfer_mode(&self, mode: i64)

The manner in which to send packets to the target_peer. See TransferMode.