Struct gdnative::api::physics_body_2d::PhysicsBody2D

pub struct PhysicsBody2D { /* fields omitted */ }

core class PhysicsBody2D inherits CollisionObject2D (unsafe).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Class hierarchy

PhysicsBody2D inherits methods from:

CollisionObject2D

Node2D

CanvasItem

Node

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl PhysicsBody2D

pub fn add_collision_exception_with(&self, body: impl AsArg<Node>)

Adds a body to the list of bodies that this body can’t collide with.

pub fn get_collision_exceptions(&self) -> VariantArray<Shared>

Returns an array of nodes that were added as collision exceptions for this body.

pub fn collision_layer(&self) -> i64

The physics layers this area is in. Collidable objects can exist in any of 32 different layers. These layers work like a tagging system, and are not visual. A collidable can use these layers to select with which objects it can collide, using the collision_mask property. A contact is detected if object A is in any of the layers that object B scans, or object B is in any layer scanned by object A. See Collision layers and masks in the documentation for more information.

pub fn get_collision_layer_bit(&self, bit: i64) -> bool

Returns an individual bit on the collision_layer.

pub fn collision_mask(&self) -> i64

The physics layers this area scans for collisions. See Collision layers and masks in the documentation for more information.

pub fn get_collision_mask_bit(&self, bit: i64) -> bool

Returns an individual bit on the collision_mask.

pub fn remove_collision_exception_with(&self, body: impl AsArg<Node>)

Removes a body from the list of bodies that this body can’t collide with.

pub fn set_collision_layer(&self, layer: i64)

The physics layers this area is in. Collidable objects can exist in any of 32 different layers. These layers work like a tagging system, and are not visual. A collidable can use these layers to select with which objects it can collide, using the collision_mask property. A contact is detected if object A is in any of the layers that object B scans, or object B is in any layer scanned by object A. See Collision layers and masks in the documentation for more information.

pub fn set_collision_layer_bit(&self, bit: i64, value: bool)

Sets individual bits on the collision_layer bitmask. Use this if you only need to change one layer’s value.

pub fn set_collision_mask(&self, mask: i64)

The physics layers this area scans for collisions. See Collision layers and masks in the documentation for more information.

pub fn set_collision_mask_bit(&self, bit: i64, value: bool)

Sets individual bits on the collision_mask bitmask. Use this if you only need to change one layer’s value.