Struct gdnative::api::resource::Resource

pub struct Resource { /* fields omitted */ }

core class Resource inherits Reference (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

Resource inherits methods from:

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Resource

pub fn new() -> Ref<Resource, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn duplicate(&self, subresources: bool) -> Option<Ref<Resource, Shared>>

Duplicates the resource, returning a new resource. By default, sub-resources are shared between resource copies for efficiency. This can be changed by passing true to the subresources argument which will copy the subresources. Note: If subresources is true, this method will only perform a shallow copy. Nested resources within subresources will not be duplicated and will still be shared.

Default Arguments

subresources - false

pub fn emit_changed(&self)

Sample code is GDScript unless otherwise noted.

Emits the changed signal. If external objects which depend on this resource should be updated, this method must be called manually whenever the state of this resource has changed (such as modification of properties). The method is equivalent to:

emit_signal("changed")

Note: This method is called automatically for built-in resources.

pub fn get_local_scene(&self) -> Option<Ref<Node, Shared>>

If [resource_local_to_scene][Self::resource_local_to_scene] is enabled and the resource was loaded from a PackedScene instantiation, returns the local scene where this resource’s unique copy is in use. Otherwise, returns null.

pub fn name(&self) -> GodotString

The name of the resource. This is an optional identifier. If [resource_name][Self::resource_name] is not empty, its value will be displayed to represent the current resource in the editor inspector. For built-in scripts, the [resource_name][Self::resource_name] will be displayed as the tab name in the script editor.

pub fn path(&self) -> GodotString

The path to the resource. In case it has its own file, it will return its filepath. If it’s tied to the scene, it will return the scene’s path, followed by the resource’s index.

pub fn get_rid(&self) -> Rid

Returns the RID of the resource (or an empty RID). Many resources (such as Texture, Mesh, etc) are high-level abstractions of resources stored in a server, so this function will return the original RID.

pub fn is_local_to_scene(&self) -> bool

If true, the resource will be made unique in each instance of its local scene. It can thus be modified in a scene instance without impacting other instances of that same scene.

pub fn set_local_to_scene(&self, enable: bool)

If true, the resource will be made unique in each instance of its local scene. It can thus be modified in a scene instance without impacting other instances of that same scene.

pub fn set_name(&self, name: impl Into<GodotString>)

The name of the resource. This is an optional identifier. If [resource_name][Self::resource_name] is not empty, its value will be displayed to represent the current resource in the editor inspector. For built-in scripts, the [resource_name][Self::resource_name] will be displayed as the tab name in the script editor.

pub fn set_path(&self, path: impl Into<GodotString>)

The path to the resource. In case it has its own file, it will return its filepath. If it’s tied to the scene, it will return the scene’s path, followed by the resource’s index.

pub fn setup_local_to_scene(&self)

This method is called when a resource with [resource_local_to_scene][Self::resource_local_to_scene] enabled is loaded from a PackedScene instantiation. Its behavior can be customized by overriding [_setup_local_to_scene][Self::_setup_local_to_scene] from script. For most resources, this method performs no base logic. ViewportTexture performs custom logic to properly set the proxy texture and flags in the local viewport.

pub fn take_over_path(&self, path: impl Into<GodotString>)

Sets the path of the resource, potentially overriding an existing cache entry for this path. This differs from setting [resource_path][Self::resource_path], as the latter would error out if another resource was already cached for the given path.