Struct gdnative::api::AStar

pub struct AStar { /* fields omitted */ }

core class AStar inherits Reference (reference-counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

AStar inherits methods from:

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl AStar

pub fn new() -> Ref<AStar, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn add_point(&self, id: i64, position: Vector3, weight_scale: f64)

Sample code is GDScript unless otherwise noted.

Adds a new point at the given position with the given identifier. The id must be 0 or larger, and the weight_scale must be 1 or larger. The weight_scale is multiplied by the result of [_compute_cost][Self::_compute_cost] when determining the overall cost of traveling across a segment from a neighboring point to this point. Thus, all else being equal, the algorithm prefers points with lower weight_scales to form a path.

var astar = AStar.new()
astar.add_point(1, Vector3(1, 0, 0), 4) # Adds the point (1, 0, 0) with weight_scale 4 and id 1

If there already exists a point for the given id, its position and weight scale are updated to the given values.

Default Arguments

weight_scale - 1.0

pub fn are_points_connected(
&self,
id: i64,
to_id: i64,
bidirectional: bool
) -> bool

Returns whether the two given points are directly connected by a segment. If bidirectional is false, returns whether movement from id to to_id is possible through this segment.

Default Arguments

bidirectional - true

pub fn clear(&self)

Clears all the points and segments.

pub fn connect_points(&self, id: i64, to_id: i64, bidirectional: bool)

Sample code is GDScript unless otherwise noted.

Creates a segment between the given points. If bidirectional is false, only movement from id to to_id is allowed, not the reverse direction.

var astar = AStar.new()
astar.add_point(1, Vector3(1, 1, 0))
astar.add_point(2, Vector3(0, 5, 0))
astar.connect_points(1, 2, false)

Default Arguments

bidirectional - true

pub fn disconnect_points(&self, id: i64, to_id: i64, bidirectional: bool)

Deletes the segment between the given points. If bidirectional is false, only movement from id to to_id is prevented, and a unidirectional segment possibly remains.

Default Arguments

bidirectional - true

pub fn get_available_point_id(&self) -> i64

Returns the next available point ID with no point associated to it.

pub fn get_closest_point(
&self,
to_position: Vector3,
include_disabled: bool
) -> i64

Returns the ID of the closest point to to_position, optionally taking disabled points into account. Returns -1 if there are no points in the points pool. Note: If several points are the closest to to_position, the one with the smallest ID will be returned, ensuring a deterministic result.

Default Arguments

include_disabled - false

pub fn get_closest_position_in_segment(&self, to_position: Vector3) -> Vector3

Sample code is GDScript unless otherwise noted.

Returns the closest position to to_position that resides inside a segment between two connected points.

var astar = AStar.new()
astar.add_point(1, Vector3(0, 0, 0))
astar.add_point(2, Vector3(0, 5, 0))
astar.connect_points(1, 2)
var res = astar.get_closest_position_in_segment(Vector3(3, 3, 0)) # Returns (0, 3, 0)

The result is in the segment that goes from y = 0 to y = 5. It’s the closest position in the segment to the given point.

pub fn get_id_path(&self, from_id: i64, to_id: i64) -> PoolArray<i32>

Sample code is GDScript unless otherwise noted.

Returns an array with the IDs of the points that form the path found by AStar between the given points. The array is ordered from the starting point to the ending point of the path.

var astar = AStar.new()
astar.add_point(1, Vector3(0, 0, 0))
astar.add_point(2, Vector3(0, 1, 0), 1) # Default weight is 1
astar.add_point(3, Vector3(1, 1, 0))
astar.add_point(4, Vector3(2, 0, 0))

astar.connect_points(1, 2, false)
astar.connect_points(2, 3, false)
astar.connect_points(4, 3, false)
astar.connect_points(1, 4, false)

var res = astar.get_id_path(1, 3) # Returns [1, 2, 3]

If you change the 2nd point’s weight to 3, then the result will be [1, 4, 3] instead, because now even though the distance is longer, it’s “easier” to get through point 4 than through point 2.

pub fn get_point_capacity(&self) -> i64

Returns the capacity of the structure backing the points, useful in conjunction with reserve_space.

pub fn get_point_connections(&self, id: i64) -> PoolArray<i32>

Sample code is GDScript unless otherwise noted.

Returns an array with the IDs of the points that form the connection with the given point.

var astar = AStar.new()
astar.add_point(1, Vector3(0, 0, 0))
astar.add_point(2, Vector3(0, 1, 0))
astar.add_point(3, Vector3(1, 1, 0))
astar.add_point(4, Vector3(2, 0, 0))

astar.connect_points(1, 2, true)
astar.connect_points(1, 3, true)

var neighbors = astar.get_point_connections(1) # Returns [2, 3]

pub fn get_point_count(&self) -> i64

Returns the number of points currently in the points pool.

pub fn get_point_path(&self, from_id: i64, to_id: i64) -> PoolArray<Vector3>

Returns an array with the points that are in the path found by AStar between the given points. The array is ordered from the starting point to the ending point of the path. Note: This method is not thread-safe. If called from a Thread, it will return an empty PoolVector3Array and will print an error message.

pub fn get_point_position(&self, id: i64) -> Vector3

Returns the position of the point associated with the given id.

pub fn get_point_weight_scale(&self, id: i64) -> f64

Returns the weight scale of the point associated with the given id.

pub fn get_points(&self) -> VariantArray<Shared>

Returns an array of all points.

pub fn has_point(&self, id: i64) -> bool

Returns whether a point associated with the given id exists.

pub fn is_point_disabled(&self, id: i64) -> bool

Returns whether a point is disabled or not for pathfinding. By default, all points are enabled.

pub fn remove_point(&self, id: i64)

Removes the point associated with the given id from the points pool.

pub fn reserve_space(&self, num_nodes: i64)

Reserves space internally for num_nodes points, useful if you’re adding a known large number of points at once, for a grid for instance. New capacity must be greater or equals to old capacity.

pub fn set_point_disabled(&self, id: i64, disabled: bool)

Disables or enables the specified point for pathfinding. Useful for making a temporary obstacle.

Default Arguments

disabled - true

pub fn set_point_position(&self, id: i64, position: Vector3)

Sets the position for the point with the given id.

pub fn set_point_weight_scale(&self, id: i64, weight_scale: f64)

Sets the weight_scale for the point with the given id. The weight_scale is multiplied by the result of [_compute_cost][Self::_compute_cost] when determining the overall cost of traveling across a segment from a neighboring point to this point.