Struct gdnative::api::AnimatedTexture

pub struct AnimatedTexture { /* fields omitted */ }

core class AnimatedTexture inherits Texture (reference-counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

AnimatedTexture inherits methods from:

Texture

Resource

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl AnimatedTexture

Constants

pub const MAX_FRAMES: i64

impl AnimatedTexture

pub fn new() -> Ref<AnimatedTexture, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn current_frame(&self) -> i64

Sets the currently visible frame of the texture.

pub fn fps(&self) -> f64

Animation speed in frames per second. This value defines the default time interval between two frames of the animation, and thus the overall duration of the animation loop based on the frames property. A value of 0 means no predefined number of frames per second, the animation will play according to each frame’s frame delay (see set_frame_delay). For example, an animation with 8 frames, no frame delay and a fps value of 2 will run for 4 seconds, with each frame lasting 0.5 seconds.

pub fn frame_delay(&self, frame: i64) -> f64

Returns the given frame’s delay value.

pub fn frame_texture(&self, frame: i64) -> Option<Ref<Texture, Shared>>

Returns the given frame’s Texture.

pub fn frames(&self) -> i64

Number of frames to use in the animation. While you can create the frames independently with set_frame_texture, you need to set this value for the animation to take new frames into account. The maximum number of frames is MAX_FRAMES.

pub fn oneshot(&self) -> bool

If true, the animation will only play once and will not loop back to the first frame after reaching the end. Note that reaching the end will not set pause to true.

pub fn pause(&self) -> bool

If true, the animation will pause where it currently is (i.e. at current_frame). The animation will continue from where it was paused when changing this property to false.

pub fn set_current_frame(&self, frame: i64)

Sets the currently visible frame of the texture.

pub fn set_fps(&self, fps: f64)

Animation speed in frames per second. This value defines the default time interval between two frames of the animation, and thus the overall duration of the animation loop based on the frames property. A value of 0 means no predefined number of frames per second, the animation will play according to each frame’s frame delay (see set_frame_delay). For example, an animation with 8 frames, no frame delay and a fps value of 2 will run for 4 seconds, with each frame lasting 0.5 seconds.

pub fn set_frame_delay(&self, frame: i64, delay: f64)

Sample code is GDScript unless otherwise noted.

Sets an additional delay (in seconds) between this frame and the next one, that will be added to the time interval defined by fps. By default, frames have no delay defined. If a delay value is defined, the final time interval between this frame and the next will be 1.0 / fps + delay. For example, for an animation with 3 frames, 2 FPS and a frame delay on the second frame of 1.2, the resulting playback will be:

Frame 0: 0.5 s (1 / fps)
Frame 1: 1.7 s (1 / fps + 1.2)
Frame 2: 0.5 s (1 / fps)
Total duration: 2.7 s

pub fn set_frame_texture(&self, frame: i64, texture: impl AsArg<Texture>)

Assigns a Texture to the given frame. Frame IDs start at 0, so the first frame has ID 0, and the last frame of the animation has ID frames - 1. You can define any number of textures up to MAX_FRAMES, but keep in mind that only frames from 0 to frames - 1 will be part of the animation.

pub fn set_frames(&self, frames: i64)

Number of frames to use in the animation. While you can create the frames independently with set_frame_texture, you need to set this value for the animation to take new frames into account. The maximum number of frames is MAX_FRAMES.

pub fn set_oneshot(&self, oneshot: bool)

If true, the animation will only play once and will not loop back to the first frame after reaching the end. Note that reaching the end will not set pause to true.

pub fn set_pause(&self, pause: bool)

If true, the animation will pause where it currently is (i.e. at current_frame). The animation will continue from where it was paused when changing this property to false.