Struct gdnative::api::Animation

pub struct Animation { /* fields omitted */ }

core class Animation inherits Resource (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

Animation inherits methods from:

Resource

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Animation

Constants

pub const INTERPOLATION_NEAREST: i64

pub const TYPE_VALUE: i64

pub const UPDATE_CONTINUOUS: i64

pub const INTERPOLATION_LINEAR: i64

pub const TYPE_TRANSFORM: i64

pub const UPDATE_DISCRETE: i64

pub const INTERPOLATION_CUBIC: i64

pub const TYPE_METHOD: i64

pub const UPDATE_TRIGGER: i64

pub const TYPE_BEZIER: i64

pub const UPDATE_CAPTURE: i64

pub const TYPE_AUDIO: i64

pub const TYPE_ANIMATION: i64

impl Animation

pub fn new() -> Ref<Animation, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn add_track(&self, _type: i64, at_position: i64) -> i64

Adds a track to the Animation.

Default Arguments

at_position - -1

pub fn animation_track_get_key_animation(
&self,
track_idx: i64,
key_idx: i64
) -> GodotString

Returns the animation name at the key identified by key_idx. The track_idx must be the index of an Animation Track.

pub fn animation_track_insert_key(
&self,
track_idx: i64,
time: f64,
animation: impl Into<GodotString>
) -> i64

Inserts a key with value animation at the given time (in seconds). The track_idx must be the index of an Animation Track.

pub fn animation_track_set_key_animation(
&self,
track_idx: i64,
key_idx: i64,
animation: impl Into<GodotString>
)

Sets the key identified by key_idx to value animation. The track_idx must be the index of an Animation Track.

pub fn audio_track_get_key_end_offset(
&self,
track_idx: i64,
key_idx: i64
) -> f64

Returns the end offset of the key identified by key_idx. The track_idx must be the index of an Audio Track. End offset is the number of seconds cut off at the ending of the audio stream.

pub fn audio_track_get_key_start_offset(
&self,
track_idx: i64,
key_idx: i64
) -> f64

Returns the start offset of the key identified by key_idx. The track_idx must be the index of an Audio Track. Start offset is the number of seconds cut off at the beginning of the audio stream.

pub fn audio_track_get_key_stream(
&self,
track_idx: i64,
key_idx: i64
) -> Option<Ref<Resource, Shared>>

Returns the audio stream of the key identified by key_idx. The track_idx must be the index of an Audio Track.

pub fn audio_track_insert_key(
&self,
track_idx: i64,
time: f64,
stream: impl AsArg<Resource>,
start_offset: f64,
end_offset: f64
) -> i64

Inserts an Audio Track key at the given time in seconds. The track_idx must be the index of an Audio Track. stream is the AudioStream resource to play. start_offset is the number of seconds cut off at the beginning of the audio stream, while end_offset is at the ending.

Default Arguments

start_offset - 0

end_offset - 0

pub fn audio_track_set_key_end_offset(
&self,
track_idx: i64,
key_idx: i64,
offset: f64
)

Sets the end offset of the key identified by key_idx to value offset. The track_idx must be the index of an Audio Track.

pub fn audio_track_set_key_start_offset(
&self,
track_idx: i64,
key_idx: i64,
offset: f64
)

Sets the start offset of the key identified by key_idx to value offset. The track_idx must be the index of an Audio Track.

pub fn audio_track_set_key_stream(
&self,
track_idx: i64,
key_idx: i64,
stream: impl AsArg<Resource>
)

Sets the stream of the key identified by key_idx to value stream. The track_idx must be the index of an Audio Track.

pub fn bezier_track_get_key_in_handle(
&self,
track_idx: i64,
key_idx: i64
) -> Vector2

Returns the in handle of the key identified by key_idx. The track_idx must be the index of a Bezier Track.

pub fn bezier_track_get_key_out_handle(
&self,
track_idx: i64,
key_idx: i64
) -> Vector2

Returns the out handle of the key identified by key_idx. The track_idx must be the index of a Bezier Track.

pub fn bezier_track_get_key_value(&self, track_idx: i64, key_idx: i64) -> f64

Returns the value of the key identified by key_idx. The track_idx must be the index of a Bezier Track.

pub fn bezier_track_insert_key(
&self,
track_idx: i64,
time: f64,
value: f64,
in_handle: Vector2,
out_handle: Vector2
) -> i64

Inserts a Bezier Track key at the given time in seconds. The track_idx must be the index of a Bezier Track. in_handle is the left-side weight of the added Bezier curve point, out_handle is the right-side one, while value is the actual value at this point.

Default Arguments

in_handle - Vector2( 0, 0 )

out_handle - Vector2( 0, 0 )

pub fn bezier_track_interpolate(&self, track_idx: i64, time: f64) -> f64

Returns the interpolated value at the given time (in seconds). The track_idx must be the index of a Bezier Track.

pub fn bezier_track_set_key_in_handle(
&self,
track_idx: i64,
key_idx: i64,
in_handle: Vector2
)

Sets the in handle of the key identified by key_idx to value in_handle. The track_idx must be the index of a Bezier Track.

pub fn bezier_track_set_key_out_handle(
&self,
track_idx: i64,
key_idx: i64,
out_handle: Vector2
)

Sets the out handle of the key identified by key_idx to value out_handle. The track_idx must be the index of a Bezier Track.

pub fn bezier_track_set_key_value(
&self,
track_idx: i64,
key_idx: i64,
value: f64
)

Sets the value of the key identified by key_idx to the given value. The track_idx must be the index of a Bezier Track.

pub fn clear(&self)

Clear the animation (clear all tracks and reset all).

pub fn copy_track(&self, track_idx: i64, to_animation: impl AsArg<Animation>)

Adds a new track that is a copy of the given track from to_animation.

pub fn find_track(&self, path: impl Into<NodePath>) -> i64

Returns the index of the specified track. If the track is not found, return -1.

pub fn length(&self) -> f64

The total length of the animation (in seconds). Note: Length is not delimited by the last key, as this one may be before or after the end to ensure correct interpolation and looping.

pub fn step(&self) -> f64

The animation step value.

pub fn get_track_count(&self) -> i64

Returns the amount of tracks in the animation.

pub fn has_loop(&self) -> bool

A flag indicating that the animation must loop. This is used for correct interpolation of animation cycles, and for hinting the player that it must restart the animation.

pub fn method_track_get_key_indices(
&self,
track_idx: i64,
time_sec: f64,
delta: f64
) -> TypedArray<i32>

Returns all the key indices of a method track, given a position and delta time.

pub fn method_track_get_name(&self, track_idx: i64, key_idx: i64) -> GodotString

Returns the method name of a method track.

pub fn method_track_get_params(
&self,
track_idx: i64,
key_idx: i64
) -> VariantArray<Shared>

Returns the arguments values to be called on a method track for a given key in a given track.

pub fn remove_track(&self, track_idx: i64)

Removes a track by specifying the track index.

pub fn set_length(&self, time_sec: f64)

The total length of the animation (in seconds). Note: Length is not delimited by the last key, as this one may be before or after the end to ensure correct interpolation and looping.

pub fn set_loop(&self, enabled: bool)

A flag indicating that the animation must loop. This is used for correct interpolation of animation cycles, and for hinting the player that it must restart the animation.

pub fn set_step(&self, size_sec: f64)

The animation step value.

pub fn track_find_key(&self, track_idx: i64, time: f64, exact: bool) -> i64

Finds the key index by time in a given track. Optionally, only find it if the exact time is given.

Default Arguments

exact - false

pub fn track_get_interpolation_loop_wrap(&self, track_idx: i64) -> bool

Returns true if the track at idx wraps the interpolation loop. New tracks wrap the interpolation loop by default.

pub fn track_get_interpolation_type(&self, track_idx: i64) -> InterpolationType

Returns the interpolation type of a given track.

pub fn track_get_key_count(&self, track_idx: i64) -> i64

Returns the amount of keys in a given track.

pub fn track_get_key_time(&self, track_idx: i64, key_idx: i64) -> f64

Returns the time at which the key is located.

pub fn track_get_key_transition(&self, track_idx: i64, key_idx: i64) -> f64

Returns the transition curve (easing) for a specific key (see the built-in math function [method @GDScript.ease]).

pub fn track_get_key_value(&self, track_idx: i64, key_idx: i64) -> Variant

Returns the value of a given key in a given track.

pub fn track_get_path(&self, track_idx: i64) -> NodePath

Gets the path of a track. For more information on the path format, see track_set_path.

pub fn track_get_type(&self, track_idx: i64) -> TrackType

Gets the type of a track.

pub fn track_insert_key(
&self,
track_idx: i64,
time: f64,
key: impl OwnedToVariant,
transition: f64
)

Insert a generic key in a given track.

Default Arguments

transition - 1

pub fn track_is_enabled(&self, track_idx: i64) -> bool

Returns true if the track at index idx is enabled.

pub fn track_is_imported(&self, track_idx: i64) -> bool

Returns true if the given track is imported. Else, return false.

pub fn track_move_down(&self, track_idx: i64)

Moves a track down.

pub fn track_move_to(&self, track_idx: i64, to_idx: i64)

Changes the index position of track idx to the one defined in to_idx.

pub fn track_move_up(&self, track_idx: i64)

Moves a track up.

pub fn track_remove_key(&self, track_idx: i64, key_idx: i64)

Removes a key by index in a given track.

pub fn track_remove_key_at_position(&self, track_idx: i64, position: f64)

Removes a key by position (seconds) in a given track.

pub fn track_set_enabled(&self, track_idx: i64, enabled: bool)

Enables/disables the given track. Tracks are enabled by default.

pub fn track_set_imported(&self, track_idx: i64, imported: bool)

Sets the given track as imported or not.

pub fn track_set_interpolation_loop_wrap(
&self,
track_idx: i64,
interpolation: bool
)

If true, the track at idx wraps the interpolation loop.

pub fn track_set_interpolation_type(&self, track_idx: i64, interpolation: i64)

Sets the interpolation type of a given track.

pub fn track_set_key_time(&self, track_idx: i64, key_idx: i64, time: f64)

Sets the time of an existing key.

pub fn track_set_key_transition(
&self,
track_idx: i64,
key_idx: i64,
transition: f64
)

Sets the transition curve (easing) for a specific key (see the built-in math function [method @GDScript.ease]).

pub fn track_set_key_value(
&self,
track_idx: i64,
key: i64,
value: impl OwnedToVariant
)

Sets the value of an existing key.

pub fn track_set_path(&self, track_idx: i64, path: impl Into<NodePath>)

Sets the path of a track. Paths must be valid scene-tree paths to a node, and must be specified starting from the parent node of the node that will reproduce the animation. Tracks that control properties or bones must append their name after the path, separated by ":". For example, "character/skeleton:ankle" or "character/mesh:transform/local".

pub fn track_swap(&self, track_idx: i64, with_idx: i64)

Swaps the track idx’s index position with the track with_idx.

pub fn transform_track_insert_key(
&self,
track_idx: i64,
time: f64,
location: Vector3,
rotation: Quat,
scale: Vector3
) -> i64

Insert a transform key for a transform track.

pub fn transform_track_interpolate(
&self,
track_idx: i64,
time_sec: f64
) -> VariantArray<Shared>

Returns the interpolated value of a transform track at a given time (in seconds). An array consisting of 3 elements: position (Vector3), rotation (Quat) and scale (Vector3).

pub fn value_track_get_key_indices(
&self,
track_idx: i64,
time_sec: f64,
delta: f64
) -> TypedArray<i32>

Returns all the key indices of a value track, given a position and delta time.

pub fn value_track_get_update_mode(&self, track_idx: i64) -> UpdateMode

Returns the update mode of a value track.

pub fn value_track_interpolate(&self, track_idx: i64, time_sec: f64) -> Variant

Returns the interpolated value at the given time (in seconds). The track_idx must be the index of a value track.

pub fn value_track_set_update_mode(&self, track_idx: i64, mode: i64)

Sets the update mode (see UpdateMode) of a value track.