Struct gdnative::api::Area2D

pub struct Area2D { /* fields omitted */ }

core class Area2D inherits CollisionObject2D (manually managed).

This class has related types in the area_2d module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

Area2D is a reference-only type. Persistent references can only exist in the unsafe Ref<Area2D> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

Area2D inherits methods from:

CollisionObject2D

Node2D

CanvasItem

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Area2D

Constants

pub const SPACE_OVERRIDE_DISABLED: i64

pub const SPACE_OVERRIDE_COMBINE: i64

pub const SPACE_OVERRIDE_COMBINE_REPLACE: i64

pub const SPACE_OVERRIDE_REPLACE: i64

pub const SPACE_OVERRIDE_REPLACE_COMBINE: i64

impl Area2D

pub fn new() -> Ref<Area2D, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn angular_damp(&self) -> f64

The rate at which objects stop spinning in this area. Represents the angular velocity lost per second. See [member ProjectSettings.physics/2d/default_angular_damp] for more details about damping.

pub fn audio_bus_name(&self) -> GodotString

The name of the area’s audio bus.

pub fn collision_layer(&self) -> i64

The area’s physics layer(s). Collidable objects can exist in any of 32 different layers. A contact is detected if object A is in any of the layers that object B scans, or object B is in any layers that object A scans. See also collision_mask. See Collision layers and masks in the documentation for more information.

pub fn get_collision_layer_bit(&self, bit: i64) -> bool

Returns an individual bit on the layer mask. Describes whether other areas will collide with this one on the given layer.

pub fn collision_mask(&self) -> i64

The physics layers this area scans to determine collision detection. See Collision layers and masks in the documentation for more information.

pub fn get_collision_mask_bit(&self, bit: i64) -> bool

Returns an individual bit on the collision mask. Describes whether this area will collide with others on the given layer.

pub fn gravity(&self) -> f64

The area’s gravity intensity (ranges from -1024 to 1024). This value multiplies the gravity vector. This is useful to alter the force of gravity without altering its direction.

pub fn gravity_distance_scale(&self) -> f64

The falloff factor for point gravity. The greater the value, the faster gravity decreases with distance.

pub fn gravity_vector(&self) -> Vector2

The area’s gravity vector (not normalized). If gravity is a point (see [gravity_point][Self::gravity_point]), this will be the point of attraction.

pub fn linear_damp(&self) -> f64

The rate at which objects stop moving in this area. Represents the linear velocity lost per second. See [member ProjectSettings.physics/2d/default_linear_damp] for more details about damping.

pub fn get_overlapping_areas(&self) -> VariantArray<Shared>

Returns a list of intersecting Area2Ds. For performance reasons (collisions are all processed at the same time) this list is modified once during the physics step, not immediately after objects are moved. Consider using signals instead.

pub fn get_overlapping_bodies(&self) -> VariantArray<Shared>

Returns a list of intersecting PhysicsBody2Ds. For performance reasons (collisions are all processed at the same time) this list is modified once during the physics step, not immediately after objects are moved. Consider using signals instead.

pub fn priority(&self) -> f64

The area’s priority. Higher priority areas are processed first.

pub fn space_override_mode(&self) -> SpaceOverride

Override mode for gravity and damping calculations within this area. See SpaceOverride for possible values.

pub fn is_gravity_a_point(&self) -> bool

If true, gravity is calculated from a point (set via [gravity_vec][Self::gravity_vec]). See also [space_override][Self::space_override].

pub fn is_monitorable(&self) -> bool

If true, other monitoring areas can detect this area.

pub fn is_monitoring(&self) -> bool

If true, the area detects bodies or areas entering and exiting it.

pub fn is_overriding_audio_bus(&self) -> bool

If true, the area’s audio bus overrides the default audio bus.

pub fn overlaps_area(&self, area: impl AsArg<Node>) -> bool

If true, the given area overlaps the Area2D. Note: The result of this test is not immediate after moving objects. For performance, list of overlaps is updated once per frame and before the physics step. Consider using signals instead.

pub fn overlaps_body(&self, body: impl AsArg<Node>) -> bool

If true, the given physics body overlaps the Area2D. Note: The result of this test is not immediate after moving objects. For performance, list of overlaps is updated once per frame and before the physics step. Consider using signals instead. The body argument can either be a PhysicsBody2D or a TileMap instance (while TileMaps are not physics body themselves, they register their tiles with collision shapes as a virtual physics body).

pub fn set_angular_damp(&self, angular_damp: f64)

The rate at which objects stop spinning in this area. Represents the angular velocity lost per second. See [member ProjectSettings.physics/2d/default_angular_damp] for more details about damping.

pub fn set_audio_bus_name(&self, name: impl Into<GodotString>)

The name of the area’s audio bus.

pub fn set_audio_bus_override(&self, enable: bool)

If true, the area’s audio bus overrides the default audio bus.

pub fn set_collision_layer(&self, collision_layer: i64)

The area’s physics layer(s). Collidable objects can exist in any of 32 different layers. A contact is detected if object A is in any of the layers that object B scans, or object B is in any layers that object A scans. See also collision_mask. See Collision layers and masks in the documentation for more information.

pub fn set_collision_layer_bit(&self, bit: i64, value: bool)

Set/clear individual bits on the layer mask. This makes getting an area in/out of only one layer easier.

pub fn set_collision_mask(&self, collision_mask: i64)

The physics layers this area scans to determine collision detection. See Collision layers and masks in the documentation for more information.

pub fn set_collision_mask_bit(&self, bit: i64, value: bool)

Set/clear individual bits on the collision mask. This makes selecting the areas scanned easier.

pub fn set_gravity(&self, gravity: f64)

The area’s gravity intensity (ranges from -1024 to 1024). This value multiplies the gravity vector. This is useful to alter the force of gravity without altering its direction.

pub fn set_gravity_distance_scale(&self, distance_scale: f64)

The falloff factor for point gravity. The greater the value, the faster gravity decreases with distance.

pub fn set_gravity_is_point(&self, enable: bool)

If true, gravity is calculated from a point (set via [gravity_vec][Self::gravity_vec]). See also [space_override][Self::space_override].

pub fn set_gravity_vector(&self, vector: Vector2)

The area’s gravity vector (not normalized). If gravity is a point (see [gravity_point][Self::gravity_point]), this will be the point of attraction.

pub fn set_linear_damp(&self, linear_damp: f64)

The rate at which objects stop moving in this area. Represents the linear velocity lost per second. See [member ProjectSettings.physics/2d/default_linear_damp] for more details about damping.

pub fn set_monitorable(&self, enable: bool)

If true, other monitoring areas can detect this area.

pub fn set_monitoring(&self, enable: bool)

If true, the area detects bodies or areas entering and exiting it.

pub fn set_priority(&self, priority: f64)

The area’s priority. Higher priority areas are processed first.

pub fn set_space_override_mode(&self, space_override_mode: i64)

Override mode for gravity and damping calculations within this area. See SpaceOverride for possible values.