Struct gdnative::api::BakedLightmap

pub struct BakedLightmap { /* fields omitted */ }

core class BakedLightmap inherits VisualInstance (unsafe).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non reference counted objects such as the ones of this type are usually owned by the engine.

BakedLightmap is a reference-only type. Persistent references can only exist in the unsafe Ref<BakedLightmap> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

BakedLightmap inherits methods from:

VisualInstance

Spatial

Node

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl BakedLightmap

Constants

pub const BAKE_ERROR_OK: i64

pub const BAKE_QUALITY_LOW: i64

pub const ENVIRONMENT_MODE_DISABLED: i64

pub const BAKE_ERROR_NO_SAVE_PATH: i64

pub const BAKE_QUALITY_MEDIUM: i64

pub const ENVIRONMENT_MODE_SCENE: i64

pub const BAKE_ERROR_NO_MESHES: i64

pub const BAKE_QUALITY_HIGH: i64

pub const ENVIRONMENT_MODE_CUSTOM_SKY: i64

pub const BAKE_ERROR_CANT_CREATE_IMAGE: i64

pub const BAKE_QUALITY_ULTRA: i64

pub const ENVIRONMENT_MODE_CUSTOM_COLOR: i64

pub const BAKE_ERROR_LIGHTMAP_SIZE: i64

pub const BAKE_ERROR_INVALID_MESH: i64

pub const BAKE_ERROR_USER_ABORTED: i64

pub const BAKE_ERROR_NO_LIGHTMAPPER: i64

impl BakedLightmap

pub fn new() -> Ref<BakedLightmap, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn bake(
&self,
from_node: impl AsArg<Node>,
data_save_path: impl Into<GodotString>
) -> BakeError

Bakes the lightmap, scanning from the given from_node root and saves the resulting BakedLightmapData in data_save_path. If no save path is provided it will try to match the path from the current light_data.

Default Arguments

from_node - null

data_save_path - ""

pub fn bake_quality(&self) -> BakeQuality

Determines the amount of samples per texel used in indrect light baking. The amount of samples for each quality level can be configured in the project settings.

pub fn bias(&self) -> f64

Raycasting bias used during baking to avoid floating point precission issues.

pub fn bounces(&self) -> i64

Number of light bounces that are taken into account during baking.

pub fn capture_cell_size(&self) -> f64

Grid size used for real-time capture information on dynamic objects.

pub fn capture_enabled(&self) -> bool

When enabled, an octree containing the scene’s lighting information will be computed. This octree will then be used to light dynamic objects in the scene.

pub fn capture_propagation(&self) -> f64

Bias value to reduce the amount of light proagation in the captured octree.

pub fn capture_quality(&self) -> BakeQuality

Bake quality of the capture data.

pub fn default_texels_per_unit(&self) -> f64

If a baked mesh doesn’t have a UV2 size hint, this value will be used to roughly compute a suitable lightmap size.

pub fn environment_custom_color(&self) -> Color

The environment color when environment_mode is set to ENVIRONMENT_MODE_CUSTOM_COLOR.

pub fn environment_custom_energy(&self) -> f64

The energy scaling factor when when environment_mode is set to ENVIRONMENT_MODE_CUSTOM_COLOR or ENVIRONMENT_MODE_CUSTOM_SKY.

pub fn environment_custom_sky(&self) -> Option<Ref<Sky, Shared>>

The Sky resource to use when environment_mode is set o ENVIRONMENT_MODE_CUSTOM_SKY.

pub fn environment_custom_sky_rotation_degrees(&self) -> Vector3

The rotation of the baked custom sky.

pub fn environment_min_light(&self) -> Color

Minimum ambient light for all the lightmap texels. This doesn’t take into account any occlusion from the scene’s geometry, it simply ensures a minimum amount of light on all the lightmap texels. Can be used for artistic control on shadow color.

pub fn environment_mode(&self) -> EnvironmentMode

Decides which environment to use during baking.

pub fn extents(&self) -> Vector3

Size of the baked lightmap. Only meshes inside this region will be included in the baked lightmap, also used as the bounds of the captured region for dynamic lighting.

pub fn image_path(&self) -> GodotString

Deprecated, in previous versions it determined the location where lightmaps were be saved.

pub fn light_data(&self) -> Option<Ref<BakedLightmapData, Shared>>

The calculated light data.

pub fn max_atlas_size(&self) -> i64

Maximum size of each lightmap layer, only used when [atlas_generate][Self::atlas_generate] is enabled.

pub fn is_generate_atlas_enabled(&self) -> bool

When enabled, the lightmapper will merge the textures for all meshes into a single large layered texture. Not supported in GLES2.

pub fn is_using_color(&self) -> bool

Store full color values in the lightmap textures. When disabled, lightmap textures will store a single brightness channel. Can be disabled to reduce disk usage if the scene contains only white lights or you don’t mind losing color information in indirect lighting.

pub fn is_using_denoiser(&self) -> bool

When enabled, a lightmap denoiser will be used to reduce the noise inherent to Monte Carlo based global illumination.

pub fn is_using_hdr(&self) -> bool

If true, stores the lightmap textures in a high dynamic range format (EXR). If false, stores the lightmap texture in a low dynamic range PNG image. This can be set to false to reduce disk usage, but light values over 1.0 will be clamped and you may see banding caused by the reduced precision. Note: Setting [use_hdr][Self::use_hdr] to true will decrease lightmap banding even when using the GLES2 backend or if [member ProjectSettings.rendering/quality/depth/hdr] is false.

pub fn set_bake_quality(&self, quality: i64)

Determines the amount of samples per texel used in indrect light baking. The amount of samples for each quality level can be configured in the project settings.

pub fn set_bias(&self, bias: f64)

Raycasting bias used during baking to avoid floating point precission issues.

pub fn set_bounces(&self, bounces: i64)

Number of light bounces that are taken into account during baking.

pub fn set_capture_cell_size(&self, capture_cell_size: f64)

Grid size used for real-time capture information on dynamic objects.

pub fn set_capture_enabled(&self, enabled: bool)

When enabled, an octree containing the scene’s lighting information will be computed. This octree will then be used to light dynamic objects in the scene.

pub fn set_capture_propagation(&self, propagation: f64)

Bias value to reduce the amount of light proagation in the captured octree.

pub fn set_capture_quality(&self, capture_quality: i64)

Bake quality of the capture data.

pub fn set_default_texels_per_unit(&self, texels: f64)

If a baked mesh doesn’t have a UV2 size hint, this value will be used to roughly compute a suitable lightmap size.

pub fn set_environment_custom_color(&self, color: Color)

The environment color when environment_mode is set to ENVIRONMENT_MODE_CUSTOM_COLOR.

pub fn set_environment_custom_energy(&self, energy: f64)

The energy scaling factor when when environment_mode is set to ENVIRONMENT_MODE_CUSTOM_COLOR or ENVIRONMENT_MODE_CUSTOM_SKY.

pub fn set_environment_custom_sky(&self, sky: impl AsArg<Sky>)

The Sky resource to use when environment_mode is set o ENVIRONMENT_MODE_CUSTOM_SKY.

pub fn set_environment_custom_sky_rotation_degrees(&self, rotation: Vector3)

The rotation of the baked custom sky.

pub fn set_environment_min_light(&self, min_light: Color)

Minimum ambient light for all the lightmap texels. This doesn’t take into account any occlusion from the scene’s geometry, it simply ensures a minimum amount of light on all the lightmap texels. Can be used for artistic control on shadow color.

pub fn set_environment_mode(&self, mode: i64)

Decides which environment to use during baking.

pub fn set_extents(&self, extents: Vector3)

Size of the baked lightmap. Only meshes inside this region will be included in the baked lightmap, also used as the bounds of the captured region for dynamic lighting.

pub fn set_generate_atlas(&self, enabled: bool)

When enabled, the lightmapper will merge the textures for all meshes into a single large layered texture. Not supported in GLES2.

pub fn set_image_path(&self, image_path: impl Into<GodotString>)

Deprecated, in previous versions it determined the location where lightmaps were be saved.

pub fn set_light_data(&self, data: impl AsArg<BakedLightmapData>)

The calculated light data.

pub fn set_max_atlas_size(&self, max_atlas_size: i64)

Maximum size of each lightmap layer, only used when [atlas_generate][Self::atlas_generate] is enabled.

pub fn set_use_color(&self, use_denoiser: bool)

Store full color values in the lightmap textures. When disabled, lightmap textures will store a single brightness channel. Can be disabled to reduce disk usage if the scene contains only white lights or you don’t mind losing color information in indirect lighting.

pub fn set_use_denoiser(&self, use_denoiser: bool)

When enabled, a lightmap denoiser will be used to reduce the noise inherent to Monte Carlo based global illumination.

pub fn set_use_hdr(&self, use_denoiser: bool)

If true, stores the lightmap textures in a high dynamic range format (EXR). If false, stores the lightmap texture in a low dynamic range PNG image. This can be set to false to reduce disk usage, but light values over 1.0 will be clamped and you may see banding caused by the reduced precision. Note: Setting [use_hdr][Self::use_hdr] to true will decrease lightmap banding even when using the GLES2 backend or if [member ProjectSettings.rendering/quality/depth/hdr] is false.