Struct gdnative::api::BulletPhysicsServer

pub struct BulletPhysicsServer { /* private fields */ }

core class BulletPhysicsServer inherits PhysicsServer (manually managed).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Class hierarchy

BulletPhysicsServer inherits methods from:

PhysicsServer

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.