Struct gdnative::api::CPUParticles

pub struct CPUParticles { /* fields omitted */ }

core class CPUParticles inherits GeometryInstance (manually managed).

This class has related types in the cpu_particles module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

CPUParticles is a reference-only type. Persistent references can only exist in the unsafe Ref<CPUParticles> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

CPUParticles inherits methods from:

GeometryInstance

VisualInstance

Spatial

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl CPUParticles

Constants

pub const DRAW_ORDER_INDEX: i64

pub const EMISSION_SHAPE_POINT: i64

pub const FLAG_ALIGN_Y_TO_VELOCITY: i64

pub const PARAM_INITIAL_LINEAR_VELOCITY: i64

pub const DRAW_ORDER_LIFETIME: i64

pub const EMISSION_SHAPE_SPHERE: i64

pub const FLAG_ROTATE_Y: i64

pub const PARAM_ANGULAR_VELOCITY: i64

pub const DRAW_ORDER_VIEW_DEPTH: i64

pub const EMISSION_SHAPE_BOX: i64

pub const FLAG_DISABLE_Z: i64

pub const PARAM_ORBIT_VELOCITY: i64

pub const EMISSION_SHAPE_POINTS: i64

pub const FLAG_MAX: i64

pub const PARAM_LINEAR_ACCEL: i64

pub const EMISSION_SHAPE_DIRECTED_POINTS: i64

pub const PARAM_RADIAL_ACCEL: i64

pub const EMISSION_SHAPE_MAX: i64

pub const PARAM_TANGENTIAL_ACCEL: i64

pub const PARAM_DAMPING: i64

pub const PARAM_ANGLE: i64

pub const PARAM_SCALE: i64

pub const PARAM_HUE_VARIATION: i64

pub const PARAM_ANIM_SPEED: i64

pub const PARAM_ANIM_OFFSET: i64

pub const PARAM_MAX: i64

impl CPUParticles

pub fn new() -> Ref<CPUParticles, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn convert_from_particles(&self, particles: impl AsArg<Node>)

Sets this node’s properties to match a given Particles node with an assigned ParticlesMaterial.

pub fn amount(&self) -> i64

The number of particles emitted in one emission cycle (corresponding to the lifetime). Note: Changing amount will reset the particle emission, therefore removing all particles that were already emitted before changing amount.

pub fn color(&self) -> Color

Unused for 3D particles.

pub fn color_ramp(&self) -> Option<Ref<Gradient, Shared>>

Unused for 3D particles.

pub fn direction(&self) -> Vector3

Unit vector specifying the particles’ emission direction.

pub fn draw_order(&self) -> DrawOrder

Particle draw order. Uses DrawOrder values.

pub fn emission_box_extents(&self) -> Vector3

The rectangle’s extents if emission_shape is set to EMISSION_SHAPE_BOX.

pub fn emission_colors(&self) -> PoolArray<Color>

Sets the Colors to modulate particles by when using EMISSION_SHAPE_POINTS or EMISSION_SHAPE_DIRECTED_POINTS.

pub fn emission_normals(&self) -> PoolArray<Vector3>

Sets the direction the particles will be emitted in when using EMISSION_SHAPE_DIRECTED_POINTS.

pub fn emission_points(&self) -> PoolArray<Vector3>

Sets the initial positions to spawn particles when using EMISSION_SHAPE_POINTS or EMISSION_SHAPE_DIRECTED_POINTS.

pub fn emission_shape(&self) -> EmissionShape

Particles will be emitted inside this region. See EmissionShape for possible values.

pub fn emission_sphere_radius(&self) -> f64

The sphere’s radius if EmissionShape is set to EMISSION_SHAPE_SPHERE.

pub fn explosiveness_ratio(&self) -> f64

How rapidly particles in an emission cycle are emitted. If greater than 0, there will be a gap in emissions before the next cycle begins.

pub fn fixed_fps(&self) -> i64

The particle system’s frame rate is fixed to a value. For instance, changing the value to 2 will make the particles render at 2 frames per second. Note this does not slow down the particle system itself.

pub fn flatness(&self) -> f64

Amount of spread in Y/Z plane. A value of 1 restricts particles to X/Z plane.

pub fn fractional_delta(&self) -> bool

If true, results in fractional delta calculation which has a smoother particles display effect.

pub fn gravity(&self) -> Vector3

Gravity applied to every particle.

pub fn lifetime(&self) -> f64

The amount of time each particle will exist (in seconds).

pub fn lifetime_randomness(&self) -> f64

Particle lifetime randomness ratio.

pub fn mesh(&self) -> Option<Ref<Mesh, Shared>>

The Mesh used for each particle. If null, particles will be spheres.

pub fn one_shot(&self) -> bool

If true, only one emission cycle occurs. If set true during a cycle, emission will stop at the cycle’s end.

pub fn param(&self, param: i64) -> f64

Tangential acceleration applied to each particle. Tangential acceleration is perpendicular to the particle’s velocity giving the particles a swirling motion.

pub fn param_curve(&self, param: i64) -> Option<Ref<Curve, Shared>>

Each particle’s tangential acceleration will vary along this Curve.

pub fn param_randomness(&self, param: i64) -> f64

Tangential acceleration randomness ratio.

pub fn particle_flag(&self, flag: i64) -> bool

If true, particles rotate around Y axis by [angle][Self::angle].

pub fn pre_process_time(&self) -> f64

Particle system starts as if it had already run for this many seconds.

pub fn randomness_ratio(&self) -> f64

Emission lifetime randomness ratio.

pub fn speed_scale(&self) -> f64

Particle system’s running speed scaling ratio. A value of 0 can be used to pause the particles.

pub fn spread(&self) -> f64

Each particle’s initial direction range from +spread to -spread degrees. Applied to X/Z plane and Y/Z planes.

pub fn use_local_coordinates(&self) -> bool

If true, particles use the parent node’s coordinate space. If false, they use global coordinates.

pub fn is_emitting(&self) -> bool

If true, particles are being emitted.

pub fn restart(&self)

Restarts the particle emitter.

pub fn set_amount(&self, amount: i64)

The number of particles emitted in one emission cycle (corresponding to the lifetime). Note: Changing amount will reset the particle emission, therefore removing all particles that were already emitted before changing amount.

pub fn set_color(&self, color: Color)

Unused for 3D particles.

pub fn set_color_ramp(&self, ramp: impl AsArg<Gradient>)

Unused for 3D particles.

pub fn set_direction(&self, direction: Vector3)

Unit vector specifying the particles’ emission direction.

pub fn set_draw_order(&self, order: i64)

Particle draw order. Uses DrawOrder values.

pub fn set_emission_box_extents(&self, extents: Vector3)

The rectangle’s extents if emission_shape is set to EMISSION_SHAPE_BOX.

pub fn set_emission_colors(&self, array: PoolArray<Color>)

Sets the Colors to modulate particles by when using EMISSION_SHAPE_POINTS or EMISSION_SHAPE_DIRECTED_POINTS.

pub fn set_emission_normals(&self, array: PoolArray<Vector3>)

Sets the direction the particles will be emitted in when using EMISSION_SHAPE_DIRECTED_POINTS.

pub fn set_emission_points(&self, array: PoolArray<Vector3>)

Sets the initial positions to spawn particles when using EMISSION_SHAPE_POINTS or EMISSION_SHAPE_DIRECTED_POINTS.

pub fn set_emission_shape(&self, shape: i64)

Particles will be emitted inside this region. See EmissionShape for possible values.

pub fn set_emission_sphere_radius(&self, radius: f64)

The sphere’s radius if EmissionShape is set to EMISSION_SHAPE_SPHERE.

pub fn set_emitting(&self, emitting: bool)

If true, particles are being emitted.

pub fn set_explosiveness_ratio(&self, ratio: f64)

How rapidly particles in an emission cycle are emitted. If greater than 0, there will be a gap in emissions before the next cycle begins.

pub fn set_fixed_fps(&self, fps: i64)

The particle system’s frame rate is fixed to a value. For instance, changing the value to 2 will make the particles render at 2 frames per second. Note this does not slow down the particle system itself.

pub fn set_flatness(&self, amount: f64)

Amount of spread in Y/Z plane. A value of 1 restricts particles to X/Z plane.

pub fn set_fractional_delta(&self, enable: bool)

If true, results in fractional delta calculation which has a smoother particles display effect.

pub fn set_gravity(&self, accel_vec: Vector3)

Gravity applied to every particle.

pub fn set_lifetime(&self, secs: f64)

The amount of time each particle will exist (in seconds).

pub fn set_lifetime_randomness(&self, random: f64)

Particle lifetime randomness ratio.

pub fn set_mesh(&self, mesh: impl AsArg<Mesh>)

The Mesh used for each particle. If null, particles will be spheres.

pub fn set_one_shot(&self, enable: bool)

If true, only one emission cycle occurs. If set true during a cycle, emission will stop at the cycle’s end.

pub fn set_param(&self, param: i64, value: f64)

Tangential acceleration applied to each particle. Tangential acceleration is perpendicular to the particle’s velocity giving the particles a swirling motion.

pub fn set_param_curve(&self, param: i64, curve: impl AsArg<Curve>)

Each particle’s tangential acceleration will vary along this Curve.

pub fn set_param_randomness(&self, param: i64, randomness: f64)

Tangential acceleration randomness ratio.

pub fn set_particle_flag(&self, flag: i64, enable: bool)

If true, particles rotate around Y axis by [angle][Self::angle].

pub fn set_pre_process_time(&self, secs: f64)

Particle system starts as if it had already run for this many seconds.

pub fn set_randomness_ratio(&self, ratio: f64)

Emission lifetime randomness ratio.

pub fn set_speed_scale(&self, scale: f64)

Particle system’s running speed scaling ratio. A value of 0 can be used to pause the particles.

pub fn set_spread(&self, degrees: f64)

Each particle’s initial direction range from +spread to -spread degrees. Applied to X/Z plane and Y/Z planes.

pub fn set_use_local_coordinates(&self, enable: bool)

If true, particles use the parent node’s coordinate space. If false, they use global coordinates.