Struct gdnative::api::CameraTexture

pub struct CameraTexture { /* fields omitted */ }

core class CameraTexture inherits Texture (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

CameraTexture inherits methods from:

Texture

Resource

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl CameraTexture

pub fn new() -> Ref<CameraTexture, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn camera_active(&self) -> bool

Convenience property that gives access to the active property of the CameraFeed.

pub fn camera_feed_id(&self) -> i64

The ID of the CameraFeed for which we want to display the image.

pub fn which_feed(&self) -> FeedImage

Which image within the CameraFeed we want access to, important if the camera image is split in a Y and CbCr component.

pub fn set_camera_active(&self, active: bool)

Convenience property that gives access to the active property of the CameraFeed.

pub fn set_camera_feed_id(&self, feed_id: i64)

The ID of the CameraFeed for which we want to display the image.

pub fn set_which_feed(&self, which_feed: i64)

Which image within the CameraFeed we want access to, important if the camera image is split in a Y and CbCr component.