Struct gdnative::api::ClippedCamera

pub struct ClippedCamera { /* fields omitted */ }

core class ClippedCamera inherits Camera (unsafe).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non reference counted objects such as the ones of this type are usually owned by the engine.

ClippedCamera is a reference-only type. Persistent references can only exist in the unsafe Ref<ClippedCamera> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

ClippedCamera inherits methods from:

Camera

Spatial

Node

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl ClippedCamera

Constants

pub const CLIP_PROCESS_PHYSICS: i64

pub const CLIP_PROCESS_IDLE: i64

impl ClippedCamera

pub fn new() -> Ref<ClippedCamera, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn add_exception(&self, node: impl AsArg<Object>)

Adds a collision exception so the camera does not collide with the specified node.

pub fn add_exception_rid(&self, rid: Rid)

Adds a collision exception so the camera does not collide with the specified RID.

pub fn clear_exceptions(&self)

Removes all collision exceptions.

pub fn get_clip_offset(&self) -> f64

Returns the distance the camera has been offset due to a collision.

pub fn collision_mask(&self) -> i64

The camera’s collision mask. Only objects in at least one collision layer matching the mask will be detected. See Collision layers and masks in the documentation for more information.

pub fn get_collision_mask_bit(&self, bit: i64) -> bool

Returns true if the specified bit index is on. Note: Bit indices range from 0-19.

pub fn margin(&self) -> f64

The camera’s collision margin. The camera can’t get closer than this distance to a colliding object.

pub fn process_mode(&self) -> ProcessMode

The camera’s process callback. See ProcessMode.

pub fn is_clip_to_areas_enabled(&self) -> bool

If true, the camera stops on contact with Areas.

pub fn is_clip_to_bodies_enabled(&self) -> bool

If true, the camera stops on contact with PhysicsBodys.

pub fn remove_exception(&self, node: impl AsArg<Object>)

Removes a collision exception with the specified node.

pub fn remove_exception_rid(&self, rid: Rid)

Removes a collision exception with the specified RID.

pub fn set_clip_to_areas(&self, enable: bool)

If true, the camera stops on contact with Areas.

pub fn set_clip_to_bodies(&self, enable: bool)

If true, the camera stops on contact with PhysicsBodys.

pub fn set_collision_mask(&self, mask: i64)

The camera’s collision mask. Only objects in at least one collision layer matching the mask will be detected. See Collision layers and masks in the documentation for more information.

pub fn set_collision_mask_bit(&self, bit: i64, value: bool)

Sets the specified bit index to the value. Note: Bit indices range from 0-19.

pub fn set_margin(&self, margin: f64)

The camera’s collision margin. The camera can’t get closer than this distance to a colliding object.

pub fn set_process_mode(&self, process_mode: i64)

The camera’s process callback. See ProcessMode.