Struct gdnative::api::DampedSpringJoint2D

pub struct DampedSpringJoint2D { /* fields omitted */ }

core class DampedSpringJoint2D inherits Joint2D (manually managed).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

DampedSpringJoint2D is a reference-only type. Persistent references can only exist in the unsafe Ref<DampedSpringJoint2D> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

DampedSpringJoint2D inherits methods from:

Joint2D

Node2D

CanvasItem

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl DampedSpringJoint2D

pub fn new() -> Ref<DampedSpringJoint2D, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn damping(&self) -> f64

The spring joint’s damping ratio. A value between 0 and 1. When the two bodies move into different directions the system tries to align them to the spring axis again. A high damping value forces the attached bodies to align faster.

pub fn length(&self) -> f64

The spring joint’s maximum length. The two attached bodies cannot stretch it past this value.

pub fn rest_length(&self) -> f64

When the bodies attached to the spring joint move they stretch or squash it. The joint always tries to resize towards this length.

pub fn stiffness(&self) -> f64

The higher the value, the less the bodies attached to the joint will deform it. The joint applies an opposing force to the bodies, the product of the stiffness multiplied by the size difference from its resting length.

pub fn set_damping(&self, damping: f64)

The spring joint’s damping ratio. A value between 0 and 1. When the two bodies move into different directions the system tries to align them to the spring axis again. A high damping value forces the attached bodies to align faster.

pub fn set_length(&self, length: f64)

The spring joint’s maximum length. The two attached bodies cannot stretch it past this value.

pub fn set_rest_length(&self, rest_length: f64)

When the bodies attached to the spring joint move they stretch or squash it. The joint always tries to resize towards this length.

pub fn set_stiffness(&self, stiffness: f64)

The higher the value, the less the bodies attached to the joint will deform it. The joint applies an opposing force to the bodies, the product of the stiffness multiplied by the size difference from its resting length.