Struct gdnative::api::DynamicFont

pub struct DynamicFont { /* fields omitted */ }

core class DynamicFont inherits Font (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

DynamicFont inherits methods from:

Font

Resource

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl DynamicFont

Constants

pub const SPACING_TOP: i64

pub const SPACING_BOTTOM: i64

pub const SPACING_CHAR: i64

pub const SPACING_SPACE: i64

impl DynamicFont

pub fn new() -> Ref<DynamicFont, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn add_fallback(&self, data: impl AsArg<DynamicFontData>)

Adds a fallback font.

pub fn get_available_chars(&self) -> GodotString

Returns a string containing all the characters available in the main and all the fallback fonts. If a given character is included in more than one font, it appears only once in the returned string.

pub fn get_fallback(&self, idx: i64) -> Option<Ref<DynamicFontData, Shared>>

Returns the fallback font at index idx.

pub fn get_fallback_count(&self) -> i64

Returns the number of fallback fonts.

pub fn font_data(&self) -> Option<Ref<DynamicFontData, Shared>>

The font data.

pub fn outline_color(&self) -> Color

The font outline’s color. Note: It’s recommended to leave this at the default value so that you can adjust it in individual controls. For example, if the outline is made black here, it won’t be possible to change its color using a Label’s font outline modulate theme item.

pub fn outline_size(&self) -> i64

The font outline’s thickness in pixels (not relative to the font size).

pub fn size(&self) -> i64

The font size in pixels.

pub fn spacing(&self, _type: i64) -> i64

Extra spacing at the top in pixels.

pub fn use_filter(&self) -> bool

If true, filtering is used. This makes the font blurry instead of pixelated when scaling it if font oversampling is disabled or ineffective. It’s recommended to enable this when using the font in a control whose size changes over time, unless a pixel art aesthetic is desired.

pub fn use_mipmaps(&self) -> bool

If true, mipmapping is used. This improves the font’s appearance when downscaling it if font oversampling is disabled or ineffective.

pub fn remove_fallback(&self, idx: i64)

Removes the fallback font at index idx.

pub fn set_fallback(&self, idx: i64, data: impl AsArg<DynamicFontData>)

Sets the fallback font at index idx.

pub fn set_font_data(&self, data: impl AsArg<DynamicFontData>)

The font data.

pub fn set_outline_color(&self, color: Color)

The font outline’s color. Note: It’s recommended to leave this at the default value so that you can adjust it in individual controls. For example, if the outline is made black here, it won’t be possible to change its color using a Label’s font outline modulate theme item.

pub fn set_outline_size(&self, size: i64)

The font outline’s thickness in pixels (not relative to the font size).

pub fn set_size(&self, data: i64)

The font size in pixels.

pub fn set_spacing(&self, _type: i64, value: i64)

Extra spacing at the top in pixels.

pub fn set_use_filter(&self, enable: bool)

If true, filtering is used. This makes the font blurry instead of pixelated when scaling it if font oversampling is disabled or ineffective. It’s recommended to enable this when using the font in a control whose size changes over time, unless a pixel art aesthetic is desired.

pub fn set_use_mipmaps(&self, enable: bool)

If true, mipmapping is used. This improves the font’s appearance when downscaling it if font oversampling is disabled or ineffective.