Struct gdnative::api::EditorInterface

pub struct EditorInterface { /* fields omitted */ }

tools class EditorInterface inherits Node (manually managed).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Class hierarchy

EditorInterface inherits methods from:

Node

Object

Tool

This class is used to interact with Godot’s editor.

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl EditorInterface

pub fn edit_node(&self, node: impl AsArg<Node>)

Edits the given Node. The node will be also selected if it’s inside the scene tree.

pub fn edit_resource(&self, resource: impl AsArg<Resource>)

Edits the given Resource.

pub fn get_base_control(&self) -> Option<Ref<Control, Shared>>

Returns the main container of Godot editor’s window. For example, you can use it to retrieve the size of the container and place your controls accordingly.

pub fn get_current_path(&self) -> GodotString

Returns the current path being viewed in the FileSystemDock.

pub fn get_edited_scene_root(&self) -> Option<Ref<Node, Shared>>

Returns the edited (current) scene’s root Node.

pub fn get_editor_scale(&self) -> f64

Returns the actual scale of the editor UI (1.0 being 100% scale). This can be used to adjust position and dimensions of the UI added by plugins. Note: This value is set via the interface/editor/display_scale and interface/editor/custom_display_scale editor settings. Editor must be restarted for changes to be properly applied.

pub fn get_editor_settings(&self) -> Option<Ref<EditorSettings, Shared>>

Returns the editor’s EditorSettings instance.

pub fn get_editor_viewport(&self) -> Option<Ref<Control, Shared>>

Returns the main editor control. Use this as a parent for main screens. Note: This returns the main editor control containing the whole editor, not the 2D or 3D viewports specifically.

pub fn get_file_system_dock(&self) -> Option<Ref<FileSystemDock, Shared>>

Returns the editor’s FileSystemDock instance.

pub fn get_inspector(&self) -> Option<Ref<EditorInspector, Shared>>

Returns the editor’s EditorInspector instance.

pub fn get_open_scenes(&self) -> VariantArray<Shared>

Returns an Array with the file paths of the currently opened scenes.

pub fn get_playing_scene(&self) -> GodotString

Returns the name of the scene that is being played. If no scene is currently being played, returns an empty string.

pub fn get_resource_filesystem(&self) -> Option<Ref<EditorFileSystem, Shared>>

Returns the editor’s EditorFileSystem instance.

pub fn get_resource_previewer(
&self
) -> Option<Ref<EditorResourcePreview, Shared>>

Returns the editor’s EditorResourcePreview instance.

pub fn get_script_editor(&self) -> Option<Ref<ScriptEditor, Shared>>

Returns the editor’s ScriptEditor instance.

pub fn get_selected_path(&self) -> GodotString

Returns the path of the directory currently selected in the FileSystemDock. If a file is selected, its base directory will be returned using String.get_base_dir instead.

pub fn get_selection(&self) -> Option<Ref<EditorSelection, Shared>>

Returns the editor’s EditorSelection instance.

pub fn inspect_object(
&self,
object: impl AsArg<Object>,
for_property: impl Into<GodotString>,
inspector_only: bool
)

Shows the given property on the given object in the editor’s Inspector dock. If inspector_only is true, plugins will not attempt to edit object.

Default Arguments

for_property - ""

inspector_only - false

pub fn is_distraction_free_mode_enabled(&self) -> bool

If true, enables distraction-free mode which hides side docks to increase the space available for the main view.

pub fn is_playing_scene(&self) -> bool

Returns true if a scene is currently being played, false otherwise. Paused scenes are considered as being played.

pub fn is_plugin_enabled(&self, plugin: impl Into<GodotString>) -> bool

Returns true if the specified plugin is enabled. The plugin name is the same as its directory name.

pub fn make_mesh_previews(
&self,
meshes: VariantArray<Shared>,
preview_size: i64
) -> VariantArray<Shared>

Returns mesh previews rendered at the given size as an Array of Textures.

pub fn open_scene_from_path(&self, scene_filepath: impl Into<GodotString>)

Opens the scene at the given path.

pub fn play_current_scene(&self)

Plays the currently active scene.

pub fn play_custom_scene(&self, scene_filepath: impl Into<GodotString>)

Plays the scene specified by its filepath.

pub fn play_main_scene(&self)

Plays the main scene.

pub fn reload_scene_from_path(&self, scene_filepath: impl Into<GodotString>)

Reloads the scene at the given path.

pub fn save_scene(&self) -> Result<(), GodotError>

Saves the scene. Returns either OK or ERR_CANT_CREATE (see [@GlobalScope] constants).

pub fn save_scene_as(&self, path: impl Into<GodotString>, with_preview: bool)

Saves the scene as a file at path.

Default Arguments

with_preview - true

pub fn select_file(&self, file: impl Into<GodotString>)

Selects the file, with the path provided by file, in the FileSystem dock.

pub fn set_distraction_free_mode(&self, enter: bool)

If true, enables distraction-free mode which hides side docks to increase the space available for the main view.

pub fn set_main_screen_editor(&self, name: impl Into<GodotString>)

Sets the editor’s current main screen to the one specified in name. name must match the text of the tab in question exactly (2D, 3D, Script, AssetLib).

pub fn set_plugin_enabled(&self, plugin: impl Into<GodotString>, enabled: bool)

Sets the enabled status of a plugin. The plugin name is the same as its directory name.

pub fn stop_playing_scene(&self)

Stops the scene that is currently playing.