Struct gdnative::api::Environment

pub struct Environment { /* private fields */ }

core class Environment inherits Resource (reference-counted).

This class has related types in the environment module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

Environment inherits methods from:

Resource

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Environment

Constants

pub const BG_CLEAR_COLOR: i64 = 0i64

pub const DOF_BLUR_QUALITY_LOW: i64 = 0i64

pub const GLOW_BLEND_MODE_ADDITIVE: i64 = 0i64

pub const SSAO_BLUR_DISABLED: i64 = 0i64

pub const SSAO_QUALITY_LOW: i64 = 0i64

pub const TONE_MAPPER_LINEAR: i64 = 0i64

pub const BG_COLOR: i64 = 1i64

pub const DOF_BLUR_QUALITY_MEDIUM: i64 = 1i64

pub const GLOW_BLEND_MODE_SCREEN: i64 = 1i64

pub const SSAO_BLUR_1x1: i64 = 1i64

pub const SSAO_QUALITY_MEDIUM: i64 = 1i64

pub const TONE_MAPPER_REINHARDT: i64 = 1i64

pub const BG_SKY: i64 = 2i64

pub const DOF_BLUR_QUALITY_HIGH: i64 = 2i64

pub const GLOW_BLEND_MODE_SOFTLIGHT: i64 = 2i64

pub const SSAO_BLUR_2x2: i64 = 2i64

pub const SSAO_QUALITY_HIGH: i64 = 2i64

pub const TONE_MAPPER_FILMIC: i64 = 2i64

pub const BG_COLOR_SKY: i64 = 3i64

pub const GLOW_BLEND_MODE_REPLACE: i64 = 3i64

pub const SSAO_BLUR_3x3: i64 = 3i64

pub const TONE_MAPPER_ACES: i64 = 3i64

pub const BG_CANVAS: i64 = 4i64

pub const TONE_MAPPER_ACES_FITTED: i64 = 4i64

pub const BG_KEEP: i64 = 5i64

pub const BG_CAMERA_FEED: i64 = 6i64

pub const BG_MAX: i64 = 7i64

impl Environment

pub fn new() -> Ref<Environment, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn adjustment_brightness(&self) -> f64

The global brightness value of the rendered scene. Effective only if adjustment_enabled is true.

pub fn adjustment_color_correction(&self) -> Option<Ref<Texture, Shared>>

Applies the provided Texture resource to affect the global color aspect of the rendered scene. Effective only if adjustment_enabled is true.

pub fn adjustment_contrast(&self) -> f64

The global contrast value of the rendered scene (default value is 1). Effective only if adjustment_enabled is true.

pub fn adjustment_saturation(&self) -> f64

The global color saturation value of the rendered scene (default value is 1). Effective only if adjustment_enabled is true.

pub fn ambient_light_color(&self) -> Color

The ambient light’s Color.

pub fn ambient_light_energy(&self) -> f64

The ambient light’s energy. The higher the value, the stronger the light.

pub fn ambient_light_sky_contribution(&self) -> f64

Defines the amount of light that the sky brings on the scene. A value of 0.0 means that the sky’s light emission has no effect on the scene illumination, thus all ambient illumination is provided by the ambient light. On the contrary, a value of 1.0 means that all the light that affects the scene is provided by the sky, thus the ambient light parameter has no effect on the scene. Note: ambient_light_sky_contribution is internally clamped between 0.0 and 1.0 (inclusive).

pub fn background(&self) -> BgMode

The background mode. See [BGMode][BGMode] for possible values.

pub fn bg_color(&self) -> Color

The Color displayed for clear areas of the scene. Only effective when using the BG_COLOR or BG_COLOR_SKY background modes).

pub fn bg_energy(&self) -> f64

The power of the light emitted by the background.

pub fn camera_feed_id(&self) -> i64

The ID of the camera feed to show in the background.

pub fn canvas_max_layer(&self) -> i64

The maximum layer ID to display. Only effective when using the BG_CANVAS background mode.

pub fn dof_blur_far_amount(&self) -> f64

The amount of far blur for the depth-of-field effect.

pub fn dof_blur_far_distance(&self) -> f64

The distance from the camera where the far blur effect affects the rendering.

pub fn dof_blur_far_quality(&self) -> DofBlurQuality

The depth-of-field far blur’s quality. Higher values can mitigate the visible banding effect seen at higher strengths, but are much slower.

pub fn dof_blur_far_transition(&self) -> f64

The length of the transition between the no-blur area and far blur.

pub fn dof_blur_near_amount(&self) -> f64

The amount of near blur for the depth-of-field effect.

pub fn dof_blur_near_distance(&self) -> f64

Distance from the camera where the near blur effect affects the rendering.

pub fn dof_blur_near_quality(&self) -> DofBlurQuality

The depth-of-field near blur’s quality. Higher values can mitigate the visible banding effect seen at higher strengths, but are much slower.

pub fn dof_blur_near_transition(&self) -> f64

The length of the transition between the near blur and no-blur area.

pub fn fog_color(&self) -> Color

The fog’s Color.

pub fn fog_depth_begin(&self) -> f64

The fog’s depth starting distance from the camera.

pub fn fog_depth_curve(&self) -> f64

The fog depth’s intensity curve. A number of presets are available in the Inspector by right-clicking the curve.

pub fn fog_depth_end(&self) -> f64

The fog’s depth end distance from the camera. If this value is set to 0, it will be equal to the current camera’s [Camera.far][Camera::far] value.

pub fn fog_height_curve(&self) -> f64

The height fog’s intensity. A number of presets are available in the Inspector by right-clicking the curve.

pub fn fog_height_max(&self) -> f64

The Y coordinate where the height fog will be the most intense. If this value is greater than fog_height_min, fog will be displayed from bottom to top. Otherwise, it will be displayed from top to bottom.

pub fn fog_height_min(&self) -> f64

The Y coordinate where the height fog will be the least intense. If this value is greater than fog_height_max, fog will be displayed from top to bottom. Otherwise, it will be displayed from bottom to top.

pub fn fog_sun_amount(&self) -> f64

The intensity of the depth fog color transition when looking towards the sun. The sun’s direction is determined automatically using the DirectionalLight node in the scene.

pub fn fog_sun_color(&self) -> Color

The depth fog’s Color when looking towards the sun.

pub fn fog_transmit_curve(&self) -> f64

The intensity of the fog light transmittance effect. Amount of light that the fog transmits.

pub fn glow_blend_mode(&self) -> GlowBlendMode

The glow blending mode.

pub fn glow_bloom(&self) -> f64

The bloom’s intensity. If set to a value higher than 0, this will make glow visible in areas darker than the [glow_hdr_threshold][Self::glow_hdr_threshold].

pub fn glow_hdr_bleed_scale(&self) -> f64

The bleed scale of the HDR glow.

pub fn glow_hdr_bleed_threshold(&self) -> f64

The lower threshold of the HDR glow. When using the GLES2 renderer (which doesn’t support HDR), this needs to be below 1.0 for glow to be visible. A value of 0.9 works well in this case.

pub fn glow_hdr_luminance_cap(&self) -> f64

The higher threshold of the HDR glow. Areas brighter than this threshold will be clamped for the purposes of the glow effect.

pub fn glow_intensity(&self) -> f64

The glow intensity. When using the GLES2 renderer, this should be increased to 1.5 to compensate for the lack of HDR rendering.

pub fn glow_strength(&self) -> f64

The glow strength. When using the GLES2 renderer, this should be increased to 1.3 to compensate for the lack of HDR rendering.

pub fn sky(&self) -> Option<Ref<Sky, Shared>>

The Sky resource defined as background.

pub fn sky_custom_fov(&self) -> f64

The Sky resource’s custom field of view.

pub fn sky_orientation(&self) -> Basis

The Sky resource’s rotation expressed as a Basis.

pub fn sky_rotation(&self) -> Vector3

The Sky resource’s rotation expressed as Euler angles in radians.

pub fn sky_rotation_degrees(&self) -> Vector3

The Sky resource’s rotation expressed as Euler angles in degrees.

pub fn ssao_ao_channel_affect(&self) -> f64

The screen-space ambient occlusion intensity on materials that have an AO texture defined. Values higher than 0 will make the SSAO effect visible in areas darkened by AO textures.

pub fn ssao_bias(&self) -> f64

The screen-space ambient occlusion bias. This should be kept high enough to prevent “smooth” curves from being affected by ambient occlusion.

pub fn ssao_blur(&self) -> SsaoBlur

The screen-space ambient occlusion blur quality. See [SSAOBlur][SSAOBlur] for possible values.

pub fn ssao_color(&self) -> Color

The screen-space ambient occlusion color.

pub fn ssao_direct_light_affect(&self) -> f64

The screen-space ambient occlusion intensity in direct light. In real life, ambient occlusion only applies to indirect light, which means its effects can’t be seen in direct light. Values higher than 0 will make the SSAO effect visible in direct light.

pub fn ssao_edge_sharpness(&self) -> f64

The screen-space ambient occlusion edge sharpness.

pub fn ssao_intensity(&self) -> f64

The primary screen-space ambient occlusion intensity. See also ssao_radius.

pub fn ssao_intensity2(&self) -> f64

The secondary screen-space ambient occlusion intensity. See also ssao_radius2.

pub fn ssao_quality(&self) -> SsaoQuality

The screen-space ambient occlusion quality. Higher qualities will make better use of small objects for ambient occlusion, but are slower.

pub fn ssao_radius(&self) -> f64

The primary screen-space ambient occlusion radius.

pub fn ssao_radius2(&self) -> f64

The secondary screen-space ambient occlusion radius. If set to a value higher than 0, enables the secondary screen-space ambient occlusion effect which can be used to improve the effect’s appearance (at the cost of performance).

pub fn ssr_depth_tolerance(&self) -> f64

The depth tolerance for screen-space reflections.

pub fn ssr_fade_in(&self) -> f64

The fade-in distance for screen-space reflections. Affects the area from the reflected material to the screen-space reflection).

pub fn ssr_fade_out(&self) -> f64

The fade-out distance for screen-space reflections. Affects the area from the screen-space reflection to the “global” reflection.

pub fn ssr_max_steps(&self) -> i64

The maximum number of steps for screen-space reflections. Higher values are slower.

pub fn tonemap_auto_exposure(&self) -> bool

If true, enables the tonemapping auto exposure mode of the scene renderer. If true, the renderer will automatically determine the exposure setting to adapt to the scene’s illumination and the observed light.

pub fn tonemap_auto_exposure_grey(&self) -> f64

The scale of the auto exposure effect. Affects the intensity of auto exposure.

pub fn tonemap_auto_exposure_max(&self) -> f64

The maximum luminance value for the auto exposure.

pub fn tonemap_auto_exposure_min(&self) -> f64

The minimum luminance value for the auto exposure.

pub fn tonemap_auto_exposure_speed(&self) -> f64

The speed of the auto exposure effect. Affects the time needed for the camera to perform auto exposure.

pub fn tonemap_exposure(&self) -> f64

The default exposure used for tonemapping.

pub fn tonemap_white(&self) -> f64

The white reference value for tonemapping. Only effective if the [tonemap_mode][Self::tonemap_mode] isn’t set to TONE_MAPPER_LINEAR.

pub fn tonemapper(&self) -> ToneMapper

The tonemapping mode to use. Tonemapping is the process that “converts” HDR values to be suitable for rendering on a SDR display. (Godot doesn’t support rendering on HDR displays yet.)

pub fn is_adjustment_enabled(&self) -> bool

If true, enables the adjustment_* properties provided by this resource. If false, modifications to the adjustment_* properties will have no effect on the rendered scene.

pub fn is_dof_blur_far_enabled(&self) -> bool

If true, enables the depth-of-field far blur effect.

pub fn is_dof_blur_near_enabled(&self) -> bool

If true, enables the depth-of-field near blur effect.

pub fn is_fog_depth_enabled(&self) -> bool

If true, the depth fog effect is enabled. When enabled, fog will appear in the distance (relative to the camera).

pub fn is_fog_enabled(&self) -> bool

If true, fog effects are enabled. [fog_height_enabled][Self::fog_height_enabled] and/or [fog_depth_enabled][Self::fog_depth_enabled] must be set to true to actually display fog.

pub fn is_fog_height_enabled(&self) -> bool

If true, the height fog effect is enabled. When enabled, fog will appear in a defined height range, regardless of the distance from the camera. This can be used to simulate “deep water” effects with a lower performance cost compared to a dedicated shader.

pub fn is_fog_transmit_enabled(&self) -> bool

Enables fog’s light transmission effect. If true, light will be more visible in the fog to simulate light scattering as in real life.

pub fn is_glow_bicubic_upscale_enabled(&self) -> bool

Smooths out the blockiness created by sampling higher levels, at the cost of performance. Note: When using the GLES2 renderer, this is only available if the GPU supports the GL_EXT_gpu_shader4 extension.

pub fn is_glow_enabled(&self) -> bool

If true, the glow effect is enabled. Note: Only effective if [member ProjectSettings.rendering/quality/intended_usage/framebuffer_allocation] is 3D (not 3D Without Effects). On mobile, [member ProjectSettings.rendering/quality/intended_usage/framebuffer_allocation] defaults to 3D Without Effects by default, so its .mobile override needs to be changed to 3D. Note: When using GLES3 on mobile, HDR rendering is disabled by default for performance reasons. This means glow will only be visible if [glow_hdr_threshold][Self::glow_hdr_threshold] is decreased below 1.0 or if glow_bloom is increased above 0.0. Also consider increasing glow_intensity to 1.5. If you want glow to behave on mobile like it does on desktop (at a performance cost), enable [member ProjectSettings.rendering/quality/depth/hdr]’s .mobile override.

pub fn is_glow_high_quality_enabled(&self) -> bool

Takes more samples during downsample pass of glow. This ensures that single pixels are captured by glow which makes the glow look smoother and more stable during movement. However, it is very expensive and makes the glow post process take twice as long.

pub fn is_glow_level_enabled(&self, idx: i64) -> bool

Returns true if the glow level idx is specified, false otherwise.

pub fn is_ssao_enabled(&self) -> bool

If true, the screen-space ambient occlusion effect is enabled. This darkens objects’ corners and cavities to simulate ambient light not reaching the entire object as in real life. This works well for small, dynamic objects, but baked lighting or ambient occlusion textures will do a better job at displaying ambient occlusion on large static objects. This is a costly effect and should be disabled first when running into performance issues.

pub fn is_ssr_enabled(&self) -> bool

If true, screen-space reflections are enabled. Screen-space reflections are more accurate than reflections from GIProbes or ReflectionProbes, but are slower and can’t reflect surfaces occluded by others.

pub fn is_ssr_rough(&self) -> bool

If true, screen-space reflections will take the material roughness into account.

pub fn set_adjustment_brightness(&self, brightness: f64)

The global brightness value of the rendered scene. Effective only if adjustment_enabled is true.

pub fn set_adjustment_color_correction( &self, color_correction: impl AsArg<Texture> )

Applies the provided Texture resource to affect the global color aspect of the rendered scene. Effective only if adjustment_enabled is true.

pub fn set_adjustment_contrast(&self, contrast: f64)

The global contrast value of the rendered scene (default value is 1). Effective only if adjustment_enabled is true.

pub fn set_adjustment_enable(&self, enabled: bool)

If true, enables the adjustment_* properties provided by this resource. If false, modifications to the adjustment_* properties will have no effect on the rendered scene.

pub fn set_adjustment_saturation(&self, saturation: f64)

The global color saturation value of the rendered scene (default value is 1). Effective only if adjustment_enabled is true.

pub fn set_ambient_light_color(&self, color: Color)

The ambient light’s Color.

pub fn set_ambient_light_energy(&self, energy: f64)

The ambient light’s energy. The higher the value, the stronger the light.

pub fn set_ambient_light_sky_contribution(&self, energy: f64)

Defines the amount of light that the sky brings on the scene. A value of 0.0 means that the sky’s light emission has no effect on the scene illumination, thus all ambient illumination is provided by the ambient light. On the contrary, a value of 1.0 means that all the light that affects the scene is provided by the sky, thus the ambient light parameter has no effect on the scene. Note: ambient_light_sky_contribution is internally clamped between 0.0 and 1.0 (inclusive).

pub fn set_background(&self, mode: i64)

The background mode. See [BGMode][BGMode] for possible values.

pub fn set_bg_color(&self, color: Color)

The Color displayed for clear areas of the scene. Only effective when using the BG_COLOR or BG_COLOR_SKY background modes).

pub fn set_bg_energy(&self, energy: f64)

The power of the light emitted by the background.

pub fn set_camera_feed_id(&self, camera_feed_id: i64)

The ID of the camera feed to show in the background.

pub fn set_canvas_max_layer(&self, layer: i64)

The maximum layer ID to display. Only effective when using the BG_CANVAS background mode.

pub fn set_dof_blur_far_amount(&self, intensity: f64)

The amount of far blur for the depth-of-field effect.

pub fn set_dof_blur_far_distance(&self, intensity: f64)

The distance from the camera where the far blur effect affects the rendering.

pub fn set_dof_blur_far_enabled(&self, enabled: bool)

If true, enables the depth-of-field far blur effect.

pub fn set_dof_blur_far_quality(&self, intensity: i64)

The depth-of-field far blur’s quality. Higher values can mitigate the visible banding effect seen at higher strengths, but are much slower.

pub fn set_dof_blur_far_transition(&self, intensity: f64)

The length of the transition between the no-blur area and far blur.

pub fn set_dof_blur_near_amount(&self, intensity: f64)

The amount of near blur for the depth-of-field effect.

pub fn set_dof_blur_near_distance(&self, intensity: f64)

Distance from the camera where the near blur effect affects the rendering.

pub fn set_dof_blur_near_enabled(&self, enabled: bool)

If true, enables the depth-of-field near blur effect.

pub fn set_dof_blur_near_quality(&self, level: i64)

The depth-of-field near blur’s quality. Higher values can mitigate the visible banding effect seen at higher strengths, but are much slower.

pub fn set_dof_blur_near_transition(&self, intensity: f64)

The length of the transition between the near blur and no-blur area.

pub fn set_fog_color(&self, color: Color)

The fog’s Color.

pub fn set_fog_depth_begin(&self, distance: f64)

The fog’s depth starting distance from the camera.

pub fn set_fog_depth_curve(&self, curve: f64)

The fog depth’s intensity curve. A number of presets are available in the Inspector by right-clicking the curve.

pub fn set_fog_depth_enabled(&self, enabled: bool)

If true, the depth fog effect is enabled. When enabled, fog will appear in the distance (relative to the camera).

pub fn set_fog_depth_end(&self, distance: f64)

The fog’s depth end distance from the camera. If this value is set to 0, it will be equal to the current camera’s [Camera.far][Camera::far] value.

pub fn set_fog_enabled(&self, enabled: bool)

If true, fog effects are enabled. [fog_height_enabled][Self::fog_height_enabled] and/or [fog_depth_enabled][Self::fog_depth_enabled] must be set to true to actually display fog.

pub fn set_fog_height_curve(&self, curve: f64)

The height fog’s intensity. A number of presets are available in the Inspector by right-clicking the curve.

pub fn set_fog_height_enabled(&self, enabled: bool)

If true, the height fog effect is enabled. When enabled, fog will appear in a defined height range, regardless of the distance from the camera. This can be used to simulate “deep water” effects with a lower performance cost compared to a dedicated shader.

pub fn set_fog_height_max(&self, height: f64)

The Y coordinate where the height fog will be the most intense. If this value is greater than fog_height_min, fog will be displayed from bottom to top. Otherwise, it will be displayed from top to bottom.

pub fn set_fog_height_min(&self, height: f64)

The Y coordinate where the height fog will be the least intense. If this value is greater than fog_height_max, fog will be displayed from top to bottom. Otherwise, it will be displayed from bottom to top.

pub fn set_fog_sun_amount(&self, amount: f64)

The intensity of the depth fog color transition when looking towards the sun. The sun’s direction is determined automatically using the DirectionalLight node in the scene.

pub fn set_fog_sun_color(&self, color: Color)

The depth fog’s Color when looking towards the sun.

pub fn set_fog_transmit_curve(&self, curve: f64)

The intensity of the fog light transmittance effect. Amount of light that the fog transmits.

pub fn set_fog_transmit_enabled(&self, enabled: bool)

Enables fog’s light transmission effect. If true, light will be more visible in the fog to simulate light scattering as in real life.

pub fn set_glow_bicubic_upscale(&self, enabled: bool)

Smooths out the blockiness created by sampling higher levels, at the cost of performance. Note: When using the GLES2 renderer, this is only available if the GPU supports the GL_EXT_gpu_shader4 extension.

pub fn set_glow_blend_mode(&self, mode: i64)

The glow blending mode.

pub fn set_glow_bloom(&self, amount: f64)

The bloom’s intensity. If set to a value higher than 0, this will make glow visible in areas darker than the [glow_hdr_threshold][Self::glow_hdr_threshold].

pub fn set_glow_enabled(&self, enabled: bool)

If true, the glow effect is enabled. Note: Only effective if [member ProjectSettings.rendering/quality/intended_usage/framebuffer_allocation] is 3D (not 3D Without Effects). On mobile, [member ProjectSettings.rendering/quality/intended_usage/framebuffer_allocation] defaults to 3D Without Effects by default, so its .mobile override needs to be changed to 3D. Note: When using GLES3 on mobile, HDR rendering is disabled by default for performance reasons. This means glow will only be visible if [glow_hdr_threshold][Self::glow_hdr_threshold] is decreased below 1.0 or if glow_bloom is increased above 0.0. Also consider increasing glow_intensity to 1.5. If you want glow to behave on mobile like it does on desktop (at a performance cost), enable [member ProjectSettings.rendering/quality/depth/hdr]’s .mobile override.

pub fn set_glow_hdr_bleed_scale(&self, scale: f64)

The bleed scale of the HDR glow.

pub fn set_glow_hdr_bleed_threshold(&self, threshold: f64)

The lower threshold of the HDR glow. When using the GLES2 renderer (which doesn’t support HDR), this needs to be below 1.0 for glow to be visible. A value of 0.9 works well in this case.

pub fn set_glow_hdr_luminance_cap(&self, amount: f64)

The higher threshold of the HDR glow. Areas brighter than this threshold will be clamped for the purposes of the glow effect.

pub fn set_glow_high_quality(&self, enabled: bool)

Takes more samples during downsample pass of glow. This ensures that single pixels are captured by glow which makes the glow look smoother and more stable during movement. However, it is very expensive and makes the glow post process take twice as long.

pub fn set_glow_intensity(&self, intensity: f64)

The glow intensity. When using the GLES2 renderer, this should be increased to 1.5 to compensate for the lack of HDR rendering.

pub fn set_glow_level(&self, idx: i64, enabled: bool)

Enables or disables the glow level at index idx. Each level relies on the previous level. This means that enabling higher glow levels will slow down the glow effect rendering, even if previous levels aren’t enabled.

pub fn set_glow_strength(&self, strength: f64)

The glow strength. When using the GLES2 renderer, this should be increased to 1.3 to compensate for the lack of HDR rendering.

pub fn set_sky(&self, sky: impl AsArg<Sky>)

The Sky resource defined as background.

pub fn set_sky_custom_fov(&self, scale: f64)

The Sky resource’s custom field of view.

pub fn set_sky_orientation(&self, orientation: Basis)

The Sky resource’s rotation expressed as a Basis.

pub fn set_sky_rotation(&self, euler_radians: Vector3)

The Sky resource’s rotation expressed as Euler angles in radians.

pub fn set_sky_rotation_degrees(&self, euler_degrees: Vector3)

The Sky resource’s rotation expressed as Euler angles in degrees.

pub fn set_ssao_ao_channel_affect(&self, amount: f64)

The screen-space ambient occlusion intensity on materials that have an AO texture defined. Values higher than 0 will make the SSAO effect visible in areas darkened by AO textures.

pub fn set_ssao_bias(&self, bias: f64)

The screen-space ambient occlusion bias. This should be kept high enough to prevent “smooth” curves from being affected by ambient occlusion.

pub fn set_ssao_blur(&self, mode: i64)

The screen-space ambient occlusion blur quality. See [SSAOBlur][SSAOBlur] for possible values.

pub fn set_ssao_color(&self, color: Color)

The screen-space ambient occlusion color.

pub fn set_ssao_direct_light_affect(&self, amount: f64)

The screen-space ambient occlusion intensity in direct light. In real life, ambient occlusion only applies to indirect light, which means its effects can’t be seen in direct light. Values higher than 0 will make the SSAO effect visible in direct light.

pub fn set_ssao_edge_sharpness(&self, edge_sharpness: f64)

The screen-space ambient occlusion edge sharpness.

pub fn set_ssao_enabled(&self, enabled: bool)

If true, the screen-space ambient occlusion effect is enabled. This darkens objects’ corners and cavities to simulate ambient light not reaching the entire object as in real life. This works well for small, dynamic objects, but baked lighting or ambient occlusion textures will do a better job at displaying ambient occlusion on large static objects. This is a costly effect and should be disabled first when running into performance issues.

pub fn set_ssao_intensity(&self, intensity: f64)

The primary screen-space ambient occlusion intensity. See also ssao_radius.

pub fn set_ssao_intensity2(&self, intensity: f64)

The secondary screen-space ambient occlusion intensity. See also ssao_radius2.

pub fn set_ssao_quality(&self, quality: i64)

The screen-space ambient occlusion quality. Higher qualities will make better use of small objects for ambient occlusion, but are slower.

pub fn set_ssao_radius(&self, radius: f64)

The primary screen-space ambient occlusion radius.

pub fn set_ssao_radius2(&self, radius: f64)

The secondary screen-space ambient occlusion radius. If set to a value higher than 0, enables the secondary screen-space ambient occlusion effect which can be used to improve the effect’s appearance (at the cost of performance).

pub fn set_ssr_depth_tolerance(&self, depth_tolerance: f64)

The depth tolerance for screen-space reflections.

pub fn set_ssr_enabled(&self, enabled: bool)

If true, screen-space reflections are enabled. Screen-space reflections are more accurate than reflections from GIProbes or ReflectionProbes, but are slower and can’t reflect surfaces occluded by others.

pub fn set_ssr_fade_in(&self, fade_in: f64)

The fade-in distance for screen-space reflections. Affects the area from the reflected material to the screen-space reflection).

pub fn set_ssr_fade_out(&self, fade_out: f64)

The fade-out distance for screen-space reflections. Affects the area from the screen-space reflection to the “global” reflection.

pub fn set_ssr_max_steps(&self, max_steps: i64)

The maximum number of steps for screen-space reflections. Higher values are slower.

pub fn set_ssr_rough(&self, rough: bool)

If true, screen-space reflections will take the material roughness into account.

pub fn set_tonemap_auto_exposure(&self, auto_exposure: bool)

If true, enables the tonemapping auto exposure mode of the scene renderer. If true, the renderer will automatically determine the exposure setting to adapt to the scene’s illumination and the observed light.

pub fn set_tonemap_auto_exposure_grey(&self, exposure_grey: f64)

The scale of the auto exposure effect. Affects the intensity of auto exposure.

pub fn set_tonemap_auto_exposure_max(&self, exposure_max: f64)

The maximum luminance value for the auto exposure.

pub fn set_tonemap_auto_exposure_min(&self, exposure_min: f64)

The minimum luminance value for the auto exposure.

pub fn set_tonemap_auto_exposure_speed(&self, exposure_speed: f64)

The speed of the auto exposure effect. Affects the time needed for the camera to perform auto exposure.

pub fn set_tonemap_exposure(&self, exposure: f64)

The default exposure used for tonemapping.

pub fn set_tonemap_white(&self, white: f64)

The white reference value for tonemapping. Only effective if the [tonemap_mode][Self::tonemap_mode] isn’t set to TONE_MAPPER_LINEAR.

pub fn set_tonemapper(&self, mode: i64)

The tonemapping mode to use. Tonemapping is the process that “converts” HDR values to be suitable for rendering on a SDR display. (Godot doesn’t support rendering on HDR displays yet.)