Struct gdnative::api::File

pub struct File { /* fields omitted */ }

core class File inherits Reference (reference-counted).

This class has related types in the file module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

File inherits methods from:

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl File

Constants

pub const COMPRESSION_FASTLZ: i64

pub const COMPRESSION_DEFLATE: i64

pub const READ: i64

pub const COMPRESSION_ZSTD: i64

pub const WRITE: i64

pub const COMPRESSION_GZIP: i64

pub const READ_WRITE: i64

pub const WRITE_READ: i64

impl File

pub fn new() -> Ref<File, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn close(&self)

Closes the currently opened file and prevents subsequent read/write operations. Use flush to persist the data to disk without closing the file.

pub fn eof_reached(&self) -> bool

Returns true if the file cursor has read past the end of the file. Note: This function will still return false while at the end of the file and only activates when reading past it. This can be confusing but it conforms to how low-level file access works in all operating systems. There is always get_len and get_position to implement a custom logic.

pub fn file_exists(&self, path: impl Into<GodotString>) -> bool

Returns true if the file exists in the given path. Note: Many resources types are imported (e.g. textures or sound files), and their source asset will not be included in the exported game, as only the imported version is used. See ResourceLoader.exists for an alternative approach that takes resource remapping into account.

pub fn flush(&self)

Writes the file’s buffer to disk. Flushing is automatically performed when the file is closed. This means you don’t need to call flush manually before closing a file using close. Still, calling flush can be used to ensure the data is safe even if the project crashes instead of being closed gracefully. Note: Only call flush when you actually need it. Otherwise, it will decrease performance due to constant disk writes.

pub fn get_16(&self) -> i64

Returns the next 16 bits from the file as an integer. See store_16 for details on what values can be stored and retrieved this way.

pub fn get_32(&self) -> i64

Returns the next 32 bits from the file as an integer. See store_32 for details on what values can be stored and retrieved this way.

pub fn get_64(&self) -> i64

Returns the next 64 bits from the file as an integer. See store_64 for details on what values can be stored and retrieved this way.

pub fn get_8(&self) -> i64

Returns the next 8 bits from the file as an integer. See store_8 for details on what values can be stored and retrieved this way.

pub fn get_as_text(&self) -> GodotString

Returns the whole file as a String. Text is interpreted as being UTF-8 encoded.

pub fn get_buffer(&self, len: i64) -> PoolArray<u8>

Returns next len bytes of the file as a PoolByteArray.

pub fn get_csv_line(
&self,
delim: impl Into<GodotString>
) -> PoolArray<GodotString>

Returns the next value of the file in CSV (Comma-Separated Values) format. You can pass a different delimiter delim to use other than the default "," (comma). This delimiter must be one-character long. Text is interpreted as being UTF-8 encoded.

Default Arguments

delim - ","

pub fn get_double(&self) -> f64

Returns the next 64 bits from the file as a floating-point number.

pub fn endian_swap(&self) -> bool

If true, the file is read with big-endian [endianness]endianness. If false, the file is read with little-endian endianness. If in doubt, leave this to false as most files are written with little-endian endianness. Note: endian_swap is only about the file format, not the CPU type. The CPU endianness doesn’t affect the default endianness for files written. Note: This is always reset to false whenever you open the file. Therefore, you must set endian_swap after opening the file, not before.

pub fn get_error(&self) -> Result<(), GodotError>

Returns the last error that happened when trying to perform operations. Compare with the ERR_FILE_* constants from Error.

pub fn get_float(&self) -> f64

Returns the next 32 bits from the file as a floating-point number.

pub fn get_len(&self) -> i64

Returns the size of the file in bytes.

pub fn get_line(&self) -> GodotString

Returns the next line of the file as a String. Text is interpreted as being UTF-8 encoded.

pub fn get_md5(&self, path: impl Into<GodotString>) -> GodotString

Returns an MD5 String representing the file at the given path or an empty String on failure.

pub fn get_modified_time(&self, file: impl Into<GodotString>) -> i64

Returns the last time the file was modified in unix timestamp format or returns a String “ERROR IN file”. This unix timestamp can be converted to datetime by using OS.get_datetime_from_unix_time.

pub fn get_pascal_string(&self) -> GodotString

Returns a String saved in Pascal format from the file. Text is interpreted as being UTF-8 encoded.

pub fn get_path(&self) -> GodotString

Returns the path as a String for the current open file.

pub fn get_path_absolute(&self) -> GodotString

Returns the absolute path as a String for the current open file.

pub fn get_position(&self) -> i64

Returns the file cursor’s position.

pub fn get_real(&self) -> f64

Returns the next bits from the file as a floating-point number.

pub fn get_sha256(&self, path: impl Into<GodotString>) -> GodotString

Returns a SHA-256 String representing the file at the given path or an empty String on failure.

pub fn get_var(&self, allow_objects: bool) -> Variant

Returns the next Variant value from the file. If allow_objects is true, decoding objects is allowed. Warning: Deserialized objects can contain code which gets executed. Do not use this option if the serialized object comes from untrusted sources to avoid potential security threats such as remote code execution.

Default Arguments

allow_objects - false

pub fn is_open(&self) -> bool

Returns true if the file is currently opened.

pub fn open(
&self,
path: impl Into<GodotString>,
flags: i64
) -> Result<(), GodotError>

Opens the file for writing or reading, depending on the flags.

pub fn open_compressed(
&self,
path: impl Into<GodotString>,
mode_flags: i64,
compression_mode: i64
) -> Result<(), GodotError>

Opens a compressed file for reading or writing.

Default Arguments

compression_mode - 0

pub fn open_encrypted(
&self,
path: impl Into<GodotString>,
mode_flags: i64,
key: PoolArray<u8>
) -> Result<(), GodotError>

Opens an encrypted file in write or read mode. You need to pass a binary key to encrypt/decrypt it. Note: The provided key must be 32 bytes long.

pub fn open_encrypted_with_pass(
&self,
path: impl Into<GodotString>,
mode_flags: i64,
pass: impl Into<GodotString>
) -> Result<(), GodotError>

Opens an encrypted file in write or read mode. You need to pass a password to encrypt/decrypt it.

pub fn seek(&self, position: i64)

Changes the file reading/writing cursor to the specified position (in bytes from the beginning of the file).

pub fn seek_end(&self, position: i64)

Changes the file reading/writing cursor to the specified position (in bytes from the end of the file). Note: This is an offset, so you should use negative numbers or the cursor will be at the end of the file.

Default Arguments

position - 0

pub fn set_endian_swap(&self, enable: bool)

If true, the file is read with big-endian [endianness]endianness. If false, the file is read with little-endian endianness. If in doubt, leave this to false as most files are written with little-endian endianness. Note: endian_swap is only about the file format, not the CPU type. The CPU endianness doesn’t affect the default endianness for files written. Note: This is always reset to false whenever you open the file. Therefore, you must set endian_swap after opening the file, not before.

pub fn store_16(&self, value: i64)

Sample code is GDScript unless otherwise noted.

Stores an integer as 16 bits in the file. Note: The value should lie in the interval [0, 2^16 - 1]. Any other value will overflow and wrap around. To store a signed integer, use store_64 or store a signed integer from the interval [-2^15, 2^15 - 1] (i.e. keeping one bit for the signedness) and compute its sign manually when reading. For example:

const MAX_15B = 1 << 15
const MAX_16B = 1 << 16

func unsigned16_to_signed(unsigned):
    return (unsigned + MAX_15B) % MAX_16B - MAX_15B

func _ready():
    var f = File.new()
    f.open("user://file.dat", File.WRITE_READ)
    f.store_16(-42) # This wraps around and stores 65494 (2^16 - 42).
    f.store_16(121) # In bounds, will store 121.
    f.seek(0) # Go back to start to read the stored value.
    var read1 = f.get_16() # 65494
    var read2 = f.get_16() # 121
    var converted1 = unsigned16_to_signed(read1) # -42
    var converted2 = unsigned16_to_signed(read2) # 121

pub fn store_32(&self, value: i64)

Stores an integer as 32 bits in the file. Note: The value should lie in the interval [0, 2^32 - 1]. Any other value will overflow and wrap around. To store a signed integer, use store_64, or convert it manually (see store_16 for an example).

pub fn store_64(&self, value: i64)

Stores an integer as 64 bits in the file. Note: The value must lie in the interval [-2^63, 2^63 - 1] (i.e. be a valid [int][int] value).

pub fn store_8(&self, value: i64)

Stores an integer as 8 bits in the file. Note: The value should lie in the interval [0, 255]. Any other value will overflow and wrap around. To store a signed integer, use store_64, or convert it manually (see store_16 for an example).

pub fn store_buffer(&self, buffer: PoolArray<u8>)

Stores the given array of bytes in the file.

pub fn store_csv_line(
&self,
values: PoolArray<GodotString>,
delim: impl Into<GodotString>
)

Store the given PoolStringArray in the file as a line formatted in the CSV (Comma-Separated Values) format. You can pass a different delimiter delim to use other than the default "," (comma). This delimiter must be one-character long. Text will be encoded as UTF-8.

Default Arguments

delim - ","

pub fn store_double(&self, value: f64)

Stores a floating-point number as 64 bits in the file.

pub fn store_float(&self, value: f64)

Stores a floating-point number as 32 bits in the file.

pub fn store_line(&self, line: impl Into<GodotString>)

Appends line to the file followed by a line return character (\n), encoding the text as UTF-8.

pub fn store_pascal_string(&self, string: impl Into<GodotString>)

Stores the given String as a line in the file in Pascal format (i.e. also store the length of the string). Text will be encoded as UTF-8.

pub fn store_real(&self, value: f64)

Stores a floating-point number in the file.

pub fn store_string(&self, string: impl Into<GodotString>)

Appends string to the file without a line return, encoding the text as UTF-8.

pub fn store_var(&self, value: impl OwnedToVariant, full_objects: bool)

Stores any Variant value in the file. If full_objects is true, encoding objects is allowed (and can potentially include code).

Default Arguments

full_objects - false