Struct gdnative::api::GIProbe

pub struct GIProbe { /* fields omitted */ }

core class GIProbe inherits VisualInstance (manually managed).

This class has related types in the gi_probe module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

GIProbe is a reference-only type. Persistent references can only exist in the unsafe Ref<GIProbe> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

GIProbe inherits methods from:

VisualInstance

Spatial

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl GIProbe

Constants

pub const SUBDIV_64: i64

pub const SUBDIV_128: i64

pub const SUBDIV_256: i64

pub const SUBDIV_512: i64

pub const SUBDIV_MAX: i64

impl GIProbe

pub fn new() -> Ref<GIProbe, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn bake(&self, from_node: impl AsArg<Node>, create_visual_debug: bool)

Bakes the effect from all GeometryInstances marked with [GeometryInstance.use_in_baked_light][GeometryInstance::use_in_baked_light] and Lights marked with either Light.BAKE_INDIRECT or Light.BAKE_ALL. If create_visual_debug is true, after baking the light, this will generate a MultiMesh that has a cube representing each solid cell with each cube colored to the cell’s albedo color. This can be used to visualize the GIProbe’s data and debug any issues that may be occurring.

Default Arguments

from_node - null

create_visual_debug - false

pub fn debug_bake(&self)

Calls bake with create_visual_debug enabled.

pub fn bias(&self) -> f64

Offsets the lookup of the light contribution from the GIProbe. This can be used to avoid self-shadowing, but may introduce light leaking at higher values. This and normal_bias should be played around with to minimize self-shadowing and light leaking. Note: bias should usually be above 1.0 as that is the size of the voxels.

pub fn dynamic_range(&self) -> i64

The maximum brightness that the GIProbe will recognize. Brightness will be scaled within this range.

pub fn energy(&self) -> f64

Energy multiplier. Makes the lighting contribution from the GIProbe brighter.

pub fn extents(&self) -> Vector3

The size of the area covered by the GIProbe. If you make the extents larger without increasing the subdivisions with subdiv, the size of each cell will increase and result in lower detailed lighting.

pub fn normal_bias(&self) -> f64

Offsets the lookup into the GIProbe based on the object’s normal direction. Can be used to reduce some self-shadowing artifacts.

pub fn probe_data(&self) -> Option<Ref<GIProbeData, Shared>>

The GIProbeData resource that holds the data for this GIProbe.

pub fn propagation(&self) -> f64

How much light propagates through the probe internally. A higher value allows light to spread further.

pub fn subdiv(&self) -> Subdiv

Number of times to subdivide the grid that the GIProbe operates on. A higher number results in finer detail and thus higher visual quality, while lower numbers result in better performance.

pub fn is_compressed(&self) -> bool

If true, the data for this GIProbe will be compressed. Compression saves space, but results in far worse visual quality.

pub fn is_interior(&self) -> bool

If true, ignores the sky contribution when calculating lighting.

pub fn set_bias(&self, max: f64)

Offsets the lookup of the light contribution from the GIProbe. This can be used to avoid self-shadowing, but may introduce light leaking at higher values. This and normal_bias should be played around with to minimize self-shadowing and light leaking. Note: bias should usually be above 1.0 as that is the size of the voxels.

pub fn set_compress(&self, enable: bool)

If true, the data for this GIProbe will be compressed. Compression saves space, but results in far worse visual quality.

pub fn set_dynamic_range(&self, max: i64)

The maximum brightness that the GIProbe will recognize. Brightness will be scaled within this range.

pub fn set_energy(&self, max: f64)

Energy multiplier. Makes the lighting contribution from the GIProbe brighter.

pub fn set_extents(&self, extents: Vector3)

The size of the area covered by the GIProbe. If you make the extents larger without increasing the subdivisions with subdiv, the size of each cell will increase and result in lower detailed lighting.

pub fn set_interior(&self, enable: bool)

If true, ignores the sky contribution when calculating lighting.

pub fn set_normal_bias(&self, max: f64)

Offsets the lookup into the GIProbe based on the object’s normal direction. Can be used to reduce some self-shadowing artifacts.

pub fn set_probe_data(&self, data: impl AsArg<GIProbeData>)

The GIProbeData resource that holds the data for this GIProbe.

pub fn set_propagation(&self, max: f64)

How much light propagates through the probe internally. A higher value allows light to spread further.

pub fn set_subdiv(&self, subdiv: i64)

Number of times to subdivide the grid that the GIProbe operates on. A higher number results in finer detail and thus higher visual quality, while lower numbers result in better performance.