Struct gdnative::api::GeometryInstance

pub struct GeometryInstance { /* fields omitted */ }

core class GeometryInstance inherits VisualInstance (manually managed).

This class has related types in the geometry_instance module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Class hierarchy

GeometryInstance inherits methods from:

VisualInstance

CullInstance

Spatial

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl GeometryInstance

Constants

pub const FLAG_USE_BAKED_LIGHT: i64

pub const LIGHTMAP_SCALE_1X: i64

pub const SHADOW_CASTING_SETTING_OFF: i64

pub const FLAG_DRAW_NEXT_FRAME_IF_VISIBLE: i64

pub const LIGHTMAP_SCALE_2X: i64

pub const SHADOW_CASTING_SETTING_ON: i64

pub const FLAG_MAX: i64

pub const LIGHTMAP_SCALE_4X: i64

pub const SHADOW_CASTING_SETTING_DOUBLE_SIDED: i64

pub const LIGHTMAP_SCALE_8X: i64

pub const SHADOW_CASTING_SETTING_SHADOWS_ONLY: i64

pub const LIGHTMAP_SCALE_MAX: i64

impl GeometryInstance

pub fn cast_shadows_setting(&self) -> ShadowCastingSetting

The selected shadow casting flag. See ShadowCastingSetting for possible values.

pub fn extra_cull_margin(&self) -> f64

The extra distance added to the GeometryInstance’s bounding box (AABB) to increase its cull box.

pub fn flag(&self, flag: i64) -> bool

If true, this GeometryInstance will be used when baking lights using a GIProbe or BakedLightmap.

pub fn generate_lightmap(&self) -> bool

When disabled, the mesh will be taken into account when computing indirect lighting, but the resulting lightmap will not be saved. Useful for emissive only materials or shadow casters.

pub fn lightmap_scale(&self) -> LightmapScale

Scale factor for the generated baked lightmap. Useful for adding detail to certain mesh instances.

pub fn lod_max_distance(&self) -> f64

The GeometryInstance’s max LOD distance. Note: This property currently has no effect.

pub fn lod_max_hysteresis(&self) -> f64

The GeometryInstance’s max LOD margin. Note: This property currently has no effect.

pub fn lod_min_distance(&self) -> f64

The GeometryInstance’s min LOD distance. Note: This property currently has no effect.

pub fn lod_min_hysteresis(&self) -> f64

The GeometryInstance’s min LOD margin. Note: This property currently has no effect.

pub fn material_override(&self) -> Option<Ref<Material, Shared>>

The material override for the whole geometry. If a material is assigned to this property, it will be used instead of any material set in any material slot of the mesh.

pub fn set_cast_shadows_setting(&self, shadow_casting_setting: i64)

The selected shadow casting flag. See ShadowCastingSetting for possible values.

pub fn set_custom_aabb(&self, aabb: Aabb)

Overrides the bounding box of this node with a custom one. To remove it, set an AABB with all fields set to zero.

pub fn set_extra_cull_margin(&self, margin: f64)

The extra distance added to the GeometryInstance’s bounding box (AABB) to increase its cull box.

pub fn set_flag(&self, flag: i64, value: bool)

If true, this GeometryInstance will be used when baking lights using a GIProbe or BakedLightmap.

pub fn set_generate_lightmap(&self, enabled: bool)

When disabled, the mesh will be taken into account when computing indirect lighting, but the resulting lightmap will not be saved. Useful for emissive only materials or shadow casters.

pub fn set_lightmap_scale(&self, scale: i64)

Scale factor for the generated baked lightmap. Useful for adding detail to certain mesh instances.

pub fn set_lod_max_distance(&self, mode: f64)

The GeometryInstance’s max LOD distance. Note: This property currently has no effect.

pub fn set_lod_max_hysteresis(&self, mode: f64)

The GeometryInstance’s max LOD margin. Note: This property currently has no effect.

pub fn set_lod_min_distance(&self, mode: f64)

The GeometryInstance’s min LOD distance. Note: This property currently has no effect.

pub fn set_lod_min_hysteresis(&self, mode: f64)

The GeometryInstance’s min LOD margin. Note: This property currently has no effect.

pub fn set_material_override(&self, material: impl AsArg<Material>)

The material override for the whole geometry. If a material is assigned to this property, it will be used instead of any material set in any material slot of the mesh.