Struct gdnative::api::HTTPClient

pub struct HTTPClient { /* fields omitted */ }

core class HTTPClient inherits Reference (reference-counted).

This class has related types in the http_client module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

HTTPClient inherits methods from:

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl HTTPClient

Constants

pub const METHOD_GET: i64

pub const STATUS_DISCONNECTED: i64

pub const METHOD_HEAD: i64

pub const STATUS_RESOLVING: i64

pub const METHOD_POST: i64

pub const STATUS_CANT_RESOLVE: i64

pub const METHOD_PUT: i64

pub const STATUS_CONNECTING: i64

pub const METHOD_DELETE: i64

pub const STATUS_CANT_CONNECT: i64

pub const METHOD_OPTIONS: i64

pub const STATUS_CONNECTED: i64

pub const METHOD_TRACE: i64

pub const STATUS_REQUESTING: i64

pub const METHOD_CONNECT: i64

pub const STATUS_BODY: i64

pub const METHOD_PATCH: i64

pub const STATUS_CONNECTION_ERROR: i64

pub const METHOD_MAX: i64

pub const STATUS_SSL_HANDSHAKE_ERROR: i64

pub const RESPONSE_CONTINUE: i64

pub const RESPONSE_SWITCHING_PROTOCOLS: i64

pub const RESPONSE_PROCESSING: i64

pub const RESPONSE_OK: i64

pub const RESPONSE_CREATED: i64

pub const RESPONSE_ACCEPTED: i64

pub const RESPONSE_NON_AUTHORITATIVE_INFORMATION: i64

pub const RESPONSE_NO_CONTENT: i64

pub const RESPONSE_RESET_CONTENT: i64

pub const RESPONSE_PARTIAL_CONTENT: i64

pub const RESPONSE_MULTI_STATUS: i64

pub const RESPONSE_ALREADY_REPORTED: i64

pub const RESPONSE_IM_USED: i64

pub const RESPONSE_MULTIPLE_CHOICES: i64

pub const RESPONSE_MOVED_PERMANENTLY: i64

pub const RESPONSE_FOUND: i64

pub const RESPONSE_SEE_OTHER: i64

pub const RESPONSE_NOT_MODIFIED: i64

pub const RESPONSE_USE_PROXY: i64

pub const RESPONSE_SWITCH_PROXY: i64

pub const RESPONSE_TEMPORARY_REDIRECT: i64

pub const RESPONSE_PERMANENT_REDIRECT: i64

pub const RESPONSE_BAD_REQUEST: i64

pub const RESPONSE_UNAUTHORIZED: i64

pub const RESPONSE_PAYMENT_REQUIRED: i64

pub const RESPONSE_FORBIDDEN: i64

pub const RESPONSE_NOT_FOUND: i64

pub const RESPONSE_METHOD_NOT_ALLOWED: i64

pub const RESPONSE_NOT_ACCEPTABLE: i64

pub const RESPONSE_PROXY_AUTHENTICATION_REQUIRED: i64

pub const RESPONSE_REQUEST_TIMEOUT: i64

pub const RESPONSE_CONFLICT: i64

pub const RESPONSE_GONE: i64

pub const RESPONSE_LENGTH_REQUIRED: i64

pub const RESPONSE_PRECONDITION_FAILED: i64

pub const RESPONSE_REQUEST_ENTITY_TOO_LARGE: i64

pub const RESPONSE_REQUEST_URI_TOO_LONG: i64

pub const RESPONSE_UNSUPPORTED_MEDIA_TYPE: i64

pub const RESPONSE_REQUESTED_RANGE_NOT_SATISFIABLE: i64

pub const RESPONSE_EXPECTATION_FAILED: i64

pub const RESPONSE_IM_A_TEAPOT: i64

pub const RESPONSE_MISDIRECTED_REQUEST: i64

pub const RESPONSE_UNPROCESSABLE_ENTITY: i64

pub const RESPONSE_LOCKED: i64

pub const RESPONSE_FAILED_DEPENDENCY: i64

pub const RESPONSE_UPGRADE_REQUIRED: i64

pub const RESPONSE_PRECONDITION_REQUIRED: i64

pub const RESPONSE_TOO_MANY_REQUESTS: i64

pub const RESPONSE_REQUEST_HEADER_FIELDS_TOO_LARGE: i64

pub const RESPONSE_UNAVAILABLE_FOR_LEGAL_REASONS: i64

pub const RESPONSE_INTERNAL_SERVER_ERROR: i64

pub const RESPONSE_NOT_IMPLEMENTED: i64

pub const RESPONSE_BAD_GATEWAY: i64

pub const RESPONSE_SERVICE_UNAVAILABLE: i64

pub const RESPONSE_GATEWAY_TIMEOUT: i64

pub const RESPONSE_HTTP_VERSION_NOT_SUPPORTED: i64

pub const RESPONSE_VARIANT_ALSO_NEGOTIATES: i64

pub const RESPONSE_INSUFFICIENT_STORAGE: i64

pub const RESPONSE_LOOP_DETECTED: i64

pub const RESPONSE_NOT_EXTENDED: i64

pub const RESPONSE_NETWORK_AUTH_REQUIRED: i64

impl HTTPClient

pub fn new() -> Ref<HTTPClient, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn close(&self)

Closes the current connection, allowing reuse of this HTTPClient.

pub fn connect_to_host(
&self,
host: impl Into<GodotString>,
port: i64,
use_ssl: bool,
verify_host: bool
) -> Result<(), GodotError>

Connects to a host. This needs to be done before any requests are sent. The host should not have http:// prepended but will strip the protocol identifier if provided. If no port is specified (or -1 is used), it is automatically set to 80 for HTTP and 443 for HTTPS (if use_ssl is enabled). verify_host will check the SSL identity of the host if set to true.

Default Arguments

port - -1

use_ssl - false

verify_host - true

pub fn connection(&self) -> Option<Ref<StreamPeer, Shared>>

The connection to use for this client.

pub fn read_chunk_size(&self) -> i64

The size of the buffer used and maximum bytes to read per iteration. See read_response_body_chunk.

pub fn get_response_body_length(&self) -> i64

Returns the response’s body length. Note: Some Web servers may not send a body length. In this case, the value returned will be -1. If using chunked transfer encoding, the body length will also be -1.

pub fn get_response_code(&self) -> i64

Returns the response’s HTTP status code.

pub fn get_response_headers(&self) -> PoolArray<GodotString>

Returns the response headers.

pub fn get_response_headers_as_dictionary(&self) -> Dictionary<Shared>

Sample code is GDScript unless otherwise noted.

Returns all response headers as a Dictionary of structure { "key": "value1; value2" } where the case-sensitivity of the keys and values is kept like the server delivers it. A value is a simple String, this string can have more than one value where “; “ is used as separator. Example:

{
    "content-length": 12,
    "Content-Type": "application/json; charset=UTF-8",
}

pub fn get_status(&self) -> Status

Returns a Status constant. Need to call poll in order to get status updates.

pub fn has_response(&self) -> bool

If true, this HTTPClient has a response available.

pub fn is_blocking_mode_enabled(&self) -> bool

If true, execution will block until all data is read from the response.

pub fn is_response_chunked(&self) -> bool

If true, this HTTPClient has a response that is chunked.

pub fn poll(&self) -> Result<(), GodotError>

This needs to be called in order to have any request processed. Check results with get_status.

pub fn query_string_from_dict(&self, fields: Dictionary<Shared>) -> GodotString

Sample code is GDScript unless otherwise noted.

Generates a GET/POST application/x-www-form-urlencoded style query string from a provided dictionary, e.g.:

var fields = {"username": "user", "password": "pass"}
var query_string = http_client.query_string_from_dict(fields)

Furthermore, if a key has a null value, only the key itself is added, without equal sign and value. If the value is an array, for each value in it a pair with the same key is added.

var fields = {"single": 123, "not_valued": null, "multiple": [22, 33, 44]}
var query_string = http_client.query_string_from_dict(fields)

pub fn read_response_body_chunk(&self) -> PoolArray<u8>

Reads one chunk from the response.

pub fn request(
&self,
method: i64,
url: impl Into<GodotString>,
headers: PoolArray<GodotString>,
body: impl Into<GodotString>
) -> Result<(), GodotError>

Sample code is GDScript unless otherwise noted.

Sends a request to the connected host. The URL parameter is just the part after the host, so for http://somehost.com/index.php, it is index.php. Headers are HTTP request headers. For available HTTP methods, see Method. To create a POST request with query strings to push to the server, do:

var fields = {"username" : "user", "password" : "pass"}
var query_string = http_client.query_string_from_dict(fields)
var headers = ["Content-Type: application/x-www-form-urlencoded", "Content-Length: " + str(query_string.length())]
var result = http_client.request(http_client.METHOD_POST, "index.php", headers, query_string)

Note: The request_data parameter is ignored if method is HTTPClient.METHOD_GET. This is because GET methods can’t contain request data. As a workaround, you can pass request data as a query string in the URL. See String.http_escape for an example.

Default Arguments

body - ""

pub fn request_raw(
&self,
method: i64,
url: impl Into<GodotString>,
headers: PoolArray<GodotString>,
body: PoolArray<u8>
) -> Result<(), GodotError>

Sends a raw request to the connected host. The URL parameter is just the part after the host, so for http://somehost.com/index.php, it is index.php. Headers are HTTP request headers. For available HTTP methods, see Method. Sends the body data raw, as a byte array and does not encode it in any way.

pub fn set_blocking_mode(&self, enabled: bool)

If true, execution will block until all data is read from the response.

pub fn set_connection(&self, connection: impl AsArg<StreamPeer>)

The connection to use for this client.

pub fn set_read_chunk_size(&self, bytes: i64)

The size of the buffer used and maximum bytes to read per iteration. See read_response_body_chunk.