Struct gdnative::api::Input

pub struct Input { /* fields omitted */ }

core singleton class Input inherits Object (manually managed).

This class has related types in the input module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Class hierarchy

Input inherits methods from:

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Input

Constants

pub const CURSOR_ARROW: i64

pub const MOUSE_MODE_VISIBLE: i64

pub const CURSOR_IBEAM: i64

pub const MOUSE_MODE_HIDDEN: i64

pub const CURSOR_POINTING_HAND: i64

pub const MOUSE_MODE_CAPTURED: i64

pub const CURSOR_CROSS: i64

pub const MOUSE_MODE_CONFINED: i64

pub const CURSOR_WAIT: i64

pub const CURSOR_BUSY: i64

pub const CURSOR_DRAG: i64

pub const CURSOR_CAN_DROP: i64

pub const CURSOR_FORBIDDEN: i64

pub const CURSOR_VSIZE: i64

pub const CURSOR_HSIZE: i64

pub const CURSOR_BDIAGSIZE: i64

pub const CURSOR_FDIAGSIZE: i64

pub const CURSOR_MOVE: i64

pub const CURSOR_VSPLIT: i64

pub const CURSOR_HSPLIT: i64

pub const CURSOR_HELP: i64

impl Input

pub fn godot_singleton() -> &'static Input

Returns a reference to the singleton instance.

pub fn action_press(&self, action: impl Into<GodotString>, strength: f64)

This will simulate pressing the specified action. The strength can be used for non-boolean actions, it’s ranged between 0 and 1 representing the intensity of the given action. Note: This method will not cause any [Node._input][Node::_input] calls. It is intended to be used with is_action_pressed and is_action_just_pressed. If you want to simulate _input, use parse_input_event instead.

Default Arguments

strength - 1.0

pub fn action_release(&self, action: impl Into<GodotString>)

If the specified action is already pressed, this will release it.

pub fn add_joy_mapping(
&self,
mapping: impl Into<GodotString>,
update_existing: bool
)

Adds a new mapping entry (in SDL2 format) to the mapping database. Optionally update already connected devices.

Default Arguments

update_existing - false

pub fn get_accelerometer(&self) -> Vector3

Returns the acceleration of the device’s accelerometer sensor, if the device has one. Otherwise, the method returns Vector3.ZERO. Note this method returns an empty Vector3 when running from the editor even when your device has an accelerometer. You must export your project to a supported device to read values from the accelerometer. Note: This method only works on iOS, Android, and UWP. On other platforms, it always returns Vector3.ZERO. On Android the unit of measurement for each axis is m/s² while on iOS and UWP it’s a multiple of the Earth’s gravitational acceleration g (~9.81 m/s²).

pub fn get_action_strength(&self, action: impl Into<GodotString>) -> f64

Returns a value between 0 and 1 representing the intensity of the given action. In a joypad, for example, the further away the axis (analog sticks or L2, R2 triggers) is from the dead zone, the closer the value will be to 1. If the action is mapped to a control that has no axis as the keyboard, the value returned will be 0 or 1.

pub fn get_connected_joypads(&self) -> VariantArray<Shared>

Returns an Array containing the device IDs of all currently connected joypads.

pub fn get_current_cursor_shape(&self) -> CursorShape

Returns the currently assigned cursor shape (see CursorShape).

pub fn get_gravity(&self) -> Vector3

Returns the gravity of the device’s accelerometer sensor, if the device has one. Otherwise, the method returns Vector3.ZERO. Note: This method only works on Android and iOS. On other platforms, it always returns Vector3.ZERO. On Android the unit of measurement for each axis is m/s² while on iOS it’s a multiple of the Earth’s gravitational acceleration g (~9.81 m/s²).

pub fn get_gyroscope(&self) -> Vector3

Returns the rotation rate in rad/s around a device’s X, Y, and Z axes of the gyroscope sensor, if the device has one. Otherwise, the method returns Vector3.ZERO. Note: This method only works on Android and iOS. On other platforms, it always returns Vector3.ZERO.

pub fn get_joy_axis(&self, device: i64, axis: i64) -> f64

Returns the current value of the joypad axis at given index (see [JoystickList][JoystickList]).

pub fn get_joy_axis_index_from_string(
&self,
axis: impl Into<GodotString>
) -> i64

Returns the index of the provided axis name.

pub fn get_joy_axis_string(&self, axis_index: i64) -> GodotString

Receives a [JoystickList][JoystickList] axis and returns its equivalent name as a string.

pub fn get_joy_button_index_from_string(
&self,
button: impl Into<GodotString>
) -> i64

Returns the index of the provided button name.

pub fn get_joy_button_string(&self, button_index: i64) -> GodotString

Receives a gamepad button from [JoystickList][JoystickList] and returns its equivalent name as a string.

pub fn get_joy_guid(&self, device: i64) -> GodotString

Returns a SDL2-compatible device GUID on platforms that use gamepad remapping. Returns "Default Gamepad" otherwise.

pub fn get_joy_name(&self, device: i64) -> GodotString

Returns the name of the joypad at the specified device index.

pub fn get_joy_vibration_duration(&self, device: i64) -> f64

Returns the duration of the current vibration effect in seconds.

pub fn get_joy_vibration_strength(&self, device: i64) -> Vector2

Returns the strength of the joypad vibration: x is the strength of the weak motor, and y is the strength of the strong motor.

pub fn get_last_mouse_speed(&self) -> Vector2

Returns the mouse speed for the last time the cursor was moved, and this until the next frame where the mouse moves. This means that even if the mouse is not moving, this function will still return the value of the last motion.

pub fn get_magnetometer(&self) -> Vector3

Returns the magnetic field strength in micro-Tesla for all axes of the device’s magnetometer sensor, if the device has one. Otherwise, the method returns Vector3.ZERO. Note: This method only works on Android, iOS and UWP. On other platforms, it always returns Vector3.ZERO.

pub fn get_mouse_button_mask(&self) -> i64

Returns mouse buttons as a bitmask. If multiple mouse buttons are pressed at the same time, the bits are added together.

pub fn get_mouse_mode(&self) -> MouseMode

Returns the mouse mode. See the constants for more information.

pub fn is_action_just_pressed(&self, action: impl Into<GodotString>) -> bool

Returns true when the user starts pressing the action event, meaning it’s true only on the frame that the user pressed down the button. This is useful for code that needs to run only once when an action is pressed, instead of every frame while it’s pressed.

pub fn is_action_just_released(&self, action: impl Into<GodotString>) -> bool

Returns true when the user stops pressing the action event, meaning it’s true only on the frame that the user released the button.

pub fn is_action_pressed(&self, action: impl Into<GodotString>) -> bool

Returns true if you are pressing the action event. Note that if an action has multiple buttons assigned and more than one of them is pressed, releasing one button will release the action, even if some other button assigned to this action is still pressed.

pub fn is_joy_button_pressed(&self, device: i64, button: i64) -> bool

Returns true if you are pressing the joypad button (see [JoystickList][JoystickList]).

pub fn is_joy_known(&self, device: i64) -> bool

Returns true if the system knows the specified device. This means that it sets all button and axis indices exactly as defined in [JoystickList][JoystickList]. Unknown joypads are not expected to match these constants, but you can still retrieve events from them.

pub fn is_key_pressed(&self, scancode: i64) -> bool

Returns true if you are pressing the key. You can pass a [KeyList][KeyList] constant.

pub fn is_mouse_button_pressed(&self, button: i64) -> bool

Returns true if you are pressing the mouse button specified with [ButtonList][ButtonList].

pub fn joy_connection_changed(
&self,
device: i64,
connected: bool,
name: impl Into<GodotString>,
guid: impl Into<GodotString>
)

Notifies the Input singleton that a connection has changed, to update the state for the device index. This is used internally and should not have to be called from user scripts. See joy_connection_changed for the signal emitted when this is triggered internally.

pub fn parse_input_event(&self, event: impl AsArg<InputEvent>)

Sample code is GDScript unless otherwise noted.

Feeds an InputEvent to the game. Can be used to artificially trigger input events from code. Also generates [Node._input][Node::_input] calls. Example:

var a = InputEventAction.new()
a.action = "ui_cancel"
a.pressed = true
Input.parse_input_event(a)

pub fn remove_joy_mapping(&self, guid: impl Into<GodotString>)

Removes all mappings from the internal database that match the given GUID.

pub fn set_custom_mouse_cursor(
&self,
image: impl AsArg<Resource>,
shape: i64,
hotspot: Vector2
)

Sets a custom mouse cursor image, which is only visible inside the game window. The hotspot can also be specified. Passing null to the image parameter resets to the system cursor. See CursorShape for the list of shapes. image’s size must be lower than 256×256. hotspot must be within image’s size. Note: AnimatedTextures aren’t supported as custom mouse cursors. If using an AnimatedTexture, only the first frame will be displayed. Note: Only images imported with the Lossless, Lossy or Uncompressed compression modes are supported. The Video RAM compression mode can’t be used for custom cursors.

Default Arguments

shape - 0

hotspot - Vector2( 0, 0 )

pub fn set_default_cursor_shape(&self, shape: i64)

Sets the default cursor shape to be used in the viewport instead of CURSOR_ARROW. Note: If you want to change the default cursor shape for Control’s nodes, use [Control.mouse_default_cursor_shape][Control::mouse_default_cursor_shape] instead. Note: This method generates an InputEventMouseMotion to update cursor immediately.

Default Arguments

shape - 0

pub fn set_mouse_mode(&self, mode: i64)

Sets the mouse mode. See the constants for more information.

pub fn set_use_accumulated_input(&self, enable: bool)

Enables or disables the accumulation of similar input events sent by the operating system. When input accumulation is enabled, all input events generated during a frame will be merged and emitted when the frame is done rendering. Therefore, this limits the number of input method calls per second to the rendering FPS. Input accumulation is enabled by default. It can be disabled to get slightly more precise/reactive input at the cost of increased CPU usage. In applications where drawing freehand lines is required, input accumulation should generally be disabled while the user is drawing the line to get results that closely follow the actual input.

pub fn start_joy_vibration(
&self,
device: i64,
weak_magnitude: f64,
strong_magnitude: f64,
duration: f64
)

Starts to vibrate the joypad. Joypads usually come with two rumble motors, a strong and a weak one. weak_magnitude is the strength of the weak motor (between 0 and 1) and strong_magnitude is the strength of the strong motor (between 0 and 1). duration is the duration of the effect in seconds (a duration of 0 will try to play the vibration indefinitely). Note: Not every hardware is compatible with long effect durations; it is recommended to restart an effect if it has to be played for more than a few seconds.

Default Arguments

duration - 0

pub fn stop_joy_vibration(&self, device: i64)

Stops the vibration of the joypad.

pub fn vibrate_handheld(&self, duration_ms: i64)

Vibrate Android and iOS devices. Note: It needs VIBRATE permission for Android at export settings. iOS does not support duration.

Default Arguments

duration_ms - 500

pub fn warp_mouse_position(&self, to: Vector2)

Sets the mouse position to the specified vector.