Struct gdnative::api::InputEventScreenTouch

pub struct InputEventScreenTouch { /* fields omitted */ }

core class InputEventScreenTouch inherits InputEvent (reference-counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

InputEventScreenTouch inherits methods from:

InputEvent

Resource

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl InputEventScreenTouch

pub fn new() -> Ref<InputEventScreenTouch, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn index(&self) -> i64

The touch index in the case of a multi-touch event. One index = one finger.

pub fn position(&self) -> Vector2

The touch position.

pub fn set_index(&self, index: i64)

The touch index in the case of a multi-touch event. One index = one finger.

pub fn set_position(&self, position: Vector2)

The touch position.

pub fn set_pressed(&self, pressed: bool)

If true, the touch’s state is pressed. If false, the touch’s state is released.