Struct gdnative::api::Joint

pub struct Joint { /* fields omitted */ }

core class Joint inherits Spatial (manually managed).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Class hierarchy

Joint inherits methods from:

Spatial

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Joint

pub fn exclude_nodes_from_collision(&self) -> bool

If true, the two bodies of the nodes are not able to collide with each other.

pub fn node_a(&self) -> NodePath

The node attached to the first side (A) of the joint.

pub fn node_b(&self) -> NodePath

The node attached to the second side (B) of the joint.

pub fn solver_priority(&self) -> i64

The priority used to define which solver is executed first for multiple joints. The lower the value, the higher the priority.

pub fn set_exclude_nodes_from_collision(&self, enable: bool)

If true, the two bodies of the nodes are not able to collide with each other.

pub fn set_node_a(&self, node: impl Into<NodePath>)

The node attached to the first side (A) of the joint.

pub fn set_node_b(&self, node: impl Into<NodePath>)

The node attached to the second side (B) of the joint.

pub fn set_solver_priority(&self, priority: i64)

The priority used to define which solver is executed first for multiple joints. The lower the value, the higher the priority.