Struct gdnative::api::KinematicCollision2D

pub struct KinematicCollision2D { /* private fields */ }

core class KinematicCollision2D inherits Reference (reference-counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

KinematicCollision2D inherits methods from:

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl KinematicCollision2D

pub fn new() -> Ref<KinematicCollision2D, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn get_angle(&self, up_direction: Vector2) -> f64

The collision angle according to up_direction, which is Vector2.UP by default. This value is always positive.

Default Arguments

up_direction - Vector2( 0, -1 )

pub fn collider(&self) -> Option<Ref<Object>>

The colliding body.

pub fn collider_id(&self) -> i64

The colliding body’s unique instance ID. See Object.get_instance_id.

pub fn collider_metadata(&self) -> Variant

The colliding body’s metadata. See Object.

pub fn collider_rid(&self) -> Rid

The colliding body’s RID used by the Physics2DServer.

pub fn collider_shape(&self) -> Option<Ref<Object>>

The colliding body’s shape.

pub fn collider_shape_index(&self) -> i64

The colliding shape’s index. See CollisionObject2D.

pub fn collider_velocity(&self) -> Vector2

The colliding object’s velocity.

pub fn local_shape(&self) -> Option<Ref<Object>>

The moving object’s colliding shape.

pub fn normal(&self) -> Vector2

The colliding body’s shape’s normal at the point of collision.

pub fn position(&self) -> Vector2

The point of collision, in global coordinates.

pub fn remainder(&self) -> Vector2

The moving object’s remaining movement vector.

pub fn travel(&self) -> Vector2

The distance the moving object traveled before collision.