Struct gdnative::api::Listener

pub struct Listener { /* fields omitted */ }

core class Listener inherits Spatial (manually managed).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

Listener is a reference-only type. Persistent references can only exist in the unsafe Ref<Listener> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

Listener inherits methods from:

Spatial

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Listener

pub fn new() -> Ref<Listener, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn clear_current(&self)

Disables the listener to use the current camera’s listener instead.

pub fn get_listener_transform(&self) -> Transform

Returns the listener’s global orthonormalized Transform.

pub fn is_current(&self) -> bool

Returns true if the listener was made current using make_current, false otherwise. Note: There may be more than one Listener marked as “current” in the scene tree, but only the one that was made current last will be used.

pub fn make_current(&self)

Enables the listener. This will override the current camera’s listener.