Struct gdnative::api::MeshInstance

pub struct MeshInstance { /* fields omitted */ }

core class MeshInstance inherits GeometryInstance (manually managed).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

MeshInstance is a reference-only type. Persistent references can only exist in the unsafe Ref<MeshInstance> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

MeshInstance inherits methods from:

GeometryInstance

VisualInstance

CullInstance

Spatial

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl MeshInstance

pub fn new() -> Ref<MeshInstance, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn create_convex_collision(&self, clean: bool, simplify: bool)

This helper creates a StaticBody child node with a ConvexPolygonShape collision shape calculated from the mesh geometry. It’s mainly used for testing.

pub fn create_debug_tangents(&self)

This helper creates a MeshInstance child node with gizmos at every vertex calculated from the mesh geometry. It’s mainly used for testing.

pub fn create_multiple_convex_collisions(&self)

pub fn create_trimesh_collision(&self)

This helper creates a StaticBody child node with a ConcavePolygonShape collision shape calculated from the mesh geometry. It’s mainly used for testing.

pub fn get_active_material(&self, surface: i64) -> Option<Ref<Material, Shared>>

Returns the Material that will be used by the Mesh when drawing. This can return the GeometryInstance.material_override, the surface override Material defined in this MeshInstance, or the surface Material defined in the Mesh. For example, if GeometryInstance.material_override is used, all surfaces will return the override material.

pub fn mesh(&self) -> Option<Ref<Mesh, Shared>>

The Mesh resource for the instance.

pub fn skeleton_path(&self) -> NodePath

NodePath to the Skeleton associated with the instance.

pub fn skin(&self) -> Option<Ref<Skin, Shared>>

Sets the skin to be used by this instance.

pub fn get_surface_material(
&self,
surface: i64
) -> Option<Ref<Material, Shared>>

Returns the Material for a surface of the Mesh resource.

pub fn get_surface_material_count(&self) -> i64

Returns the number of surface materials.

pub fn is_software_skinning_transform_normals_enabled(&self) -> bool

If true, normals are transformed when software skinning is used. Set to false when normals are not needed for better performance. See [member ProjectSettings.rendering/quality/skinning/software_skinning_fallback] for details about how software skinning is enabled.

pub fn set_mesh(&self, mesh: impl AsArg<Mesh>)

The Mesh resource for the instance.

pub fn set_skeleton_path(&self, skeleton_path: impl Into<NodePath>)

NodePath to the Skeleton associated with the instance.

pub fn set_skin(&self, skin: impl AsArg<Skin>)

Sets the skin to be used by this instance.

pub fn set_software_skinning_transform_normals(&self, enabled: bool)

If true, normals are transformed when software skinning is used. Set to false when normals are not needed for better performance. See [member ProjectSettings.rendering/quality/skinning/software_skinning_fallback] for details about how software skinning is enabled.

pub fn set_surface_material(&self, surface: i64, material: impl AsArg<Material>)

Sets the Material for a surface of the Mesh resource.