Struct gdnative::api::MeshLibrary

pub struct MeshLibrary { /* fields omitted */ }

core class MeshLibrary inherits Resource (reference-counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

MeshLibrary inherits methods from:

Resource

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl MeshLibrary

pub fn new() -> Ref<MeshLibrary, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn clear(&self)

Clears the library.

pub fn create_item(&self, id: i64)

Creates a new item in the library with the given ID. You can get an unused ID from get_last_unused_item_id.

pub fn find_item_by_name(&self, name: impl Into<GodotString>) -> i64

Returns the first item with the given name.

pub fn get_item_list(&self) -> TypedArray<i32>

Returns the list of item IDs in use.

pub fn get_item_mesh(&self, id: i64) -> Option<Ref<Mesh, Shared>>

Returns the item’s mesh.

pub fn get_item_name(&self, id: i64) -> GodotString

Returns the item’s name.

pub fn get_item_navmesh(&self, id: i64) -> Option<Ref<NavigationMesh, Shared>>

Returns the item’s navigation mesh.

pub fn get_item_navmesh_transform(&self, id: i64) -> Transform

Returns the transform applied to the item’s navigation mesh.

pub fn get_item_preview(&self, id: i64) -> Option<Ref<Texture, Shared>>

When running in the editor, returns a generated item preview (a 3D rendering in isometric perspective). When used in a running project, returns the manually-defined item preview which can be set using set_item_preview. Returns an empty Texture if no preview was manually set in a running project.

pub fn get_item_shapes(&self, id: i64) -> VariantArray<Shared>

Returns an item’s collision shapes. The array consists of each Shape followed by its Transform.

pub fn get_last_unused_item_id(&self) -> i64

Gets an unused ID for a new item.

pub fn remove_item(&self, id: i64)

Removes the item.

pub fn set_item_mesh(&self, id: i64, mesh: impl AsArg<Mesh>)

Sets the item’s mesh.

pub fn set_item_name(&self, id: i64, name: impl Into<GodotString>)

Sets the item’s name. This name is shown in the editor. It can also be used to look up the item later using find_item_by_name.

pub fn set_item_navmesh(&self, id: i64, navmesh: impl AsArg<NavigationMesh>)

Sets the item’s navigation mesh.

pub fn set_item_navmesh_transform(&self, id: i64, navmesh: Transform)

Sets the transform to apply to the item’s navigation mesh.

pub fn set_item_preview(&self, id: i64, texture: impl AsArg<Texture>)

Sets a texture to use as the item’s preview icon in the editor.

pub fn set_item_shapes(&self, id: i64, shapes: VariantArray<Shared>)

Sets an item’s collision shapes. The array should consist of Shape objects, each followed by a Transform that will be applied to it. For shapes that should not have a transform, use [Transform.IDENTITY][Transform::IDENTITY].