Struct gdnative::api::NetworkedMultiplayerENet

pub struct NetworkedMultiplayerENet { /* private fields */ }

core class NetworkedMultiplayerENet inherits NetworkedMultiplayerPeer (reference-counted).

This class has related types in the networked_multiplayer_enet module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

NetworkedMultiplayerENet inherits methods from:

NetworkedMultiplayerPeer

PacketPeer

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl NetworkedMultiplayerENet

Constants

pub const COMPRESS_NONE: i64 = 0i64

pub const COMPRESS_RANGE_CODER: i64 = 1i64

pub const COMPRESS_FASTLZ: i64 = 2i64

pub const COMPRESS_ZLIB: i64 = 3i64

pub const COMPRESS_ZSTD: i64 = 4i64

impl NetworkedMultiplayerENet

pub fn new() -> Ref<NetworkedMultiplayerENet, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn close_connection(&self, wait_usec: i64)

pub fn create_client( &self, address: impl Into<GodotString>, port: i64, in_bandwidth: i64, out_bandwidth: i64, client_port: i64 ) -> Result<(), GodotError>

pub fn create_server( &self, port: i64, max_clients: i64, in_bandwidth: i64, out_bandwidth: i64 ) -> Result<(), GodotError>

pub fn disconnect_peer(&self, id: i64, now: bool)

pub fn channel_count(&self) -> i64

pub fn compression_mode(&self) -> CompressionMode

pub fn dtls_hostname(&self) -> GodotString

pub fn get_last_packet_channel(&self) -> i64

pub fn get_packet_channel(&self) -> i64

pub fn get_peer_address(&self, id: i64) -> GodotString

pub fn get_peer_port(&self, id: i64) -> i64

pub fn transfer_channel(&self) -> i64

pub fn is_always_ordered(&self) -> bool

pub fn is_dtls_enabled(&self) -> bool

pub fn is_dtls_verify_enabled(&self) -> bool

pub fn is_server_relay_enabled(&self) -> bool

pub fn set_always_ordered(&self, ordered: bool)

pub fn set_bind_ip(&self, ip: impl Into<GodotString>)

pub fn set_channel_count(&self, channels: i64)

pub fn set_compression_mode(&self, mode: i64)

pub fn set_dtls_certificate(&self, certificate: impl AsArg<X509Certificate>)

pub fn set_dtls_enabled(&self, enabled: bool)

pub fn set_dtls_hostname(&self, hostname: impl Into<GodotString>)

pub fn set_dtls_key(&self, key: impl AsArg<CryptoKey>)

pub fn set_dtls_verify_enabled(&self, enabled: bool)

pub fn set_peer_timeout( &self, id: i64, timeout_limit: i64, timeout_min: i64, timeout_max: i64 )

pub fn set_server_relay_enabled(&self, enabled: bool)

pub fn set_transfer_channel(&self, channel: i64)