Struct gdnative::api::Node2D

pub struct Node2D { /* fields omitted */ }

core class Node2D inherits CanvasItem (manually managed).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

Node2D is a reference-only type. Persistent references can only exist in the unsafe Ref<Node2D> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

Node2D inherits methods from:

CanvasItem

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Node2D

pub fn new() -> Ref<Node2D, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn apply_scale(&self, ratio: Vector2)

Multiplies the current scale by the ratio vector.

pub fn get_angle_to(&self, point: Vector2) -> f64

Returns the angle between the node and the point in radians.

pub fn global_position(&self) -> Vector2

Global position.

pub fn global_rotation(&self) -> f64

Global rotation in radians.

pub fn global_rotation_degrees(&self) -> f64

Global rotation in degrees.

pub fn global_scale(&self) -> Vector2

Global scale.

pub fn position(&self) -> Vector2

Position, relative to the node’s parent.

pub fn get_relative_transform_to_parent(
&self,
parent: impl AsArg<Node>
) -> Transform2D

Returns the Transform2D relative to this node’s parent.

pub fn rotation(&self) -> f64

Rotation in radians, relative to the node’s parent.

pub fn rotation_degrees(&self) -> f64

Rotation in degrees, relative to the node’s parent.

pub fn scale(&self) -> Vector2

The node’s scale. Unscaled value: (1, 1).

pub fn z_index(&self) -> i64

Z index. Controls the order in which the nodes render. A node with a higher Z index will display in front of others.

pub fn global_translate(&self, offset: Vector2)

Adds the offset vector to the node’s global position.

pub fn is_z_relative(&self) -> bool

If true, the node’s Z index is relative to its parent’s Z index. If this node’s Z index is 2 and its parent’s effective Z index is 3, then this node’s effective Z index will be 2 + 3 = 5.

pub fn look_at(&self, point: Vector2)

Rotates the node so it points towards the point, which is expected to use global coordinates.

pub fn move_local_x(&self, delta: f64, scaled: bool)

Applies a local translation on the node’s X axis based on the [Node._process][Node::_process]’s delta. If scaled is false, normalizes the movement.

Default Arguments

scaled - false

pub fn move_local_y(&self, delta: f64, scaled: bool)

Applies a local translation on the node’s Y axis based on the [Node._process][Node::_process]’s delta. If scaled is false, normalizes the movement.

Default Arguments

scaled - false

pub fn rotate(&self, radians: f64)

Applies a rotation to the node, in radians, starting from its current rotation.

pub fn set_global_position(&self, position: Vector2)

Global position.

pub fn set_global_rotation(&self, radians: f64)

Global rotation in radians.

pub fn set_global_rotation_degrees(&self, degrees: f64)

Global rotation in degrees.

pub fn set_global_scale(&self, scale: Vector2)

Global scale.

pub fn set_global_transform(&self, xform: Transform2D)

Global Transform2D.

pub fn set_position(&self, position: Vector2)

Position, relative to the node’s parent.

pub fn set_rotation(&self, radians: f64)

Rotation in radians, relative to the node’s parent.

pub fn set_rotation_degrees(&self, degrees: f64)

Rotation in degrees, relative to the node’s parent.

pub fn set_scale(&self, scale: Vector2)

The node’s scale. Unscaled value: (1, 1).

pub fn set_transform(&self, xform: Transform2D)

Local Transform2D.

pub fn set_z_as_relative(&self, enable: bool)

If true, the node’s Z index is relative to its parent’s Z index. If this node’s Z index is 2 and its parent’s effective Z index is 3, then this node’s effective Z index will be 2 + 3 = 5.

pub fn set_z_index(&self, z_index: i64)

Z index. Controls the order in which the nodes render. A node with a higher Z index will display in front of others.

pub fn to_global(&self, local_point: Vector2) -> Vector2

Transforms the provided local position into a position in global coordinate space. The input is expected to be local relative to the Node2D it is called on. e.g. Applying this method to the positions of child nodes will correctly transform their positions into the global coordinate space, but applying it to a node’s own position will give an incorrect result, as it will incorporate the node’s own transformation into its global position.

pub fn to_local(&self, global_point: Vector2) -> Vector2

Transforms the provided global position into a position in local coordinate space. The output will be local relative to the Node2D it is called on. e.g. It is appropriate for determining the positions of child nodes, but it is not appropriate for determining its own position relative to its parent.

pub fn translate(&self, offset: Vector2)

Translates the node by the given offset in local coordinates.