Struct gdnative::api::Object

pub struct Object { /* private fields */ }

The base class of all classes in the Godot hierarchy.

This class has related types in the object module.

Official documentation

See the documentation of this class in the Godot engines official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

Object is a reference-only type. Persistent references can only exist in the unsafe Ref<Object> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Object

Constants

pub const NOTIFICATION_POSTINITIALIZE: i64

pub const CONNECT_DEFERRED: i64

pub const NOTIFICATION_PREDELETE: i64

pub const CONNECT_PERSIST: i64

pub const CONNECT_ONESHOT: i64

pub const CONNECT_REFERENCE_COUNTED: i64

impl Object

pub fn new() -> Ref<Object, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn add_user_signal(
&self,
signal: impl Into<GodotString>,
arguments: VariantArray<Shared>
)

Adds a user-defined signal. Arguments are optional, but can be added as an Array of dictionaries, each containing name: String and type: int (see [enum Variant.Type]) entries.

Default Arguments

arguments - [ ]

pub unsafe fn call(
&self,
method: impl Into<GodotString>,
varargs: &[Variant]
) -> Variant

Sample code is GDScript unless otherwise noted.

Calls the method on the object and returns the result. This method supports a variable number of arguments, so parameters are passed as a comma separated list. Example:

call("set", "position", Vector2(42.0, 0.0))

Note: In C#, the method name must be specified as snake_case if it is defined by a built-in Godot node. This doesnt apply to user-defined methods where you should use the same convention as in the C# source (typically PascalCase).

Safety

This function bypasses Rusts static type checks (aliasing, thread boundaries, calls to free(), ).

pub unsafe fn call_deferred(
&self,
method: impl Into<GodotString>,
varargs: &[Variant]
) -> Variant

Sample code is GDScript unless otherwise noted.

Calls the method on the object during idle time. This method supports a variable number of arguments, so parameters are passed as a comma separated list. Example:

call_deferred("set", "position", Vector2(42.0, 0.0))

Note: In C#, the method name must be specified as snake_case if it is defined by a built-in Godot node. This doesnt apply to user-defined methods where you should use the same convention as in the C# source (typically PascalCase).

Safety

This function bypasses Rusts static type checks (aliasing, thread boundaries, calls to free(), ).

pub unsafe fn callv(
&self,
method: impl Into<GodotString>,
arg_array: VariantArray<Shared>
) -> Variant

Sample code is GDScript unless otherwise noted.

Calls the method on the object and returns the result. Contrarily to call, this method does not support a variable number of arguments but expects all parameters to be via a single Array.

callv("set", [ "position", Vector2(42.0, 0.0) ])

Safety

This function bypasses Rusts static type checks (aliasing, thread boundaries, calls to free(), ).

pub fn can_translate_messages(&self) -> bool

Returns true if the object can translate strings. See set_message_translation and tr.

pub fn connect(
&self,
signal: impl Into<GodotString>,
target: impl AsArg<Object>,
method: impl Into<GodotString>,
binds: VariantArray<Shared>,
flags: i64
) -> Result<(), GodotError>

Sample code is GDScript unless otherwise noted.

Connects a signal to a method on a target object. Pass optional binds to the call as an Array of parameters. These parameters will be passed to the method after any parameter used in the call to emit_signal. Use flags to set deferred or one-shot connections. See ConnectFlags constants. A signal can only be connected once to a method. It will throw an error if already connected, unless the signal was connected with CONNECT_REFERENCE_COUNTED. To avoid this, first, use is_connected to check for existing connections. If the target is destroyed in the games lifecycle, the connection will be lost. Examples:

connect("pressed", self, "_on_Button_pressed") # BaseButton signal
connect("text_entered", self, "_on_LineEdit_text_entered") # LineEdit signal
connect("hit", self, "_on_Player_hit", [ weapon_type, damage ]) # User-defined signal

An example of the relationship between binds passed to connect and parameters used when calling emit_signal:

connect("hit", self, "_on_Player_hit", [ weapon_type, damage ]) # weapon_type and damage are passed last
emit_signal("hit", "Dark lord", 5) # "Dark lord" and 5 are passed first
func _on_Player_hit(hit_by, level, weapon_type, damage):
    print("Hit by %s (lvl %d) with weapon %s for %d damage" % [hit_by, level, weapon_type, damage])

Default Arguments

binds - [ ]

flags - 0

pub fn disconnect(
&self,
signal: impl Into<GodotString>,
target: impl AsArg<Object>,
method: impl Into<GodotString>
)

Disconnects a signal from a method on the given target. If you try to disconnect a connection that does not exist, the method will throw an error. Use is_connected to ensure that the connection exists.

pub fn emit_signal(
&self,
signal: impl Into<GodotString>,
varargs: &[Variant]
) -> Variant

Sample code is GDScript unless otherwise noted.

Emits the given signal. The signal must exist, so it should be a built-in signal of this class or one of its parent classes, or a user-defined signal. This method supports a variable number of arguments, so parameters are passed as a comma separated list. Example:

emit_signal("hit", weapon_type, damage)
emit_signal("game_over")

pub fn get(&self, property: impl Into<GodotString>) -> Variant

Returns the Variant value of the given property. If the property doesnt exist, this will return null. Note: In C#, the property name must be specified as snake_case if it is defined by a built-in Godot node. This doesnt apply to user-defined properties where you should use the same convention as in the C# source (typically PascalCase).

pub fn get_class(&self) -> GodotString

Returns the objects class as a String. See also is_class. Note: get_class does not take class_name declarations into account. If the object has a class_name defined, the base class name will be returned instead.

pub fn get_incoming_connections(&self) -> VariantArray<Shared>

Returns an Array of dictionaries with information about signals that are connected to the object. Each Dictionary contains three String entries:

source is a reference to the signal emitter.

signal_name is the name of the connected signal.

method_name is the name of the method to which the signal is connected.

pub fn get_indexed(&self, property: impl Into<NodePath>) -> Variant

Gets the objects property indexed by the given NodePath. The node path should be relative to the current object and can use the colon character (:) to access nested properties. Examples: "position:x" or "material:next_pass:blend_mode".

pub fn get_instance_id(&self) -> i64

Returns the objects unique instance ID. This ID can be saved in EncodedObjectAsID, and can be used to retrieve the object instance with [method @GDScript.instance_from_id].

pub fn get_meta(&self, name: impl Into<GodotString>) -> Variant

Returns the objects metadata entry for the given name.

pub fn get_meta_list(&self) -> PoolArray<GodotString>

Returns the objects metadata as a PoolStringArray.

pub fn get_method_list(&self) -> VariantArray<Shared>

Returns the objects methods and their signatures as an Array.

pub fn get_property_list(&self) -> VariantArray<Shared>

Returns the objects property list as an Array of dictionaries. Each propertys Dictionary contain at least name: String and type: int (see [enum Variant.Type]) entries. Optionally, it can also include hint: int (see [PropertyHint][PropertyHint]), hint_string: String, and usage: int (see [PropertyUsageFlags][PropertyUsageFlags]).

pub fn get_script(&self) -> Option<Ref<Reference, Shared>>

Returns the objects Script instance, or null if none is assigned.

pub fn get_signal_connection_list(
&self,
signal: impl Into<GodotString>
) -> VariantArray<Shared>

Returns an Array of connections for the given signal.

pub fn get_signal_list(&self) -> VariantArray<Shared>

Returns the list of signals as an Array of dictionaries.

pub fn has_meta(&self, name: impl Into<GodotString>) -> bool

Returns true if a metadata entry is found with the given name.

pub fn has_method(&self, method: impl Into<GodotString>) -> bool

Returns true if the object contains the given method.

pub fn has_signal(&self, signal: impl Into<GodotString>) -> bool

Returns true if the given signal exists.

pub fn has_user_signal(&self, signal: impl Into<GodotString>) -> bool

Returns true if the given user-defined signal exists. Only signals added using add_user_signal are taken into account.

pub fn is_blocking_signals(&self) -> bool

Returns true if signal emission blocking is enabled.

pub fn is_class(&self, class: impl Into<GodotString>) -> bool

Returns true if the object inherits from the given class. See also get_class. Note: is_class does not take class_name declarations into account. If the object has a class_name defined, is_class will return false for that name.

pub fn is_connected(
&self,
signal: impl Into<GodotString>,
target: impl AsArg<Object>,
method: impl Into<GodotString>
) -> bool

Returns true if a connection exists for a given signal, target, and method.

pub fn is_queued_for_deletion(&self) -> bool

Returns true if the Node.queue_free method was called for the object.

pub fn notification(&self, what: i64, reversed: bool)

Send a given notification to the object, which will also trigger a call to the [_notification][Self::_notification] method of all classes that the object inherits from. If reversed is true, [_notification][Self::_notification] is called first on the objects own class, and then up to its successive parent classes. If reversed is false, [_notification][Self::_notification] is called first on the highest ancestor (Object itself), and then down to its successive inheriting classes.

Default Arguments

reversed - false

pub fn property_list_changed_notify(&self)

Notify the editor that the property list has changed, so that editor plugins can take the new values into account. Does nothing on export builds.

pub fn remove_meta(&self, name: impl Into<GodotString>)

Removes a given entry from the objects metadata. See also set_meta.

pub fn set(&self, property: impl Into<GodotString>, value: impl OwnedToVariant)

Assigns a new value to the given property. If the property does not exist, nothing will happen. Note: In C#, the property name must be specified as snake_case if it is defined by a built-in Godot node. This doesnt apply to user-defined properties where you should use the same convention as in the C# source (typically PascalCase).

pub fn set_block_signals(&self, enable: bool)

If set to true, signal emission is blocked.

pub fn set_deferred(
&self,
property: impl Into<GodotString>,
value: impl OwnedToVariant
)

Assigns a new value to the given property, after the current frames physics step. This is equivalent to calling set via call_deferred, i.e. call_deferred("set", property, value). Note: In C#, the property name must be specified as snake_case if it is defined by a built-in Godot node. This doesnt apply to user-defined properties where you should use the same convention as in the C# source (typically PascalCase).

pub fn set_indexed(
&self,
property: impl Into<NodePath>,
value: impl OwnedToVariant
)

Sample code is GDScript unless otherwise noted.

Assigns a new value to the property identified by the NodePath. The node path should be relative to the current object and can use the colon character (:) to access nested properties. Example:

set_indexed("position", Vector2(42, 0))
set_indexed("position:y", -10)
print(position) # (42, -10)

pub fn set_message_translation(&self, enable: bool)

Defines whether the object can translate strings (with calls to tr). Enabled by default.

pub fn set_meta(&self, name: impl Into<GodotString>, value: impl OwnedToVariant)

Adds, changes or removes a given entry in the objects metadata. Metadata are serialized and can take any Variant value. To remove a given entry from the objects metadata, use remove_meta. Metadata is also removed if its value is set to null. This means you can also use set_meta("name", null) to remove metadata for "name".

pub fn set_script(&self, script: impl AsArg<Reference>)

Assigns a script to the object. Each object can have a single script assigned to it, which are used to extend its functionality. If the object already had a script, the previous script instance will be freed and its variables and state will be lost. The new scripts [_init][Self::_init] method will be called.

pub fn to_string(&self) -> GodotString

Returns a String representing the object. If not overridden, defaults to "[ClassName:RID]". Override the method [_to_string][Self::_to_string] to customize the String representation.

pub fn tr(&self, message: impl Into<GodotString>) -> GodotString

Translates a message using translation catalogs configured in the Project Settings. Only works if message translation is enabled (which it is by default), otherwise it returns the message unchanged. See set_message_translation.