Struct gdnative::api::OccluderShapePolygon

pub struct OccluderShapePolygon { /* private fields */ }

core class OccluderShapePolygon inherits OccluderShape (reference-counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

OccluderShapePolygon inherits methods from:

OccluderShape

Resource

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl OccluderShapePolygon

pub fn new() -> Ref<OccluderShapePolygon, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn hole_points(&self) -> PoolArray<Vector2>

Allows changing the hole geometry from code.

pub fn polygon_points(&self) -> PoolArray<Vector2>

Allows changing the polygon geometry from code.

pub fn is_two_way(&self) -> bool

Specifies whether the occluder should operate from both sides. If false, the occluder will operate one way only.

pub fn set_hole_point(&self, index: i64, position: Vector2)

Sets an individual hole point position.

pub fn set_hole_points(&self, points: PoolArray<Vector2>)

Allows changing the hole geometry from code.

pub fn set_polygon_point(&self, index: i64, position: Vector2)

Sets an individual polygon point position.

pub fn set_polygon_points(&self, points: PoolArray<Vector2>)

Allows changing the polygon geometry from code.

pub fn set_two_way(&self, two_way: bool)

Specifies whether the occluder should operate from both sides. If false, the occluder will operate one way only.