Struct gdnative::api::PacketPeerUDP

pub struct PacketPeerUDP { /* fields omitted */ }

core class PacketPeerUDP inherits PacketPeer (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

PacketPeerUDP inherits methods from:

PacketPeer

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl PacketPeerUDP

pub fn new() -> Ref<PacketPeerUDP, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn close(&self)

Closes the UDP socket the PacketPeerUDP is currently listening on.

pub fn connect_to_host(
&self,
host: impl Into<GodotString>,
port: i64
) -> Result<(), GodotError>

Calling this method connects this UDP peer to the given host/port pair. UDP is in reality connectionless, so this option only means that incoming packets from different addresses are automatically discarded, and that outgoing packets are always sent to the connected address (future calls to set_dest_address are not allowed). This method does not send any data to the remote peer, to do that, use PacketPeer.put_var or PacketPeer.put_packet as usual. See also UDPServer. Note: Connecting to the remote peer does not help to protect from malicious attacks like IP spoofing, etc. Think about using an encryption technique like SSL or DTLS if you feel like your application is transferring sensitive information.

pub fn get_packet_ip(&self) -> GodotString

Returns the IP of the remote peer that sent the last packet(that was received with PacketPeer.get_packet or PacketPeer.get_var).

pub fn get_packet_port(&self) -> i64

Returns the port of the remote peer that sent the last packet(that was received with PacketPeer.get_packet or PacketPeer.get_var).

pub fn is_connected_to_host(&self) -> bool

Returns true if the UDP socket is open and has been connected to a remote address. See connect_to_host.

pub fn is_listening(&self) -> bool

Returns whether this PacketPeerUDP is listening.

pub fn join_multicast_group(
&self,
multicast_address: impl Into<GodotString>,
interface_name: impl Into<GodotString>
) -> Result<(), GodotError>

Joins the multicast group specified by multicast_address using the interface identified by interface_name. You can join the same multicast group with multiple interfaces. Use IP.get_local_interfaces to know which are available. Note: Some Android devices might require the CHANGE_WIFI_MULTICAST_STATE permission for multicast to work.

pub fn leave_multicast_group(
&self,
multicast_address: impl Into<GodotString>,
interface_name: impl Into<GodotString>
) -> Result<(), GodotError>

Removes the interface identified by interface_name from the multicast group specified by multicast_address.

pub fn listen(
&self,
port: i64,
bind_address: impl Into<GodotString>,
recv_buf_size: i64
) -> Result<(), GodotError>

Makes this PacketPeerUDP listen on the port binding to bind_address with a buffer size recv_buf_size. If bind_address is set to "*" (default), the peer will listen on all available addresses (both IPv4 and IPv6). If bind_address is set to "0.0.0.0" (for IPv4) or "::" (for IPv6), the peer will listen on all available addresses matching that IP type. If bind_address is set to any valid address (e.g. "192.168.1.101", "::1", etc), the peer will only listen on the interface with that addresses (or fail if no interface with the given address exists).

Default Arguments

bind_address - "*"

recv_buf_size - 65536

pub fn set_broadcast_enabled(&self, enabled: bool)

Enable or disable sending of broadcast packets (e.g. set_dest_address("255.255.255.255", 4343). This option is disabled by default. Note: Some Android devices might require the CHANGE_WIFI_MULTICAST_STATE permission and this option to be enabled to receive broadcast packets too.

pub fn set_dest_address(
&self,
host: impl Into<GodotString>,
port: i64
) -> Result<(), GodotError>

Sets the destination address and port for sending packets and variables. A hostname will be resolved using DNS if needed. Note: set_broadcast_enabled must be enabled before sending packets to a broadcast address (e.g. 255.255.255.255).

pub fn wait(&self) -> Result<(), GodotError>

Sample code is GDScript unless otherwise noted.

Waits for a packet to arrive on the listening port. See listen. Note: wait can’t be interrupted once it has been called. This can be worked around by allowing the other party to send a specific “death pill” packet like this:

socket.set_dest_address("127.0.0.1", 789)
socket.put_packet("Time to stop".to_ascii())

while socket.wait() == OK:
    var data = socket.get_packet().get_string_from_ascii()
    if data == "Time to stop":
        return