Struct gdnative::api::Particles

pub struct Particles { /* fields omitted */ }

core class Particles inherits GeometryInstance (manually managed).

This class has related types in the particles module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

Particles is a reference-only type. Persistent references can only exist in the unsafe Ref<Particles> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

Particles inherits methods from:

GeometryInstance

VisualInstance

Spatial

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Particles

Constants

pub const DRAW_ORDER_INDEX: i64

pub const DRAW_ORDER_LIFETIME: i64

pub const DRAW_ORDER_VIEW_DEPTH: i64

pub const MAX_DRAW_PASSES: i64

impl Particles

pub fn new() -> Ref<Particles, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn capture_aabb(&self) -> Aabb

Returns the axis-aligned bounding box that contains all the particles that are active in the current frame.

pub fn amount(&self) -> i64

The number of particles emitted in one emission cycle (corresponding to the lifetime). Note: Changing amount will reset the particle emission, therefore removing all particles that were already emitted before changing amount.

pub fn draw_order(&self) -> DrawOrder

Particle draw order. Uses DrawOrder values.

pub fn draw_pass_mesh(&self, pass: i64) -> Option<Ref<Mesh, Shared>>

Mesh that is drawn for the fourth draw pass.

pub fn draw_passes(&self) -> i64

The number of draw passes when rendering particles.

pub fn explosiveness_ratio(&self) -> f64

Time ratio between each emission. If 0, particles are emitted continuously. If 1, all particles are emitted simultaneously.

pub fn fixed_fps(&self) -> i64

The particle system’s frame rate is fixed to a value. For instance, changing the value to 2 will make the particles render at 2 frames per second. Note this does not slow down the simulation of the particle system itself.

pub fn fractional_delta(&self) -> bool

If true, results in fractional delta calculation which has a smoother particles display effect.

pub fn lifetime(&self) -> f64

The amount of time each particle will exist (in seconds).

pub fn one_shot(&self) -> bool

If true, only amount particles will be emitted.

pub fn pre_process_time(&self) -> f64

Amount of time to preprocess the particles before animation starts. Lets you start the animation some time after particles have started emitting.

pub fn process_material(&self) -> Option<Ref<Material, Shared>>

Material for processing particles. Can be a ParticlesMaterial or a ShaderMaterial.

pub fn randomness_ratio(&self) -> f64

Emission randomness ratio.

pub fn speed_scale(&self) -> f64

Speed scaling ratio. A value of 0 can be used to pause the particles.

pub fn use_local_coordinates(&self) -> bool

If true, particles use the parent node’s coordinate space. If false, they use global coordinates.

pub fn visibility_aabb(&self) -> Aabb

The AABB that determines the node’s region which needs to be visible on screen for the particle system to be active. Grow the box if particles suddenly appear/disappear when the node enters/exits the screen. The AABB can be grown via code or with the Particles → Generate AABB editor tool. Note: If the ParticlesMaterial in use is configured to cast shadows, you may want to enlarge this AABB to ensure the shadow is updated when particles are off-screen.

pub fn is_emitting(&self) -> bool

If true, particles are being emitted.

pub fn restart(&self)

Restarts the particle emission, clearing existing particles.

pub fn set_amount(&self, amount: i64)

The number of particles emitted in one emission cycle (corresponding to the lifetime). Note: Changing amount will reset the particle emission, therefore removing all particles that were already emitted before changing amount.

pub fn set_draw_order(&self, order: i64)

Particle draw order. Uses DrawOrder values.

pub fn set_draw_pass_mesh(&self, pass: i64, mesh: impl AsArg<Mesh>)

Mesh that is drawn for the fourth draw pass.

pub fn set_draw_passes(&self, passes: i64)

The number of draw passes when rendering particles.

pub fn set_emitting(&self, emitting: bool)

If true, particles are being emitted.

pub fn set_explosiveness_ratio(&self, ratio: f64)

Time ratio between each emission. If 0, particles are emitted continuously. If 1, all particles are emitted simultaneously.

pub fn set_fixed_fps(&self, fps: i64)

The particle system’s frame rate is fixed to a value. For instance, changing the value to 2 will make the particles render at 2 frames per second. Note this does not slow down the simulation of the particle system itself.

pub fn set_fractional_delta(&self, enable: bool)

If true, results in fractional delta calculation which has a smoother particles display effect.

pub fn set_lifetime(&self, secs: f64)

The amount of time each particle will exist (in seconds).

pub fn set_one_shot(&self, enable: bool)

If true, only amount particles will be emitted.

pub fn set_pre_process_time(&self, secs: f64)

Amount of time to preprocess the particles before animation starts. Lets you start the animation some time after particles have started emitting.

pub fn set_process_material(&self, material: impl AsArg<Material>)

Material for processing particles. Can be a ParticlesMaterial or a ShaderMaterial.

pub fn set_randomness_ratio(&self, ratio: f64)

Emission randomness ratio.

pub fn set_speed_scale(&self, scale: f64)

Speed scaling ratio. A value of 0 can be used to pause the particles.

pub fn set_use_local_coordinates(&self, enable: bool)

If true, particles use the parent node’s coordinate space. If false, they use global coordinates.

pub fn set_visibility_aabb(&self, aabb: Aabb)

The AABB that determines the node’s region which needs to be visible on screen for the particle system to be active. Grow the box if particles suddenly appear/disappear when the node enters/exits the screen. The AABB can be grown via code or with the Particles → Generate AABB editor tool. Note: If the ParticlesMaterial in use is configured to cast shadows, you may want to enlarge this AABB to ensure the shadow is updated when particles are off-screen.