Struct gdnative::api::SceneState

pub struct SceneState { /* fields omitted */ }

core class SceneState inherits Reference (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

SceneState inherits methods from:

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl SceneState

Constants

pub const GEN_EDIT_STATE_DISABLED: i64

pub const GEN_EDIT_STATE_INSTANCE: i64

pub const GEN_EDIT_STATE_MAIN: i64

impl SceneState

pub fn get_connection_binds(&self, idx: i64) -> VariantArray<Shared>

Returns the list of bound parameters for the signal at idx.

pub fn get_connection_count(&self) -> i64

Returns the number of signal connections in the scene. The idx argument used to query connection metadata in other get_connection_* methods in the interval [0, get_connection_count() - 1].

pub fn get_connection_flags(&self, idx: i64) -> i64

Returns the connection flags for the signal at idx. See [enum Object.ConnectFlags] constants.

pub fn get_connection_method(&self, idx: i64) -> GodotString

Returns the method connected to the signal at idx.

pub fn get_connection_signal(&self, idx: i64) -> GodotString

Returns the name of the signal at idx.

pub fn get_connection_source(&self, idx: i64) -> NodePath

Returns the path to the node that owns the signal at idx, relative to the root node.

pub fn get_connection_target(&self, idx: i64) -> NodePath

Returns the path to the node that owns the method connected to the signal at idx, relative to the root node.

pub fn get_node_count(&self) -> i64

Returns the number of nodes in the scene. The idx argument used to query node data in other get_node_* methods in the interval [0, get_node_count() - 1].

pub fn get_node_groups(&self, idx: i64) -> TypedArray<GodotString>

Returns the list of group names associated with the node at idx.

pub fn get_node_index(&self, idx: i64) -> i64

Returns the node’s index, which is its position relative to its siblings. This is only relevant and saved in scenes for cases where new nodes are added to an instanced or inherited scene among siblings from the base scene. Despite the name, this index is not related to the idx argument used here and in other methods.

pub fn get_node_instance(&self, idx: i64) -> Option<Ref<PackedScene, Shared>>

Returns a PackedScene for the node at idx (i.e. the whole branch starting at this node, with its child nodes and resources), or null if the node is not an instance.

pub fn get_node_instance_placeholder(&self, idx: i64) -> GodotString

Returns the path to the represented scene file if the node at idx is an InstancePlaceholder.

pub fn get_node_name(&self, idx: i64) -> GodotString

Returns the name of the node at idx.

pub fn get_node_owner_path(&self, idx: i64) -> NodePath

Returns the path to the owner of the node at idx, relative to the root node.

pub fn get_node_path(&self, idx: i64, for_parent: bool) -> NodePath

Returns the path to the node at idx. If for_parent is true, returns the path of the idx node’s parent instead.

Default Arguments

for_parent - false

pub fn get_node_property_count(&self, idx: i64) -> i64

Returns the number of exported or overridden properties for the node at idx. The prop_idx argument used to query node property data in other get_node_property_* methods in the interval [0, get_node_property_count() - 1].

pub fn get_node_property_name(&self, idx: i64, prop_idx: i64) -> GodotString

Returns the name of the property at prop_idx for the node at idx.

pub fn get_node_property_value(&self, idx: i64, prop_idx: i64) -> Variant

Returns the value of the property at prop_idx for the node at idx.

pub fn get_node_type(&self, idx: i64) -> GodotString

Returns the type of the node at idx.

pub fn is_node_instance_placeholder(&self, idx: i64) -> bool

Returns true if the node at idx is an InstancePlaceholder.