Struct gdnative::api::SoftBody

pub struct SoftBody { /* fields omitted */ }

core class SoftBody inherits MeshInstance (unsafe).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non reference counted objects such as the ones of this type are usually owned by the engine.

SoftBody is a reference-only type. Persistent references can only exist in the unsafe Ref<SoftBody> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

SoftBody inherits methods from:

MeshInstance

GeometryInstance

VisualInstance

Spatial

Node

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl SoftBody

pub fn new() -> Ref<SoftBody, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn add_collision_exception_with(&self, body: impl AsArg<Node>)

Adds a body to the list of bodies that this body can’t collide with.

pub fn areaAngular_stiffness(&self) -> f64

pub fn get_collision_exceptions(&self) -> VariantArray<Shared>

Returns an array of nodes that were added as collision exceptions for this body.

pub fn collision_layer(&self) -> i64

The physics layers this SoftBody is in. Collidable objects can exist in any of 32 different layers. These layers work like a tagging system, and are not visual. A collidable can use these layers to select with which objects it can collide, using the collision_mask property. A contact is detected if object A is in any of the layers that object B scans, or object B is in any layer scanned by object A. See Collision layers and masks in the documentation for more information.

pub fn get_collision_layer_bit(&self, bit: i64) -> bool

Returns an individual bit on the collision mask.

pub fn collision_mask(&self) -> i64

The physics layers this SoftBody scans for collisions. See Collision layers and masks in the documentation for more information.

pub fn get_collision_mask_bit(&self, bit: i64) -> bool

Returns an individual bit on the collision mask.

pub fn damping_coefficient(&self) -> f64

pub fn drag_coefficient(&self) -> f64

pub fn linear_stiffness(&self) -> f64

pub fn parent_collision_ignore(&self) -> NodePath

NodePath to a CollisionObject this SoftBody should avoid clipping.

pub fn pose_matching_coefficient(&self) -> f64

pub fn pressure_coefficient(&self) -> f64

pub fn simulation_precision(&self) -> i64

Increasing this value will improve the resulting simulation, but can affect performance. Use with care.

pub fn total_mass(&self) -> f64

The SoftBody’s mass.

pub fn volume_stiffness(&self) -> f64

pub fn is_ray_pickable(&self) -> bool

If true, the SoftBody will respond to RayCasts.

pub fn remove_collision_exception_with(&self, body: impl AsArg<Node>)

Removes a body from the list of bodies that this body can’t collide with.

pub fn set_areaAngular_stiffness(&self, areaAngular_stiffness: f64)

pub fn set_collision_layer(&self, collision_layer: i64)

The physics layers this SoftBody is in. Collidable objects can exist in any of 32 different layers. These layers work like a tagging system, and are not visual. A collidable can use these layers to select with which objects it can collide, using the collision_mask property. A contact is detected if object A is in any of the layers that object B scans, or object B is in any layer scanned by object A. See Collision layers and masks in the documentation for more information.

pub fn set_collision_layer_bit(&self, bit: i64, value: bool)

Sets individual bits on the layer mask. Use this if you only need to change one layer’s value.

pub fn set_collision_mask(&self, collision_mask: i64)

The physics layers this SoftBody scans for collisions. See Collision layers and masks in the documentation for more information.

pub fn set_collision_mask_bit(&self, bit: i64, value: bool)

Sets individual bits on the collision mask. Use this if you only need to change one layer’s value.

pub fn set_damping_coefficient(&self, damping_coefficient: f64)

pub fn set_drag_coefficient(&self, drag_coefficient: f64)

pub fn set_linear_stiffness(&self, linear_stiffness: f64)

pub fn set_parent_collision_ignore(
&self,
parent_collision_ignore: impl Into<NodePath>
)

NodePath to a CollisionObject this SoftBody should avoid clipping.

pub fn set_pose_matching_coefficient(&self, pose_matching_coefficient: f64)

pub fn set_pressure_coefficient(&self, pressure_coefficient: f64)

pub fn set_ray_pickable(&self, ray_pickable: bool)

If true, the SoftBody will respond to RayCasts.

pub fn set_simulation_precision(&self, simulation_precision: i64)

Increasing this value will improve the resulting simulation, but can affect performance. Use with care.

pub fn set_total_mass(&self, mass: f64)

The SoftBody’s mass.

pub fn set_volume_stiffness(&self, volume_stiffness: f64)