Struct gdnative::api::SpatialMaterial

pub struct SpatialMaterial { /* fields omitted */ }

core class SpatialMaterial inherits Material (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

SpatialMaterial inherits methods from:

Material

Resource

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl SpatialMaterial

Constants

pub const BILLBOARD_DISABLED: i64

pub const BLEND_MODE_MIX: i64

pub const CULL_BACK: i64

pub const DEPTH_DRAW_OPAQUE_ONLY: i64

pub const DETAIL_UV_1: i64

pub const DIFFUSE_BURLEY: i64

pub const DISTANCE_FADE_DISABLED: i64

pub const EMISSION_OP_ADD: i64

pub const FEATURE_TRANSPARENT: i64

pub const FLAG_UNSHADED: i64

pub const SPECULAR_SCHLICK_GGX: i64

pub const TEXTURE_ALBEDO: i64

pub const TEXTURE_CHANNEL_RED: i64

pub const BILLBOARD_ENABLED: i64

pub const BLEND_MODE_ADD: i64

pub const CULL_FRONT: i64

pub const DEPTH_DRAW_ALWAYS: i64

pub const DETAIL_UV_2: i64

pub const DIFFUSE_LAMBERT: i64

pub const DISTANCE_FADE_PIXEL_ALPHA: i64

pub const EMISSION_OP_MULTIPLY: i64

pub const FEATURE_EMISSION: i64

pub const FLAG_USE_VERTEX_LIGHTING: i64

pub const SPECULAR_BLINN: i64

pub const TEXTURE_CHANNEL_GREEN: i64

pub const TEXTURE_METALLIC: i64

pub const BILLBOARD_FIXED_Y: i64

pub const BLEND_MODE_SUB: i64

pub const CULL_DISABLED: i64

pub const DEPTH_DRAW_DISABLED: i64

pub const DIFFUSE_LAMBERT_WRAP: i64

pub const DISTANCE_FADE_PIXEL_DITHER: i64

pub const FEATURE_NORMAL_MAPPING: i64

pub const FLAG_DISABLE_DEPTH_TEST: i64

pub const SPECULAR_PHONG: i64

pub const TEXTURE_CHANNEL_BLUE: i64

pub const TEXTURE_ROUGHNESS: i64

pub const BILLBOARD_PARTICLES: i64

pub const BLEND_MODE_MUL: i64

pub const DEPTH_DRAW_ALPHA_OPAQUE_PREPASS: i64

pub const DIFFUSE_OREN_NAYAR: i64

pub const DISTANCE_FADE_OBJECT_DITHER: i64

pub const FEATURE_RIM: i64

pub const FLAG_ALBEDO_FROM_VERTEX_COLOR: i64

pub const SPECULAR_TOON: i64

pub const TEXTURE_CHANNEL_ALPHA: i64

pub const TEXTURE_EMISSION: i64

pub const DIFFUSE_TOON: i64

pub const FEATURE_CLEARCOAT: i64

pub const FLAG_SRGB_VERTEX_COLOR: i64

pub const SPECULAR_DISABLED: i64

pub const TEXTURE_CHANNEL_GRAYSCALE: i64

pub const TEXTURE_NORMAL: i64

pub const FEATURE_ANISOTROPY: i64

pub const FLAG_USE_POINT_SIZE: i64

pub const TEXTURE_RIM: i64

pub const FEATURE_AMBIENT_OCCLUSION: i64

pub const FLAG_FIXED_SIZE: i64

pub const TEXTURE_CLEARCOAT: i64

pub const FEATURE_DEPTH_MAPPING: i64

pub const FLAG_BILLBOARD_KEEP_SCALE: i64

pub const TEXTURE_FLOWMAP: i64

pub const FEATURE_SUBSURACE_SCATTERING: i64

pub const FLAG_UV1_USE_TRIPLANAR: i64

pub const TEXTURE_AMBIENT_OCCLUSION: i64

pub const FEATURE_TRANSMISSION: i64

pub const FLAG_UV2_USE_TRIPLANAR: i64

pub const TEXTURE_DEPTH: i64

pub const FEATURE_REFRACTION: i64

pub const FLAG_TRIPLANAR_USE_WORLD: i64

pub const TEXTURE_SUBSURFACE_SCATTERING: i64

pub const FEATURE_DETAIL: i64

pub const FLAG_AO_ON_UV2: i64

pub const TEXTURE_TRANSMISSION: i64

pub const FEATURE_MAX: i64

pub const FLAG_EMISSION_ON_UV2: i64

pub const TEXTURE_REFRACTION: i64

pub const FLAG_USE_ALPHA_SCISSOR: i64

pub const TEXTURE_DETAIL_MASK: i64

pub const FLAG_ALBEDO_TEXTURE_FORCE_SRGB: i64

pub const TEXTURE_DETAIL_ALBEDO: i64

pub const FLAG_DONT_RECEIVE_SHADOWS: i64

pub const TEXTURE_DETAIL_NORMAL: i64

pub const FLAG_ENSURE_CORRECT_NORMALS: i64

pub const TEXTURE_MAX: i64

pub const FLAG_DISABLE_AMBIENT_LIGHT: i64

pub const FLAG_USE_SHADOW_TO_OPACITY: i64

pub const FLAG_MAX: i64

impl SpatialMaterial

pub fn new() -> Ref<SpatialMaterial, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn albedo(&self) -> Color

The material’s base color.

pub fn alpha_scissor_threshold(&self) -> f64

Threshold at which the alpha scissor will discard values.

pub fn anisotropy(&self) -> f64

The strength of the anisotropy effect.

pub fn ao_light_affect(&self) -> f64

Amount that ambient occlusion affects lighting from lights. If 0, ambient occlusion only affects ambient light. If 1, ambient occlusion affects lights just as much as it affects ambient light. This can be used to impact the strength of the ambient occlusion effect, but typically looks unrealistic.

pub fn ao_texture_channel(&self) -> TextureChannel

Specifies the channel of the [ao_texture][Self::ao_texture] in which the ambient occlusion information is stored. This is useful when you store the information for multiple effects in a single texture. For example if you stored metallic in the red channel, roughness in the blue, and ambient occlusion in the green you could reduce the number of textures you use.

pub fn billboard_mode(&self) -> BillboardMode

Controls how the object faces the camera. See BillboardMode. Note: Billboard mode is not suitable for VR because the left-right vector of the camera is not horizontal when the screen is attached to your head instead of on the table. See GitHub issue #41567 for details.

pub fn blend_mode(&self) -> BlendMode

The material’s blend mode. Note: Values other than Mix force the object into the transparent pipeline. See BlendMode.

pub fn clearcoat(&self) -> f64

Sets the strength of the clearcoat effect. Setting to 0 looks the same as disabling the clearcoat effect.

pub fn clearcoat_gloss(&self) -> f64

Sets the roughness of the clearcoat pass. A higher value results in a smoother clearcoat while a lower value results in a rougher clearcoat.

pub fn cull_mode(&self) -> CullMode

Which side of the object is not drawn when backfaces are rendered. See CullMode.

pub fn depth_deep_parallax_flip_binormal(&self) -> bool

If true, direction of the binormal is flipped before using in the depth effect. This may be necessary if you have encoded your binormals in a way that is conflicting with the depth effect.

pub fn depth_deep_parallax_flip_tangent(&self) -> bool

If true, direction of the tangent is flipped before using in the depth effect. This may be necessary if you have encoded your tangents in a way that is conflicting with the depth effect.

pub fn depth_deep_parallax_max_layers(&self) -> i64

Number of layers to use when using [depth_deep_parallax][Self::depth_deep_parallax] and the view direction is perpendicular to the surface of the object. A higher number will be more performance demanding while a lower number may not look as crisp.

pub fn depth_deep_parallax_min_layers(&self) -> i64

Number of layers to use when using [depth_deep_parallax][Self::depth_deep_parallax] and the view direction is parallel to the surface of the object. A higher number will be more performance demanding while a lower number may not look as crisp.

pub fn depth_draw_mode(&self) -> DepthDrawMode

Determines when depth rendering takes place. See DepthDrawMode. See also [flags_transparent][Self::flags_transparent].

pub fn depth_scale(&self) -> f64

Scales the depth offset effect. A higher number will create a larger depth.

pub fn detail_blend_mode(&self) -> BlendMode

Specifies how the [detail_albedo][Self::detail_albedo] should blend with the current ALBEDO. See BlendMode for options.

pub fn detail_uv(&self) -> DetailUv

Specifies whether to use UV or UV2 for the detail layer. See [DetailUV][DetailUV] for options.

pub fn diffuse_mode(&self) -> DiffuseMode

The algorithm used for diffuse light scattering. See DiffuseMode.

pub fn distance_fade(&self) -> DistanceFadeMode

Specifies which type of fade to use. Can be any of the DistanceFadeModes.

pub fn distance_fade_max_distance(&self) -> f64

Distance at which the object appears fully opaque. Note: If distance_fade_max_distance is less than distance_fade_min_distance, the behavior will be reversed. The object will start to fade away at distance_fade_max_distance and will fully disappear once it reaches distance_fade_min_distance.

pub fn distance_fade_min_distance(&self) -> f64

Distance at which the object starts to become visible. If the object is less than this distance away, it will be invisible. Note: If distance_fade_min_distance is greater than distance_fade_max_distance, the behavior will be reversed. The object will start to fade away at distance_fade_max_distance and will fully disappear once it reaches distance_fade_min_distance.

pub fn emission(&self) -> Color

The emitted light’s color. See [emission_enabled][Self::emission_enabled].

pub fn emission_energy(&self) -> f64

The emitted light’s strength. See [emission_enabled][Self::emission_enabled].

pub fn emission_operator(&self) -> EmissionOperator

Sets how emission interacts with [emission_texture][Self::emission_texture]. Can either add or multiply. See EmissionOperator for options.

pub fn feature(&self, feature: i64) -> bool

If true, the transmission effect is enabled.

pub fn flag(&self, flag: i64) -> bool

If true, the vertex color is used as albedo color.

pub fn grow(&self) -> f64

Grows object vertices in the direction of their normals.

pub fn line_width(&self) -> f64

Currently unimplemented in Godot.

pub fn metallic(&self) -> f64

A high value makes the material appear more like a metal. Non-metals use their albedo as the diffuse color and add diffuse to the specular reflection. With non-metals, the reflection appears on top of the albedo color. Metals use their albedo as a multiplier to the specular reflection and set the diffuse color to black resulting in a tinted reflection. Materials work better when fully metal or fully non-metal, values between 0 and 1 should only be used for blending between metal and non-metal sections. To alter the amount of reflection use roughness.

pub fn metallic_texture_channel(&self) -> TextureChannel

Specifies the channel of the [metallic_texture][Self::metallic_texture] in which the metallic information is stored. This is useful when you store the information for multiple effects in a single texture. For example if you stored metallic in the red channel, roughness in the blue, and ambient occlusion in the green you could reduce the number of textures you use.

pub fn normal_scale(&self) -> f64

The strength of the normal map’s effect.

pub fn particles_anim_h_frames(&self) -> i64

The number of horizontal frames in the particle sprite sheet. Only enabled when using BILLBOARD_PARTICLES. See [params_billboard_mode][Self::params_billboard_mode].

pub fn particles_anim_loop(&self) -> bool

If true, particle animations are looped. Only enabled when using BILLBOARD_PARTICLES. See [params_billboard_mode][Self::params_billboard_mode].

pub fn particles_anim_v_frames(&self) -> i64

The number of vertical frames in the particle sprite sheet. Only enabled when using BILLBOARD_PARTICLES. See [params_billboard_mode][Self::params_billboard_mode].

pub fn point_size(&self) -> f64

The point size in pixels. See [flags_use_point_size][Self::flags_use_point_size].

pub fn proximity_fade_distance(&self) -> f64

Distance over which the fade effect takes place. The larger the distance the longer it takes for an object to fade.

pub fn refraction(&self) -> f64

The strength of the refraction effect.

pub fn refraction_texture_channel(&self) -> TextureChannel

Specifies the channel of the [ao_texture][Self::ao_texture] in which the ambient occlusion information is stored. This is useful when you store the information for multiple effects in a single texture. For example if you stored metallic in the red channel, roughness in the blue, and ambient occlusion in the green you could reduce the number of textures you use.

pub fn rim(&self) -> f64

Sets the strength of the rim lighting effect.

pub fn rim_tint(&self) -> f64

The amount of to blend light and albedo color when rendering rim effect. If 0 the light color is used, while 1 means albedo color is used. An intermediate value generally works best.

pub fn roughness(&self) -> f64

Surface reflection. A value of 0 represents a perfect mirror while a value of 1 completely blurs the reflection. See also metallic.

pub fn roughness_texture_channel(&self) -> TextureChannel

Specifies the channel of the [ao_texture][Self::ao_texture] in which the ambient occlusion information is stored. This is useful when you store the information for multiple effects in a single texture. For example if you stored metallic in the red channel, roughness in the blue, and ambient occlusion in the green you could reduce the number of textures you use.

pub fn specular(&self) -> f64

Sets the size of the specular lobe. The specular lobe is the bright spot that is reflected from light sources. Note: unlike metallic, this is not energy-conserving, so it should be left at 0.5 in most cases. See also roughness.

pub fn specular_mode(&self) -> SpecularMode

The method for rendering the specular blob. See SpecularMode.

pub fn subsurface_scattering_strength(&self) -> f64

The strength of the subsurface scattering effect.

pub fn texture(&self, param: i64) -> Option<Ref<Texture, Shared>>

Texture used to control the transmission effect per-pixel. Added to transmission.

pub fn transmission(&self) -> Color

The color used by the transmission effect. Represents the light passing through an object.

pub fn uv1_offset(&self) -> Vector3

How much to offset the UV coordinates. This amount will be added to UV in the vertex function. This can be used to offset a texture.

pub fn uv1_scale(&self) -> Vector3

How much to scale the UV coordinates. This is multiplied by UV in the vertex function.

pub fn uv1_triplanar_blend_sharpness(&self) -> f64

A lower number blends the texture more softly while a higher number blends the texture more sharply.

pub fn uv2_offset(&self) -> Vector3

How much to offset the UV2 coordinates. This amount will be added to UV2 in the vertex function. This can be used to offset a texture.

pub fn uv2_scale(&self) -> Vector3

How much to scale the UV2 coordinates. This is multiplied by UV2 in the vertex function.

pub fn uv2_triplanar_blend_sharpness(&self) -> f64

A lower number blends the texture more softly while a higher number blends the texture more sharply.

pub fn is_depth_deep_parallax_enabled(&self) -> bool

If true, the shader will read depth texture at multiple points along the view ray to determine occlusion and parrallax. This can be very performance demanding, but results in more realistic looking depth mapping.

pub fn is_grow_enabled(&self) -> bool

If true, enables the vertex grow setting. See [params_grow_amount][Self::params_grow_amount].

pub fn is_proximity_fade_enabled(&self) -> bool

If true, the proximity fade effect is enabled. The proximity fade effect fades out each pixel based on its distance to another object.

pub fn set_albedo(&self, albedo: Color)

The material’s base color.

pub fn set_alpha_scissor_threshold(&self, threshold: f64)

Threshold at which the alpha scissor will discard values.

pub fn set_anisotropy(&self, anisotropy: f64)

The strength of the anisotropy effect.

pub fn set_ao_light_affect(&self, amount: f64)

Amount that ambient occlusion affects lighting from lights. If 0, ambient occlusion only affects ambient light. If 1, ambient occlusion affects lights just as much as it affects ambient light. This can be used to impact the strength of the ambient occlusion effect, but typically looks unrealistic.

pub fn set_ao_texture_channel(&self, channel: i64)

Specifies the channel of the [ao_texture][Self::ao_texture] in which the ambient occlusion information is stored. This is useful when you store the information for multiple effects in a single texture. For example if you stored metallic in the red channel, roughness in the blue, and ambient occlusion in the green you could reduce the number of textures you use.

pub fn set_billboard_mode(&self, mode: i64)

Controls how the object faces the camera. See BillboardMode. Note: Billboard mode is not suitable for VR because the left-right vector of the camera is not horizontal when the screen is attached to your head instead of on the table. See GitHub issue #41567 for details.

pub fn set_blend_mode(&self, blend_mode: i64)

The material’s blend mode. Note: Values other than Mix force the object into the transparent pipeline. See BlendMode.

pub fn set_clearcoat(&self, clearcoat: f64)

Sets the strength of the clearcoat effect. Setting to 0 looks the same as disabling the clearcoat effect.

pub fn set_clearcoat_gloss(&self, clearcoat_gloss: f64)

Sets the roughness of the clearcoat pass. A higher value results in a smoother clearcoat while a lower value results in a rougher clearcoat.

pub fn set_cull_mode(&self, cull_mode: i64)

Which side of the object is not drawn when backfaces are rendered. See CullMode.

pub fn set_depth_deep_parallax(&self, enable: bool)

If true, the shader will read depth texture at multiple points along the view ray to determine occlusion and parrallax. This can be very performance demanding, but results in more realistic looking depth mapping.

pub fn set_depth_deep_parallax_flip_binormal(&self, flip: bool)

If true, direction of the binormal is flipped before using in the depth effect. This may be necessary if you have encoded your binormals in a way that is conflicting with the depth effect.

pub fn set_depth_deep_parallax_flip_tangent(&self, flip: bool)

If true, direction of the tangent is flipped before using in the depth effect. This may be necessary if you have encoded your tangents in a way that is conflicting with the depth effect.

pub fn set_depth_deep_parallax_max_layers(&self, layer: i64)

Number of layers to use when using [depth_deep_parallax][Self::depth_deep_parallax] and the view direction is perpendicular to the surface of the object. A higher number will be more performance demanding while a lower number may not look as crisp.

pub fn set_depth_deep_parallax_min_layers(&self, layer: i64)

Number of layers to use when using [depth_deep_parallax][Self::depth_deep_parallax] and the view direction is parallel to the surface of the object. A higher number will be more performance demanding while a lower number may not look as crisp.

pub fn set_depth_draw_mode(&self, depth_draw_mode: i64)

Determines when depth rendering takes place. See DepthDrawMode. See also [flags_transparent][Self::flags_transparent].

pub fn set_depth_scale(&self, depth_scale: f64)

Scales the depth offset effect. A higher number will create a larger depth.

pub fn set_detail_blend_mode(&self, detail_blend_mode: i64)

Specifies how the [detail_albedo][Self::detail_albedo] should blend with the current ALBEDO. See BlendMode for options.

pub fn set_detail_uv(&self, detail_uv: i64)

Specifies whether to use UV or UV2 for the detail layer. See [DetailUV][DetailUV] for options.

pub fn set_diffuse_mode(&self, diffuse_mode: i64)

The algorithm used for diffuse light scattering. See DiffuseMode.

pub fn set_distance_fade(&self, mode: i64)

Specifies which type of fade to use. Can be any of the DistanceFadeModes.

pub fn set_distance_fade_max_distance(&self, distance: f64)

Distance at which the object appears fully opaque. Note: If distance_fade_max_distance is less than distance_fade_min_distance, the behavior will be reversed. The object will start to fade away at distance_fade_max_distance and will fully disappear once it reaches distance_fade_min_distance.

pub fn set_distance_fade_min_distance(&self, distance: f64)

Distance at which the object starts to become visible. If the object is less than this distance away, it will be invisible. Note: If distance_fade_min_distance is greater than distance_fade_max_distance, the behavior will be reversed. The object will start to fade away at distance_fade_max_distance and will fully disappear once it reaches distance_fade_min_distance.

pub fn set_emission(&self, emission: Color)

The emitted light’s color. See [emission_enabled][Self::emission_enabled].

pub fn set_emission_energy(&self, emission_energy: f64)

The emitted light’s strength. See [emission_enabled][Self::emission_enabled].

pub fn set_emission_operator(&self, operator: i64)

Sets how emission interacts with [emission_texture][Self::emission_texture]. Can either add or multiply. See EmissionOperator for options.

pub fn set_feature(&self, feature: i64, enable: bool)

If true, the transmission effect is enabled.

pub fn set_flag(&self, flag: i64, enable: bool)

If true, the vertex color is used as albedo color.

pub fn set_grow(&self, amount: f64)

Grows object vertices in the direction of their normals.

pub fn set_grow_enabled(&self, enable: bool)

If true, enables the vertex grow setting. See [params_grow_amount][Self::params_grow_amount].

pub fn set_line_width(&self, line_width: f64)

Currently unimplemented in Godot.

pub fn set_metallic(&self, metallic: f64)

A high value makes the material appear more like a metal. Non-metals use their albedo as the diffuse color and add diffuse to the specular reflection. With non-metals, the reflection appears on top of the albedo color. Metals use their albedo as a multiplier to the specular reflection and set the diffuse color to black resulting in a tinted reflection. Materials work better when fully metal or fully non-metal, values between 0 and 1 should only be used for blending between metal and non-metal sections. To alter the amount of reflection use roughness.

pub fn set_metallic_texture_channel(&self, channel: i64)

Specifies the channel of the [metallic_texture][Self::metallic_texture] in which the metallic information is stored. This is useful when you store the information for multiple effects in a single texture. For example if you stored metallic in the red channel, roughness in the blue, and ambient occlusion in the green you could reduce the number of textures you use.

pub fn set_normal_scale(&self, normal_scale: f64)

The strength of the normal map’s effect.

pub fn set_particles_anim_h_frames(&self, frames: i64)

The number of horizontal frames in the particle sprite sheet. Only enabled when using BILLBOARD_PARTICLES. See [params_billboard_mode][Self::params_billboard_mode].

pub fn set_particles_anim_loop(&self, _loop: bool)

If true, particle animations are looped. Only enabled when using BILLBOARD_PARTICLES. See [params_billboard_mode][Self::params_billboard_mode].

pub fn set_particles_anim_v_frames(&self, frames: i64)

The number of vertical frames in the particle sprite sheet. Only enabled when using BILLBOARD_PARTICLES. See [params_billboard_mode][Self::params_billboard_mode].

pub fn set_point_size(&self, point_size: f64)

The point size in pixels. See [flags_use_point_size][Self::flags_use_point_size].

pub fn set_proximity_fade(&self, enabled: bool)

If true, the proximity fade effect is enabled. The proximity fade effect fades out each pixel based on its distance to another object.

pub fn set_proximity_fade_distance(&self, distance: f64)

Distance over which the fade effect takes place. The larger the distance the longer it takes for an object to fade.

pub fn set_refraction(&self, refraction: f64)

The strength of the refraction effect.

pub fn set_refraction_texture_channel(&self, channel: i64)

Specifies the channel of the [ao_texture][Self::ao_texture] in which the ambient occlusion information is stored. This is useful when you store the information for multiple effects in a single texture. For example if you stored metallic in the red channel, roughness in the blue, and ambient occlusion in the green you could reduce the number of textures you use.

pub fn set_rim(&self, rim: f64)

Sets the strength of the rim lighting effect.

pub fn set_rim_tint(&self, rim_tint: f64)

The amount of to blend light and albedo color when rendering rim effect. If 0 the light color is used, while 1 means albedo color is used. An intermediate value generally works best.

pub fn set_roughness(&self, roughness: f64)

Surface reflection. A value of 0 represents a perfect mirror while a value of 1 completely blurs the reflection. See also metallic.

pub fn set_roughness_texture_channel(&self, channel: i64)

Specifies the channel of the [ao_texture][Self::ao_texture] in which the ambient occlusion information is stored. This is useful when you store the information for multiple effects in a single texture. For example if you stored metallic in the red channel, roughness in the blue, and ambient occlusion in the green you could reduce the number of textures you use.

pub fn set_specular(&self, specular: f64)

Sets the size of the specular lobe. The specular lobe is the bright spot that is reflected from light sources. Note: unlike metallic, this is not energy-conserving, so it should be left at 0.5 in most cases. See also roughness.

pub fn set_specular_mode(&self, specular_mode: i64)

The method for rendering the specular blob. See SpecularMode.

pub fn set_subsurface_scattering_strength(&self, strength: f64)

The strength of the subsurface scattering effect.

pub fn set_texture(&self, param: i64, texture: impl AsArg<Texture>)

Texture used to control the transmission effect per-pixel. Added to transmission.

pub fn set_transmission(&self, transmission: Color)

The color used by the transmission effect. Represents the light passing through an object.

pub fn set_uv1_offset(&self, offset: Vector3)

How much to offset the UV coordinates. This amount will be added to UV in the vertex function. This can be used to offset a texture.

pub fn set_uv1_scale(&self, scale: Vector3)

How much to scale the UV coordinates. This is multiplied by UV in the vertex function.

pub fn set_uv1_triplanar_blend_sharpness(&self, sharpness: f64)

A lower number blends the texture more softly while a higher number blends the texture more sharply.

pub fn set_uv2_offset(&self, offset: Vector3)

How much to offset the UV2 coordinates. This amount will be added to UV2 in the vertex function. This can be used to offset a texture.

pub fn set_uv2_scale(&self, scale: Vector3)

How much to scale the UV2 coordinates. This is multiplied by UV2 in the vertex function.

pub fn set_uv2_triplanar_blend_sharpness(&self, sharpness: f64)

A lower number blends the texture more softly while a higher number blends the texture more sharply.