Struct gdnative::api::SpinBox

pub struct SpinBox { /* fields omitted */ }

core class SpinBox inherits Range (manually managed).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

SpinBox is a reference-only type. Persistent references can only exist in the unsafe Ref<SpinBox> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

SpinBox inherits methods from:

Range

Control

CanvasItem

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl SpinBox

pub fn new() -> Ref<SpinBox, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn apply(&self)

Applies the current value of this SpinBox.

pub fn align(&self) -> Align

Sets the text alignment of the SpinBox.

pub fn get_line_edit(&self) -> Option<Ref<LineEdit, Shared>>

Returns the LineEdit instance from this SpinBox. You can use it to access properties and methods of LineEdit.

pub fn prefix(&self) -> GodotString

Adds the specified prefix string before the numerical value of the SpinBox.

pub fn suffix(&self) -> GodotString

Adds the specified suffix string after the numerical value of the SpinBox.

pub fn is_editable(&self) -> bool

If true, the SpinBox will be editable. Otherwise, it will be read only.

pub fn set_align(&self, align: i64)

Sets the text alignment of the SpinBox.

pub fn set_editable(&self, editable: bool)

If true, the SpinBox will be editable. Otherwise, it will be read only.

pub fn set_prefix(&self, prefix: impl Into<GodotString>)

Adds the specified prefix string before the numerical value of the SpinBox.

pub fn set_suffix(&self, suffix: impl Into<GodotString>)

Adds the specified suffix string after the numerical value of the SpinBox.