Struct gdnative::api::StaticBody2D

pub struct StaticBody2D { /* fields omitted */ }

core class StaticBody2D inherits PhysicsBody2D (manually managed).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

StaticBody2D is a reference-only type. Persistent references can only exist in the unsafe Ref<StaticBody2D> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

StaticBody2D inherits methods from:

PhysicsBody2D

CollisionObject2D

Node2D

CanvasItem

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl StaticBody2D

pub fn new() -> Ref<StaticBody2D, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn bounce(&self) -> f64

The body’s bounciness. Values range from 0 (no bounce) to 1 (full bounciness). Deprecated, use PhysicsMaterial.bounce instead via physics_material_override.

pub fn constant_angular_velocity(&self) -> f64

The body’s constant angular velocity. This does not rotate the body, but affects colliding bodies, as if it were rotating.

pub fn constant_linear_velocity(&self) -> Vector2

The body’s constant linear velocity. This does not move the body, but affects colliding bodies, as if it were moving.

pub fn friction(&self) -> f64

The body’s friction. Values range from 0 (no friction) to 1 (full friction). Deprecated, use PhysicsMaterial.friction instead via physics_material_override.

pub fn physics_material_override(&self) -> Option<Ref<PhysicsMaterial, Shared>>

The physics material override for the body. If a material is assigned to this property, it will be used instead of any other physics material, such as an inherited one.

pub fn set_bounce(&self, bounce: f64)

The body’s bounciness. Values range from 0 (no bounce) to 1 (full bounciness). Deprecated, use PhysicsMaterial.bounce instead via physics_material_override.

pub fn set_constant_angular_velocity(&self, vel: f64)

The body’s constant angular velocity. This does not rotate the body, but affects colliding bodies, as if it were rotating.

pub fn set_constant_linear_velocity(&self, vel: Vector2)

The body’s constant linear velocity. This does not move the body, but affects colliding bodies, as if it were moving.

pub fn set_friction(&self, friction: f64)

The body’s friction. Values range from 0 (no friction) to 1 (full friction). Deprecated, use PhysicsMaterial.friction instead via physics_material_override.

pub fn set_physics_material_override(
&self,
physics_material_override: impl AsArg<PhysicsMaterial>
)

The physics material override for the body. If a material is assigned to this property, it will be used instead of any other physics material, such as an inherited one.