Struct gdnative::api::StreamPeer

pub struct StreamPeer { /* fields omitted */ }

core class StreamPeer inherits Reference (reference-counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

StreamPeer inherits methods from:

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl StreamPeer

pub fn get_16(&self) -> i64

Gets a signed 16-bit value from the stream.

pub fn get_32(&self) -> i64

Gets a signed 32-bit value from the stream.

pub fn get_64(&self) -> i64

Gets a signed 64-bit value from the stream.

pub fn get_8(&self) -> i64

Gets a signed byte from the stream.

pub fn get_available_bytes(&self) -> i64

Returns the amount of bytes this StreamPeer has available.

pub fn get_data(&self, bytes: i64) -> VariantArray<Shared>

Returns a chunk data with the received bytes. The amount of bytes to be received can be requested in the bytes argument. If not enough bytes are available, the function will block until the desired amount is received. This function returns two values, an [enum @GlobalScope.Error] code and a data array.

pub fn get_double(&self) -> f64

Gets a double-precision float from the stream.

pub fn get_float(&self) -> f64

Gets a single-precision float from the stream.

pub fn get_partial_data(&self, bytes: i64) -> VariantArray<Shared>

Returns a chunk data with the received bytes. The amount of bytes to be received can be requested in the “bytes” argument. If not enough bytes are available, the function will return how many were actually received. This function returns two values, an [enum @GlobalScope.Error] code, and a data array.

pub fn get_string(&self, bytes: i64) -> GodotString

Gets a string with byte-length bytes from the stream. If bytes is negative (default) the length will be read from the stream using the reverse process of put_string.

Default Arguments

bytes - -1

pub fn get_u16(&self) -> i64

Gets an unsigned 16-bit value from the stream.

pub fn get_u32(&self) -> i64

Gets an unsigned 32-bit value from the stream.

pub fn get_u64(&self) -> i64

Gets an unsigned 64-bit value from the stream.

pub fn get_u8(&self) -> i64

Gets an unsigned byte from the stream.

pub fn get_utf8_string(&self, bytes: i64) -> GodotString

Gets an UTF-8 string with byte-length bytes from the stream (this decodes the string sent as UTF-8). If bytes is negative (default) the length will be read from the stream using the reverse process of put_utf8_string.

Default Arguments

bytes - -1

pub fn get_var(&self, allow_objects: bool) -> Variant

Gets a Variant from the stream. If allow_objects is true, decoding objects is allowed. Warning: Deserialized objects can contain code which gets executed. Do not use this option if the serialized object comes from untrusted sources to avoid potential security threats such as remote code execution.

Default Arguments

allow_objects - false

pub fn is_big_endian_enabled(&self) -> bool

If true, this StreamPeer will using big-endian format for encoding and decoding.

pub fn put_16(&self, value: i64)

Puts a signed 16-bit value into the stream.

pub fn put_32(&self, value: i64)

Puts a signed 32-bit value into the stream.

pub fn put_64(&self, value: i64)

Puts a signed 64-bit value into the stream.

pub fn put_8(&self, value: i64)

Puts a signed byte into the stream.

pub fn put_data(&self, data: PoolArray<u8>) -> Result<(), GodotError>

Sends a chunk of data through the connection, blocking if necessary until the data is done sending. This function returns an [enum @GlobalScope.Error] code.

pub fn put_double(&self, value: f64)

Puts a double-precision float into the stream.

pub fn put_float(&self, value: f64)

Puts a single-precision float into the stream.

pub fn put_partial_data(&self, data: PoolArray<u8>) -> VariantArray<Shared>

Sends a chunk of data through the connection. If all the data could not be sent at once, only part of it will. This function returns two values, an [enum @GlobalScope.Error] code and an integer, describing how much data was actually sent.

pub fn put_string(&self, value: impl Into<GodotString>)

Sample code is GDScript unless otherwise noted.

Puts a zero-terminated ASCII string into the stream prepended by a 32-bit unsigned integer representing its size. Note: To put an ASCII string without prepending its size, you can use put_data:

put_data("Hello world".to_ascii())

pub fn put_u16(&self, value: i64)

Puts an unsigned 16-bit value into the stream.

pub fn put_u32(&self, value: i64)

Puts an unsigned 32-bit value into the stream.

pub fn put_u64(&self, value: i64)

Puts an unsigned 64-bit value into the stream.

pub fn put_u8(&self, value: i64)

Puts an unsigned byte into the stream.

pub fn put_utf8_string(&self, value: impl Into<GodotString>)

Sample code is GDScript unless otherwise noted.

Puts a zero-terminated UTF-8 string into the stream prepended by a 32 bits unsigned integer representing its size. Note: To put an UTF-8 string without prepending its size, you can use put_data:

put_data("Hello world".to_utf8())

pub fn put_var(&self, value: impl OwnedToVariant, full_objects: bool)

Puts a Variant into the stream. If full_objects is true encoding objects is allowed (and can potentially include code).

Default Arguments

full_objects - false

pub fn set_big_endian(&self, enable: bool)

If true, this StreamPeer will using big-endian format for encoding and decoding.