Struct gdnative::api::StyleBoxTexture

pub struct StyleBoxTexture { /* private fields */ }

core class StyleBoxTexture inherits StyleBox (reference-counted).

This class has related types in the style_box_texture module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

StyleBoxTexture inherits methods from:

StyleBox

Resource

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl StyleBoxTexture

Constants

pub const AXIS_STRETCH_MODE_STRETCH: i64 = 0i64

pub const AXIS_STRETCH_MODE_TILE: i64 = 1i64

pub const AXIS_STRETCH_MODE_TILE_FIT: i64 = 2i64

impl StyleBoxTexture

pub fn new() -> Ref<StyleBoxTexture, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn expand_margin_size(&self, margin: i64) -> f64

Returns the size of the given margin’s expand margin. See Margin for possible values.

pub fn h_axis_stretch_mode(&self) -> AxisStretchMode

Controls how the stylebox’s texture will be stretched or tiled horizontally. See AxisStretchMode for possible values.

pub fn margin_size(&self, margin: i64) -> f64

Returns the size of the given margin. See Margin for possible values.

pub fn modulate(&self) -> Color

Modulates the color of the texture when this style box is drawn.

pub fn normal_map(&self) -> Option<Ref<Texture, Shared>>

The normal map to use when drawing this style box. Note: Godot expects the normal map to use X+, Y-, and Z+ coordinates. See this page for a comparison of normal map coordinates expected by popular engines.

pub fn region_rect(&self) -> Rect2

Species a sub-region of the texture to use. This is equivalent to first wrapping the texture in an AtlasTexture with the same region.

pub fn texture(&self) -> Option<Ref<Texture, Shared>>

The texture to use when drawing this style box.

pub fn v_axis_stretch_mode(&self) -> AxisStretchMode

Controls how the stylebox’s texture will be stretched or tiled vertically. See AxisStretchMode for possible values.

pub fn is_draw_center_enabled(&self) -> bool

If true, the nine-patch texture’s center tile will be drawn.

pub fn set_draw_center(&self, enable: bool)

If true, the nine-patch texture’s center tile will be drawn.

pub fn set_expand_margin_all(&self, size: f64)

Sets the expand margin to size pixels for all margins.

pub fn set_expand_margin_individual( &self, size_left: f64, size_top: f64, size_right: f64, size_bottom: f64 )

Sets the expand margin for each margin to size_left, size_top, size_right, and size_bottom pixels.

pub fn set_expand_margin_size(&self, margin: i64, size: f64)

Sets the expand margin to size pixels for the given margin. See Margin for possible values.

pub fn set_h_axis_stretch_mode(&self, mode: i64)

Controls how the stylebox’s texture will be stretched or tiled horizontally. See AxisStretchMode for possible values.

pub fn set_margin_size(&self, margin: i64, size: f64)

Sets the margin to size pixels for the given margin. See Margin for possible values.

pub fn set_modulate(&self, color: Color)

Modulates the color of the texture when this style box is drawn.

pub fn set_normal_map(&self, normal_map: impl AsArg<Texture>)

The normal map to use when drawing this style box. Note: Godot expects the normal map to use X+, Y-, and Z+ coordinates. See this page for a comparison of normal map coordinates expected by popular engines.

pub fn set_region_rect(&self, region: Rect2)

Species a sub-region of the texture to use. This is equivalent to first wrapping the texture in an AtlasTexture with the same region.

pub fn set_texture(&self, texture: impl AsArg<Texture>)

The texture to use when drawing this style box.

pub fn set_v_axis_stretch_mode(&self, mode: i64)

Controls how the stylebox’s texture will be stretched or tiled vertically. See AxisStretchMode for possible values.

pub fn expand_margin_bottom(&self) -> f64

Expands the bottom margin of this style box when drawing, causing it to be drawn larger than requested.

pub fn set_expand_margin_bottom(&self, value: f64)

Expands the bottom margin of this style box when drawing, causing it to be drawn larger than requested.

pub fn expand_margin_left(&self) -> f64

Expands the left margin of this style box when drawing, causing it to be drawn larger than requested.

pub fn set_expand_margin_left(&self, value: f64)

Expands the left margin of this style box when drawing, causing it to be drawn larger than requested.

pub fn expand_margin_right(&self) -> f64

Expands the right margin of this style box when drawing, causing it to be drawn larger than requested.

pub fn set_expand_margin_right(&self, value: f64)

Expands the right margin of this style box when drawing, causing it to be drawn larger than requested.

pub fn expand_margin_top(&self) -> f64

Expands the top margin of this style box when drawing, causing it to be drawn larger than requested.

pub fn set_expand_margin_top(&self, value: f64)

Expands the top margin of this style box when drawing, causing it to be drawn larger than requested.

pub fn margin_bottom(&self) -> f64

Increases the bottom margin of the 3×3 texture box. A higher value means more of the source texture is considered to be part of the bottom border of the 3×3 box. This is also the value used as fallback for StyleBox.content_margin_bottom if it is negative.

pub fn set_margin_bottom(&self, value: f64)

Increases the bottom margin of the 3×3 texture box. A higher value means more of the source texture is considered to be part of the bottom border of the 3×3 box. This is also the value used as fallback for StyleBox.content_margin_bottom if it is negative.

pub fn margin_left(&self) -> f64

Increases the left margin of the 3×3 texture box. A higher value means more of the source texture is considered to be part of the left border of the 3×3 box. This is also the value used as fallback for StyleBox.content_margin_left if it is negative.

pub fn set_margin_left(&self, value: f64)

Increases the left margin of the 3×3 texture box. A higher value means more of the source texture is considered to be part of the left border of the 3×3 box. This is also the value used as fallback for StyleBox.content_margin_left if it is negative.

pub fn margin_right(&self) -> f64

Increases the right margin of the 3×3 texture box. A higher value means more of the source texture is considered to be part of the right border of the 3×3 box. This is also the value used as fallback for StyleBox.content_margin_right if it is negative.

pub fn set_margin_right(&self, value: f64)

Increases the right margin of the 3×3 texture box. A higher value means more of the source texture is considered to be part of the right border of the 3×3 box. This is also the value used as fallback for StyleBox.content_margin_right if it is negative.

pub fn margin_top(&self) -> f64

Increases the top margin of the 3×3 texture box. A higher value means more of the source texture is considered to be part of the top border of the 3×3 box. This is also the value used as fallback for StyleBox.content_margin_top if it is negative.

pub fn set_margin_top(&self, value: f64)

Increases the top margin of the 3×3 texture box. A higher value means more of the source texture is considered to be part of the top border of the 3×3 box. This is also the value used as fallback for StyleBox.content_margin_top if it is negative.