Struct gdnative::api::SurfaceTool

pub struct SurfaceTool { /* fields omitted */ }

core class SurfaceTool inherits Reference (reference-counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

SurfaceTool inherits methods from:

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl SurfaceTool

pub fn new() -> Ref<SurfaceTool, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn add_bones(&self, bones: PoolArray<i32>)

Specifies an array of bones to use for the next vertex. bones must contain 4 integers.

pub fn add_color(&self, color: Color)

Specifies a Color to use for the next vertex. If every vertex needs to have this information set and you fail to submit it for the first vertex, this information may not be used at all. Note: The material must have [SpatialMaterial.vertex_color_use_as_albedo][SpatialMaterial::vertex_color_use_as_albedo] enabled for the vertex color to be visible.

pub fn add_index(&self, index: i64)

Adds an index to index array if you are using indexed vertices. Does not need to be called before adding vertices.

pub fn add_normal(&self, normal: Vector3)

Specifies a normal to use for the next vertex. If every vertex needs to have this information set and you fail to submit it for the first vertex, this information may not be used at all.

pub fn add_smooth_group(&self, smooth: bool)

Specifies whether the current vertex (if using only vertex arrays) or current index (if also using index arrays) should use smooth normals for normal calculation.

pub fn add_tangent(&self, tangent: Plane)

Specifies a tangent to use for the next vertex. If every vertex needs to have this information set and you fail to submit it for the first vertex, this information may not be used at all.

pub fn add_triangle_fan(
&self,
vertices: PoolArray<Vector3>,
uvs: PoolArray<Vector2>,
colors: PoolArray<Color>,
uv2s: PoolArray<Vector2>,
normals: PoolArray<Vector3>,
tangents: VariantArray<Shared>
)

Inserts a triangle fan made of array data into Mesh being constructed. Requires the primitive type be set to Mesh.PRIMITIVE_TRIANGLES.

Default Arguments

uvs - PoolVector2Array( )

colors - PoolColorArray( )

uv2s - PoolVector2Array( )

normals - PoolVector3Array( )

tangents - [ ]

pub fn add_uv(&self, uv: Vector2)

Specifies a set of UV coordinates to use for the next vertex. If every vertex needs to have this information set and you fail to submit it for the first vertex, this information may not be used at all.

pub fn add_uv2(&self, uv2: Vector2)

Specifies an optional second set of UV coordinates to use for the next vertex. If every vertex needs to have this information set and you fail to submit it for the first vertex, this information may not be used at all.

pub fn add_vertex(&self, vertex: Vector3)

Specifies the position of current vertex. Should be called after specifying other vertex properties (e.g. Color, UV).

pub fn add_weights(&self, weights: PoolArray<f32>)

Specifies weight values to use for the next vertex. weights must contain 4 values. If every vertex needs to have this information set and you fail to submit it for the first vertex, this information may not be used at all.

pub fn append_from(
&self,
existing: impl AsArg<Mesh>,
surface: i64,
transform: Transform
)

Append vertices from a given Mesh surface onto the current vertex array with specified Transform.

pub fn begin(&self, primitive: i64)

Called before adding any vertices. Takes the primitive type as an argument (e.g. Mesh.PRIMITIVE_TRIANGLES).

pub fn clear(&self)

Clear all information passed into the surface tool so far.

pub fn commit(
&self,
existing: impl AsArg<ArrayMesh>,
flags: i64
) -> Option<Ref<ArrayMesh, Shared>>

Returns a constructed ArrayMesh from current information passed in. If an existing ArrayMesh is passed in as an argument, will add an extra surface to the existing ArrayMesh. Default flag is Mesh.ARRAY_COMPRESS_DEFAULT. See ARRAY_COMPRESS_* constants in [enum Mesh.ArrayFormat] for other flags.

Default Arguments

existing - null

flags - 97280

pub fn commit_to_arrays(&self) -> VariantArray<Shared>

Commits the data to the same format used by ArrayMesh.add_surface_from_arrays. This way you can further process the mesh data using the ArrayMesh API.

pub fn create_from(&self, existing: impl AsArg<Mesh>, surface: i64)

Creates a vertex array from an existing Mesh.

pub fn create_from_blend_shape(
&self,
existing: impl AsArg<Mesh>,
surface: i64,
blend_shape: impl Into<GodotString>
)

Creates a vertex array from the specified blend shape of an existing Mesh. This can be used to extract a specific pose from a blend shape.

pub fn deindex(&self)

Removes the index array by expanding the vertex array.

pub fn generate_normals(&self, flip: bool)

Generates normals from vertices so you do not have to do it manually. If flip is true, the resulting normals will be inverted. generate_normals should be called after generating geometry and before committing the mesh using commit or commit_to_arrays. Note: generate_normals only works if the primitive type to be set to Mesh.PRIMITIVE_TRIANGLES.

Default Arguments

flip - false

pub fn generate_tangents(&self)

Generates a tangent vector for each vertex. Requires that each vertex have UVs and normals set already.

pub fn index(&self)

Shrinks the vertex array by creating an index array. This can improve performance by avoiding vertex reuse.

pub fn set_material(&self, material: impl AsArg<Material>)

Sets Material to be used by the Mesh you are constructing.