Struct gdnative::api::TileMap

pub struct TileMap { /* fields omitted */ }

core class TileMap inherits Node2D (manually managed).

This class has related types in the tile_map module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

TileMap is a reference-only type. Persistent references can only exist in the unsafe Ref<TileMap> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

TileMap inherits methods from:

Node2D

CanvasItem

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl TileMap

Constants

pub const INVALID_CELL: i64

pub const HALF_OFFSET_X: i64

pub const MODE_SQUARE: i64

pub const TILE_ORIGIN_TOP_LEFT: i64

pub const HALF_OFFSET_Y: i64

pub const MODE_ISOMETRIC: i64

pub const TILE_ORIGIN_CENTER: i64

pub const HALF_OFFSET_DISABLED: i64

pub const MODE_CUSTOM: i64

pub const TILE_ORIGIN_BOTTOM_LEFT: i64

pub const HALF_OFFSET_NEGATIVE_X: i64

pub const HALF_OFFSET_NEGATIVE_Y: i64

impl TileMap

pub fn new() -> Ref<TileMap, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn clear(&self)

Clears all cells.

pub fn fix_invalid_tiles(&self)

Clears cells that do not exist in the tileset.

pub fn get_cell(&self, x: i64, y: i64) -> i64

Returns the tile index of the given cell. If no tile exists in the cell, returns INVALID_CELL.

pub fn get_cell_autotile_coord(&self, x: i64, y: i64) -> Vector2

Returns the coordinate (subtile column and row) of the autotile variation in the tileset. Returns a zero vector if the cell doesn’t have autotiling.

pub fn cell_size(&self) -> Vector2

The TileMap’s cell size.

pub fn get_cellv(&self, position: Vector2) -> i64

Returns the tile index of the cell given by a Vector2. If no tile exists in the cell, returns INVALID_CELL.

pub fn clip_uv(&self) -> bool

If true, the cell’s UVs will be clipped.

pub fn collision_bounce(&self) -> f64

Bounce value for static body collisions (see collision_use_kinematic).

pub fn collision_friction(&self) -> f64

Friction value for static body collisions (see collision_use_kinematic).

pub fn collision_layer(&self) -> i64

The collision layer(s) for all colliders in the TileMap. See Collision layers and masks in the documentation for more information.

pub fn get_collision_layer_bit(&self, bit: i64) -> bool

Returns true if the given collision layer bit is set.

pub fn collision_mask(&self) -> i64

The collision mask(s) for all colliders in the TileMap. See Collision layers and masks in the documentation for more information.

pub fn get_collision_mask_bit(&self, bit: i64) -> bool

Returns true if the given collision mask bit is set.

pub fn collision_use_kinematic(&self) -> bool

If true, TileMap collisions will be handled as a kinematic body. If false, collisions will be handled as static body.

pub fn collision_use_parent(&self) -> bool

If true, this tilemap’s collision shape will be added to the collision shape of the parent. The parent has to be a CollisionObject2D.

pub fn custom_transform(&self) -> Transform2D

The custom Transform2D to be applied to the TileMap’s cells.

pub fn half_offset(&self) -> HalfOffset

Amount to offset alternating tiles. See HalfOffset for possible values.

pub fn mode(&self) -> Mode

The TileMap orientation mode. See Mode for possible values.

pub fn occluder_light_mask(&self) -> i64

The light mask assigned to all light occluders in the TileMap. The TileSet’s light occluders will cast shadows only from Light2D(s) that have the same light mask(s).

pub fn quadrant_size(&self) -> i64

The TileMap’s quadrant size. Optimizes drawing by batching, using chunks of this size.

pub fn tile_origin(&self) -> TileOrigin

Position for tile origin. See TileOrigin for possible values.

pub fn tileset(&self) -> Option<Ref<TileSet, Shared>>

The assigned TileSet.

pub fn get_used_cells(&self) -> VariantArray<Shared>

Returns a Vector2 array with the positions of all cells containing a tile from the tileset (i.e. a tile index different from -1).

pub fn get_used_cells_by_id(&self, id: i64) -> VariantArray<Shared>

Returns an array of all cells with the given tile index specified in id.

pub fn get_used_rect(&self) -> Rect2

Returns a rectangle enclosing the used (non-empty) tiles of the map.

pub fn is_cell_transposed(&self, x: i64, y: i64) -> bool

Returns true if the given cell is transposed, i.e. the X and Y axes are swapped.

pub fn is_cell_x_flipped(&self, x: i64, y: i64) -> bool

Returns true if the given cell is flipped in the X axis.

pub fn is_cell_y_flipped(&self, x: i64, y: i64) -> bool

Returns true if the given cell is flipped in the Y axis.

pub fn is_centered_textures_enabled(&self) -> bool

If true, the textures will be centered in the middle of each tile. This is useful for certain isometric or top-down modes when textures are made larger or smaller than the tiles (e.g. to avoid flickering on tile edges). The offset is still applied, but from the center of the tile. If used, [compatibility_mode][Self::compatibility_mode] is ignored. If false, the texture position start in the top-left corner unless [compatibility_mode][Self::compatibility_mode] is enabled.

pub fn is_compatibility_mode_enabled(&self) -> bool

If true, the compatibility with the tilemaps made in Godot 3.1 or earlier is maintained (textures move when the tile origin changes and rotate if the texture size is not homogeneous). This mode presents problems when doing flip_h, flip_v and transpose tile operations on non-homogeneous isometric tiles (e.g. 2:1), in which the texture could not coincide with the collision, thus it is not recommended for isometric or non-square tiles. If false, the textures do not move when doing flip_h, flip_v operations if no offset is used, nor when changing the tile origin. The compatibility mode doesn’t work with the [centered_textures][Self::centered_textures] option, because displacing textures with the [cell_tile_origin][Self::cell_tile_origin] option or in irregular tiles is not relevant when centering those textures.

pub fn is_show_collision_enabled(&self) -> bool

If true, collision shapes are shown in the editor and at run-time. Requires Visible Collision Shapes to be enabled in the Debug menu for collision shapes to be visible at run-time.

pub fn is_y_sort_mode_enabled(&self) -> bool

If true, the TileMap’s direct children will be drawn in order of their Y coordinate.

pub fn map_to_world(
&self,
map_position: Vector2,
ignore_half_ofs: bool
) -> Vector2

Sample code is GDScript unless otherwise noted.

Returns the local position of the top left corner of the cell corresponding to the given tilemap (grid-based) coordinates. To get the global position, use Node2D.to_global:

var local_position = my_tilemap.map_to_world(map_position)
var global_position = my_tilemap.to_global(local_position)

Optionally, the tilemap’s half offset can be ignored.

Default Arguments

ignore_half_ofs - false

pub fn set_cell(
&self,
x: i64,
y: i64,
tile: i64,
flip_x: bool,
flip_y: bool,
transpose: bool,
autotile_coord: Vector2
)

Sample code is GDScript unless otherwise noted.

Sets the tile index for the cell given by a Vector2. An index of -1 clears the cell. Optionally, the tile can also be flipped, transposed, or given autotile coordinates. The autotile coordinate refers to the column and row of the subtile. Note: Data such as navigation polygons and collision shapes are not immediately updated for performance reasons. If you need these to be immediately updated, you can call update_dirty_quadrants. Overriding this method also overrides it internally, allowing custom logic to be implemented when tiles are placed/removed:

func set_cell(x, y, tile, flip_x=false, flip_y=false, transpose=false, autotile_coord=Vector2()):
    .set_cell(x, y, tile, flip_x, flip_y, transpose, autotile_coord)

Default Arguments

flip_x - false

flip_y - false

transpose - false

autotile_coord - Vector2( 0, 0 )

pub fn set_cell_size(&self, size: Vector2)

The TileMap’s cell size.

pub fn set_cellv(
&self,
position: Vector2,
tile: i64,
flip_x: bool,
flip_y: bool,
transpose: bool
)

Sets the tile index for the given cell. An index of -1 clears the cell. Optionally, the tile can also be flipped or transposed. Note: Data such as navigation polygons and collision shapes are not immediately updated for performance reasons. If you need these to be immediately updated, you can call update_dirty_quadrants.

Default Arguments

flip_x - false

flip_y - false

transpose - false

pub fn set_centered_textures(&self, enable: bool)

If true, the textures will be centered in the middle of each tile. This is useful for certain isometric or top-down modes when textures are made larger or smaller than the tiles (e.g. to avoid flickering on tile edges). The offset is still applied, but from the center of the tile. If used, [compatibility_mode][Self::compatibility_mode] is ignored. If false, the texture position start in the top-left corner unless [compatibility_mode][Self::compatibility_mode] is enabled.

pub fn set_clip_uv(&self, enable: bool)

If true, the cell’s UVs will be clipped.

pub fn set_collision_bounce(&self, value: f64)

Bounce value for static body collisions (see collision_use_kinematic).

pub fn set_collision_friction(&self, value: f64)

Friction value for static body collisions (see collision_use_kinematic).

pub fn set_collision_layer(&self, layer: i64)

The collision layer(s) for all colliders in the TileMap. See Collision layers and masks in the documentation for more information.

pub fn set_collision_layer_bit(&self, bit: i64, value: bool)

Sets the given collision layer bit.

pub fn set_collision_mask(&self, mask: i64)

The collision mask(s) for all colliders in the TileMap. See Collision layers and masks in the documentation for more information.

pub fn set_collision_mask_bit(&self, bit: i64, value: bool)

Sets the given collision mask bit.

pub fn set_collision_use_kinematic(&self, use_kinematic: bool)

If true, TileMap collisions will be handled as a kinematic body. If false, collisions will be handled as static body.

pub fn set_collision_use_parent(&self, use_parent: bool)

If true, this tilemap’s collision shape will be added to the collision shape of the parent. The parent has to be a CollisionObject2D.

pub fn set_compatibility_mode(&self, enable: bool)

If true, the compatibility with the tilemaps made in Godot 3.1 or earlier is maintained (textures move when the tile origin changes and rotate if the texture size is not homogeneous). This mode presents problems when doing flip_h, flip_v and transpose tile operations on non-homogeneous isometric tiles (e.g. 2:1), in which the texture could not coincide with the collision, thus it is not recommended for isometric or non-square tiles. If false, the textures do not move when doing flip_h, flip_v operations if no offset is used, nor when changing the tile origin. The compatibility mode doesn’t work with the [centered_textures][Self::centered_textures] option, because displacing textures with the [cell_tile_origin][Self::cell_tile_origin] option or in irregular tiles is not relevant when centering those textures.

pub fn set_custom_transform(&self, custom_transform: Transform2D)

The custom Transform2D to be applied to the TileMap’s cells.

pub fn set_half_offset(&self, half_offset: i64)

Amount to offset alternating tiles. See HalfOffset for possible values.

pub fn set_mode(&self, mode: i64)

The TileMap orientation mode. See Mode for possible values.

pub fn set_occluder_light_mask(&self, mask: i64)

The light mask assigned to all light occluders in the TileMap. The TileSet’s light occluders will cast shadows only from Light2D(s) that have the same light mask(s).

pub fn set_quadrant_size(&self, size: i64)

The TileMap’s quadrant size. Optimizes drawing by batching, using chunks of this size.

pub fn set_show_collision(&self, enable: bool)

If true, collision shapes are shown in the editor and at run-time. Requires Visible Collision Shapes to be enabled in the Debug menu for collision shapes to be visible at run-time.

pub fn set_tile_origin(&self, origin: i64)

Position for tile origin. See TileOrigin for possible values.

pub fn set_tileset(&self, tileset: impl AsArg<TileSet>)

The assigned TileSet.

pub fn set_y_sort_mode(&self, enable: bool)

If true, the TileMap’s direct children will be drawn in order of their Y coordinate.

pub fn update_bitmask_area(&self, position: Vector2)

Applies autotiling rules to the cell (and its adjacent cells) referenced by its grid-based X and Y coordinates.

pub fn update_bitmask_region(&self, start: Vector2, end: Vector2)

Applies autotiling rules to the cells in the given region (specified by grid-based X and Y coordinates). Calling with invalid (or missing) parameters applies autotiling rules for the entire tilemap.

Default Arguments

start - Vector2( 0, 0 )

end - Vector2( 0, 0 )

pub fn update_dirty_quadrants(&self)

Updates the tile map’s quadrants, allowing things such as navigation and collision shapes to be immediately used if modified.

pub fn world_to_map(&self, world_position: Vector2) -> Vector2

Sample code is GDScript unless otherwise noted.

Returns the tilemap (grid-based) coordinates corresponding to the given local position. To use this with a global position, first determine the local position with Node2D.to_local:

var local_position = my_tilemap.to_local(global_position)
var map_position = my_tilemap.world_to_map(local_position)