Struct gdnative::api::VehicleBody

pub struct VehicleBody { /* private fields */ }

core class VehicleBody inherits RigidBody (manually managed).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

VehicleBody is a reference-only type. Persistent references can only exist in the unsafe Ref<VehicleBody> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

VehicleBody inherits methods from:

RigidBody

PhysicsBody

CollisionObject

Spatial

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl VehicleBody

pub fn new() -> Ref<VehicleBody, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn brake(&self) -> f64

Slows down the vehicle by applying a braking force. The vehicle is only slowed down if the wheels are in contact with a surface. The force you need to apply to adequately slow down your vehicle depends on the RigidBody.mass of the vehicle. For a vehicle with a mass set to 1000, try a value in the 25 - 30 range for hard braking.

pub fn engine_force(&self) -> f64

Accelerates the vehicle by applying an engine force. The vehicle is only sped up if the wheels that have [VehicleWheel.use_as_traction][VehicleWheel::use_as_traction] set to true and are in contact with a surface. The RigidBody.mass of the vehicle has an effect on the acceleration of the vehicle. For a vehicle with a mass set to 1000, try a value in the 25 - 50 range for acceleration. Note: The simulation does not take the effect of gears into account, you will need to add logic for this if you wish to simulate gears. A negative value will result in the vehicle reversing.

pub fn steering(&self) -> f64

The steering angle for the vehicle. Setting this to a non-zero value will result in the vehicle turning when it’s moving. Wheels that have [VehicleWheel.use_as_steering][VehicleWheel::use_as_steering] set to true will automatically be rotated.

pub fn set_brake(&self, brake: f64)

Slows down the vehicle by applying a braking force. The vehicle is only slowed down if the wheels are in contact with a surface. The force you need to apply to adequately slow down your vehicle depends on the RigidBody.mass of the vehicle. For a vehicle with a mass set to 1000, try a value in the 25 - 30 range for hard braking.

pub fn set_engine_force(&self, engine_force: f64)

Accelerates the vehicle by applying an engine force. The vehicle is only sped up if the wheels that have [VehicleWheel.use_as_traction][VehicleWheel::use_as_traction] set to true and are in contact with a surface. The RigidBody.mass of the vehicle has an effect on the acceleration of the vehicle. For a vehicle with a mass set to 1000, try a value in the 25 - 50 range for acceleration. Note: The simulation does not take the effect of gears into account, you will need to add logic for this if you wish to simulate gears. A negative value will result in the vehicle reversing.

pub fn set_steering(&self, steering: f64)

The steering angle for the vehicle. Setting this to a non-zero value will result in the vehicle turning when it’s moving. Wheels that have [VehicleWheel.use_as_steering][VehicleWheel::use_as_steering] set to true will automatically be rotated.