Struct gdnative::api::VideoPlayer

pub struct VideoPlayer { /* private fields */ }

core class VideoPlayer inherits Control (manually managed).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

VideoPlayer is a reference-only type. Persistent references can only exist in the unsafe Ref<VideoPlayer> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

VideoPlayer inherits methods from:

Control

CanvasItem

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl VideoPlayer

pub fn new() -> Ref<VideoPlayer, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn audio_track(&self) -> i64

The embedded audio track to play.

pub fn buffering_msec(&self) -> i64

Amount of time in milliseconds to store in buffer while playing.

pub fn bus(&self) -> GodotString

Audio bus to use for sound playback.

pub fn stream(&self) -> Option<Ref<VideoStream, Shared>>

The assigned video stream. See description for supported formats.

pub fn get_stream_name(&self) -> GodotString

Returns the video stream’s name, or "<No Stream>" if no video stream is assigned.

pub fn stream_position(&self) -> f64

The current position of the stream, in seconds. Note: Changing this value won’t have any effect as seeking is not implemented yet, except in video formats implemented by a GDNative add-on.

pub fn get_video_texture(&self) -> Option<Ref<Texture, Shared>>

Returns the current frame as a Texture.

pub fn volume(&self) -> f64

Audio volume as a linear value.

pub fn volume_db(&self) -> f64

Audio volume in dB.

pub fn has_autoplay(&self) -> bool

If true, playback starts when the scene loads.

pub fn has_expand(&self) -> bool

If true, the video scales to the control size. Otherwise, the control minimum size will be automatically adjusted to match the video stream’s dimensions.

pub fn is_paused(&self) -> bool

If true, the video is paused.

pub fn is_playing(&self) -> bool

Returns true if the video is playing. Note: The video is still considered playing if paused during playback.

pub fn play(&self)

Starts the video playback from the beginning. If the video is paused, this will not unpause the video.

pub fn set_audio_track(&self, track: i64)

The embedded audio track to play.

pub fn set_autoplay(&self, enabled: bool)

If true, playback starts when the scene loads.

pub fn set_buffering_msec(&self, msec: i64)

Amount of time in milliseconds to store in buffer while playing.

pub fn set_bus(&self, bus: impl Into<GodotString>)

Audio bus to use for sound playback.

pub fn set_expand(&self, enable: bool)

If true, the video scales to the control size. Otherwise, the control minimum size will be automatically adjusted to match the video stream’s dimensions.

pub fn set_paused(&self, paused: bool)

If true, the video is paused.

pub fn set_stream(&self, stream: impl AsArg<VideoStream>)

The assigned video stream. See description for supported formats.

pub fn set_stream_position(&self, position: f64)

The current position of the stream, in seconds. Note: Changing this value won’t have any effect as seeking is not implemented yet, except in video formats implemented by a GDNative add-on.

pub fn set_volume(&self, volume: f64)

Audio volume as a linear value.

pub fn set_volume_db(&self, db: f64)

Audio volume in dB.

pub fn stop(&self)

Stops the video playback and sets the stream position to 0. Note: Although the stream position will be set to 0, the first frame of the video stream won’t become the current frame.