Struct gdnative::api::VisualScriptBuiltinFunc

pub struct VisualScriptBuiltinFunc { /* fields omitted */ }

core class VisualScriptBuiltinFunc inherits VisualScriptNode (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

VisualScriptBuiltinFunc inherits methods from:

VisualScriptNode

Resource

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl VisualScriptBuiltinFunc

Constants

pub const MATH_SIN: i64

pub const MATH_COS: i64

pub const MATH_TAN: i64

pub const MATH_SINH: i64

pub const MATH_COSH: i64

pub const MATH_TANH: i64

pub const MATH_ASIN: i64

pub const MATH_ACOS: i64

pub const MATH_ATAN: i64

pub const MATH_ATAN2: i64

pub const MATH_SQRT: i64

pub const MATH_FMOD: i64

pub const MATH_FPOSMOD: i64

pub const MATH_FLOOR: i64

pub const MATH_CEIL: i64

pub const MATH_ROUND: i64

pub const MATH_ABS: i64

pub const MATH_SIGN: i64

pub const MATH_POW: i64

pub const MATH_LOG: i64

pub const MATH_EXP: i64

pub const MATH_ISNAN: i64

pub const MATH_ISINF: i64

pub const MATH_EASE: i64

pub const MATH_DECIMALS: i64

pub const MATH_STEPIFY: i64

pub const MATH_LERP: i64

pub const MATH_INVERSE_LERP: i64

pub const MATH_RANGE_LERP: i64

pub const MATH_MOVE_TOWARD: i64

pub const MATH_DECTIME: i64

pub const MATH_RANDOMIZE: i64

pub const MATH_RAND: i64

pub const MATH_RANDF: i64

pub const MATH_RANDOM: i64

pub const MATH_SEED: i64

pub const MATH_RANDSEED: i64

pub const MATH_DEG2RAD: i64

pub const MATH_RAD2DEG: i64

pub const MATH_LINEAR2DB: i64

pub const MATH_DB2LINEAR: i64

pub const MATH_POLAR2CARTESIAN: i64

pub const MATH_CARTESIAN2POLAR: i64

pub const MATH_WRAP: i64

pub const MATH_WRAPF: i64

pub const LOGIC_MAX: i64

pub const LOGIC_MIN: i64

pub const LOGIC_CLAMP: i64

pub const LOGIC_NEAREST_PO2: i64

pub const OBJ_WEAKREF: i64

pub const FUNC_FUNCREF: i64

pub const TYPE_CONVERT: i64

pub const TYPE_OF: i64

pub const TYPE_EXISTS: i64

pub const TEXT_CHAR: i64

pub const TEXT_STR: i64

pub const TEXT_PRINT: i64

pub const TEXT_PRINTERR: i64

pub const TEXT_PRINTRAW: i64

pub const VAR_TO_STR: i64

pub const STR_TO_VAR: i64

pub const VAR_TO_BYTES: i64

pub const BYTES_TO_VAR: i64

pub const COLORN: i64

pub const MATH_SMOOTHSTEP: i64

pub const MATH_POSMOD: i64

pub const MATH_LERP_ANGLE: i64

pub const TEXT_ORD: i64

pub const FUNC_MAX: i64

impl VisualScriptBuiltinFunc

pub fn new() -> Ref<VisualScriptBuiltinFunc, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn func(&self) -> BuiltinFunc

pub fn set_func(&self, which: i64)