Struct gdnative::api::VisualShaderNodeBooleanConstant

pub struct VisualShaderNodeBooleanConstant { /* private fields */ }

core class VisualShaderNodeBooleanConstant inherits VisualShaderNode (reference-counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

VisualShaderNodeBooleanConstant inherits methods from:

VisualShaderNode

Resource

Reference

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl VisualShaderNodeBooleanConstant

pub fn new() -> Ref<VisualShaderNodeBooleanConstant, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn constant(&self) -> bool

A boolean constant which represents a state of this node.

pub fn set_constant(&self, value: bool)

A boolean constant which represents a state of this node.