Struct gdnative::api::WebXRInterface

pub struct WebXRInterface { /* fields omitted */ }

core class WebXRInterface inherits ARVRInterface (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

WebXRInterface inherits methods from:

ARVRInterface

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl WebXRInterface

pub fn bounds_geometry(&self) -> TypedArray<Vector3>

pub fn get_controller(
&self,
controller_id: i64
) -> Option<Ref<ARVRPositionalTracker, Shared>>

pub fn optional_features(&self) -> GodotString

pub fn reference_space_type(&self) -> GodotString

pub fn requested_reference_space_types(&self) -> GodotString

pub fn required_features(&self) -> GodotString

pub fn session_mode(&self) -> GodotString

pub fn visibility_state(&self) -> GodotString

pub fn is_session_supported(&self, session_mode: impl Into<GodotString>)

pub fn set_optional_features(&self, optional_features: impl Into<GodotString>)

pub fn set_requested_reference_space_types(
&self,
requested_reference_space_types: impl Into<GodotString>
)

pub fn set_required_features(&self, required_features: impl Into<GodotString>)

pub fn set_session_mode(&self, session_mode: impl Into<GodotString>)