Struct gdnative::api::theme::Theme

pub struct Theme { /* fields omitted */ }

core class Theme inherits Resource (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

Theme inherits methods from:

Resource

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Theme

pub fn new() -> Ref<Theme, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn clear(&self)

Clears all values on the theme.

pub fn clear_color(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
)

Clears the Color at name if the theme has node_type.

pub fn clear_constant(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
)

Clears the constant at name if the theme has node_type.

pub fn clear_font(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
)

Clears the Font at name if the theme has node_type.

pub fn clear_icon(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
)

Clears the icon at name if the theme has node_type.

pub fn clear_stylebox(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
)

Clears StyleBox at name if the theme has node_type.

pub fn copy_default_theme(&self)

Sets the theme’s values to a copy of the default theme values.

pub fn copy_theme(&self, other: impl AsArg<Theme>)

Sets the theme’s values to a copy of a given theme.

pub fn get_color(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
) -> Color

Returns the Color at name if the theme has node_type.

pub fn get_color_list(
&self,
node_type: impl Into<GodotString>
) -> TypedArray<GodotString>

Returns all the Colors as a PoolStringArray filled with each Color’s name, for use in get_color, if the theme has node_type.

pub fn get_constant(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
) -> i64

Returns the constant at name if the theme has node_type.

pub fn get_constant_list(
&self,
node_type: impl Into<GodotString>
) -> TypedArray<GodotString>

Returns all the constants as a PoolStringArray filled with each constant’s name, for use in get_constant, if the theme has node_type.

pub fn default_font(&self) -> Option<Ref<Font, Shared>>

The theme’s default font.

pub fn get_font(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
) -> Option<Ref<Font, Shared>>

Returns the Font at name if the theme has node_type.

pub fn get_font_list(
&self,
node_type: impl Into<GodotString>
) -> TypedArray<GodotString>

Returns all the Fonts as a PoolStringArray filled with each Font’s name, for use in get_font, if the theme has node_type.

pub fn get_icon(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
) -> Option<Ref<Texture, Shared>>

Returns the icon Texture at name if the theme has node_type.

pub fn get_icon_list(
&self,
node_type: impl Into<GodotString>
) -> TypedArray<GodotString>

Returns all the icons as a PoolStringArray filled with each Texture’s name, for use in get_icon, if the theme has node_type.

pub fn get_stylebox(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
) -> Option<Ref<StyleBox, Shared>>

Returns the StyleBox at name if the theme has node_type. Valid names may be found using get_stylebox_list. Valid node_types may be found using get_stylebox_types.

pub fn get_stylebox_list(
&self,
node_type: impl Into<GodotString>
) -> TypedArray<GodotString>

Returns all the StyleBoxs as a PoolStringArray filled with each StyleBox’s name, for use in get_stylebox, if the theme has node_type. Valid node_types may be found using get_stylebox_types.

pub fn get_stylebox_types(&self) -> TypedArray<GodotString>

Returns all the StyleBox types as a PoolStringArray filled with each StyleBox’s type, for use in get_stylebox and/or get_stylebox_list, if the theme has node_type.

pub fn get_type_list(
&self,
node_type: impl Into<GodotString>
) -> TypedArray<GodotString>

Returns all the types in node_type as a PoolStringArray for use in any of the get_* functions, if the theme has node_type.

pub fn has_color(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
) -> bool

Returns true if Color with name is in node_type. Returns false if the theme does not have node_type.

pub fn has_constant(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
) -> bool

Returns true if constant with name is in node_type. Returns false if the theme does not have node_type.

pub fn has_font(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
) -> bool

Returns true if Font with name is in node_type. Returns false if the theme does not have node_type.

pub fn has_icon(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
) -> bool

Returns true if icon Texture with name is in node_type. Returns false if the theme does not have node_type.

pub fn has_stylebox(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>
) -> bool

Returns true if StyleBox with name is in node_type. Returns false if the theme does not have node_type.

pub fn set_color(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>,
color: Color
)

Sets the theme’s Color to color at name in node_type. Does nothing if the theme does not have node_type.

pub fn set_constant(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>,
constant: i64
)

Sets the theme’s constant to constant at name in node_type. Does nothing if the theme does not have node_type.

pub fn set_default_font(&self, font: impl AsArg<Font>)

The theme’s default font.

pub fn set_font(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>,
font: impl AsArg<Font>
)

Sets the theme’s Font to font at name in node_type. Does nothing if the theme does not have node_type.

pub fn set_icon(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>,
texture: impl AsArg<Texture>
)

Sets the theme’s icon Texture to texture at name in node_type. Does nothing if the theme does not have node_type.

pub fn set_stylebox(
&self,
name: impl Into<GodotString>,
node_type: impl Into<GodotString>,
texture: impl AsArg<StyleBox>
)

Sets theme’s StyleBox to stylebox at name in node_type. Does nothing if the theme does not have node_type.