Function gdnative::api::utils::autoload

pub unsafe fn autoload<'a, T>(name: &str) -> Option<TRef<'a, T, Shared>> where
    T: SubClass<Node>, 

Convenience method to obtain a reference to an “auto-load” node, that is a child of the root node. Returns None if the node does not exist or is not of the correct type.

Safety

This method accesses the scene tree. As a result, any calls to this function must follow the official thread-safety guidelines. assume_safe invariants must be observed for the resulting node during 'a, if any.