Struct gdnative::api::web_rtc_data_channel::WebRTCDataChannel

pub struct WebRTCDataChannel { /* fields omitted */ }

core class WebRTCDataChannel inherits PacketPeer (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

WebRTCDataChannel inherits methods from:

PacketPeer

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl WebRTCDataChannel

Constants

pub const STATE_CONNECTING: i64

pub const WRITE_MODE_TEXT: i64

pub const STATE_OPEN: i64

pub const WRITE_MODE_BINARY: i64

pub const STATE_CLOSING: i64

pub const STATE_CLOSED: i64

impl WebRTCDataChannel

pub fn close(&self)

pub fn get_id(&self) -> i64

pub fn get_label(&self) -> GodotString

pub fn get_max_packet_life_time(&self) -> i64

pub fn get_max_retransmits(&self) -> i64

pub fn get_protocol(&self) -> GodotString

pub fn get_ready_state(&self) -> ChannelState

pub fn write_mode(&self) -> WriteMode

pub fn is_negotiated(&self) -> bool

pub fn is_ordered(&self) -> bool

pub fn poll(&self) -> Result<(), GodotError>

pub fn set_write_mode(&self, write_mode: i64)

pub fn was_string_packet(&self) -> bool