Enum gdnative::core_types::VariantDispatch

#[repr(u32)]
pub enum VariantDispatch {

    Nil,
    Bool(bool),
    I64(i64),
    F64(f64),
    GodotString(GodotString),
    Vector2(Vector2),
    Rect2(Rect2),
    Vector3(Vector3),
    Transform2D(Transform2D),
    Plane(Plane),
    Quat(Quat),
    Aabb(Aabb),
    Basis(Basis),
    Transform(Transform),
    Color(Color),
    NodePath(NodePath),
    Rid(Rid),
    Object(Variant),
    Dictionary(Dictionary<Shared>),
    VariantArray(VariantArray<Shared>),
    ByteArray(PoolArray<u8>),
    Int32Array(PoolArray<i32>),
    Float32Array(PoolArray<f32>),
    StringArray(PoolArray<GodotString>),
    Vector2Array(PoolArray<Vector2>),
    Vector3Array(PoolArray<Vector3>),
    ColorArray(PoolArray<Color>),

}

Rust enum associating each primitive variant type to its value.

For Variants containing objects, the original Variant is returned unchanged, due to the limitations of statically-determined memory management.

Variants

Nil

Bool(bool)

I64(i64)

F64(f64)

GodotString(GodotString)

Vector2(Vector2)

Rect2(Rect2)

Vector3(Vector3)

Transform2D(Transform2D)

Plane(Plane)

Quat(Quat)

Aabb(Aabb)

Basis(Basis)

Transform(Transform)

Color(Color)

NodePath(NodePath)

Rid(Rid)

Object(Variant)

Dictionary(Dictionary<Shared>)

VariantArray(VariantArray<Shared>)

ByteArray(PoolArray<u8>)

Int32Array(PoolArray<i32>)

Float32Array(PoolArray<f32>)

StringArray(PoolArray<GodotString>)

Vector2Array(PoolArray<Vector2>)

Vector3Array(PoolArray<Vector3>)

ColorArray(PoolArray<Color>)