Struct gdnative::core_types::PoolArray

pub struct PoolArray<T>
where
    T: PoolElement,
{ /* private fields */ }

A reference-counted CoW typed vector using Godot’s pool allocator, generic over possible element types.

PoolArray unifies all the different Pool*Array types exported by Godot. It can be used in exported Rust methods as parameter and return types, as well as in exported properties. However, it is limited to the element types, for which a Pool*Array exists in GDScript, i.e. it cannot contain user-defined types. If you need other types, look into VariantArray or directly use Vec<T> for type safety.

This type is CoW (copy-on-write). The Clone implementation of this type creates a new reference without copying the contents.

If you need to read elements, e.g. for iteration or conversion to another collection, the read() method provides a view that dereferences to &[T]. Analogously, write() provides a writable view that dereferences to &mut [T].

For element mutations, it’s usually recommended to do process them in batch using write() or the append() methods, as opposed to push() or set(), because the latter ones trigger CoW behavior each time they are called.

Implementations

impl<T> PoolArray<T>
where T: PoolElement,

pub fn new() -> PoolArray<T>

Creates an empty array.

pub fn from_variant_array(array: &VariantArray) -> PoolArray<T>

Creates from a VariantArray by making a best effort to convert each variant.

pub fn from_vec(src: Vec<T>) -> PoolArray<T>

Creates a PoolArray moving elements from src.

If your source type isn’t precisely a Vec<T>, keep in mind that PoolElement implements the FromIterator trait, which allows it to be constructed from iterators, typically through collect().

For example:

use gdnative::core_types::PoolArray;

// Collect from range
let arr = (0i32..4).collect::<PoolArray<_>>();

// Type conversion
let vec: Vec<u32> = vec![1, 1, 2, 3, 5]; // note: unsigned
let arr = vec.iter().map(|&e| e as i32).collect::<PoolArray<_>>();

pub fn to_vec(&self) -> Vec<T>
where T: Clone,

Copies all elements to a Vec, leaving this instance untouched.

Equivalent to self.read().to_vec(). Only use this if your destination type is precisely a Vec<T>. Otherwise, call read() which can be used as a slice.

pub fn push(&mut self, val: T)

Appends an element to the end of the array.

Calling push() triggers copy-on-write behavior. To insert a large number of elements, consider using append(), resize() or write().

pub fn push_ref(&mut self, val: &T)

Appends an element to the end of the array by reference.

Calling push_ref() triggers copy-on-write behavior. To insert a large number of elements, consider using append(), resize() or write().

pub fn append(&mut self, src: &PoolArray<T>)

Copies and appends all values in src to the end of the array.

pub fn append_vec(&mut self, src: &mut Vec<T>)

Moves all the elements from src into self, leaving src empty.

Panics

If the resulting length would not fit in i32.

pub fn insert(&mut self, offset: i32, val: T) -> bool

Inserts an element at the given offset and returns true if successful.

pub fn insert_ref(&mut self, offset: i32, val: &T) -> bool

Inserts an element by reference at the given offset and returns true if successful.

pub fn invert(&mut self)

Inverts the order of the elements in the array.

pub fn remove(&mut self, idx: i32)

Removes an element at the given offset.

pub fn resize(&mut self, size: i32)

Changes the size of the array, possibly removing elements or pushing default values.

pub fn get(&self, idx: i32) -> T

Returns a copy of the element at the given offset.

pub fn set(&mut self, idx: i32, val: T)

Sets the value of the element at the given offset.

pub fn set_ref(&mut self, idx: i32, val: &T)

Sets the value of the element at the given offset by reference.

pub fn len(&self) -> i32

Returns the number of elements in the array.

pub fn is_empty(&self) -> bool

Returns true if the container is empty.

pub fn read(&self) -> Aligned<ReadGuard<'_, T>>

Returns a scoped read-only view into this array.

The returned read guard implements Deref with target type [T], i.e. can be dereferenced to &[T]. This means all non-mutating (&self) slice methods can be used, see here.

pub fn write(&mut self) -> Aligned<WriteGuard<'_, T>>

Returns a scoped read-write view into this array. This triggers CoW once per lock, instead of once each mutation.

The returned write guard implements DerefMut with target type [T], i.e. can be dereferenced to &mut [T]. This means all mutating and read-only slice methods can be used, see here.

impl<T> PoolArray<T>
where T: PoolElement + Copy,

pub fn from_slice(src: &[T]) -> PoolArray<T>

Creates a new PoolArray by copying from src.

Equivalent to a new object created with new(), followed by a subsequent append_slice().

Panics

If the length of src does not fit in i32.

pub fn append_slice(&mut self, src: &[T])

Copies and appends all values in src to the end of the array.

Panics

If the resulting length would not fit in i32.