Struct gdnative::core_types::Rid

pub struct Rid(/* private fields */);

A RID (โresource IDโ) is an opaque handle that refers to a Godot Resource.

RIDs do not grant access to the resource itself. Instead, they can be used in lower-level resource APIs such as the servers. See also Godot API docs for RID.

Note that RIDs are highly unsafe to work with (especially with a Release build of Godot):

They are untyped, i.e. Godot does not recognize if they represent the correct resource type.

The internal handle is interpreted as a raw pointer by Godot, meaning that passing an invalid or wrongly typed RID is instant undefined behavior.

For this reason, GDNative methods accepting Rid parameters are marked unsafe.

Implementations

impl Rid

pub fn new() -> Rid

Creates an empty, invalid RID.

pub unsafe fn get_id(self) -> i32

Returns the ID of the referenced resource.

Panics

When this instance is empty, i.e. self.is_occupied() is false.

Safety

RIDs are untyped and interpreted as raw pointers by the engine. If this method is called on an invalid resource ID, the behavior is undefined. This can happen when the resource behind the RID is no longer alive.

pub fn is_occupied(self) -> bool

Check if this RID is non-empty. This does not mean itโs valid or safe to use!

This simply checks if the handle has not been initialized with the empty default. It does not give any indication about whether it points to a valid resource.