Struct gdnative::core_types::Variant

pub struct Variant(_);

A Variant can represent all Godot values (core types or Object class instances).

The underlying data is either stored inline or reference-counted on the heap, depending on the size of the type and whether the it is trivially copyable.

If you compile godot-rust with the serde feature enabled, you will have access to serialization/deserialization support: the traits Serializable and Deserializable will be automatically implemented on VariantDispatch as well as most of the types in core_types.

Implementations

impl Variant

pub fn from_vector2(val: &Vector2) -> Variant

Creates a Variant wrapping a Vector2.

pub fn from_vector3(val: &Vector3) -> Variant

Creates a Variant wrapping a Vector3.

pub fn from_quat(val: &Quat) -> Variant

Creates a Variant wrapping a Quat.

pub fn from_plane(val: &Plane) -> Variant

Creates a Variant wrapping a Plane.

pub fn from_rect2(val: &Rect2) -> Variant

Creates a Variant wrapping a Rect2.

pub fn from_transform(val: &Transform) -> Variant

Creates a Variant wrapping a Transform.

pub fn from_transform2d(val: &Transform2D) -> Variant

Creates a Variant wrapping a Transform2D.

pub fn from_basis(val: &Basis) -> Variant

Creates a Variant wrapping a Basis.

pub fn from_color(val: &Color) -> Variant

Creates a Variant wrapping a Color.

pub fn from_aabb(val: &Aabb) -> Variant

Creates a Variant wrapping an Aabb.

pub fn from_rid(val: &Rid) -> Variant

Creates a Variant wrapping an Rid.

pub fn from_node_path(val: &NodePath) -> Variant

Creates a Variant wrapping a NodePath.

pub fn from_godot_string(val: &GodotString) -> Variant

Creates a Variant wrapping a GodotString.

pub fn from_array(val: &VariantArray<Shared>) -> Variant

Creates an Variant wrapping an array of variants.

pub fn from_byte_array(val: &PoolArray<u8>) -> Variant

Creates a Variant wrapping a byte array.

pub fn from_int32_array(val: &PoolArray<i32>) -> Variant

Creates a Variant wrapping an array of 32bit signed integers.

pub fn from_float32_array(val: &PoolArray<f32>) -> Variant

Creates a Variant wrapping an array of 32bit floats.

pub fn from_string_array(val: &PoolArray<GodotString>) -> Variant

Creates a Variant wrapping an array of godot strings.

pub fn from_vector2_array(val: &PoolArray<Vector2>) -> Variant

Creates a Variant wrapping an array of 2d vectors.

pub fn from_vector3_array(val: &PoolArray<Vector3>) -> Variant

Creates a Variant wrapping an array of 3d vectors.

pub fn from_color_array(val: &PoolArray<Color>) -> Variant

Creates a Variant wrapping an array of colors.

pub fn from_dictionary(val: &Dictionary<Shared>) -> Variant

Creates a Variant wrapping a dictionary.

pub fn new() -> Variant

Creates an empty Variant.

pub fn from_str<S>(s: S) -> Variant where
S: AsRef<str>,

Creates a Variant wrapping a string.

pub fn from_object<R>(val: R) -> Variant where
R: AsVariant,

Creates a Variant wrapping a Godot object.

pub fn from_i64(v: i64) -> Variant

Creates a Variant wrapping a signed integer value.

pub fn from_u64(v: u64) -> Variant

Creates a Variant wrapping an unsigned integer value.

pub fn from_f64(v: f64) -> Variant

Creates a Variant wrapping a double-precision float value.

pub fn from_bool(v: bool) -> Variant

Creates a Variant wrapping an boolean.

pub fn to_vector2(&self) -> Vector2

Do a best effort to create a Vector2 out of the variant, possibly returning a default value.

pub fn try_to_vector2(&self) -> Option<Vector2>

Returns Some(Vector2) if this variant is one, None otherwise.

pub fn to_vector3(&self) -> Vector3

Do a best effort to create a Vector3 out of the variant, possibly returning a default value.

pub fn try_to_vector3(&self) -> Option<Vector3>

Returns Some(Vector3) if this variant is one, None otherwise.

pub fn to_quat(&self) -> Quat

Do a best effort to create a Quat out of the variant, possibly returning a default value.

pub fn try_to_quat(&self) -> Option<Quat>

Returns Some(Quat) if this variant is one, None otherwise.

pub fn to_rect2(&self) -> Rect2

Do a best effort to create a Rect2 out of the variant, possibly returning a default value.

pub fn try_to_rect2(&self) -> Option<Rect2>

Returns Some(Rect2) if this variant is one, None otherwise.

pub fn to_transform2d(&self) -> Transform2D

Do a best effort to create a Transform2D out of the variant, possibly returning a default value.

pub fn try_to_transform2d(&self) -> Option<Transform2D>

Returns Some(Transform2D) if this variant is one, None otherwise.

pub fn to_f64(&self) -> f64

Do a best effort to create a f64 out of the variant, possibly returning a default value.

pub fn try_to_f64(&self) -> Option<f64>

Returns Some(f64) if this variant is one, None otherwise.

pub fn to_i64(&self) -> i64

Do a best effort to create an i64 out of the variant, possibly returning a default value.

pub fn try_to_i64(&self) -> Option<i64>

Returns Some(i64) if this variant is one, None otherwise.

pub fn to_bool(&self) -> bool

Do a best effort to create a bool out of the variant, possibly returning a default value.

pub fn try_to_bool(&self) -> Option<bool>

Returns Some(bool) if this variant is one, None otherwise.

pub fn to_u64(&self) -> u64

Do a best effort to create a u64 out of the variant, possibly returning a default value.

pub fn try_to_u64(&self) -> Option<u64>

Returns Some(u64) if this variant is one, None otherwise.

pub fn to_plane(&self) -> Plane

Do a best effort to create a Plane out of the variant, possibly returning a default value.

pub fn try_to_plane(&self) -> Option<Plane>

Returns Some(Plane) if this variant is one, None otherwise.

pub fn to_transform(&self) -> Transform

Do a best effort to create a Transform out of the variant, possibly returning a default value.

pub fn try_to_transform(&self) -> Option<Transform>

Returns Some(Transform) if this variant is one, None otherwise.

pub fn to_color(&self) -> Color

Do a best effort to create a Color out of the variant, possibly returning a default value.

pub fn try_to_color(&self) -> Option<Color>

Returns Some(Color) if this variant is one, None otherwise.

pub fn to_basis(&self) -> Basis

Do a best effort to create a Basis out of the variant, possibly returning a default value.

pub fn try_to_basis(&self) -> Option<Basis>

Returns Some(Basis) if this variant is one, None otherwise.

pub fn to_aabb(&self) -> Aabb

Do a best effort to create an Aabb out of the variant, possibly returning a default value.

pub fn try_to_aabb(&self) -> Option<Aabb>

Returns Some(Aabb) if this variant is one, None otherwise.

pub fn to_node_path(&self) -> NodePath

Do a best effort to create a NodePath out of the variant, possibly returning a default value.

pub fn try_to_node_path(&self) -> Option<NodePath>

Returns Some(NodePath) if this variant is one, None otherwise.

pub fn to_godot_string(&self) -> GodotString

Do a best effort to create a GodotString out of the variant, possibly returning a default value.

pub fn try_to_godot_string(&self) -> Option<GodotString>

Returns Some(GodotString) if this variant is one, None otherwise.

pub fn to_rid(&self) -> Rid

Do a best effort to create a Rid out of the variant, possibly returning a default value.

pub fn try_to_rid(&self) -> Option<Rid>

Returns Some(Rid) if this variant is one, None otherwise.

pub fn to_array(&self) -> VariantArray<Shared>

Do a best effort to create a VariantArray out of the variant, possibly returning a default value.

pub fn try_to_array(&self) -> Option<VariantArray<Shared>>

Returns Some(VariantArray) if this variant is one, None otherwise.

pub fn to_byte_array(&self) -> PoolArray<u8>

Do a best effort to create a ByteArray out of the variant, possibly returning a default value.

pub fn try_to_byte_array(&self) -> Option<PoolArray<u8>>

Returns Some(ByteArray) if this variant is one, None otherwise.

pub fn to_int32_array(&self) -> PoolArray<i32>

Do a best effort to create an Int32Array out of the variant, possibly returning a default value.

pub fn try_to_int32_array(&self) -> Option<PoolArray<i32>>

Returns Some(Int32Array) if this variant is one, None otherwise.

pub fn to_float32_array(&self) -> PoolArray<f32>

Do a best effort to create a Float32Array out of the variant, possibly returning a default value.

pub fn try_to_float32_array(&self) -> Option<PoolArray<f32>>

Returns Some(Float32Array) if this variant is one, None otherwise.

pub fn to_string_array(&self) -> PoolArray<GodotString>

Do a best effort to create a StringArray out of the variant, possibly returning a default value.

pub fn try_to_string_array(&self) -> Option<PoolArray<GodotString>>

Returns Some(StringArray) if this variant is one, None otherwise.

pub fn to_vector2_array(&self) -> PoolArray<Vector2>

Do a best effort to create a Vector2Array out of the variant, possibly returning a default value.

pub fn try_to_vector2_array(&self) -> Option<PoolArray<Vector2>>

Returns Some(Vector2Array) if this variant is one, None otherwise.

pub fn to_vector3_array(&self) -> PoolArray<Vector3>

Do a best effort to create a Vector3Array out of the variant, possibly returning a default value.

pub fn try_to_vector3_array(&self) -> Option<PoolArray<Vector3>>

Returns Some(Vector3Array) if this variant is one, None otherwise.

pub fn to_color_array(&self) -> PoolArray<Color>

Do a best effort to create a ColorArray out of the variant, possibly returning a default value.

pub fn try_to_color_array(&self) -> Option<PoolArray<Color>>

Returns Some(ColorArray) if this variant is one, None otherwise.

pub fn to_dictionary(&self) -> Dictionary<Shared>

Do a best effort to create a Dictionary out of the variant, possibly returning a default value.

pub fn try_to_dictionary(&self) -> Option<Dictionary<Shared>>

Returns Some(Dictionary) if this variant is one, None otherwise.

pub fn try_to_object<T>(&self) -> Option<Ref<T, Shared>> where
T: GodotObject,

pub fn try_to_object_with_error<T>(
&self
) -> Result<Ref<T, Shared>, FromVariantError> where
T: GodotObject,

pub fn try_to_string(&self) -> Option<String>

pub fn get_type(&self) -> VariantType

Returns this variant’s type.

pub fn dispatch(&self) -> VariantDispatch

Converts this variant to a primitive value depending on its type.

Examples

let variant = 42.to_variant();
let number_as_float = match variant.dispatch() {
    VariantDispatch::I64(i) => i as f64,
    VariantDispatch::F64(f) => f,
    _ => panic!("not a number"),
};
approx::assert_relative_eq!(42.0, number_as_float);

pub fn is_nil(&self) -> bool

Returns true if this is an empty variant.

pub fn has_method(&self, method: impl Into<GodotString>) -> bool

pub unsafe fn call(
&mut self,
method: impl Into<GodotString>,
args: &[Variant]
) -> Result<Variant, CallError>

Invokes a method on the held object.

Safety

This method may invoke [Object::call()] internally, which is unsafe, as it allows execution of arbitrary code (including user-defined code in GDScript or unsafe Rust).

pub fn evaluate(
&self,
op: VariantOperator,
rhs: &Variant
) -> Result<Variant, InvalidOp>

Evaluates a variant operator on self and rhs and returns the result on success.

Errors

Returns Err(InvalidOp) if the result is not valid.

pub fn forget(self) -> godot_variant

Returns the internal ffi representation of the variant and consumes the rust object without running the destructor.

This should be only used when certain that the receiving side is responsible for running the destructor for the object, otherwise it is leaked.