Module gdnative::export::hint

Strongly typed property hints.

Structs

ArrayHint

Array hints optionally with an element hint.

EnumHint

Hints that an integer, float or string property is an enumerated value to pick in a list.

ExpEasingHint

Hints that a float property should be edited via an exponential easing function.

RangeHint

Hints that an integer or float property should be within an inclusive range.

Enums

ColorHint

Possible hints for Color.

FloatHint

Possible hints for floats.

IntHint

Possible hints for integers.

StringHint

Possible hints for strings.