Enum gdnative::export::hint::StringHint

#[non_exhaustive]
pub enum StringHint {
    Enum(EnumHint),
    File(EnumHint),
    GlobalFile(EnumHint),
    Dir,
    GlobalDir,
    Multiline,
    Placeholder {
        placeholder: String,
    },
}

Possible hints for strings.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Enum(EnumHint)

Hints that an integer, float or string property is an enumerated value to pick in a list.

File(EnumHint)

Hints that a string property is a path to a file.

GlobalFile(EnumHint)

Hints that a string property is an absolute path to a file outside the project folder.

Dir

Hints that a string property is a path to a directory.

GlobalDir

Hints that a string property is an absolute path to a directory outside the project folder.

Multiline

Hints that a string property is text with line breaks.

Placeholder

Fields

placeholder: String

Hints that a string property should have a placeholder text visible on its input field, whenever the property is empty.

Implementations

impl StringHint

pub fn export_info(self) -> ExportInfo