Struct gdnative::export::hint::ArrayHint

pub struct ArrayHint { /* private fields */ }

Array hints optionally with an element hint.

Implementations

impl ArrayHint

pub fn new() -> ArrayHint

Returns an ArrayHint without a element hint.

pub fn with_element<T>() -> ArrayHint
where T: Export,

Returns an ArrayHint with an element hint for type T, but without a hint for that type.

pub fn with_element_hint<T>(hint: <T as Export>::Hint) -> ArrayHint
where T: Export,

Returns an ArrayHint with an element hint for type T, and a hint for that type.

pub fn with_maybe_element_hint<T>( hint: Option<<T as Export>::Hint>, ) -> ArrayHint
where T: Export,

Returns an ArrayHint with an element hint for type T, and optionally a hint for that type.

impl ArrayHint

pub fn export_info(self) -> ExportInfo