Trait gdnative::export::user_data::UserData

pub unsafe trait UserData: Clone {
    type Target: NativeClass;

    fn new(val: Self::Target) -> Self;
    fn into_user_data(self) -> *const c_void;
    unsafe fn consume_user_data_unchecked(ptr: *const c_void) -> Self;
    unsafe fn clone_from_user_data_unchecked(ptr: *const c_void) -> Self;
}

Trait for customizable user-data wrappers.

See module-level documentation for detailed explanation on user-data.

Required Associated Types

type Target: NativeClass

Required Methods

fn new(val: Self::Target) -> Self

Creates a new owned wrapper from a NativeClass instance.

This operation must never fail.

fn into_user_data(self) -> *const c_void

Takes a native instance and returns an opaque pointer that can be used to recover it.

This gives “ownership” to the engine.

This operation must never fail.

unsafe fn consume_user_data_unchecked(ptr: *const c_void) -> Self

Takes an opaque pointer produced by into_user_data and “consumes” it to produce the original instance, keeping the reference count.

This should be used when “ownership” is taken from the engine, i.e. destructors. Use elsewhere can lead to premature drops of the instance contained inside.

ptr is guaranteed to be non-null.

This operation must never fail.

Safety

ptr must be pointing to valid data of the correct type.

unsafe fn clone_from_user_data_unchecked(ptr: *const c_void) -> Self

Takes an opaque pointer produced by into_user_data and “clones” it to produce the original instance, increasing the reference count.

This should be used when user data is “borrowed” from the engine.

ptr is guaranteed to be non-null.

This operation must never fail.

Safety

ptr must be pointing to valid data of the correct type.