Function gdnative::globalscope::inverse_lerp

pub fn inverse_lerp(range: RangeInclusive<f32>, value: f32) -> f32

Find linear interpolation weight from interpolated values.

Returns an interpolation or extrapolation factor considering the range specified in range.start() and range.end(), and the interpolated value specified in weight.

The returned value will be between 0.0 and 1.0 if weight is between range.start() and range.end() (inclusive).

If weight is located outside this range, then an extrapolation factor will be returned (return value lower than 0.0 or greater than 1.0).

Examples:

use gdnative::globalscope::*;

assert_eq!(inverse_lerp(20.0..=30.0, 27.5), 0.75);