Function gdnative::globalscope::lerp

pub fn lerp(range: RangeInclusive<f32>, weight: f32) -> f32

Linearly interpolates between two values, by the factor defined in weight.

To perform interpolation, weight should be between 0.0 and 1.0 (inclusive). However, values outside this range are allowed and can be used to perform extrapolation.

use gdnative::globalscope::*;
assert_eq!(lerp(0.0..=4.0, 0.75), 3.0);