Function gdnative::globalscope::nearest_po2

pub fn nearest_po2(value: i32) -> u32

Returns the nearest equal or larger power of 2 for an integer value.

In other words, returns the smallest value a where a = pow(2, n) such that value <= a for some non-negative integer n.

This behaves like u32::next_power_of_two() for value >= 1.

Warning: This function returns 0 rather than 1 for non-positive values of value (in reality, 1 is the smallest integer power of 2).

Examples:

use gdnative::globalscope::*;

assert_eq!(nearest_po2(3), 4);
assert_eq!(nearest_po2(4), 4);
assert_eq!(nearest_po2(5), 8);
assert_eq!(nearest_po2(0), 0);
assert_eq!(nearest_po2(-1), 0);