Function gdnative::globalscope::smoothstep

pub fn smoothstep(from: f32, to: f32, s: f32) -> f32

Smooth (Hermite) interpolation.

Returns the result of smoothly interpolating the value of s between 0 and 1, based on where s lies with respect to the edges from and to.

The return value is 0 if s <= from, and 1 if s >= to.

If s lies between from and to, the returned value follows an S-shaped curve that maps s between 0 and 1.
This S-shaped curve is the cubic Hermite interpolator, given by f(y) = 3*y^2 - 2*y^3 where y = (x-from) / (to-from).

Compared to ease() with a curve value of -1.6521, smoothstep() returns the smoothest possible curve with no sudden changes in the derivative.

If you need to perform more advanced transitions, use Tween or AnimationPlayer.

Examples:

use gdnative::globalscope::*;

assert_eq!(smoothstep(0.0, 2.0, -5.0), 0.0);
assert_eq!(smoothstep(0.0, 2.0, 0.5), 0.15625);
assert_eq!(smoothstep(0.0, 2.0, 1.0), 0.5);
assert_eq!(smoothstep(0.0, 2.0, 2.0), 1.0);