Macro gdnative::godot_error

macro_rules! godot_error {
    ($($args:tt)*) => { ... };
}

Print an error using the engine’s logging system (visible in the editor).

Guarantees

It’s guaranteed that the expansion result of this macro may only panic if:

Any of the arguments for the message panicked in fmt.

The formatted message contains the NUL byte (\0) anywhere.