Module gdnative::nativescript::init::method

Method registration

Structs

ArgBuilder

Builder for providing additional argument information for error reporting.

ArgumentError

Error during argument parsing.

MethodBuilder

Builder type used to register a method on a NativeClass.

ScriptMethodDeprecated

ScriptMethodAttributesDeprecated

StaticArgs

Adapter for methods whose arguments are statically determined. If the arguments would fail to type check, the method will print the errors to Godot’s debug console and return null.

Varargs

Safe interface to a list of borrowed method arguments with a convenient API for common operations with them. Can also be used as an iterator.

Enums

RpcMode

Traits

FromVarargs

Trait for structures that can be parsed from Varargs.

Method

Safe low-level trait for stateful, variadic methods that can be called on a native script type.

StaticArgsMethod

Trait for methods whose argument lists are known at compile time. Not to be confused with a “static method”.

Type Definitions

ScriptMethodFnDeprecated