Module gdnative::nativescript::init::property

Property registration.

Modules

hint

Strongly typed property hints.

Structs

ArrayHint

Array hints optionally with an element hint.

EnumHint

Hints that an integer, float or string property is an enumerated value to pick in a list.

ExpEasingHint

Hints that a float property should be edited via an exponential easing function.

ExportInfo

Metadata about the exported property.

PropertyBuilder

Builder type used to register a property on a NativeClass.

RangeHint

Hints that an integer or float property should be within an inclusive range.

Usage

Enums

ColorHint

Possible hints for Color.

FloatHint

Possible hints for floats.

IntHint

Possible hints for integers.

StringHint

Possible hints for strings.

Traits

Export

Trait for exportable types.