Module gdnative::object::memory

Marker types to express the memory management method of Godot types.

Enums

ManuallyManaged

Marker that indicates that a type is manually managed.

RefCounted

Marker that indicates that a type is reference-counted.

Traits

Memory

Trait to parameterize over the memory management markers ManuallyManaged and RefCounted.