Module gdnative::object::ownership

Typestates to express ownership and thread safety of Godot types.

Structs

Shared

Marker that indicates that a value currently might be shared in the same or over multiple threads.

ThreadLocal

Marker that indicates that a value can currently only be shared in the same thread.

Unique

Marker that indicates that a value currently only has a single unique reference.

Traits

LocalThreadOwnership

Trait to parametrize over the ownership markers that are local to the current thread: Unique and ThreadLocal.

NonUniqueOwnership

Trait to parametrize over the ownership markers that are not unique: Shared and ThreadLocal.

Ownership

Trait to parametrize over the ownership markers Unique, Shared and ThreadLocal.