Struct gdnative::object::TRef

pub struct TRef<'a, T, Own = Shared>
where
    T: GodotObject,
    Own: Ownership,
{ /* private fields */ }

A temporary safe pointer to Godot objects that tracks thread access status. TRef can be coerced into bare references with Deref.

See the type-level documentation on Ref for detailed documentation on the reference system of godot-rust.

Using as method arguments or return values

TRef<T, Shared> can be passed into methods.

Using as owner arguments in NativeScript methods

Itโs possible to use TRef as the owner argument in NativeScript methods. This can make passing owner to methods easier.

Implementations

impl<'a, T, Own> TRef<'a, T, Own>
where T: GodotObject, Own: Ownership,

pub fn as_ref(self) -> &'a T

Returns the underlying reference without thread access.

pub fn cast<U>(self) -> Option<TRef<'a, U, Own>>
where U: GodotObject + SubClass<T>,

Performs a dynamic reference cast to target type, keeping the thread access info.

pub fn upcast<U>(&self) -> TRef<'a, U, Own>
where U: GodotObject, T: SubClass<U>,

Performs a static reference upcast to a supertype that is guaranteed to be valid, keeping the thread access info.

This is guaranteed to be a no-op at runtime.

pub fn cast_instance<C>(self) -> Option<TInstance<'a, C, Own>>
where C: NativeClass<Base = T>,

Convenience method to downcast to TInstance where self is the base object.

impl<'a, Kind, T, Own> TRef<'a, T, Own>
where Kind: Memory, T: GodotObject<Memory = Kind>, Own: NonUniqueOwnership,

pub fn claim(self) -> Ref<T, Own>

Persists this reference into a persistent Ref with the same thread access.

This is only available for non-Unique accesses.

impl<'a, T> TRef<'a, T>
where T: GodotObject,

pub unsafe fn try_from_instance_id(id: i64) -> Option<TRef<'a, T>>

Recovers a instance ID previously returned by Object::get_instance_id if the object is still alive.

Safety

During the entirety of 'a, the thread from which try_from_instance_id is called must have exclusive access to the underlying object, if it is still alive.

pub unsafe fn from_instance_id(id: i64) -> TRef<'a, T>

Recovers a instance ID previously returned by Object::get_instance_id if the object is still alive, and panics otherwise. This does NOT guarantee that the resulting reference is safe to use.

Panics

Panics if the given id refers to a destroyed object. For a non-panicking version, see try_from_instance_id.

Safety

During the entirety of 'a, the thread from which try_from_instance_id is called must have exclusive access to the underlying object, if it is still alive.