Attribute Macro gdnative::prelude::methods

#[methods]

Collects method signatures of all functions in a NativeClass that have the #[export] attribute and registers them with Godot.

For example, in the following class

use gdnative::prelude::*;

#[derive(NativeClass)]
#[inherit(Reference)]
#[no_constructor]
struct Foo {}

#[methods]
impl Foo {
    #[export]
    fn foo(&self, _owner: &Reference, bar: i64) -> i64 {
        bar
    }
}

Will expand to

use gdnative::prelude::*;
struct Foo{}
impl NativeClass for Foo {
    type Base = gdnative::api::Reference;
    type UserData = gdnative::nativescript::user_data::LocalCellData<Self>;
    fn class_name() -> &'static str {
        "Foo"
    }
}
impl gdnative::nativescript::NativeClassMethods for Foo {
    fn register(builder: &ClassBuilder<Self>) {
        use gdnative::nativescript::export::*;
        builder.build_method("foo", gdnative::macros::godot_wrap_method!(Foo, fn foo(&self, _owner: &Reference, bar: i64) -> i64))
            .with_rpc_mode(RpcMode::Disabled)
            .done_stateless();
    }
}
impl Foo {
    fn foo(&self, _owner: &Reference, bar: i64) -> i64 {
        bar
    }
}

Important: Only one impl block per struct may be attributed with #[methods].

For more context, please refer to gdnative::derive::NativeClass.