Module gdnative::prelude::core_types

Types that represent core-datatypes of Godot.

Modules

dictionary

error

vector3

Structs

Aabb

Axis-aligned bounding box.

Aligned

An pool array access that is (assumed to be) aligned.

Basis

A 3x3 matrix.

Color

RGBA color with 32 bits floating point components.

Dictionary

A reference-counted Dictionary of Variant key-value pairs.

GodotChar

Type representing a character in Godot’s native encoding. Can be converted to and from char. Depending on the platform, this might not always be able to represent a full code point.

GodotString

Godot’s reference-counted string type.

IntoIter

InvalidOp

Error indicating that an operator result is invalid.

Iter

MaybeNot

Wrapper type around a FromVariant result that may not be a success

MaybeUnaligned

An pool array access that may be unaligned.

NodePath

A reference-counted relative or absolute path in a scene tree, for use with Node.get_node() and similar functions. It can reference a node, a resource within a node, or a property of a node or resource.

Owned

An pool array write access with an owned aligned copy. The data is written back when this is dropped.

Plane

Plane in hessian form.

Quat

Rect2

Rid

The RID type is used to access the unique integer ID of a resource. They are opaque, so they do not grant access to the associated resource by themselves.

StringName

Transform

3D Transformation (3x4 matrix) Using basis + origin representation.

Transform2D

TypedArray

A reference-counted CoW typed vector using Godot’s pool allocator, generic over possible element types.

Variant

A Variant can represent many of godot’s core types.

VariantArray

A reference-counted Variant vector. Godot’s generic array data type. Negative indices can be used to count from the right.

Vector2

Vector3

Enums

Axis

CallError

FromVariantError

Error type returned by FromVariant::from_variant.

GodotCharError

Error indicating that a GodotChar cannot be converted to a char.

GodotError

Error codes used in various Godot APIs.

VariantDispatch

Rust enum associating each primitive variant type to its value.

VariantEnumRepr

VariantOperator

Godot variant operator kind.

VariantStructRepr

VariantType

Traits

Element

Trait for element types that can be contained in TypedArray. This trait is sealed and has no public interface.

FromVariant

Types that can be converted from a Variant.

Guard

Trait for array access guards

OwnedToVariant

Types that can only be safely converted to a Variant as owned values. Such types cannot implement ToVariant in general, but can still be passed to API methods as arguments, or used as return values. Notably, this includes Unique arrays, dictionaries, and references to Godot objects and instances.

ToVariant

Types that can be converted to a Variant.

ToVariantEq

Trait for types whose ToVariant implementations preserve equivalence.

WritePtr

Marker trait for write access guards

Type Definitions

ByteArray

A reference-counted vector of u8 that uses Godot’s pool allocator.

ColorArray

A reference-counted vector of Color that uses Godot’s pool allocator.

Float32Array

A reference-counted vector of f32 that uses Godot’s pool allocator.

Int32Array

A reference-counted vector of i32 that uses Godot’s pool allocator.

StringArray

A reference-counted vector of GodotString that uses Godot’s pool allocator.

Vector2Array

A reference-counted vector of Vector2 that uses Godot’s pool allocator.

Vector3Array

A reference-counted vector of Vector3 that uses Godot’s pool allocator.