Enum gdnative::prelude::core_types::VariantDispatch

#[repr(u32)]
pub enum VariantDispatch {

    Nil,
    Bool(bool),
    I64(i64),
    F64(f64),
    GodotString(GodotString),
    Vector2(Vector2),
    Rect2(Rect2),
    Vector3(Vector3),
    Transform2D(Transform2D),
    Plane(Plane),
    Quat(Quat),
    Aabb(Aabb),
    Basis(Basis),
    Transform(Transform),
    Color(Color),
    NodePath(NodePath),
    Rid(Rid),
    Object(Variant),
    Dictionary(Dictionary<Shared>),
    VariantArray(VariantArray<Shared>),
    ByteArray(TypedArray<u8>),
    Int32Array(TypedArray<i32>),
    Float32Array(TypedArray<f32>),
    StringArray(TypedArray<GodotString>),
    Vector2Array(TypedArray<Vector2>),
    Vector3Array(TypedArray<Vector3>),
    ColorArray(TypedArray<Color>),

}

Rust enum associating each primitive variant type to its value.

For Variants containing objects, the original Variant is returned unchanged, due to the limitations of statically-determined memory management.

Variants

Nil

Bool(bool)

Tuple Fields of Bool

0: bool

I64(i64)

Tuple Fields of I64

0: i64

F64(f64)

Tuple Fields of F64

0: f64

GodotString(GodotString)

Tuple Fields of GodotString

0: GodotString

Vector2(Vector2)

Tuple Fields of Vector2

0: Vector2

Rect2(Rect2)

Tuple Fields of Rect2

0: Rect2

Vector3(Vector3)

Tuple Fields of Vector3

0: Vector3

Transform2D(Transform2D)

Tuple Fields of Transform2D

0: Transform2D

Plane(Plane)

Tuple Fields of Plane

0: Plane

Quat(Quat)

Tuple Fields of Quat

0: Quat

Aabb(Aabb)

Tuple Fields of Aabb

0: Aabb

Basis(Basis)

Tuple Fields of Basis

0: Basis

Transform(Transform)

Tuple Fields of Transform

0: Transform

Color(Color)

Tuple Fields of Color

0: Color

NodePath(NodePath)

Tuple Fields of NodePath

0: NodePath

Rid(Rid)

Tuple Fields of Rid

0: Rid

Object(Variant)

Tuple Fields of Object

0: Variant

Dictionary(Dictionary<Shared>)

Tuple Fields of Dictionary

0: Dictionary<Shared>

VariantArray(VariantArray<Shared>)

Tuple Fields of VariantArray

0: VariantArray<Shared>

ByteArray(TypedArray<u8>)

Tuple Fields of ByteArray

0: TypedArray<u8>

Int32Array(TypedArray<i32>)

Tuple Fields of Int32Array

0: TypedArray<i32>

Float32Array(TypedArray<f32>)

Tuple Fields of Float32Array

0: TypedArray<f32>

StringArray(TypedArray<GodotString>)

Tuple Fields of StringArray

0: TypedArray<GodotString>

Vector2Array(TypedArray<Vector2>)

Tuple Fields of Vector2Array

0: TypedArray<Vector2>

Vector3Array(TypedArray<Vector3>)

Tuple Fields of Vector3Array

0: TypedArray<Vector3>

ColorArray(TypedArray<Color>)

Tuple Fields of ColorArray

0: TypedArray<Color>