Struct gdnative::prelude::core_types::Color

#[repr(C)]
pub struct Color {
    pub r: f32,
    pub g: f32,
    pub b: f32,
    pub a: f32,
}

RGBA color with 32 bits floating point components.

Fields
r: f32g: f32b: f32a: f32
Implementations

impl Color

pub fn rgba(r: f32, g: f32, b: f32, a: f32) -> Color

👎 Deprecated

pub fn rgb(r: f32, g: f32, b: f32) -> Color

👎 Deprecated

pub fn from_rgba(r: f32, g: f32, b: f32, a: f32) -> Color

pub fn from_rgb(r: f32, g: f32, b: f32) -> Color

pub fn from_hsv(h: f32, s: f32, v: f32) -> Color

pub fn from_hsva(h: f32, s: f32, v: f32, a: f32) -> Color

pub fn h(&self) -> f32

pub fn s(&self) -> f32

pub fn v(&self) -> f32

pub fn lerp(&self, other: Color, weight: f32) -> Color

pub fn blend(&self, other: &Color) -> Color

pub fn contrasted(&self) -> Color

pub fn darkened(&self, amount: f32) -> Color

pub fn gray(&self) -> f32

pub fn inverted(&self) -> Color

pub fn to_html(self, with_alpha: bool) -> GodotString

pub fn to_abgr32(self) -> u32

Returns the reverse of the RGBA32 byte representation for this color where each byte represents a component of the ABGR profile. This is the byte information used when storing this color as a part of a texture.

Endianness

On big endian architecture this is stored in ABGR byte order On little endian machines this is stored in RGBA byte order

Example

0x00FF7FFF would be the equivalent to Color::from_rgba(1.0, 0.5, 1.0, 0.0)

pub fn to_abgr64(self) -> u64

Returns the reverse of the RGBA64 byte representation for this color where each word represents represents a component of the ABGR profile. This is the byte information used when storing this color as a part of a texture.

Endianness

On big endian architecture this is stored in ABGR word order On little endian machines this is stored in RGBA word order

Example

0x0000FFFF7FFFFFFF would be the equivalent to Color::from_rgba(0.0, 1.0, 0.5, 1.0)

pub fn to_argb32(self) -> u32

Returns the ARGB32 format representation representation for this color where each byte represents a component of the ARGB profile. This is the byte information used when storing this color as a part of a texture.

Endianness

On big endian architecture this is stored in the order ARGB byte order On little endian machines this is stored in the order BGRA byte order 0x0000FFFF7FFFFFFF would be the equivalent to Color::from_rgba(1.0, 0.5, 1.0, 0.0)

pub fn to_argb64(self) -> u64

Returns the ARGB64 format representation for this color where each word represents a component of the ARGB profile. This is the byte information used when storing this color as a part of a texture.

Endianness

On big endian architecture this is stored in the order ARGB word order On little endian machines this is stored in the order BGRA word order

Example

0x0000FFFF7FFFFFFF would be the equivalent to Color::from_rgba(1.0, 0.5, 1.0, 0.0)

pub fn to_rgba32(self) -> u32

Returns the OpenGL Texture format byte representation for this color where each byte represents a component of the RGBA profile. This is the byte information used when storing this color as a part of a texture.

Endianness

On big endian architecture this is stored in RGBA byte order On little endian machines this is stored in ABGR byte order

Example

0x00FF7FFF would be the equivalent to Color::from_rgba(0.0, 1.0, 0.5, 1.0)

pub fn to_rgba64(self) -> u64

Returns the OpenGL Texture format byte representation for this color where each byte represents a component of the RGBA profile. This is the byte information used when storing this color as a part of a texture.

Endianness

On big endian architecture this is stored in RGBA word order On little endian machines this is stored in ABGR word order

Example

0x0000FFFF7FFFFFFF would be the equivalent to Color::from_rgba(0.0, 1.0, 0.5, 1.0)