Module gdnative::prelude::nativescript

Types and functions related to the NativeScript extension of GDNative.

Modules

class

init

Types and functionalities to declare and initialize gdnative classes.

profiling

Interface to Godot’s built-in profiler.

type_tag

user_data

Customizable user-data wrappers.

Structs

ClassBuilder

ExportInfo

Metadata about the exported property.

InitHandle

A handle that can register new classes to the engine during initialization.

Instance

A persistent reference to a GodotObject with a rust NativeClass attached.

MethodBuilder

Builder type used to register a method on a NativeClass.

PropertyBuilder

Builder type used to register a property on a NativeClass.

PropertyUsage

RefInstance

A reference to a GodotObject with a rust NativeClass attached that is assumed safe during a certain lifetime.

ScriptMethodDeprecated

ScriptMethodAttributesDeprecated

Signal

SignalArgument

Varargs

Safe interface to a list of borrowed method arguments with a convenient API for common operations with them. Can also be used as an iterator.

Enums

RpcMode

Traits

Export

Trait for exportable types.

Map

Trait for wrappers that can be mapped immutably.

MapMut

Trait for wrappers that can be mapped mutably.

MapOwned

Trait for wrappers that can be mapped once.

Method

Safe low-level trait for stateful, variadic methods that can be called on a native script type.

NativeClass

Trait used for describing and initializing a Godot script class.

NativeClassMethods

Trait used to provide information of Godot-exposed methods of a script class.

OwnerArg

Trait for types that can be used as the owner arguments of exported methods. This trait is sealed and has no public interface.

UserData

Trait for customizable user-data wrappers.

Type Definitions

ScriptMethodFnDeprecated