Module gdnative::prelude::nativescript::export

Low-level API to register and export GDNative classes, methods and properties.

Init and exit hooks

Three endpoints are automatically invoked by the engine during startup and shutdown:

godot_gdnative_init,

godot_nativescript_init,

godot_gdnative_terminate,

All three must be present. To quickly define all three endpoints using the default names, use godot_init.

Registering script classes

InitHandle is the registry of all your exported symbols. To register script classes, call InitHandle::add_class or InitHandle::add_tool_class in your godot_nativescript_init or godot_init callback:

use gdnative::prelude::*;

fn init(handle: InitHandle) {
    handle.add_class::<HelloWorld>();
}

godot_init!(init);

For full examples, see examples in the godot-rust repository.

Modules

method

Method registration

property

Property registration.

Structs

ClassBuilder

ExportInfo

Metadata about the exported property.

InitHandle

A handle that can register new classes to the engine during initialization.

MethodBuilder

Builder type used to register a method on a NativeClass.

PropertyBuilder

Builder type used to register a property on a NativeClass.

PropertyUsage

ScriptMethodDeprecated

ScriptMethodAttributesDeprecated

Signal

SignalArgument

Varargs

Safe interface to a list of borrowed method arguments with a convenient API for common operations with them. Can also be used as an iterator.

Enums

RpcMode

Traits

Export

Trait for exportable types.

Method

Safe low-level trait for stateful, variadic methods that can be called on a native script type.

Type Definitions

ScriptMethodFnDeprecated