Trait gdnative::prelude::nativescript::export::Export

pub trait Export: ToVariant {
    type Hint;
    fn export_info(hint: Option<Self::Hint>) -> ExportInfo;
}

Trait for exportable types.

Associated Types

type Hint

A type-specific hint type that is valid for the type being exported.

If this type shows up as NoHint, a private, uninhabitable type indicating that there are no hints available for the time being, users must use None for properties of this type. This ensures that it will not be a breaking change to add a hint for the type later, since it supports no operations and cannot be named directly in user code.

Required methods

fn export_info(hint: Option<Self::Hint>) -> ExportInfo

Returns ExportInfo given an optional typed hint.