Trait gdnative::prelude::nativescript::OwnerArg

pub trait OwnerArg<'a, T, Access>: Sealed where
    Access: 'static + ThreadAccess,
    T: GodotObject, { }

Trait for types that can be used as the owner arguments of exported methods. This trait is sealed and has no public interface.

Safety

Whenever a NativeScript methods is called, it’s assumed that the owner is safe to use. When calling a method that may call non-thread-safe methods on its owner from non-Rust code, the official thread-safety guidelines must be followed to prevent undefined behavior.