Struct gdnative::prelude::CanvasItem

pub struct CanvasItem { /* fields omitted */ }

core class CanvasItem inherits Node (unsafe).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Class hierarchy

CanvasItem inherits methods from:

Node

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl CanvasItem

Constants

pub const BLEND_MODE_MIX: i64

pub const BLEND_MODE_ADD: i64

pub const BLEND_MODE_SUB: i64

pub const BLEND_MODE_MUL: i64

pub const BLEND_MODE_PREMULT_ALPHA: i64

pub const BLEND_MODE_DISABLED: i64

pub const NOTIFICATION_DRAW: i64

pub const NOTIFICATION_VISIBILITY_CHANGED: i64

pub const NOTIFICATION_ENTER_CANVAS: i64

pub const NOTIFICATION_EXIT_CANVAS: i64

pub const NOTIFICATION_TRANSFORM_CHANGED: i64

impl CanvasItem

pub fn draw_arc(
&self,
center: Vector2,
radius: f64,
start_angle: f64,
end_angle: f64,
point_count: i64,
color: Color,
width: f64,
antialiased: bool
)

Draws an arc between the given angles. The larger the value of point_count, the smoother the curve.

Default Arguments

width - 1.0

antialiased - false

pub fn draw_char(
&self,
font: impl AsArg<Font>,
position: Vector2,
char: impl Into<GodotString>,
next: impl Into<GodotString>,
modulate: Color
) -> f64

Draws a string character using a custom font. Returns the advance, depending on the character width and kerning with an optional next character.

Default Arguments

modulate - Color( 1, 1, 1, 1 )

pub fn draw_circle(&self, position: Vector2, radius: f64, color: Color)

Draws a colored circle.

pub fn draw_colored_polygon(
&self,
points: TypedArray<Vector2>,
color: Color,
uvs: TypedArray<Vector2>,
texture: impl AsArg<Texture>,
normal_map: impl AsArg<Texture>,
antialiased: bool
)

Draws a colored polygon of any amount of points, convex or concave.

Default Arguments

uvs - PoolVector2Array( )

texture - null

normal_map - null

antialiased - false

pub fn draw_line(
&self,
from: Vector2,
to: Vector2,
color: Color,
width: f64,
antialiased: bool
)

Draws a line from a 2D point to another, with a given color and width. It can be optionally antialiased.

Default Arguments

width - 1.0

antialiased - false

pub fn draw_mesh(
&self,
mesh: impl AsArg<Mesh>,
texture: impl AsArg<Texture>,
normal_map: impl AsArg<Texture>,
transform: Transform2D,
modulate: Color
)

Draws a Mesh in 2D, using the provided texture. See MeshInstance2D for related documentation.

Default Arguments

normal_map - null

transform - Transform2D( 1, 0, 0, 1, 0, 0 )

modulate - Color( 1, 1, 1, 1 )

pub fn draw_multiline(
&self,
points: TypedArray<Vector2>,
color: Color,
width: f64,
antialiased: bool
)

Draws multiple, parallel lines with a uniform color. Note: width and antialiased are currently not implemented and have no effect.

Default Arguments

width - 1.0

antialiased - false

pub fn draw_multiline_colors(
&self,
points: TypedArray<Vector2>,
colors: TypedArray<Color>,
width: f64,
antialiased: bool
)

Draws multiple, parallel lines with a uniform width and segment-by-segment coloring. Colors assigned to line segments match by index between points and colors. Note: width and antialiased are currently not implemented and have no effect.

Default Arguments

width - 1.0

antialiased - false

pub fn draw_multimesh(
&self,
multimesh: impl AsArg<MultiMesh>,
texture: impl AsArg<Texture>,
normal_map: impl AsArg<Texture>
)

Draws a MultiMesh in 2D with the provided texture. See MultiMeshInstance2D for related documentation.

Default Arguments

normal_map - null

pub fn draw_polygon(
&self,
points: TypedArray<Vector2>,
colors: TypedArray<Color>,
uvs: TypedArray<Vector2>,
texture: impl AsArg<Texture>,
normal_map: impl AsArg<Texture>,
antialiased: bool
)

Draws a polygon of any amount of points, convex or concave.

Default Arguments

uvs - PoolVector2Array( )

texture - null

normal_map - null

antialiased - false

pub fn draw_polyline(
&self,
points: TypedArray<Vector2>,
color: Color,
width: f64,
antialiased: bool
)

Draws interconnected line segments with a uniform color and width and optional antialiasing.

Default Arguments

width - 1.0

antialiased - false

pub fn draw_polyline_colors(
&self,
points: TypedArray<Vector2>,
colors: TypedArray<Color>,
width: f64,
antialiased: bool
)

Draws interconnected line segments with a uniform width, segment-by-segment coloring, and optional antialiasing. Colors assigned to line segments match by index between points and colors.

Default Arguments

width - 1.0

antialiased - false

pub fn draw_primitive(
&self,
points: TypedArray<Vector2>,
colors: TypedArray<Color>,
uvs: TypedArray<Vector2>,
texture: impl AsArg<Texture>,
width: f64,
normal_map: impl AsArg<Texture>
)

Draws a custom primitive. 1 point for a point, 2 points for a line, 3 points for a triangle and 4 points for a quad.

Default Arguments

texture - null

width - 1.0

normal_map - null

pub fn draw_rect(
&self,
rect: Rect2,
color: Color,
filled: bool,
width: f64,
antialiased: bool
)

Draws a rectangle. If filled is true, the rectangle will be filled with the color specified. If filled is false, the rectangle will be drawn as a stroke with the color and width specified. If antialiased is true, the lines will be antialiased. Note: width and antialiased are only effective if filled is false.

Default Arguments

filled - true

width - 1.0

antialiased - false

pub fn draw_set_transform(
&self,
position: Vector2,
rotation: f64,
scale: Vector2
)

Sets a custom transform for drawing via components. Anything drawn afterwards will be transformed by this.

pub fn draw_set_transform_matrix(&self, xform: Transform2D)

Sets a custom transform for drawing via matrix. Anything drawn afterwards will be transformed by this.

pub fn draw_string(
&self,
font: impl AsArg<Font>,
position: Vector2,
text: impl Into<GodotString>,
modulate: Color,
clip_w: i64
)

Sample code is GDScript unless otherwise noted.

Draws text using the specified font at the position (bottom-left corner using the baseline of the font). The text will have its color multiplied by modulate. If clip_w is greater than or equal to 0, the text will be clipped if it exceeds the specified width. Example using the default project font:

var default_font = Control.new().get_font("font")
draw_string(default_font, Vector2(64, 64), "Hello world")

See also Font.draw.

Default Arguments

modulate - Color( 1, 1, 1, 1 )

clip_w - -1

pub fn draw_style_box(&self, style_box: impl AsArg<StyleBox>, rect: Rect2)

Draws a styled rectangle.

pub fn draw_texture(
&self,
texture: impl AsArg<Texture>,
position: Vector2,
modulate: Color,
normal_map: impl AsArg<Texture>
)

Draws a texture at a given position.

Default Arguments

modulate - Color( 1, 1, 1, 1 )

normal_map - null

pub fn draw_texture_rect(
&self,
texture: impl AsArg<Texture>,
rect: Rect2,
tile: bool,
modulate: Color,
transpose: bool,
normal_map: impl AsArg<Texture>
)

Draws a textured rectangle at a given position, optionally modulated by a color. If transpose is true, the texture will have its X and Y coordinates swapped.

Default Arguments

modulate - Color( 1, 1, 1, 1 )

transpose - false

normal_map - null

pub fn draw_texture_rect_region(
&self,
texture: impl AsArg<Texture>,
rect: Rect2,
src_rect: Rect2,
modulate: Color,
transpose: bool,
normal_map: impl AsArg<Texture>,
clip_uv: bool
)

Draws a textured rectangle region at a given position, optionally modulated by a color. If transpose is true, the texture will have its X and Y coordinates swapped.

Default Arguments

modulate - Color( 1, 1, 1, 1 )

transpose - false

normal_map - null

clip_uv - true

pub fn force_update_transform(&self)

Forces the transform to update. Transform changes in physics are not instant for performance reasons. Transforms are accumulated and then set. Use this if you need an up-to-date transform when doing physics operations.

pub fn get_canvas(&self) -> Rid

Returns the RID of the World2D canvas where this item is in.

pub fn get_canvas_item(&self) -> Rid

Returns the canvas item RID used by VisualServer for this item.

pub fn get_canvas_transform(&self) -> Transform2D

Returns the transform matrix of this item’s canvas.

pub fn get_global_mouse_position(&self) -> Vector2

Returns the global position of the mouse.

pub fn get_global_transform(&self) -> Transform2D

Returns the global transform matrix of this item.

pub fn get_global_transform_with_canvas(&self) -> Transform2D

Returns the global transform matrix of this item in relation to the canvas.

pub fn light_mask(&self) -> i64

The rendering layers in which this CanvasItem responds to Light2D nodes.

pub fn get_local_mouse_position(&self) -> Vector2

Returns the mouse position relative to this item’s position.

pub fn material(&self) -> Option<Ref<Material, Shared>>

The material applied to textures on this CanvasItem.

pub fn modulate(&self) -> Color

The color applied to textures on this CanvasItem.

pub fn self_modulate(&self) -> Color

The color applied to textures on this CanvasItem. This is not inherited by children CanvasItems.

pub fn get_transform(&self) -> Transform2D

Returns the transform matrix of this item.

pub fn use_parent_material(&self) -> bool

If true, the parent CanvasItem’s material property is used as this one’s material.

pub fn get_viewport_rect(&self) -> Rect2

Returns the viewport’s boundaries as a Rect2.

pub fn get_viewport_transform(&self) -> Transform2D

Returns this item’s transform in relation to the viewport.

pub fn get_world_2d(&self) -> Option<Ref<World2D, Shared>>

Returns the World2D where this item is in.

pub fn hide(&self)

Hide the CanvasItem if it’s currently visible.

pub fn is_draw_behind_parent_enabled(&self) -> bool

If true, the object draws behind its parent.

pub fn is_local_transform_notification_enabled(&self) -> bool

Returns true if local transform notifications are communicated to children.

pub fn is_set_as_toplevel(&self) -> bool

Returns true if the node is set as top-level. See set_as_toplevel.

pub fn is_transform_notification_enabled(&self) -> bool

Returns true if global transform notifications are communicated to children.

pub fn is_visible(&self) -> bool

If true, this CanvasItem is drawn. The node is only visible if all of its antecedents are visible as well (in other words, is_visible_in_tree must return true). Note: For controls that inherit Popup, the correct way to make them visible is to call one of the multiple popup*() functions instead.

pub fn is_visible_in_tree(&self) -> bool

Returns true if the node is present in the SceneTree, its [visible][Self::visible] property is true and all its antecedents are also visible. If any antecedent is hidden, this node will not be visible in the scene tree.

pub fn make_canvas_position_local(&self, screen_point: Vector2) -> Vector2

Assigns screen_point as this node’s new local transform.

pub fn make_input_local(
&self,
event: impl AsArg<InputEvent>
) -> Option<Ref<InputEvent, Shared>>

Transformations issued by event’s inputs are applied in local space instead of global space.

pub fn set_as_toplevel(&self, enable: bool)

If enable is true, the node won’t inherit its transform from parent canvas items.

pub fn set_draw_behind_parent(&self, enable: bool)

If true, the object draws behind its parent.

pub fn set_light_mask(&self, light_mask: i64)

The rendering layers in which this CanvasItem responds to Light2D nodes.

pub fn set_material(&self, material: impl AsArg<Material>)

The material applied to textures on this CanvasItem.

pub fn set_modulate(&self, modulate: Color)

The color applied to textures on this CanvasItem.

pub fn set_notify_local_transform(&self, enable: bool)

If enable is true, children will be updated with local transform data.

pub fn set_notify_transform(&self, enable: bool)

If enable is true, children will be updated with global transform data.

pub fn set_self_modulate(&self, self_modulate: Color)

The color applied to textures on this CanvasItem. This is not inherited by children CanvasItems.

pub fn set_use_parent_material(&self, enable: bool)

If true, the parent CanvasItem’s material property is used as this one’s material.

pub fn set_visible(&self, visible: bool)

If true, this CanvasItem is drawn. The node is only visible if all of its antecedents are visible as well (in other words, is_visible_in_tree must return true). Note: For controls that inherit Popup, the correct way to make them visible is to call one of the multiple popup*() functions instead.

pub fn show(&self)

Show the CanvasItem if it’s currently hidden. For controls that inherit Popup, the correct way to make them visible is to call one of the multiple popup*() functions instead.

pub fn update(&self)

Queue the CanvasItem for update. NOTIFICATION_DRAW will be called on idle time to request redraw.