Struct gdnative::prelude::InputEvent

pub struct InputEvent { /* fields omitted */ }

core class InputEvent inherits Resource (reference counted).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

InputEvent inherits methods from:

Resource

Reference

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl InputEvent

pub fn accumulate(&self, with_event: impl AsArg<InputEvent>) -> bool

Returns true if the given input event and this input event can be added together (only for events of type InputEventMouseMotion). The given input event’s position, global position and speed will be copied. The resulting relative is a sum of both events. Both events’ modifiers have to be identical.

pub fn as_text(&self) -> GodotString

Returns a String representation of the event.

pub fn get_action_strength(&self, action: impl Into<GodotString>) -> f64

Returns a value between 0.0 and 1.0 depending on the given actions’ state. Useful for getting the value of events of type InputEventJoypadMotion.

pub fn device(&self) -> i64

The event’s device ID. Note: This device ID will always be -1 for emulated mouse input from a touchscreen. This can be used to distinguish emulated mouse input from physical mouse input.

pub fn is_action(&self, action: impl Into<GodotString>) -> bool

Returns true if this input event matches a pre-defined action of any type.

pub fn is_action_pressed(
&self,
action: impl Into<GodotString>,
allow_echo: bool
) -> bool

Returns true if the given action is being pressed (and is not an echo event for InputEventKey events, unless allow_echo is true). Not relevant for events of type InputEventMouseMotion or InputEventScreenDrag.

Default Arguments

allow_echo - false

pub fn is_action_released(&self, action: impl Into<GodotString>) -> bool

Returns true if the given action is released (i.e. not pressed). Not relevant for events of type InputEventMouseMotion or InputEventScreenDrag.

pub fn is_action_type(&self) -> bool

Returns true if this input event’s type is one that can be assigned to an input action.

pub fn is_echo(&self) -> bool

Returns true if this input event is an echo event (only for events of type InputEventKey).

pub fn is_pressed(&self) -> bool

Returns true if this input event is pressed. Not relevant for events of type InputEventMouseMotion or InputEventScreenDrag.

pub fn set_device(&self, device: i64)

The event’s device ID. Note: This device ID will always be -1 for emulated mouse input from a touchscreen. This can be used to distinguish emulated mouse input from physical mouse input.

pub fn shortcut_match(&self, event: impl AsArg<InputEvent>) -> bool

Returns true if the given input event is checking for the same key (InputEventKey), button (InputEventJoypadButton) or action (InputEventAction).

pub fn xformed_by(
&self,
xform: Transform2D,
local_ofs: Vector2
) -> Option<Ref<InputEvent, Shared>>

Returns a copy of the given input event which has been offset by local_ofs and transformed by xform. Relevant for events of type InputEventMouseButton, InputEventMouseMotion, InputEventScreenTouch, InputEventScreenDrag, InputEventMagnifyGesture and InputEventPanGesture.

Default Arguments

local_ofs - Vector2( 0, 0 )