Struct gdnative::prelude::KinematicBody

pub struct KinematicBody { /* fields omitted */ }

core class KinematicBody inherits PhysicsBody (unsafe).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non reference counted objects such as the ones of this type are usually owned by the engine.

KinematicBody is a reference-only type. Persistent references can only exist in the unsafe Ref<KinematicBody> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

KinematicBody inherits methods from:

PhysicsBody

CollisionObject

Spatial

Node

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl KinematicBody

pub fn new() -> Ref<KinematicBody, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn axis_lock(&self, axis: i64) -> bool

Lock the body’s Z axis movement.

pub fn get_floor_normal(&self) -> Vector3

Returns the surface normal of the floor at the last collision point. Only valid after calling move_and_slide or move_and_slide_with_snap and when is_on_floor returns true.

pub fn get_floor_velocity(&self) -> Vector3

Returns the linear velocity of the floor at the last collision point. Only valid after calling move_and_slide or move_and_slide_with_snap and when is_on_floor returns true.

pub fn safe_margin(&self) -> f64

Extra margin used for collision recovery in motion functions (see move_and_collide, move_and_slide, move_and_slide_with_snap). If the body is at least this close to another body, it will consider them to be colliding and will be pushed away before performing the actual motion. A higher value means it’s more flexible for detecting collision, which helps with consistently detecting walls and floors. A lower value forces the collision algorithm to use more exact detection, so it can be used in cases that specifically require precision, e.g at very low scale to avoid visible jittering, or for stability with a stack of kinematic bodies.

pub fn get_slide_collision(
&self,
slide_idx: i64
) -> Option<Ref<KinematicCollision, Shared>>

Returns a KinematicCollision, which contains information about a collision that occurred during the last call to move_and_slide or move_and_slide_with_snap. Since the body can collide several times in a single call to move_and_slide, you must specify the index of the collision in the range 0 to (get_slide_count - 1).

pub fn get_slide_count(&self) -> i64

Returns the number of times the body collided and changed direction during the last call to move_and_slide or move_and_slide_with_snap.

pub fn is_on_ceiling(&self) -> bool

Returns true if the body is on the ceiling. Only updates when calling move_and_slide or move_and_slide_with_snap.

pub fn is_on_floor(&self) -> bool

Returns true if the body is on the floor. Only updates when calling move_and_slide or move_and_slide_with_snap.

pub fn is_on_wall(&self) -> bool

Returns true if the body is on a wall. Only updates when calling move_and_slide or move_and_slide_with_snap.

pub fn move_and_collide(
&self,
rel_vec: Vector3,
infinite_inertia: bool,
exclude_raycast_shapes: bool,
test_only: bool
) -> Option<Ref<KinematicCollision, Shared>>

Moves the body along the vector rel_vec. The body will stop if it collides. Returns a KinematicCollision, which contains information about the collision. If test_only is true, the body does not move but the would-be collision information is given.

Default Arguments

infinite_inertia - true

exclude_raycast_shapes - true

test_only - false

pub fn move_and_slide(
&self,
linear_velocity: Vector3,
up_direction: Vector3,
stop_on_slope: bool,
max_slides: i64,
floor_max_angle: f64,
infinite_inertia: bool
) -> Vector3

Moves the body along a vector. If the body collides with another, it will slide along the other body rather than stop immediately. If the other body is a KinematicBody or RigidBody, it will also be affected by the motion of the other body. You can use this to make moving and rotating platforms, or to make nodes push other nodes. This method should be used in [Node._physics_process][Node::_physics_process] (or in a method called by [Node._physics_process][Node::_physics_process]), as it uses the physics step’s delta value automatically in calculations. Otherwise, the simulation will run at an incorrect speed. linear_velocity is the velocity vector (typically meters per second). Unlike in move_and_collide, you should not multiply it by delta — the physics engine handles applying the velocity. up_direction is the up direction, used to determine what is a wall and what is a floor or a ceiling. If set to the default value of Vector3(0, 0, 0), everything is considered a wall. If stop_on_slope is true, body will not slide on slopes when you include gravity in linear_velocity and the body is standing still. If the body collides, it will change direction a maximum of max_slides times before it stops. floor_max_angle is the maximum angle (in radians) where a slope is still considered a floor (or a ceiling), rather than a wall. The default value equals 45 degrees. If infinite_inertia is true, body will be able to push RigidBody nodes, but it won’t also detect any collisions with them. If false, it will interact with RigidBody nodes like with StaticBody. Returns the linear_velocity vector, rotated and/or scaled if a slide collision occurred. To get detailed information about collisions that occurred, use get_slide_collision.

Default Arguments

up_direction - Vector3( 0, 0, 0 )

stop_on_slope - false

max_slides - 4

floor_max_angle - 0.785398

infinite_inertia - true

pub fn move_and_slide_with_snap(
&self,
linear_velocity: Vector3,
snap: Vector3,
up_direction: Vector3,
stop_on_slope: bool,
max_slides: i64,
floor_max_angle: f64,
infinite_inertia: bool
) -> Vector3

Moves the body while keeping it attached to slopes. Similar to move_and_slide. As long as the snap vector is in contact with the ground, the body will remain attached to the surface. This means you must disable snap in order to jump, for example. You can do this by setting snap to (0, 0, 0) or by using move_and_slide instead.

Default Arguments

up_direction - Vector3( 0, 0, 0 )

stop_on_slope - false

max_slides - 4

floor_max_angle - 0.785398

infinite_inertia - true

pub fn set_axis_lock(&self, axis: i64, lock: bool)

Lock the body’s Z axis movement.

pub fn set_safe_margin(&self, pixels: f64)

Extra margin used for collision recovery in motion functions (see move_and_collide, move_and_slide, move_and_slide_with_snap). If the body is at least this close to another body, it will consider them to be colliding and will be pushed away before performing the actual motion. A higher value means it’s more flexible for detecting collision, which helps with consistently detecting walls and floors. A lower value forces the collision algorithm to use more exact detection, so it can be used in cases that specifically require precision, e.g at very low scale to avoid visible jittering, or for stability with a stack of kinematic bodies.

pub fn test_move(
&self,
from: Transform,
rel_vec: Vector3,
infinite_inertia: bool
) -> bool

Checks for collisions without moving the body. Virtually sets the node’s position, scale and rotation to that of the given Transform, then tries to move the body along the vector rel_vec. Returns true if a collision would occur.

Default Arguments

infinite_inertia - true