Struct gdnative::prelude::Reference

pub struct Reference { /* fields omitted */ }

Base class of all reference-counted types. Inherits Object.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Class hierarchy

Reference inherits methods from:

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Reference

pub fn new() -> Ref<Reference, Unique>

Creates a new instance of this object.

This is a reference-counted type. The returned object is automatically managed by Ref.

pub fn init_ref(&self) -> bool

Initializes the internal reference counter. Use this only if you really know what you are doing. Returns whether the initialization was successful.