Struct gdnative::prelude::SceneTree

pub struct SceneTree { /* private fields */ }

core class SceneTree inherits MainLoop (manually managed).

This class has related types in the scene_tree module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

SceneTree is a reference-only type. Persistent references can only exist in the unsafe Ref<SceneTree> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

SceneTree inherits methods from:

MainLoop

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl SceneTree

Constants

pub const GROUP_CALL_DEFAULT: i64 = 0i64

pub const STRETCH_ASPECT_IGNORE: i64 = 0i64

pub const STRETCH_MODE_DISABLED: i64 = 0i64

pub const GROUP_CALL_REVERSE: i64 = 1i64

pub const STRETCH_ASPECT_KEEP: i64 = 1i64

pub const STRETCH_MODE_2D: i64 = 1i64

pub const GROUP_CALL_REALTIME: i64 = 2i64

pub const STRETCH_ASPECT_KEEP_WIDTH: i64 = 2i64

pub const STRETCH_MODE_VIEWPORT: i64 = 2i64

pub const STRETCH_ASPECT_KEEP_HEIGHT: i64 = 3i64

pub const GROUP_CALL_UNIQUE: i64 = 4i64

pub const STRETCH_ASPECT_EXPAND: i64 = 4i64

impl SceneTree

pub fn new() -> Ref<SceneTree, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn call_group( &self, group: impl Into<GodotString>, method: impl Into<GodotString>, varargs: &[Variant] ) -> Variant

Calls method on each member of the given group. You can pass arguments to method by specifying them at the end of the method call. This method is equivalent of calling call_group_flags with GROUP_CALL_DEFAULT flag. Note: method may only have 5 arguments at most (7 arguments passed to this method in total). Note: Due to design limitations, call_group will fail silently if one of the arguments is null. Note: call_group will always call methods with an one-frame delay, in a way similar to Object.call_deferred. To call methods immediately, use call_group_flags with the GROUP_CALL_REALTIME flag.

pub fn call_group_flags( &self, flags: i64, group: impl Into<GodotString>, method: impl Into<GodotString>, varargs: &[Variant] ) -> Variant

Sample code is GDScript unless otherwise noted.

Calls method on each member of the given group, respecting the given GroupCallFlags. You can pass arguments to method by specifying them at the end of the method call. Note: method may only have 5 arguments at most (8 arguments passed to this method in total). Note: Due to design limitations, call_group_flags will fail silently if one of the arguments is null.

# Call the method immediately and in reverse order.
get_tree().call_group_flags(SceneTree.GROUP_CALL_REALTIME | SceneTree.GROUP_CALL_REVERSE, "bases", "destroy")

pub fn change_scene( &self, path: impl Into<GodotString> ) -> Result<(), GodotError>

Changes the running scene to the one at the given path, after loading it into a PackedScene and creating a new instance. Returns [OK][Self::OK] on success, [ERR_CANT_OPEN][Self::ERR_CANT_OPEN] if the path cannot be loaded into a PackedScene, or [ERR_CANT_CREATE][Self::ERR_CANT_CREATE] if that scene cannot be instantiated. Note: The scene change is deferred, which means that the new scene node is added on the next idle frame. You won’t be able to access it immediately after the change_scene call.

pub fn change_scene_to( &self, packed_scene: impl AsArg<PackedScene> ) -> Result<(), GodotError>

Changes the running scene to a new instance of the given PackedScene. Returns [OK][Self::OK] on success or [ERR_CANT_CREATE][Self::ERR_CANT_CREATE] if the scene cannot be instantiated. Note: The scene change is deferred, which means that the new scene node is added on the next idle frame. You won’t be able to access it immediately after the change_scene_to call. Note: Passing a value of null into the method will unload the current scene without loading a new one.

pub fn create_timer( &self, time_sec: f64, pause_mode_process: bool ) -> Option<Ref<SceneTreeTimer, Shared>>

Sample code is GDScript unless otherwise noted.

Returns a SceneTreeTimer which will [signal SceneTreeTimer.timeout] after the given time in seconds elapsed in this SceneTree. If pause_mode_process is set to false, pausing the SceneTree will also pause the timer. Commonly used to create a one-shot delay timer as in the following example:

func some_function():
    print("start")
    yield(get_tree().create_timer(1.0), "timeout")
    print("end")

The timer will be automatically freed after its time elapses.

Default Arguments

pause_mode_process - true

pub fn create_tween(&self) -> Option<Ref<SceneTreeTween, Shared>>

Creates and returns a new SceneTreeTween.

pub fn current_scene(&self) -> Option<Ref<Node, Shared>>

The current scene.

pub fn edited_scene_root(&self) -> Option<Ref<Node, Shared>>

The root of the edited scene.

pub fn get_frame(&self) -> i64

Returns the current frame number, i.e. the total frame count since the application started.

pub fn multiplayer(&self) -> Option<Ref<MultiplayerAPI, Shared>>

The default MultiplayerAPI instance for this SceneTree.

pub fn get_network_connected_peers(&self) -> PoolArray<i32>

Returns the peer IDs of all connected peers of this SceneTree’s network_peer.

pub fn network_peer(&self) -> Option<Ref<NetworkedMultiplayerPeer, Shared>>

The peer object to handle the RPC system (effectively enabling networking when set). Depending on the peer itself, the SceneTree will become a network server (check with is_network_server) and will set the root node’s network mode to master, or it will become a regular peer with the root node set to puppet. All child nodes are set to inherit the network mode by default. Handling of networking-related events (connection, disconnection, new clients) is done by connecting to SceneTree’s signals.

pub fn get_network_unique_id(&self) -> i64

Returns the unique peer ID of this SceneTree’s network_peer.

pub fn get_node_count(&self) -> i64

Returns the number of nodes in this SceneTree.

pub fn get_nodes_in_group( &self, group: impl Into<GodotString> ) -> VariantArray<Shared>

Returns a list of all nodes assigned to the given group.

pub fn get_processed_tweens(&self) -> VariantArray<Shared>

Returns an array of currently existing SceneTreeTweens in the SceneTree (both running and paused).

pub fn root(&self) -> Option<Ref<Viewport, Shared>>

The SceneTree’s root Viewport.

pub fn get_rpc_sender_id(&self) -> i64

Returns the sender’s peer ID for the most recently received RPC call.

pub fn has_group(&self, name: impl Into<GodotString>) -> bool

Returns true if the given group exists.

pub fn has_network_peer(&self) -> bool

Returns true if there is a network_peer set.

pub fn is_auto_accept_quit(&self) -> bool

If true, the application automatically accepts quitting. For mobile platforms, see [quit_on_go_back][Self::quit_on_go_back].

pub fn is_debugging_collisions_hint(&self) -> bool

If true, collision shapes will be visible when running the game from the editor for debugging purposes. Note: This property is not designed to be changed at run-time. Changing the value of [debug_collisions_hint][Self::debug_collisions_hint] while the project is running will not have the desired effect.

pub fn is_debugging_navigation_hint(&self) -> bool

If true, navigation polygons will be visible when running the game from the editor for debugging purposes. Note: This property is not designed to be changed at run-time. Changing the value of [debug_navigation_hint][Self::debug_navigation_hint] while the project is running will not have the desired effect.

pub fn is_input_handled(&self) -> bool

Returns true if the most recent InputEvent was marked as handled with set_input_as_handled.

pub fn is_multiplayer_poll_enabled(&self) -> bool

If true (default value), enables automatic polling of the MultiplayerAPI for this SceneTree during idle_frame. If false, you need to manually call MultiplayerAPI.poll to process network packets and deliver RPCs/RSETs. This allows running RPCs/RSETs in a different loop (e.g. physics, thread, specific time step) and for manual Mutex protection when accessing the MultiplayerAPI from threads.

pub fn is_network_server(&self) -> bool

Returns true if this SceneTree’s network_peer is in server mode (listening for connections).

pub fn is_paused(&self) -> bool

If true, the SceneTree is paused. Doing so will have the following behavior:

2D and 3D physics will be stopped. This includes signals and collision detection.

[Node._process][Node::_process], [Node._physics_process][Node::_physics_process] and [Node._input][Node::_input] will not be called anymore in nodes.

pub fn is_physics_interpolation_enabled(&self) -> bool

Although physics interpolation would normally be globally turned on and off using [member ProjectSettings.physics/common/physics_interpolation], this property allows control over interpolation at runtime.

pub fn is_quit_on_go_back(&self) -> bool

If true, the application quits automatically on going back (e.g. on Android). To handle ‘Go Back’ button when this option is disabled, use MainLoop.NOTIFICATION_WM_GO_BACK_REQUEST.

pub fn is_refusing_new_network_connections(&self) -> bool

If true, the SceneTree’s network_peer refuses new incoming connections.

pub fn is_using_font_oversampling(&self) -> bool

If true, font oversampling is enabled. This means that DynamicFonts will be rendered at higher or lower size than configured based on the viewport’s scaling ratio. For example, in a viewport scaled with a factor 1.5, a font configured with size 14 would be rendered with size 21 (14 * 1.5). Note: Font oversampling is only used if the viewport stretch mode is STRETCH_MODE_VIEWPORT, and if the stretch aspect mode is different from STRETCH_ASPECT_IGNORE. Note: This property is set automatically for the active SceneTree when the project starts based on the configuration of rendering/quality/dynamic_fonts/use_oversampling in ProjectSettings. The property can however be overridden at runtime as needed.

pub fn notify_group(&self, group: impl Into<GodotString>, notification: i64)

Sends the given notification to all members of the group.

pub fn notify_group_flags( &self, call_flags: i64, group: impl Into<GodotString>, notification: i64 )

Sends the given notification to all members of the group, respecting the given GroupCallFlags.

pub fn queue_delete(&self, obj: impl AsArg<Object>)

Queues the given object for deletion, delaying the call to [Object.free][Object::free] to after the current frame.

pub fn quit(&self, exit_code: i64)

Quits the application at the end of the current iteration. A process exit_code can optionally be passed as an argument. If this argument is 0 or greater, it will override the OS.exit_code defined before quitting the application. Note: On iOS this method doesn’t work. Instead, as recommended by the iOS Human Interface Guidelines, the user is expected to close apps via the Home button.

Default Arguments

exit_code - -1

pub fn reload_current_scene(&self) -> Result<(), GodotError>

Reloads the currently active scene. Returns [OK][Self::OK] on success, [ERR_UNCONFIGURED][Self::ERR_UNCONFIGURED] if no current_scene was defined yet, [ERR_CANT_OPEN][Self::ERR_CANT_OPEN] if current_scene cannot be loaded into a PackedScene, or [ERR_CANT_CREATE][Self::ERR_CANT_CREATE] if the scene cannot be instantiated.

pub fn set_auto_accept_quit(&self, enabled: bool)

If true, the application automatically accepts quitting. For mobile platforms, see [quit_on_go_back][Self::quit_on_go_back].

pub fn set_current_scene(&self, child_node: impl AsArg<Node>)

The current scene.

pub fn set_debug_collisions_hint(&self, enable: bool)

If true, collision shapes will be visible when running the game from the editor for debugging purposes. Note: This property is not designed to be changed at run-time. Changing the value of [debug_collisions_hint][Self::debug_collisions_hint] while the project is running will not have the desired effect.

pub fn set_debug_navigation_hint(&self, enable: bool)

If true, navigation polygons will be visible when running the game from the editor for debugging purposes. Note: This property is not designed to be changed at run-time. Changing the value of [debug_navigation_hint][Self::debug_navigation_hint] while the project is running will not have the desired effect.

pub fn set_edited_scene_root(&self, scene: impl AsArg<Node>)

The root of the edited scene.

pub fn set_group( &self, group: impl Into<GodotString>, property: impl Into<GodotString>, value: impl OwnedToVariant )

Sets the given property to value on all members of the given group.

pub fn set_group_flags( &self, call_flags: i64, group: impl Into<GodotString>, property: impl Into<GodotString>, value: impl OwnedToVariant )

Sets the given property to value on all members of the given group, respecting the given GroupCallFlags.

pub fn set_input_as_handled(&self)

Marks the most recent InputEvent as handled.

pub fn set_multiplayer(&self, multiplayer: impl AsArg<MultiplayerAPI>)

The default MultiplayerAPI instance for this SceneTree.

pub fn set_multiplayer_poll_enabled(&self, enabled: bool)

If true (default value), enables automatic polling of the MultiplayerAPI for this SceneTree during idle_frame. If false, you need to manually call MultiplayerAPI.poll to process network packets and deliver RPCs/RSETs. This allows running RPCs/RSETs in a different loop (e.g. physics, thread, specific time step) and for manual Mutex protection when accessing the MultiplayerAPI from threads.

pub fn set_network_peer(&self, peer: impl AsArg<NetworkedMultiplayerPeer>)

The peer object to handle the RPC system (effectively enabling networking when set). Depending on the peer itself, the SceneTree will become a network server (check with is_network_server) and will set the root node’s network mode to master, or it will become a regular peer with the root node set to puppet. All child nodes are set to inherit the network mode by default. Handling of networking-related events (connection, disconnection, new clients) is done by connecting to SceneTree’s signals.

pub fn set_pause(&self, enable: bool)

If true, the SceneTree is paused. Doing so will have the following behavior:

2D and 3D physics will be stopped. This includes signals and collision detection.

[Node._process][Node::_process], [Node._physics_process][Node::_physics_process] and [Node._input][Node::_input] will not be called anymore in nodes.

pub fn set_physics_interpolation_enabled(&self, enabled: bool)

Although physics interpolation would normally be globally turned on and off using [member ProjectSettings.physics/common/physics_interpolation], this property allows control over interpolation at runtime.

pub fn set_quit_on_go_back(&self, enabled: bool)

If true, the application quits automatically on going back (e.g. on Android). To handle ‘Go Back’ button when this option is disabled, use MainLoop.NOTIFICATION_WM_GO_BACK_REQUEST.

pub fn set_refuse_new_network_connections(&self, refuse: bool)

If true, the SceneTree’s network_peer refuses new incoming connections.

pub fn set_screen_stretch( &self, mode: i64, aspect: i64, minsize: Vector2, scale: f64 )

Configures screen stretching to the given StretchMode, StretchAspect, minimum size and scale.

Default Arguments

scale - 1

pub fn set_use_font_oversampling(&self, enable: bool)

If true, font oversampling is enabled. This means that DynamicFonts will be rendered at higher or lower size than configured based on the viewport’s scaling ratio. For example, in a viewport scaled with a factor 1.5, a font configured with size 14 would be rendered with size 21 (14 * 1.5). Note: Font oversampling is only used if the viewport stretch mode is STRETCH_MODE_VIEWPORT, and if the stretch aspect mode is different from STRETCH_ASPECT_IGNORE. Note: This property is set automatically for the active SceneTree when the project starts based on the configuration of rendering/quality/dynamic_fonts/use_oversampling in ProjectSettings. The property can however be overridden at runtime as needed.