Struct gdnative::prelude::Sprite

pub struct Sprite { /* fields omitted */ }

core class Sprite inherits Node2D (manually managed).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

Sprite is a reference-only type. Persistent references can only exist in the unsafe Ref<Sprite> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

Sprite inherits methods from:

Node2D

CanvasItem

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Sprite

pub fn new() -> Ref<Sprite, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn frame(&self) -> i64

Current frame to display from sprite sheet. hframes or vframes must be greater than 1.

pub fn frame_coords(&self) -> Vector2

Coordinates of the frame to display from sprite sheet. This is as an alias for the frame property. hframes or vframes must be greater than 1.

pub fn hframes(&self) -> i64

The number of columns in the sprite sheet.

pub fn normal_map(&self) -> Option<Ref<Texture, Shared>>

The normal map gives depth to the Sprite. Note: Godot expects the normal map to use X+, Y-, and Z+ coordinates. See this page for a comparison of normal map coordinates expected by popular engines.

pub fn offset(&self) -> Vector2

The texture’s drawing offset.

pub fn get_rect(&self) -> Rect2

Sample code is GDScript unless otherwise noted.

Returns a Rect2 representing the Sprite’s boundary in local coordinates. Can be used to detect if the Sprite was clicked. Example:

func _input(event):
    if event is InputEventMouseButton and event.pressed and event.button_index == BUTTON_LEFT:
        if get_rect().has_point(to_local(event.position)):
            print("A click!")

pub fn region_rect(&self) -> Rect2

The region of the atlas texture to display. [region_enabled][Self::region_enabled] must be true.

pub fn texture(&self) -> Option<Ref<Texture, Shared>>

Texture object to draw.

pub fn vframes(&self) -> i64

The number of rows in the sprite sheet.

pub fn is_centered(&self) -> bool

If true, texture is centered.

pub fn is_flipped_h(&self) -> bool

If true, texture is flipped horizontally.

pub fn is_flipped_v(&self) -> bool

If true, texture is flipped vertically.

pub fn is_pixel_opaque(&self, pos: Vector2) -> bool

Returns true, if the pixel at the given position is opaque and false in other case. Note: It also returns false, if the sprite’s texture is null or if the given position is invalid.

pub fn is_region(&self) -> bool

If true, texture is cut from a larger atlas texture. See region_rect.

pub fn is_region_filter_clip_enabled(&self) -> bool

If true, the outermost pixels get blurred out.

pub fn set_centered(&self, centered: bool)

If true, texture is centered.

pub fn set_flip_h(&self, flip_h: bool)

If true, texture is flipped horizontally.

pub fn set_flip_v(&self, flip_v: bool)

If true, texture is flipped vertically.

pub fn set_frame(&self, frame: i64)

Current frame to display from sprite sheet. hframes or vframes must be greater than 1.

pub fn set_frame_coords(&self, coords: Vector2)

Coordinates of the frame to display from sprite sheet. This is as an alias for the frame property. hframes or vframes must be greater than 1.

pub fn set_hframes(&self, hframes: i64)

The number of columns in the sprite sheet.

pub fn set_normal_map(&self, normal_map: impl AsArg<Texture>)

The normal map gives depth to the Sprite. Note: Godot expects the normal map to use X+, Y-, and Z+ coordinates. See this page for a comparison of normal map coordinates expected by popular engines.

pub fn set_offset(&self, offset: Vector2)

The texture’s drawing offset.

pub fn set_region(&self, enabled: bool)

If true, texture is cut from a larger atlas texture. See region_rect.

pub fn set_region_filter_clip(&self, enabled: bool)

If true, the outermost pixels get blurred out.

pub fn set_region_rect(&self, rect: Rect2)

The region of the atlas texture to display. [region_enabled][Self::region_enabled] must be true.

pub fn set_texture(&self, texture: impl AsArg<Texture>)

Texture object to draw.

pub fn set_vframes(&self, vframes: i64)

The number of rows in the sprite sheet.