Struct gdnative::prelude::Vector3

#[repr(C)]
pub struct Vector3 {
    pub x: f32,
    pub y: f32,
    pub z: f32,
}

3D vector class.

See also Vector3 in the Godot API doc.

Fields
x: f32y: f32z: f32
Implementations

impl Vector3

Helper methods for Vector3.

See the official Godot documentation.

pub const ZERO: Vector3

The zero vector.

pub const ONE: Vector3

A vector with all components set to 1. Typically used for scaling.

pub const INF: Vector3

A vector with all components set to +infinity.

pub const LEFT: Vector3

Unit vector in -X direction.

pub const RIGHT: Vector3

Unit vector in +X direction.

pub const UP: Vector3

Unit vector in +Y direction.

pub const DOWN: Vector3

Unit vector in -Y direction.

pub const FORWARD: Vector3

Unit vector in -Z direction.

pub const BACK: Vector3

Unit vector in +Z direction.

pub const fn new(x: f32, y: f32, z: f32) -> Vector3

Returns a Vector3 with the given components.

pub fn abs(self) -> Vector3

Returns a new vector with all components in absolute values (i.e. positive).

pub fn angle_to(self, to: Vector3) -> f32

Returns the minimum angle to the given vector, in radians.

pub fn bounce(self, n: Vector3) -> Vector3

Returns the vector “bounced off” from a plane defined by the given normal.

pub fn ceil(self) -> Vector3

Returns a new vector with all components rounded up (towards positive infinity).

pub fn cross(self, b: Vector3) -> Vector3

Returns the cross product of this vector and b.

pub fn cubic_interpolate(
self,
b: Vector3,
pre_a: Vector3,
post_b: Vector3,
t: f32
) -> Vector3

Performs a cubic interpolation between vectors pre_a, a, b, post_b (a is current), by the given amount t. t is on the range of 0.0 to 1.0, representing the amount of interpolation.

pub fn direction_to(self, other: Vector3) -> Vector3

Returns the normalized vector pointing from this vector to other.

pub fn distance_squared_to(self, other: Vector3) -> f32

Returns the squared distance to other.

This method runs faster than distance_to, so prefer it if you need to compare vectors or need the squared distance for some formula.

pub fn distance_to(self, other: Vector3) -> f32

Returns the distance to other.

pub fn dot(self, other: Vector3) -> f32

Returns the dot product of this vector and b. This can be used to compare the angle between two vectors. For example, this can be used to determine whether an enemy is facing the player.

The dot product will be 0 for a straight angle (90 degrees), greater than 0 for angles narrower than 90 degrees and lower than 0 for angles wider than 90 degrees.

When using unit (normalized) vectors, the result will always be between -1.0 (180 degree angle) when the vectors are facing opposite directions, and 1.0 (0 degree angle) when the vectors are aligned.

Note: a.dot(b) is equivalent to b.dot(a).

pub fn floor(self) -> Vector3

Returns a new vector with all components rounded down (towards negative infinity).

pub fn inverse(self) -> Vector3

Returns the inverse of the vector. This is the same as Vector3::new(1.0 / self.x, 1.0 / self.y, 1.0 / self.z).

pub fn is_equal_approx(self, v: Vector3) -> bool

Returns true if this vector and v are approximately equal, by running relative_eq on each component.

pub fn is_normalized(self) -> bool

Returns true if the vector is normalized, and false otherwise.

pub fn length(self) -> f32

Returns the length (magnitude) of this vector.

pub fn length_squared(self) -> f32

Returns the squared length (squared magnitude) of this vector.

This method runs faster than length, so prefer it if you need to compare vectors or need the squared distance for some formula.

pub fn linear_interpolate(self, b: Vector3, t: f32) -> Vector3

Returns the result of the linear interpolation between this vector and b by amount t. t is on the range of 0.0 to 1.0, representing the amount of interpolation.

pub fn max_axis(self) -> Axis

Returns the axis of the vector’s largest value. See Axis enum. If all components are equal, this method returns Axis::X.

pub fn min_axis(self) -> Axis

Returns the axis of the vector’s smallest value. See Axis enum. If all components are equal, this method returns Axis::Z.

pub fn move_toward(self, to: Vector3, delta: f32) -> Vector3

Moves this vector toward to by the fixed delta amount.

pub fn normalized(self) -> Vector3

Returns the vector scaled to unit length. Equivalent to v / v.length().

pub fn outer(self, b: Vector3) -> Basis

Returns the outer product with b.

pub fn posmod(self, rem: f32) -> Vector3

Returns a vector composed of the rem_euclid of this vector’s components and mod.

pub fn posmodv(self, remv: Vector3) -> Vector3

Returns a vector composed of the rem_euclid of this vector’s components and remv components.

pub fn project(self, b: Vector3) -> Vector3

Returns this vector projected onto another vector b.

pub fn reflect(self, n: Vector3) -> Vector3

Returns this vector reflected from a plane defined by the given normal.

pub fn rotated(self, axis: Vector3, phi: f32) -> Vector3

Rotates this vector around a given axis by phi radians. The axis must be a normalized vector.

pub fn round(self) -> Vector3

Returns this vector with all components rounded to the nearest integer, with halfway cases rounded away from zero.

pub fn sign(self) -> Vector3

Returns a vector with each component set to one or negative one, depending on the signs of this vector’s components, or zero if the component is zero, by calling signum on each component.

pub fn slerp(self, b: Vector3, t: f32) -> Vector3

Returns the result of spherical linear interpolation between this vector and b, by amount t. t is on the range of 0.0 to 1.0, representing the amount of interpolation.

Note: Both vectors must be normalized.

pub fn slide(self, n: Vector3) -> Vector3

Returns this vector slid along a plane defined by the given normal.

pub fn snapped(self, by: Vector3) -> Vector3

Returns this vector with each component snapped to the nearest multiple of step. This can also be used to round to an arbitrary number of decimals.

pub fn to_diagonal_matrix(self) -> Basis

Returns a diagonal matrix with the vector as main diagonal.

This is equivalent to a Basis with no rotation or shearing and this vector’s components set as the scale.