Trait gdnative::prelude::Method

pub trait Method<C>: 'static + Send + Syncwhere
    C: NativeClass,{
    fn call(&self, this: TInstance<'_, C, Shared>, args: Varargs<'_>) -> Variant;

    fn site() -> Option<Site<'static>> { ... }
}

Safe low-level trait for stateful, variadic methods that can be called on a native script type.

Required Methods

fn call(&self, this: TInstance<'_, C, Shared>, args: Varargs<'_>) -> Variant

Calls the method on this with args.

Provided Methods

fn site() -> Option<Site<'static>>

Returns an optional site where this method is defined. Used for logging errors in FFI wrappers.

Default implementation returns None.