Module gdnative::profiler

Interface to Godot’s built-in profiler.

Macros

profile_sig

Convenience macro to create a profiling signature with a given tag.

Structs

Signature

A string encoding information about the code being profiled for Godot’s built-in profiler.

Functions

add_data

Add a data point to Godot’s built-in profiler. The profiler only has microsecond precision. Sub-microsecond time is truncated.

profile

Times a closure and adds the measured time to Godot’s built-in profiler with the given signature, and then returns it’s return value.