Struct gdnative::api::Engine

pub struct Engine { /* private fields */ }

core singleton class Engine inherits Object (manually managed).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Class hierarchy

Engine inherits methods from:

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Engine

pub fn godot_singleton() -> &'static Engine

Returns a reference to the singleton instance.

pub fn get_author_info(&self) -> Dictionary<Shared>

Returns engine author information in a Dictionary. lead_developers - Array of Strings, lead developer names founders - Array of Strings, founder names project_managers - Array of Strings, project manager names developers - Array of Strings, developer names

pub fn get_copyright_info(&self) -> VariantArray<Shared>

Returns an Array of copyright information Dictionaries. name - String, component name parts - Array of Dictionaries {files, copyright, license} describing subsections of the component

pub fn get_donor_info(&self) -> Dictionary<Shared>

Returns a Dictionary of Arrays of donor names. {platinum_sponsors, gold_sponsors, silver_sponsors, bronze_sponsors, mini_sponsors, gold_donors, silver_donors, bronze_donors}

pub fn get_frames_drawn(&self) -> i64

Returns the total number of frames drawn. On headless platforms, or if the render loop is disabled with --disable-render-loop via command line, get_frames_drawn always returns 0. See get_idle_frames.

pub fn get_frames_per_second(&self) -> f64

Returns the frames per second of the running game.

pub fn get_idle_frames(&self) -> i64

Sample code is GDScript unless otherwise noted.

Returns the total number of frames passed since engine initialization which is advanced on each idle frame, regardless of whether the render loop is enabled. See also get_frames_drawn and get_physics_frames. get_idle_frames can be used to run expensive logic less often without relying on a Timer:

func _process(_delta):
    if Engine.get_idle_frames() % 2 == 0:
        pass  # Run expensive logic only once every 2 idle (render) frames here.

pub fn iterations_per_second(&self) -> i64

The number of fixed iterations per second. This controls how often physics simulation and [Node._physics_process][Node::_physics_process] methods are run. This value should generally always be set to 60 or above, as Godot doesn’t interpolate the physics step. As a result, values lower than 60 will look stuttery. This value can be increased to make input more reactive or work around collision tunneling issues, but keep in mind doing so will increase CPU usage. See also target_fps and [member ProjectSettings.physics/common/physics_fps]. Note: Only 8 physics ticks may be simulated per rendered frame at most. If more than 8 physics ticks have to be simulated per rendered frame to keep up with rendering, the game will appear to slow down (even if delta is used consistently in physics calculations). Therefore, it is recommended not to increase Engine.iterations_per_second above 240. Otherwise, the game will slow down when the rendering framerate goes below 30 FPS.

pub fn get_license_info(&self) -> Dictionary<Shared>

Returns Dictionary of licenses used by Godot and included third party components.

pub fn get_license_text(&self) -> GodotString

Returns Godot license text.

pub fn get_main_loop(&self) -> Option<Ref<MainLoop, Shared>>

Returns the main loop object (see MainLoop and SceneTree).

pub fn get_physics_frames(&self) -> i64

Sample code is GDScript unless otherwise noted.

Returns the total number of frames passed since engine initialization which is advanced on each physics frame. See also get_idle_frames. get_physics_frames can be used to run expensive logic less often without relying on a Timer:

func _physics_process(_delta):
    if Engine.get_physics_frames() % 2 == 0:
        pass  # Run expensive logic only once every 2 physics frames here.

pub fn get_physics_interpolation_fraction(&self) -> f64

Returns the fraction through the current physics tick we are at the time of rendering the frame. This can be used to implement fixed timestep interpolation.

pub fn physics_jitter_fix(&self) -> f64

Controls how much physics ticks are synchronized with real time. For 0 or less, the ticks are synchronized. Such values are recommended for network games, where clock synchronization matters. Higher values cause higher deviation of the in-game clock and real clock but smooth out framerate jitters. The default value of 0.5 should be fine for most; values above 2 could cause the game to react to dropped frames with a noticeable delay and are not recommended. Note: For best results, when using a custom physics interpolation solution, the physics jitter fix should be disabled by setting physics_jitter_fix to 0.

pub fn get_singleton( &self, name: impl Into<GodotString> ) -> Option<Ref<Object, Shared>>

Returns a global singleton with given name. Often used for plugins, e.g. GodotPayment on Android.

pub fn target_fps(&self) -> i64

The desired frames per second. If the hardware cannot keep up, this setting may not be respected. A value of 0 means no limit.

pub fn time_scale(&self) -> f64

Controls how fast or slow the in-game clock ticks versus the real life one. It defaults to 1.0. A value of 2.0 means the game moves twice as fast as real life, whilst a value of 0.5 means the game moves at half the regular speed.

pub fn get_version_info(&self) -> Dictionary<Shared>

Sample code is GDScript unless otherwise noted.

Returns the current engine version information in a Dictionary. major - Holds the major version number as an int minor - Holds the minor version number as an int patch - Holds the patch version number as an int hex - Holds the full version number encoded as a hexadecimal int with one byte (2 places) per number (see example below) status - Holds the status (e.g. “beta”, “rc1”, “rc2”, … “stable”) as a String build - Holds the build name (e.g. “custom_build”) as a String hash - Holds the full Git commit hash as a String year - Holds the year the version was released in as an int string - major + minor + patch + status + build in a single String The hex value is encoded as follows, from left to right: one byte for the major, one byte for the minor, one byte for the patch version. For example, “3.1.12” would be 0x03010C. Note: It’s still an int internally, and printing it will give you its decimal representation, which is not particularly meaningful. Use hexadecimal literals for easy version comparisons from code:

if Engine.get_version_info().hex >= 0x030200:
    # Do things specific to version 3.2 or later
else:
    # Do things specific to versions before 3.2

pub fn has_singleton(&self, name: impl Into<GodotString>) -> bool

Returns true if a singleton with given name exists in global scope.

pub fn is_editor_hint(&self) -> bool

Sample code is GDScript unless otherwise noted.

If true, the script is currently running inside the editor. This is useful for tool scripts to conditionally draw editor helpers, or prevent accidentally running “game” code that would affect the scene state while in the editor:

if Engine.editor_hint:
    draw_gizmos()
else:
    simulate_physics()

See Running code in the editor in the documentation for more information. Note: To detect whether the script is run from an editor build (e.g. when pressing F5), use OS.has_feature with the "editor" argument instead. OS.has_feature("editor") will evaluate to true both when the code is running in the editor and when running the project from the editor, but it will evaluate to false when the code is run from an exported project.

pub fn is_in_physics_frame(&self) -> bool

Returns true if the game is inside the fixed process and physics phase of the game loop.

pub fn is_printing_error_messages(&self) -> bool

If false, stops printing error and warning messages to the console and editor Output log. This can be used to hide error and warning messages during unit test suite runs. This property is equivalent to the [member ProjectSettings.application/run/disable_stderr] project setting. Warning: If you set this to false anywhere in the project, important error messages may be hidden even if they are emitted from other scripts. If this is set to false in a tool script, this will also impact the editor itself. Do not report bugs before ensuring error messages are enabled (as they are by default). Note: This property does not impact the editor’s Errors tab when running a project from the editor.

pub fn set_editor_hint(&self, enabled: bool)

Sample code is GDScript unless otherwise noted.

If true, the script is currently running inside the editor. This is useful for tool scripts to conditionally draw editor helpers, or prevent accidentally running “game” code that would affect the scene state while in the editor:

if Engine.editor_hint:
    draw_gizmos()
else:
    simulate_physics()

See Running code in the editor in the documentation for more information. Note: To detect whether the script is run from an editor build (e.g. when pressing F5), use OS.has_feature with the "editor" argument instead. OS.has_feature("editor") will evaluate to true both when the code is running in the editor and when running the project from the editor, but it will evaluate to false when the code is run from an exported project.

pub fn set_iterations_per_second(&self, iterations_per_second: i64)

The number of fixed iterations per second. This controls how often physics simulation and [Node._physics_process][Node::_physics_process] methods are run. This value should generally always be set to 60 or above, as Godot doesn’t interpolate the physics step. As a result, values lower than 60 will look stuttery. This value can be increased to make input more reactive or work around collision tunneling issues, but keep in mind doing so will increase CPU usage. See also target_fps and [member ProjectSettings.physics/common/physics_fps]. Note: Only 8 physics ticks may be simulated per rendered frame at most. If more than 8 physics ticks have to be simulated per rendered frame to keep up with rendering, the game will appear to slow down (even if delta is used consistently in physics calculations). Therefore, it is recommended not to increase Engine.iterations_per_second above 240. Otherwise, the game will slow down when the rendering framerate goes below 30 FPS.

pub fn set_physics_jitter_fix(&self, physics_jitter_fix: f64)

Controls how much physics ticks are synchronized with real time. For 0 or less, the ticks are synchronized. Such values are recommended for network games, where clock synchronization matters. Higher values cause higher deviation of the in-game clock and real clock but smooth out framerate jitters. The default value of 0.5 should be fine for most; values above 2 could cause the game to react to dropped frames with a noticeable delay and are not recommended. Note: For best results, when using a custom physics interpolation solution, the physics jitter fix should be disabled by setting physics_jitter_fix to 0.

pub fn set_print_error_messages(&self, enabled: bool)

If false, stops printing error and warning messages to the console and editor Output log. This can be used to hide error and warning messages during unit test suite runs. This property is equivalent to the [member ProjectSettings.application/run/disable_stderr] project setting. Warning: If you set this to false anywhere in the project, important error messages may be hidden even if they are emitted from other scripts. If this is set to false in a tool script, this will also impact the editor itself. Do not report bugs before ensuring error messages are enabled (as they are by default). Note: This property does not impact the editor’s Errors tab when running a project from the editor.

pub fn set_target_fps(&self, target_fps: i64)

The desired frames per second. If the hardware cannot keep up, this setting may not be respected. A value of 0 means no limit.

pub fn set_time_scale(&self, time_scale: f64)

Controls how fast or slow the in-game clock ticks versus the real life one. It defaults to 1.0. A value of 2.0 means the game moves twice as fast as real life, whilst a value of 0.5 means the game moves at half the regular speed.