Struct gdnative::api::Geometry

pub struct Geometry { /* private fields */ }

core singleton class Geometry inherits Object (manually managed).

This class has related types in the geometry module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Class hierarchy

Geometry inherits methods from:

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Geometry

Constants

pub const END_POLYGON: i64 = 0i64

pub const JOIN_SQUARE: i64 = 0i64

pub const OPERATION_UNION: i64 = 0i64

pub const END_JOINED: i64 = 1i64

pub const JOIN_ROUND: i64 = 1i64

pub const OPERATION_DIFFERENCE: i64 = 1i64

pub const END_BUTT: i64 = 2i64

pub const JOIN_MITER: i64 = 2i64

pub const OPERATION_INTERSECTION: i64 = 2i64

pub const END_SQUARE: i64 = 3i64

pub const OPERATION_XOR: i64 = 3i64

pub const END_ROUND: i64 = 4i64

impl Geometry

pub fn godot_singleton() -> &'static Geometry

Returns a reference to the singleton instance.

pub fn build_box_planes(&self, extents: Vector3) -> VariantArray<Shared>

Returns an array with 6 Planes that describe the sides of a box centered at the origin. The box size is defined by extents, which represents one (positive) corner of the box (i.e. half its actual size).

pub fn build_capsule_planes( &self, radius: f64, height: f64, sides: i64, lats: i64, axis: i64 ) -> VariantArray<Shared>

Returns an array of Planes closely bounding a faceted capsule centered at the origin with radius radius and height height. The parameter sides defines how many planes will be generated for the side part of the capsule, whereas lats gives the number of latitudinal steps at the bottom and top of the capsule. The parameter axis describes the axis along which the capsule is oriented (0 for X, 1 for Y, 2 for Z).

Default Arguments

axis - 2

pub fn build_cylinder_planes( &self, radius: f64, height: f64, sides: i64, axis: i64 ) -> VariantArray<Shared>

Returns an array of Planes closely bounding a faceted cylinder centered at the origin with radius radius and height height. The parameter sides defines how many planes will be generated for the round part of the cylinder. The parameter axis describes the axis along which the cylinder is oriented (0 for X, 1 for Y, 2 for Z).

Default Arguments

axis - 2

pub fn clip_polygon( &self, points: PoolArray<Vector3>, plane: Plane ) -> PoolArray<Vector3>

Clips the polygon defined by the points in points against the plane and returns the points of the clipped polygon.

pub fn clip_polygons_2d( &self, polygon_a: PoolArray<Vector2>, polygon_b: PoolArray<Vector2> ) -> VariantArray<Shared>

Clips polygon_a against polygon_b and returns an array of clipped polygons. This performs OPERATION_DIFFERENCE between polygons. Returns an empty array if polygon_b completely overlaps polygon_a. If polygon_b is enclosed by polygon_a, returns an outer polygon (boundary) and inner polygon (hole) which could be distinguished by calling is_polygon_clockwise.

pub fn clip_polyline_with_polygon_2d( &self, polyline: PoolArray<Vector2>, polygon: PoolArray<Vector2> ) -> VariantArray<Shared>

Clips polyline against polygon and returns an array of clipped polylines. This performs OPERATION_DIFFERENCE between the polyline and the polygon. This operation can be thought of as cutting a line with a closed shape.

pub fn convex_hull_2d(&self, points: PoolArray<Vector2>) -> PoolArray<Vector2>

Given an array of Vector2s, returns the convex hull as a list of points in counterclockwise order. The last point is the same as the first one.

pub fn exclude_polygons_2d( &self, polygon_a: PoolArray<Vector2>, polygon_b: PoolArray<Vector2> ) -> VariantArray<Shared>

Mutually excludes common area defined by intersection of polygon_a and polygon_b (see intersect_polygons_2d) and returns an array of excluded polygons. This performs OPERATION_XOR between polygons. In other words, returns all but common area between polygons. The operation may result in an outer polygon (boundary) and inner polygon (hole) produced which could be distinguished by calling is_polygon_clockwise.

pub fn get_closest_point_to_segment( &self, point: Vector3, s1: Vector3, s2: Vector3 ) -> Vector3

Returns the 3D point on the 3D segment (s1, s2) that is closest to point. The returned point will always be inside the specified segment.

pub fn get_closest_point_to_segment_2d( &self, point: Vector2, s1: Vector2, s2: Vector2 ) -> Vector2

Returns the 2D point on the 2D segment (s1, s2) that is closest to point. The returned point will always be inside the specified segment.

pub fn get_closest_point_to_segment_uncapped( &self, point: Vector3, s1: Vector3, s2: Vector3 ) -> Vector3

Returns the 3D point on the 3D line defined by (s1, s2) that is closest to point. The returned point can be inside the segment (s1, s2) or outside of it, i.e. somewhere on the line extending from the segment.

pub fn get_closest_point_to_segment_uncapped_2d( &self, point: Vector2, s1: Vector2, s2: Vector2 ) -> Vector2

Returns the 2D point on the 2D line defined by (s1, s2) that is closest to point. The returned point can be inside the segment (s1, s2) or outside of it, i.e. somewhere on the line extending from the segment.

pub fn get_closest_points_between_segments( &self, p1: Vector3, p2: Vector3, q1: Vector3, q2: Vector3 ) -> PoolArray<Vector3>

Given the two 3D segments (p1, p2) and (q1, q2), finds those two points on the two segments that are closest to each other. Returns a PoolVector3Array that contains this point on (p1, p2) as well the accompanying point on (q1, q2).

pub fn get_closest_points_between_segments_2d( &self, p1: Vector2, q1: Vector2, p2: Vector2, q2: Vector2 ) -> PoolArray<Vector2>

Given the two 2D segments (p1, q1) and (p2, q2), finds those two points on the two segments that are closest to each other. Returns a PoolVector2Array that contains this point on (p1, q1) as well the accompanying point on (p2, q2).

pub fn get_uv84_normal_bit(&self, normal: Vector3) -> i64

Used internally by the engine.

pub fn intersect_polygons_2d( &self, polygon_a: PoolArray<Vector2>, polygon_b: PoolArray<Vector2> ) -> VariantArray<Shared>

Intersects polygon_a with polygon_b and returns an array of intersected polygons. This performs OPERATION_INTERSECTION between polygons. In other words, returns common area shared by polygons. Returns an empty array if no intersection occurs. The operation may result in an outer polygon (boundary) and inner polygon (hole) produced which could be distinguished by calling is_polygon_clockwise.

pub fn intersect_polyline_with_polygon_2d( &self, polyline: PoolArray<Vector2>, polygon: PoolArray<Vector2> ) -> VariantArray<Shared>

Intersects polyline with polygon and returns an array of intersected polylines. This performs OPERATION_INTERSECTION between the polyline and the polygon. This operation can be thought of as chopping a line with a closed shape.

pub fn is_point_in_circle( &self, point: Vector2, circle_position: Vector2, circle_radius: f64 ) -> bool

Returns true if point is inside the circle or if it’s located exactly on the circle’s boundary, otherwise returns false.

pub fn is_point_in_polygon( &self, point: Vector2, polygon: PoolArray<Vector2> ) -> bool

Returns true if point is inside polygon or if it’s located exactly on polygon’s boundary, otherwise returns false.

pub fn is_polygon_clockwise(&self, polygon: PoolArray<Vector2>) -> bool

Returns true if polygon’s vertices are ordered in clockwise order, otherwise returns false.

pub fn line_intersects_line_2d( &self, from_a: Vector2, dir_a: Vector2, from_b: Vector2, dir_b: Vector2 ) -> Variant

Checks if the two lines (from_a, dir_a) and (from_b, dir_b) intersect. If yes, return the point of intersection as Vector2. If no intersection takes place, returns null. Note: The lines are specified using direction vectors, not end points.

pub fn make_atlas(&self, sizes: PoolArray<Vector2>) -> Dictionary<Shared>

Given an array of Vector2s representing tiles, builds an atlas. The returned dictionary has two keys: points is an array of Vector2 that specifies the positions of each tile, size contains the overall size of the whole atlas as Vector2.

pub fn merge_polygons_2d( &self, polygon_a: PoolArray<Vector2>, polygon_b: PoolArray<Vector2> ) -> VariantArray<Shared>

Merges (combines) polygon_a and polygon_b and returns an array of merged polygons. This performs OPERATION_UNION between polygons. The operation may result in an outer polygon (boundary) and multiple inner polygons (holes) produced which could be distinguished by calling is_polygon_clockwise.

pub fn offset_polygon_2d( &self, polygon: PoolArray<Vector2>, delta: f64, join_type: i64 ) -> VariantArray<Shared>

Sample code is GDScript unless otherwise noted.

Inflates or deflates polygon by delta units (pixels). If delta is positive, makes the polygon grow outward. If delta is negative, shrinks the polygon inward. Returns an array of polygons because inflating/deflating may result in multiple discrete polygons. Returns an empty array if delta is negative and the absolute value of it approximately exceeds the minimum bounding rectangle dimensions of the polygon. Each polygon’s vertices will be rounded as determined by join_type, see PolyJoinType. The operation may result in an outer polygon (boundary) and inner polygon (hole) produced which could be distinguished by calling is_polygon_clockwise. Note: To translate the polygon’s vertices specifically, use the Transform2D.xform method:

var polygon = PoolVector2Array([Vector2(0, 0), Vector2(100, 0), Vector2(100, 100), Vector2(0, 100)])
var offset = Vector2(50, 50)
polygon = Transform2D(0, offset).xform(polygon)
print(polygon) # prints [Vector2(50, 50), Vector2(150, 50), Vector2(150, 150), Vector2(50, 150)]

Default Arguments

join_type - 0

pub fn offset_polyline_2d( &self, polyline: PoolArray<Vector2>, delta: f64, join_type: i64, end_type: i64 ) -> VariantArray<Shared>

Inflates or deflates polyline by delta units (pixels), producing polygons. If delta is positive, makes the polyline grow outward. Returns an array of polygons because inflating/deflating may result in multiple discrete polygons. If delta is negative, returns an empty array. Each polygon’s vertices will be rounded as determined by join_type, see PolyJoinType. Each polygon’s endpoints will be rounded as determined by end_type, see PolyEndType. The operation may result in an outer polygon (boundary) and inner polygon (hole) produced which could be distinguished by calling is_polygon_clockwise.

Default Arguments

join_type - 0

end_type - 3

pub fn point_is_inside_triangle( &self, point: Vector2, a: Vector2, b: Vector2, c: Vector2 ) -> bool

Returns if point is inside the triangle specified by a, b and c.

pub fn ray_intersects_triangle( &self, from: Vector3, dir: Vector3, a: Vector3, b: Vector3, c: Vector3 ) -> Variant

Tests if the 3D ray starting at from with the direction of dir intersects the triangle specified by a, b and c. If yes, returns the point of intersection as Vector3. If no intersection takes place, an empty Variant is returned.

pub fn segment_intersects_circle( &self, segment_from: Vector2, segment_to: Vector2, circle_position: Vector2, circle_radius: f64 ) -> f64

Given the 2D segment (segment_from, segment_to), returns the position on the segment (as a number between 0 and 1) at which the segment hits the circle that is located at position circle_position and has radius circle_radius. If the segment does not intersect the circle, -1 is returned (this is also the case if the line extending the segment would intersect the circle, but the segment does not).

pub fn segment_intersects_convex( &self, from: Vector3, to: Vector3, planes: VariantArray<Shared> ) -> PoolArray<Vector3>

Given a convex hull defined though the Planes in the array planes, tests if the segment (from, to) intersects with that hull. If an intersection is found, returns a PoolVector3Array containing the point the intersection and the hull’s normal. If no intersecion is found, an the returned array is empty.

pub fn segment_intersects_cylinder( &self, from: Vector3, to: Vector3, height: f64, radius: f64 ) -> PoolArray<Vector3>

Checks if the segment (from, to) intersects the cylinder with height height that is centered at the origin and has radius radius. If no, returns an empty PoolVector3Array. If an intersection takes place, the returned array contains the point of intersection and the cylinder’s normal at the point of intersection.

pub fn segment_intersects_segment_2d( &self, from_a: Vector2, to_a: Vector2, from_b: Vector2, to_b: Vector2 ) -> Variant

Checks if the two segments (from_a, to_a) and (from_b, to_b) intersect. If yes, return the point of intersection as Vector2. If no intersection takes place, returns null.

pub fn segment_intersects_sphere( &self, from: Vector3, to: Vector3, sphere_position: Vector3, sphere_radius: f64 ) -> PoolArray<Vector3>

Checks if the segment (from, to) intersects the sphere that is located at sphere_position and has radius sphere_radius. If no, returns an empty PoolVector3Array. If yes, returns a PoolVector3Array containing the point of intersection and the sphere’s normal at the point of intersection.

pub fn segment_intersects_triangle( &self, from: Vector3, to: Vector3, a: Vector3, b: Vector3, c: Vector3 ) -> Variant

Tests if the segment (from, to) intersects the triangle a, b, c. If yes, returns the point of intersection as Vector3. If no intersection takes place, an empty Variant is returned.

pub fn triangulate_delaunay_2d( &self, points: PoolArray<Vector2> ) -> PoolArray<i32>

Triangulates the area specified by discrete set of points such that no point is inside the circumcircle of any resulting triangle. Returns a PoolIntArray where each triangle consists of three consecutive point indices into points (i.e. the returned array will have n * 3 elements, with n being the number of found triangles). If the triangulation did not succeed, an empty PoolIntArray is returned.

pub fn triangulate_polygon(&self, polygon: PoolArray<Vector2>) -> PoolArray<i32>

Triangulates the polygon specified by the points in polygon. Returns a PoolIntArray where each triangle consists of three consecutive point indices into polygon (i.e. the returned array will have n * 3 elements, with n being the number of found triangles). Output triangles will always be counter clockwise, and the contour will be flipped if it’s clockwise. If the triangulation did not succeed, an empty PoolIntArray is returned.