Struct gdnative::api::GlobalConstants

pub struct GlobalConstants { /* private fields */ }

core singleton class GlobalConstants (reference-counted)

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

The lifetime of this object is automatically managed through reference counting.

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl GlobalConstants

Constants

pub const KEY_MODIFIER_MASK: i64 = -16_777_216i64

pub const JOY_INVALID_OPTION: i64 = -1i64

pub const CORNER_TOP_LEFT: i64 = 0i64

pub const HALIGN_LEFT: i64 = 0i64

pub const HORIZONTAL: i64 = 0i64

pub const JOY_ANALOG_LX: i64 = 0i64

pub const JOY_AXIS_0: i64 = 0i64

pub const JOY_BUTTON_0: i64 = 0i64

pub const JOY_DS_B: i64 = 0i64

pub const JOY_OPENVR_TOUCHPADX: i64 = 0i64

pub const JOY_SONY_X: i64 = 0i64

pub const JOY_XBOX_A: i64 = 0i64

pub const MARGIN_LEFT: i64 = 0i64

pub const OK: i64 = 0i64

pub const OP_EQUAL: i64 = 0i64

pub const PROPERTY_HINT_NONE: i64 = 0i64

pub const TYPE_NIL: i64 = 0i64

pub const VALIGN_TOP: i64 = 0i64

pub const BUTTON_LEFT: i64 = 1i64

pub const BUTTON_MASK_LEFT: i64 = 1i64

pub const CORNER_TOP_RIGHT: i64 = 1i64

pub const FAILED: i64 = 1i64

pub const HALIGN_CENTER: i64 = 1i64

pub const JOY_ANALOG_LY: i64 = 1i64

pub const JOY_AXIS_1: i64 = 1i64

pub const JOY_BUTTON_1: i64 = 1i64

pub const JOY_DS_A: i64 = 1i64

pub const JOY_OCULUS_BY: i64 = 1i64

pub const JOY_OPENVR_MENU: i64 = 1i64

pub const JOY_OPENVR_TOUCHPADY: i64 = 1i64

pub const JOY_SONY_CIRCLE: i64 = 1i64

pub const JOY_XBOX_B: i64 = 1i64

pub const MARGIN_TOP: i64 = 1i64

pub const METHOD_FLAGS_DEFAULT: i64 = 1i64

pub const METHOD_FLAG_NORMAL: i64 = 1i64

pub const OP_NOT_EQUAL: i64 = 1i64

pub const PROPERTY_HINT_RANGE: i64 = 1i64

pub const PROPERTY_USAGE_STORAGE: i64 = 1i64

pub const TYPE_BOOL: i64 = 1i64

pub const VALIGN_CENTER: i64 = 1i64

pub const VERTICAL: i64 = 1i64

pub const BUTTON_MASK_RIGHT: i64 = 2i64

pub const BUTTON_RIGHT: i64 = 2i64

pub const CORNER_BOTTOM_RIGHT: i64 = 2i64

pub const ERR_UNAVAILABLE: i64 = 2i64

pub const HALIGN_RIGHT: i64 = 2i64

pub const JOY_ANALOG_RX: i64 = 2i64

pub const JOY_AXIS_2: i64 = 2i64

pub const JOY_BUTTON_2: i64 = 2i64

pub const JOY_DS_Y: i64 = 2i64

pub const JOY_SONY_SQUARE: i64 = 2i64

pub const JOY_VR_ANALOG_TRIGGER: i64 = 2i64

pub const JOY_VR_GRIP: i64 = 2i64

pub const JOY_XBOX_X: i64 = 2i64

pub const MARGIN_RIGHT: i64 = 2i64

pub const METHOD_FLAG_EDITOR: i64 = 2i64

pub const OP_LESS: i64 = 2i64

pub const PROPERTY_HINT_EXP_RANGE: i64 = 2i64

pub const PROPERTY_USAGE_EDITOR: i64 = 2i64

pub const TYPE_INT: i64 = 2i64

pub const VALIGN_BOTTOM: i64 = 2i64

pub const BUTTON_MIDDLE: i64 = 3i64

pub const CORNER_BOTTOM_LEFT: i64 = 3i64

pub const ERR_UNCONFIGURED: i64 = 3i64

pub const JOY_ANALOG_RY: i64 = 3i64

pub const JOY_AXIS_3: i64 = 3i64

pub const JOY_BUTTON_3: i64 = 3i64

pub const JOY_DS_X: i64 = 3i64

pub const JOY_OCULUS_MENU: i64 = 3i64

pub const JOY_SONY_TRIANGLE: i64 = 3i64

pub const JOY_XBOX_Y: i64 = 3i64

pub const MARGIN_BOTTOM: i64 = 3i64

pub const OP_LESS_EQUAL: i64 = 3i64

pub const PROPERTY_HINT_ENUM: i64 = 3i64

pub const TYPE_REAL: i64 = 3i64

pub const BUTTON_MASK_MIDDLE: i64 = 4i64

pub const BUTTON_WHEEL_UP: i64 = 4i64

pub const ERR_UNAUTHORIZED: i64 = 4i64

pub const JOY_AXIS_4: i64 = 4i64

pub const JOY_BUTTON_4: i64 = 4i64

pub const JOY_L: i64 = 4i64

pub const JOY_VR_ANALOG_GRIP: i64 = 4i64

pub const METHOD_FLAG_NOSCRIPT: i64 = 4i64

pub const OP_GREATER: i64 = 4i64

pub const PROPERTY_HINT_EXP_EASING: i64 = 4i64

pub const PROPERTY_USAGE_NETWORK: i64 = 4i64

pub const TYPE_STRING: i64 = 4i64

pub const BUTTON_WHEEL_DOWN: i64 = 5i64

pub const ERR_PARAMETER_RANGE_ERROR: i64 = 5i64

pub const JOY_AXIS_5: i64 = 5i64

pub const JOY_BUTTON_5: i64 = 5i64

pub const JOY_R: i64 = 5i64

pub const OP_GREATER_EQUAL: i64 = 5i64

pub const PROPERTY_HINT_LENGTH: i64 = 5i64

pub const PROPERTY_USAGE_NOEDITOR: i64 = 5i64

pub const TYPE_VECTOR2: i64 = 5i64

pub const BUTTON_WHEEL_LEFT: i64 = 6i64

pub const ERR_OUT_OF_MEMORY: i64 = 6i64

pub const JOY_ANALOG_L2: i64 = 6i64

pub const JOY_AXIS_6: i64 = 6i64

pub const JOY_BUTTON_6: i64 = 6i64

pub const JOY_L2: i64 = 6i64

pub const OP_ADD: i64 = 6i64

pub const TYPE_RECT2: i64 = 6i64

pub const BUTTON_WHEEL_RIGHT: i64 = 7i64

pub const ERR_FILE_NOT_FOUND: i64 = 7i64

pub const JOY_ANALOG_R2: i64 = 7i64

pub const JOY_AXIS_7: i64 = 7i64

pub const JOY_BUTTON_7: i64 = 7i64

pub const JOY_OCULUS_AX: i64 = 7i64

pub const JOY_R2: i64 = 7i64

pub const OP_SUBTRACT: i64 = 7i64

pub const PROPERTY_HINT_KEY_ACCEL: i64 = 7i64

pub const PROPERTY_USAGE_DEFAULT: i64 = 7i64

pub const TYPE_VECTOR3: i64 = 7i64

pub const BUTTON_XBUTTON1: i64 = 8i64

pub const ERR_FILE_BAD_DRIVE: i64 = 8i64

pub const JOY_AXIS_8: i64 = 8i64

pub const JOY_BUTTON_8: i64 = 8i64

pub const JOY_L3: i64 = 8i64

pub const METHOD_FLAG_CONST: i64 = 8i64

pub const MIDI_MESSAGE_NOTE_OFF: i64 = 8i64

pub const OP_MULTIPLY: i64 = 8i64

pub const PROPERTY_HINT_FLAGS: i64 = 8i64

pub const PROPERTY_USAGE_EDITOR_HELPER: i64 = 8i64

pub const TYPE_TRANSFORM2D: i64 = 8i64

pub const BUTTON_XBUTTON2: i64 = 9i64

pub const ERR_FILE_BAD_PATH: i64 = 9i64

pub const JOY_AXIS_9: i64 = 9i64

pub const JOY_BUTTON_9: i64 = 9i64

pub const JOY_R3: i64 = 9i64

pub const MIDI_MESSAGE_NOTE_ON: i64 = 9i64

pub const OP_DIVIDE: i64 = 9i64

pub const PROPERTY_HINT_LAYERS_2D_RENDER: i64 = 9i64

pub const TYPE_PLANE: i64 = 9i64

pub const ERR_FILE_NO_PERMISSION: i64 = 10i64

pub const JOY_AXIS_MAX: i64 = 10i64

pub const JOY_BUTTON_10: i64 = 10i64

pub const JOY_SELECT: i64 = 10i64

pub const MIDI_MESSAGE_AFTERTOUCH: i64 = 10i64

pub const OP_NEGATE: i64 = 10i64

pub const PROPERTY_HINT_LAYERS_2D_PHYSICS: i64 = 10i64

pub const TYPE_QUAT: i64 = 10i64

pub const ERR_FILE_ALREADY_IN_USE: i64 = 11i64

pub const JOY_BUTTON_11: i64 = 11i64

pub const JOY_START: i64 = 11i64

pub const MIDI_MESSAGE_CONTROL_CHANGE: i64 = 11i64

pub const OP_POSITIVE: i64 = 11i64

pub const PROPERTY_HINT_LAYERS_2D_NAVIGATION: i64 = 11i64

pub const TYPE_AABB: i64 = 11i64

pub const ERR_FILE_CANT_OPEN: i64 = 12i64

pub const JOY_BUTTON_12: i64 = 12i64

pub const JOY_DPAD_UP: i64 = 12i64

pub const MIDI_MESSAGE_PROGRAM_CHANGE: i64 = 12i64

pub const OP_MODULE: i64 = 12i64

pub const PROPERTY_HINT_LAYERS_3D_RENDER: i64 = 12i64

pub const TYPE_BASIS: i64 = 12i64

pub const ERR_FILE_CANT_WRITE: i64 = 13i64

pub const JOY_BUTTON_13: i64 = 13i64

pub const JOY_DPAD_DOWN: i64 = 13i64

pub const MIDI_MESSAGE_CHANNEL_PRESSURE: i64 = 13i64

pub const OP_STRING_CONCAT: i64 = 13i64

pub const PROPERTY_HINT_LAYERS_3D_PHYSICS: i64 = 13i64

pub const TYPE_TRANSFORM: i64 = 13i64

pub const ERR_FILE_CANT_READ: i64 = 14i64

pub const JOY_BUTTON_14: i64 = 14i64

pub const JOY_DPAD_LEFT: i64 = 14i64

pub const JOY_VR_PAD: i64 = 14i64

pub const MIDI_MESSAGE_PITCH_BEND: i64 = 14i64

pub const OP_SHIFT_LEFT: i64 = 14i64

pub const PROPERTY_HINT_LAYERS_3D_NAVIGATION: i64 = 14i64

pub const TYPE_COLOR: i64 = 14i64

pub const ERR_FILE_UNRECOGNIZED: i64 = 15i64

pub const JOY_BUTTON_15: i64 = 15i64

pub const JOY_DPAD_RIGHT: i64 = 15i64

pub const JOY_VR_TRIGGER: i64 = 15i64

pub const OP_SHIFT_RIGHT: i64 = 15i64

pub const PROPERTY_HINT_FILE: i64 = 15i64

pub const TYPE_NODE_PATH: i64 = 15i64

pub const ERR_FILE_CORRUPT: i64 = 16i64

pub const JOY_BUTTON_16: i64 = 16i64

pub const JOY_GUIDE: i64 = 16i64

pub const METHOD_FLAG_REVERSE: i64 = 16i64

pub const OP_BIT_AND: i64 = 16i64

pub const PROPERTY_HINT_DIR: i64 = 16i64

pub const PROPERTY_USAGE_CHECKABLE: i64 = 16i64

pub const TYPE_RID: i64 = 16i64

pub const ERR_FILE_MISSING_DEPENDENCIES: i64 = 17i64

pub const JOY_BUTTON_17: i64 = 17i64

pub const JOY_MISC1: i64 = 17i64

pub const OP_BIT_OR: i64 = 17i64

pub const PROPERTY_HINT_GLOBAL_FILE: i64 = 17i64

pub const TYPE_OBJECT: i64 = 17i64

pub const ERR_FILE_EOF: i64 = 18i64

pub const JOY_BUTTON_18: i64 = 18i64

pub const JOY_PADDLE1: i64 = 18i64

pub const OP_BIT_XOR: i64 = 18i64

pub const PROPERTY_HINT_GLOBAL_DIR: i64 = 18i64

pub const TYPE_DICTIONARY: i64 = 18i64

pub const ERR_CANT_OPEN: i64 = 19i64

pub const JOY_BUTTON_19: i64 = 19i64

pub const JOY_PADDLE2: i64 = 19i64

pub const OP_BIT_NEGATE: i64 = 19i64

pub const PROPERTY_HINT_RESOURCE_TYPE: i64 = 19i64

pub const TYPE_ARRAY: i64 = 19i64

pub const ERR_CANT_CREATE: i64 = 20i64

pub const JOY_BUTTON_20: i64 = 20i64

pub const JOY_PADDLE3: i64 = 20i64

pub const OP_AND: i64 = 20i64

pub const PROPERTY_HINT_MULTILINE_TEXT: i64 = 20i64

pub const TYPE_RAW_ARRAY: i64 = 20i64

pub const ERR_QUERY_FAILED: i64 = 21i64

pub const JOY_BUTTON_21: i64 = 21i64

pub const JOY_PADDLE4: i64 = 21i64

pub const OP_OR: i64 = 21i64

pub const PROPERTY_HINT_PLACEHOLDER_TEXT: i64 = 21i64

pub const TYPE_INT_ARRAY: i64 = 21i64

pub const ERR_ALREADY_IN_USE: i64 = 22i64

pub const JOY_BUTTON_22: i64 = 22i64

pub const JOY_TOUCHPAD: i64 = 22i64

pub const OP_XOR: i64 = 22i64

pub const PROPERTY_HINT_COLOR_NO_ALPHA: i64 = 22i64

pub const TYPE_REAL_ARRAY: i64 = 22i64

pub const ERR_LOCKED: i64 = 23i64

pub const OP_NOT: i64 = 23i64

pub const PROPERTY_HINT_IMAGE_COMPRESS_LOSSY: i64 = 23i64

pub const TYPE_STRING_ARRAY: i64 = 23i64

pub const ERR_TIMEOUT: i64 = 24i64

pub const OP_IN: i64 = 24i64

pub const PROPERTY_HINT_IMAGE_COMPRESS_LOSSLESS: i64 = 24i64

pub const TYPE_VECTOR2_ARRAY: i64 = 24i64

pub const ERR_CANT_CONNECT: i64 = 25i64

pub const OP_MAX: i64 = 25i64

pub const TYPE_VECTOR3_ARRAY: i64 = 25i64

pub const ERR_CANT_RESOLVE: i64 = 26i64

pub const TYPE_COLOR_ARRAY: i64 = 26i64

pub const ERR_CONNECTION_ERROR: i64 = 27i64

pub const TYPE_MAX: i64 = 27i64

pub const ERR_CANT_ACQUIRE_RESOURCE: i64 = 28i64

pub const ERR_CANT_FORK: i64 = 29i64

pub const ERR_INVALID_DATA: i64 = 30i64

pub const ERR_INVALID_PARAMETER: i64 = 31i64

pub const ERR_ALREADY_EXISTS: i64 = 32i64

pub const KEY_SPACE: i64 = 32i64

pub const METHOD_FLAG_VIRTUAL: i64 = 32i64

pub const PROPERTY_USAGE_CHECKED: i64 = 32i64

pub const ERR_DOES_NOT_EXIST: i64 = 33i64

pub const KEY_EXCLAM: i64 = 33i64

pub const ERR_DATABASE_CANT_READ: i64 = 34i64

pub const KEY_QUOTEDBL: i64 = 34i64

pub const ERR_DATABASE_CANT_WRITE: i64 = 35i64

pub const KEY_NUMBERSIGN: i64 = 35i64

pub const ERR_COMPILATION_FAILED: i64 = 36i64

pub const KEY_DOLLAR: i64 = 36i64

pub const ERR_METHOD_NOT_FOUND: i64 = 37i64

pub const KEY_PERCENT: i64 = 37i64

pub const ERR_LINK_FAILED: i64 = 38i64

pub const KEY_AMPERSAND: i64 = 38i64

pub const ERR_SCRIPT_FAILED: i64 = 39i64

pub const KEY_APOSTROPHE: i64 = 39i64

pub const PROPERTY_HINT_ENUM_SUGGESTION: i64 = 39i64

pub const ERR_CYCLIC_LINK: i64 = 40i64

pub const KEY_PARENLEFT: i64 = 40i64

pub const ERR_INVALID_DECLARATION: i64 = 41i64

pub const KEY_PARENRIGHT: i64 = 41i64

pub const ERR_DUPLICATE_SYMBOL: i64 = 42i64

pub const KEY_ASTERISK: i64 = 42i64

pub const ERR_PARSE_ERROR: i64 = 43i64

pub const KEY_PLUS: i64 = 43i64

pub const ERR_BUSY: i64 = 44i64

pub const KEY_COMMA: i64 = 44i64

pub const ERR_SKIP: i64 = 45i64

pub const KEY_MINUS: i64 = 45i64

pub const ERR_HELP: i64 = 46i64

pub const KEY_PERIOD: i64 = 46i64

pub const ERR_BUG: i64 = 47i64

pub const KEY_SLASH: i64 = 47i64

pub const ERR_PRINTER_ON_FIRE: i64 = 48i64

pub const KEY_0: i64 = 48i64

pub const KEY_1: i64 = 49i64

pub const KEY_2: i64 = 50i64

pub const KEY_3: i64 = 51i64

pub const KEY_4: i64 = 52i64

pub const KEY_5: i64 = 53i64

pub const KEY_6: i64 = 54i64

pub const KEY_7: i64 = 55i64

pub const KEY_8: i64 = 56i64

pub const KEY_9: i64 = 57i64

pub const KEY_COLON: i64 = 58i64

pub const KEY_SEMICOLON: i64 = 59i64

pub const KEY_LESS: i64 = 60i64

pub const KEY_EQUAL: i64 = 61i64

pub const KEY_GREATER: i64 = 62i64

pub const KEY_QUESTION: i64 = 63i64

pub const KEY_AT: i64 = 64i64

pub const METHOD_FLAG_FROM_SCRIPT: i64 = 64i64

pub const PROPERTY_USAGE_INTERNATIONALIZED: i64 = 64i64

pub const KEY_A: i64 = 65i64

pub const KEY_B: i64 = 66i64

pub const KEY_C: i64 = 67i64

pub const KEY_D: i64 = 68i64

pub const KEY_E: i64 = 69i64

pub const KEY_F: i64 = 70i64

pub const KEY_G: i64 = 71i64

pub const PROPERTY_USAGE_DEFAULT_INTL: i64 = 71i64

pub const KEY_H: i64 = 72i64

pub const KEY_I: i64 = 73i64

pub const KEY_J: i64 = 74i64

pub const KEY_K: i64 = 75i64

pub const KEY_L: i64 = 76i64

pub const KEY_M: i64 = 77i64

pub const KEY_N: i64 = 78i64

pub const KEY_O: i64 = 79i64

pub const KEY_P: i64 = 80i64

pub const KEY_Q: i64 = 81i64

pub const KEY_R: i64 = 82i64

pub const KEY_S: i64 = 83i64

pub const KEY_T: i64 = 84i64

pub const KEY_U: i64 = 85i64

pub const KEY_V: i64 = 86i64

pub const KEY_W: i64 = 87i64

pub const KEY_X: i64 = 88i64

pub const KEY_Y: i64 = 89i64

pub const KEY_Z: i64 = 90i64

pub const KEY_BRACKETLEFT: i64 = 91i64

pub const KEY_BACKSLASH: i64 = 92i64

pub const KEY_BRACKETRIGHT: i64 = 93i64

pub const KEY_ASCIICIRCUM: i64 = 94i64

pub const KEY_UNDERSCORE: i64 = 95i64

pub const KEY_QUOTELEFT: i64 = 96i64

pub const KEY_BRACELEFT: i64 = 123i64

pub const KEY_BAR: i64 = 124i64

pub const KEY_BRACERIGHT: i64 = 125i64

pub const KEY_ASCIITILDE: i64 = 126i64

pub const BUTTON_MASK_XBUTTON1: i64 = 128i64

pub const JOY_BUTTON_MAX: i64 = 128i64

pub const METHOD_FLAG_VARARG: i64 = 128i64

pub const PROPERTY_USAGE_GROUP: i64 = 128i64

pub const KEY_NOBREAKSPACE: i64 = 160i64

pub const KEY_EXCLAMDOWN: i64 = 161i64

pub const KEY_CENT: i64 = 162i64

pub const KEY_STERLING: i64 = 163i64

pub const KEY_CURRENCY: i64 = 164i64

pub const KEY_YEN: i64 = 165i64

pub const KEY_BROKENBAR: i64 = 166i64

pub const KEY_SECTION: i64 = 167i64

pub const KEY_DIAERESIS: i64 = 168i64

pub const KEY_COPYRIGHT: i64 = 169i64

pub const KEY_ORDFEMININE: i64 = 170i64

pub const KEY_GUILLEMOTLEFT: i64 = 171i64

pub const KEY_NOTSIGN: i64 = 172i64

pub const KEY_HYPHEN: i64 = 173i64

pub const KEY_REGISTERED: i64 = 174i64

pub const KEY_MACRON: i64 = 175i64

pub const KEY_DEGREE: i64 = 176i64

pub const KEY_PLUSMINUS: i64 = 177i64

pub const KEY_TWOSUPERIOR: i64 = 178i64

pub const KEY_THREESUPERIOR: i64 = 179i64

pub const KEY_ACUTE: i64 = 180i64

pub const KEY_MU: i64 = 181i64

pub const KEY_PARAGRAPH: i64 = 182i64

pub const KEY_PERIODCENTERED: i64 = 183i64

pub const KEY_CEDILLA: i64 = 184i64

pub const KEY_ONESUPERIOR: i64 = 185i64

pub const KEY_MASCULINE: i64 = 186i64

pub const KEY_GUILLEMOTRIGHT: i64 = 187i64

pub const KEY_ONEQUARTER: i64 = 188i64

pub const KEY_ONEHALF: i64 = 189i64

pub const KEY_THREEQUARTERS: i64 = 190i64

pub const KEY_QUESTIONDOWN: i64 = 191i64

pub const KEY_AGRAVE: i64 = 192i64

pub const KEY_AACUTE: i64 = 193i64

pub const KEY_ACIRCUMFLEX: i64 = 194i64

pub const KEY_ATILDE: i64 = 195i64

pub const KEY_ADIAERESIS: i64 = 196i64

pub const KEY_ARING: i64 = 197i64

pub const KEY_AE: i64 = 198i64

pub const KEY_CCEDILLA: i64 = 199i64

pub const KEY_EGRAVE: i64 = 200i64

pub const KEY_EACUTE: i64 = 201i64

pub const KEY_ECIRCUMFLEX: i64 = 202i64

pub const KEY_EDIAERESIS: i64 = 203i64

pub const KEY_IGRAVE: i64 = 204i64

pub const KEY_IACUTE: i64 = 205i64

pub const KEY_ICIRCUMFLEX: i64 = 206i64

pub const KEY_IDIAERESIS: i64 = 207i64

pub const KEY_ETH: i64 = 208i64

pub const KEY_NTILDE: i64 = 209i64

pub const KEY_OGRAVE: i64 = 210i64

pub const KEY_OACUTE: i64 = 211i64

pub const KEY_OCIRCUMFLEX: i64 = 212i64

pub const KEY_OTILDE: i64 = 213i64

pub const KEY_ODIAERESIS: i64 = 214i64

pub const KEY_MULTIPLY: i64 = 215i64

pub const KEY_OOBLIQUE: i64 = 216i64

pub const KEY_UGRAVE: i64 = 217i64

pub const KEY_UACUTE: i64 = 218i64

pub const KEY_UCIRCUMFLEX: i64 = 219i64

pub const KEY_UDIAERESIS: i64 = 220i64

pub const KEY_YACUTE: i64 = 221i64

pub const KEY_THORN: i64 = 222i64

pub const KEY_SSHARP: i64 = 223i64

pub const MIDI_MESSAGE_SYSTEM_EXCLUSIVE: i64 = 240i64

pub const MIDI_MESSAGE_QUARTER_FRAME: i64 = 241i64

pub const MIDI_MESSAGE_SONG_POSITION_POINTER: i64 = 242i64

pub const MIDI_MESSAGE_SONG_SELECT: i64 = 243i64

pub const MIDI_MESSAGE_TUNE_REQUEST: i64 = 246i64

pub const KEY_DIVISION: i64 = 247i64

pub const MIDI_MESSAGE_TIMING_CLOCK: i64 = 248i64

pub const MIDI_MESSAGE_START: i64 = 250i64

pub const MIDI_MESSAGE_CONTINUE: i64 = 251i64

pub const MIDI_MESSAGE_STOP: i64 = 252i64

pub const MIDI_MESSAGE_ACTIVE_SENSING: i64 = 254i64

pub const KEY_YDIAERESIS: i64 = 255i64

pub const MIDI_MESSAGE_SYSTEM_RESET: i64 = 255i64

pub const BUTTON_MASK_XBUTTON2: i64 = 256i64

pub const PROPERTY_USAGE_CATEGORY: i64 = 256i64

pub const PROPERTY_USAGE_NO_INSTANCE_STATE: i64 = 2_048i64

pub const PROPERTY_USAGE_RESTART_IF_CHANGED: i64 = 4_096i64

pub const PROPERTY_USAGE_SCRIPT_VARIABLE: i64 = 8_192i64

pub const SPKEY: i64 = 16_777_216i64

pub const KEY_ESCAPE: i64 = 16_777_217i64

pub const KEY_TAB: i64 = 16_777_218i64

pub const KEY_BACKTAB: i64 = 16_777_219i64

pub const KEY_BACKSPACE: i64 = 16_777_220i64

pub const KEY_ENTER: i64 = 16_777_221i64

pub const KEY_KP_ENTER: i64 = 16_777_222i64

pub const KEY_INSERT: i64 = 16_777_223i64

pub const KEY_DELETE: i64 = 16_777_224i64

pub const KEY_PAUSE: i64 = 16_777_225i64

pub const KEY_PRINT: i64 = 16_777_226i64

pub const KEY_SYSREQ: i64 = 16_777_227i64

pub const KEY_CLEAR: i64 = 16_777_228i64

pub const KEY_HOME: i64 = 16_777_229i64

pub const KEY_END: i64 = 16_777_230i64

pub const KEY_LEFT: i64 = 16_777_231i64

pub const KEY_UP: i64 = 16_777_232i64

pub const KEY_RIGHT: i64 = 16_777_233i64

pub const KEY_DOWN: i64 = 16_777_234i64

pub const KEY_PAGEUP: i64 = 16_777_235i64

pub const KEY_PAGEDOWN: i64 = 16_777_236i64

pub const KEY_SHIFT: i64 = 16_777_237i64

pub const KEY_CONTROL: i64 = 16_777_238i64

pub const KEY_META: i64 = 16_777_239i64

pub const KEY_ALT: i64 = 16_777_240i64

pub const KEY_CAPSLOCK: i64 = 16_777_241i64

pub const KEY_NUMLOCK: i64 = 16_777_242i64

pub const KEY_SCROLLLOCK: i64 = 16_777_243i64

pub const KEY_F1: i64 = 16_777_244i64

pub const KEY_F2: i64 = 16_777_245i64

pub const KEY_F3: i64 = 16_777_246i64

pub const KEY_F4: i64 = 16_777_247i64

pub const KEY_F5: i64 = 16_777_248i64

pub const KEY_F6: i64 = 16_777_249i64

pub const KEY_F7: i64 = 16_777_250i64

pub const KEY_F8: i64 = 16_777_251i64

pub const KEY_F9: i64 = 16_777_252i64

pub const KEY_F10: i64 = 16_777_253i64

pub const KEY_F11: i64 = 16_777_254i64

pub const KEY_F12: i64 = 16_777_255i64

pub const KEY_F13: i64 = 16_777_256i64

pub const KEY_F14: i64 = 16_777_257i64

pub const KEY_F15: i64 = 16_777_258i64

pub const KEY_F16: i64 = 16_777_259i64

pub const KEY_SUPER_L: i64 = 16_777_260i64

pub const KEY_SUPER_R: i64 = 16_777_261i64

pub const KEY_MENU: i64 = 16_777_262i64

pub const KEY_HYPER_L: i64 = 16_777_263i64

pub const KEY_HYPER_R: i64 = 16_777_264i64

pub const KEY_HELP: i64 = 16_777_265i64

pub const KEY_DIRECTION_L: i64 = 16_777_266i64

pub const KEY_DIRECTION_R: i64 = 16_777_267i64

pub const KEY_BACK: i64 = 16_777_280i64

pub const KEY_FORWARD: i64 = 16_777_281i64

pub const KEY_STOP: i64 = 16_777_282i64

pub const KEY_REFRESH: i64 = 16_777_283i64

pub const KEY_VOLUMEDOWN: i64 = 16_777_284i64

pub const KEY_VOLUMEMUTE: i64 = 16_777_285i64

pub const KEY_VOLUMEUP: i64 = 16_777_286i64

pub const KEY_BASSBOOST: i64 = 16_777_287i64

pub const KEY_BASSUP: i64 = 16_777_288i64

pub const KEY_BASSDOWN: i64 = 16_777_289i64

pub const KEY_TREBLEUP: i64 = 16_777_290i64

pub const KEY_TREBLEDOWN: i64 = 16_777_291i64

pub const KEY_MEDIAPLAY: i64 = 16_777_292i64

pub const KEY_MEDIASTOP: i64 = 16_777_293i64

pub const KEY_MEDIAPREVIOUS: i64 = 16_777_294i64

pub const KEY_MEDIANEXT: i64 = 16_777_295i64

pub const KEY_MEDIARECORD: i64 = 16_777_296i64

pub const KEY_HOMEPAGE: i64 = 16_777_297i64

pub const KEY_FAVORITES: i64 = 16_777_298i64

pub const KEY_SEARCH: i64 = 16_777_299i64

pub const KEY_STANDBY: i64 = 16_777_300i64

pub const KEY_OPENURL: i64 = 16_777_301i64

pub const KEY_LAUNCHMAIL: i64 = 16_777_302i64

pub const KEY_LAUNCHMEDIA: i64 = 16_777_303i64

pub const KEY_LAUNCH0: i64 = 16_777_304i64

pub const KEY_LAUNCH1: i64 = 16_777_305i64

pub const KEY_LAUNCH2: i64 = 16_777_306i64

pub const KEY_LAUNCH3: i64 = 16_777_307i64

pub const KEY_LAUNCH4: i64 = 16_777_308i64

pub const KEY_LAUNCH5: i64 = 16_777_309i64

pub const KEY_LAUNCH6: i64 = 16_777_310i64

pub const KEY_LAUNCH7: i64 = 16_777_311i64

pub const KEY_LAUNCH8: i64 = 16_777_312i64

pub const KEY_LAUNCH9: i64 = 16_777_313i64

pub const KEY_LAUNCHA: i64 = 16_777_314i64

pub const KEY_LAUNCHB: i64 = 16_777_315i64

pub const KEY_LAUNCHC: i64 = 16_777_316i64

pub const KEY_LAUNCHD: i64 = 16_777_317i64

pub const KEY_LAUNCHE: i64 = 16_777_318i64

pub const KEY_LAUNCHF: i64 = 16_777_319i64

pub const KEY_KP_MULTIPLY: i64 = 16_777_345i64

pub const KEY_KP_DIVIDE: i64 = 16_777_346i64

pub const KEY_KP_SUBTRACT: i64 = 16_777_347i64

pub const KEY_KP_PERIOD: i64 = 16_777_348i64

pub const KEY_KP_ADD: i64 = 16_777_349i64

pub const KEY_KP_0: i64 = 16_777_350i64

pub const KEY_KP_1: i64 = 16_777_351i64

pub const KEY_KP_2: i64 = 16_777_352i64

pub const KEY_KP_3: i64 = 16_777_353i64

pub const KEY_KP_4: i64 = 16_777_354i64

pub const KEY_KP_5: i64 = 16_777_355i64

pub const KEY_KP_6: i64 = 16_777_356i64

pub const KEY_KP_7: i64 = 16_777_357i64

pub const KEY_KP_8: i64 = 16_777_358i64

pub const KEY_KP_9: i64 = 16_777_359i64

pub const KEY_CODE_MASK: i64 = 33_554_431i64

pub const KEY_UNKNOWN: i64 = 33_554_431i64

pub const KEY_MASK_SHIFT: i64 = 33_554_432i64

pub const KEY_MASK_ALT: i64 = 67_108_864i64

pub const KEY_MASK_META: i64 = 134_217_728i64

pub const KEY_MASK_CMD: i64 = 268_435_456i64

pub const KEY_MASK_CTRL: i64 = 268_435_456i64

pub const KEY_MASK_KPAD: i64 = 536_870_912i64

pub const KEY_MASK_GROUP_SWITCH: i64 = 1_073_741_824i64

impl GlobalConstants

pub fn godot_singleton() -> &'static GlobalConstants

Returns a reference to the singleton instance.