Struct gdnative::api::OS

pub struct OS { /* fields omitted */ }

core singleton class OS inherits Object (unsafe).

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Class hierarchy

OS inherits methods from:

Object

Safety

All types in the Godot API have “interior mutability” in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Access in these types tracks whether the access is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl OS

Constants

pub const APPLICATION_HANDLE: i64

pub const DAY_SUNDAY: i64

pub const POWERSTATE_UNKNOWN: i64

pub const SCREEN_ORIENTATION_LANDSCAPE: i64

pub const SYSTEM_DIR_DESKTOP: i64

pub const VIDEO_DRIVER_GLES3: i64

pub const DAY_MONDAY: i64

pub const DISPLAY_HANDLE: i64

pub const MONTH_JANUARY: i64

pub const POWERSTATE_ON_BATTERY: i64

pub const SCREEN_ORIENTATION_PORTRAIT: i64

pub const SYSTEM_DIR_DCIM: i64

pub const VIDEO_DRIVER_GLES2: i64

pub const DAY_TUESDAY: i64

pub const MONTH_FEBRUARY: i64

pub const POWERSTATE_NO_BATTERY: i64

pub const SCREEN_ORIENTATION_REVERSE_LANDSCAPE: i64

pub const SYSTEM_DIR_DOCUMENTS: i64

pub const WINDOW_HANDLE: i64

pub const DAY_WEDNESDAY: i64

pub const MONTH_MARCH: i64

pub const POWERSTATE_CHARGING: i64

pub const SCREEN_ORIENTATION_REVERSE_PORTRAIT: i64

pub const SYSTEM_DIR_DOWNLOADS: i64

pub const WINDOW_VIEW: i64

pub const DAY_THURSDAY: i64

pub const MONTH_APRIL: i64

pub const OPENGL_CONTEXT: i64

pub const POWERSTATE_CHARGED: i64

pub const SCREEN_ORIENTATION_SENSOR_LANDSCAPE: i64

pub const SYSTEM_DIR_MOVIES: i64

pub const DAY_FRIDAY: i64

pub const MONTH_MAY: i64

pub const SCREEN_ORIENTATION_SENSOR_PORTRAIT: i64

pub const SYSTEM_DIR_MUSIC: i64

pub const DAY_SATURDAY: i64

pub const MONTH_JUNE: i64

pub const SCREEN_ORIENTATION_SENSOR: i64

pub const SYSTEM_DIR_PICTURES: i64

pub const MONTH_JULY: i64

pub const SYSTEM_DIR_RINGTONES: i64

pub const MONTH_AUGUST: i64

pub const MONTH_SEPTEMBER: i64

pub const MONTH_OCTOBER: i64

pub const MONTH_NOVEMBER: i64

pub const MONTH_DECEMBER: i64

impl OS

pub fn godot_singleton() -> &'static OS

Returns a reference to the singleton instance.

pub fn alert(&self, text: impl Into<GodotString>, title: impl Into<GodotString>)

Displays a modal dialog box using the host OS’ facilities. Execution is blocked until the dialog is closed.

Default Arguments

title - "Alert!"

pub fn can_draw(&self) -> bool

Returns true if the host OS allows drawing.

pub fn can_use_threads(&self) -> bool

Returns true if the current host platform is using multiple threads.

pub fn center_window(&self)

Centers the window on the screen if in windowed mode.

pub fn close_midi_inputs(&self)

Shuts down system MIDI driver. Note: This method is implemented on Linux, macOS and Windows.

pub fn delay_msec(&self, msec: i64)

Delay execution of the current thread by msec milliseconds. usec must be greater than or equal to 0. Otherwise, delay_msec will do nothing and will print an error message.

pub fn delay_usec(&self, usec: i64)

Delay execution of the current thread by usec microseconds. usec must be greater than or equal to 0. Otherwise, delay_usec will do nothing and will print an error message.

pub fn dump_memory_to_file(&self, file: impl Into<GodotString>)

Dumps the memory allocation ringlist to a file (only works in debug). Entry format per line: “Address - Size - Description”.

pub fn dump_resources_to_file(&self, file: impl Into<GodotString>)

Dumps all used resources to file (only works in debug). Entry format per line: “Resource Type : Resource Location”. At the end of the file is a statistic of all used Resource Types.

pub fn execute(
&self,
path: impl Into<GodotString>,
arguments: TypedArray<GodotString>,
blocking: bool,
output: VariantArray<Shared>,
read_stderr: bool
) -> i64

Sample code is GDScript unless otherwise noted.

Execute the file at the given path with the arguments passed as an array of strings. Platform path resolution will take place. The resolved file must exist and be executable. The arguments are used in the given order and separated by a space, so OS.execute("ping", ["-w", "3", "godotengine.org"], false) will resolve to ping -w 3 godotengine.org in the system’s shell. This method has slightly different behavior based on whether the blocking mode is enabled. If blocking is true, the Godot thread will pause its execution while waiting for the process to terminate. The shell output of the process will be written to the output array as a single string. When the process terminates, the Godot thread will resume execution. If blocking is false, the Godot thread will continue while the new process runs. It is not possible to retrieve the shell output in non-blocking mode, so output will be empty. The return value also depends on the blocking mode. When blocking, the method will return an exit code of the process. When non-blocking, the method returns a process ID, which you can use to monitor the process (and potentially terminate it with kill). If the process forking (non-blocking) or opening (blocking) fails, the method will return -1 or another exit code. Example of blocking mode and retrieving the shell output:

var output = []
var exit_code = OS.execute("ls", ["-l", "/tmp"], true, output)

Example of non-blocking mode, running another instance of the project and storing its process ID:

var pid = OS.execute(OS.get_executable_path(), [], false)

If you wish to access a shell built-in or perform a composite command, a platform-specific shell can be invoked. For example:

OS.execute("CMD.exe", ["/C", "cd %TEMP% && dir"], true, output)

Note: This method is implemented on Android, iOS, Linux, macOS and Windows.

Default Arguments

blocking - true

output - [ ]

read_stderr - false

pub fn find_scancode_from_string(&self, string: impl Into<GodotString>) -> i64

Returns the scancode of the given string (e.g. “Escape”).

pub fn get_audio_driver_count(&self) -> i64

Returns the total number of available audio drivers.

pub fn get_audio_driver_name(&self, driver: i64) -> GodotString

Returns the audio driver name for the given index.

pub fn borderless_window(&self) -> bool

If true, removes the window frame. Note: Setting window_borderless to false disables per-pixel transparency.

pub fn clipboard(&self) -> GodotString

The clipboard from the host OS. Might be unavailable on some platforms.

pub fn get_cmdline_args(&self) -> TypedArray<GodotString>

Sample code is GDScript unless otherwise noted.

Returns the command-line arguments passed to the engine. Command-line arguments can be written in any form, including both --key value and --key=value forms so they can be properly parsed, as long as custom command-line arguments do not conflict with engine arguments. You can also incorporate environment variables using the get_environment method. You can set editor/main_run_args in the Project Settings to define command-line arguments to be passed by the editor when running the project. Here’s a minimal example on how to parse command-line arguments into a dictionary using the --key=value form for arguments:

var arguments = {}
for argument in OS.get_cmdline_args():
    if argument.find("=") > -1:
        var key_value = argument.split("=")
        arguments[key_value[`0`][0].lstrip("--")] = key_value[`1`][1]

pub fn get_connected_midi_inputs(&self) -> TypedArray<GodotString>

Returns an array of MIDI device names. The returned array will be empty if the system MIDI driver has not previously been initialised with open_midi_inputs. Note: This method is implemented on Linux, macOS and Windows.

pub fn current_screen(&self) -> i64

The current screen index (starting from 0).

pub fn current_tablet_driver(&self) -> GodotString

The current tablet driver in use.

pub fn get_current_video_driver(&self) -> VideoDriver

Returns the currently used video driver, using one of the values from VideoDriver.

pub fn get_date(&self, utc: bool) -> Dictionary<Shared>

Returns current date as a dictionary of keys: year, month, day, weekday, dst (Daylight Savings Time).

Default Arguments

utc - false

pub fn get_datetime(&self, utc: bool) -> Dictionary<Shared>

Returns current datetime as a dictionary of keys: year, month, day, weekday, dst (Daylight Savings Time), hour, minute, second.

Default Arguments

utc - false

pub fn get_datetime_from_unix_time(
&self,
unix_time_val: i64
) -> Dictionary<Shared>

Gets a dictionary of time values corresponding to the given UNIX epoch time (in seconds). The returned Dictionary’s values will be the same as get_datetime, with the exception of Daylight Savings Time as it cannot be determined from the epoch.

pub fn get_dynamic_memory_usage(&self) -> i64

Returns the total amount of dynamic memory used (only works in debug).

pub fn get_environment(&self, variable: impl Into<GodotString>) -> GodotString

Returns the value of an environment variable. Returns an empty string if the environment variable doesn’t exist. Note: Double-check the casing of variable. Environment variable names are case-sensitive on all platforms except Windows.

pub fn get_executable_path(&self) -> GodotString

Returns the path to the current engine executable.

pub fn exit_code(&self) -> i64

The exit code passed to the OS when the main loop exits. By convention, an exit code of 0 indicates success whereas a non-zero exit code indicates an error. For portability reasons, the exit code should be set between 0 and 125 (inclusive). Note: This value will be ignored if using SceneTree.quit with an exit_code argument passed.

pub fn get_granted_permissions(&self) -> TypedArray<GodotString>

With this function you can get the list of dangerous permissions that have been granted to the Android application. Note: This method is implemented on Android.

pub fn get_ime_selection(&self) -> Vector2

Returns the IME cursor position (the currently-edited portion of the string) relative to the characters in the composition string. MainLoop.NOTIFICATION_OS_IME_UPDATE is sent to the application to notify it of changes to the IME cursor position. Note: This method is implemented on macOS.

pub fn get_ime_text(&self) -> GodotString

Returns the IME intermediate composition string. MainLoop.NOTIFICATION_OS_IME_UPDATE is sent to the application to notify it of changes to the IME composition string. Note: This method is implemented on macOS.

pub fn get_latin_keyboard_variant(&self) -> GodotString

Returns the current latin keyboard variant as a String. Possible return values are: "QWERTY", "AZERTY", "QZERTY", "DVORAK", "NEO", "COLEMAK" or "ERROR". Note: This method is implemented on Linux, macOS and Windows. Returns "QWERTY" on unsupported platforms.

pub fn get_locale(&self) -> GodotString

Returns the host OS locale.

pub fn low_processor_usage_mode_sleep_usec(&self) -> i64

The amount of sleeping between frames when the low-processor usage mode is enabled (in microseconds). Higher values will result in lower CPU usage.

pub fn max_window_size(&self) -> Vector2

The maximum size of the window (without counting window manager decorations). Does not affect fullscreen mode. Set to (0, 0) to reset to the system default value.

pub fn min_window_size(&self) -> Vector2

The minimum size of the window (without counting window manager decorations). Does not affect fullscreen mode. Set to (0, 0) to reset to the system default value.

pub fn get_model_name(&self) -> GodotString

Returns the model name of the current device. Note: This method is implemented on Android and iOS. Returns "GenericDevice" on unsupported platforms.

pub fn get_name(&self) -> GodotString

Returns the name of the host OS. Possible values are: "Android", "iOS", "HTML5", "OSX", "Server", "Windows", "UWP", "X11".

pub fn get_native_handle(&self, handle_type: i64) -> i64

Returns internal structure pointers for use in GDNative plugins. Note: This method is implemented on Linux and Windows (other OSs will soon be supported).

pub fn get_power_percent_left(&self) -> i64

Returns the amount of battery left in the device as a percentage. Returns -1 if power state is unknown. Note: This method is implemented on Linux, macOS and Windows.

pub fn get_power_seconds_left(&self) -> i64

Returns an estimate of the time left in seconds before the device runs out of battery. Returns -1 if power state is unknown. Note: This method is implemented on Linux, macOS and Windows.

pub fn get_power_state(&self) -> PowerState

Returns the current state of the device regarding battery and power. See PowerState constants. Note: This method is implemented on Linux, macOS and Windows.

pub fn get_process_id(&self) -> i64

Returns the project’s process ID. Note: This method is implemented on Android, iOS, Linux, macOS and Windows.

pub fn get_processor_count(&self) -> i64

Returns the number of threads available on the host machine.

pub fn get_real_window_size(&self) -> Vector2

Returns the window size including decorations like window borders.

pub fn get_scancode_string(&self, code: i64) -> GodotString

Returns the given scancode as a string (e.g. Return values: "Escape", "Shift+Escape"). See also InputEventKey.scancode and InputEventKey.get_scancode_with_modifiers.

pub fn get_screen_count(&self) -> i64

Returns the number of displays attached to the host machine.

pub fn get_screen_dpi(&self, screen: i64) -> i64

Sample code is GDScript unless otherwise noted.

Returns the dots per inch density of the specified screen. If screen is -1 (the default value), the current screen will be used. Note: On macOS, returned value is inaccurate if fractional display scaling mode is used. Note: On Android devices, the actual screen densities are grouped into six generalized densities:

   ldpi - 120 dpi
   mdpi - 160 dpi
   hdpi - 240 dpi
  xhdpi - 320 dpi
 xxhdpi - 480 dpi
xxxhdpi - 640 dpi

Note: This method is implemented on Android, Linux, macOS and Windows. Returns 72 on unsupported platforms.

Default Arguments

screen - -1

pub fn get_screen_max_scale(&self) -> f64

Return the greatest scale factor of all screens. Note: On macOS returned value is 2.0 if there is at least one hiDPI (Retina) screen in the system, and 1.0 in all other cases. Note: This method is implemented on macOS.

pub fn screen_orientation(&self) -> ScreenOrientation

The current screen orientation.

pub fn get_screen_position(&self, screen: i64) -> Vector2

Returns the position of the specified screen by index. If screen is -1 (the default value), the current screen will be used.

Default Arguments

screen - -1

pub fn get_screen_scale(&self, screen: i64) -> f64

Return the scale factor of the specified screen by index. If screen is -1 (the default value), the current screen will be used. Note: On macOS returned value is 2.0 for hiDPI (Retina) screen, and 1.0 for all other cases. Note: This method is implemented on macOS.

Default Arguments

screen - -1

pub fn get_screen_size(&self, screen: i64) -> Vector2

Returns the dimensions in pixels of the specified screen. If screen is -1 (the default value), the current screen will be used.

Default Arguments

screen - -1

pub fn get_splash_tick_msec(&self) -> i64

Returns the amount of time in milliseconds it took for the boot logo to appear.

pub fn get_static_memory_peak_usage(&self) -> i64

Returns the maximum amount of static memory used (only works in debug).

pub fn get_static_memory_usage(&self) -> i64

Returns the amount of static memory being used by the program in bytes.

pub fn get_system_dir(&self, dir: i64) -> GodotString

Returns the actual path to commonly used folders across different platforms. Available locations are specified in SystemDir. Note: This method is implemented on Android, Linux, macOS and Windows.

pub fn get_system_time_msecs(&self) -> i64

Returns the epoch time of the operating system in milliseconds.

pub fn get_system_time_secs(&self) -> i64

Returns the epoch time of the operating system in seconds.

pub fn get_tablet_driver_count(&self) -> i64

Returns the total number of available tablet drivers. Note: This method is implemented on Windows.

pub fn get_tablet_driver_name(&self, idx: i64) -> GodotString

Returns the tablet driver name for the given index. Note: This method is implemented on Windows.

pub fn get_thread_caller_id(&self) -> i64

Returns the ID of the current thread. This can be used in logs to ease debugging of multi-threaded applications. Note: Thread IDs are not deterministic and may be reused across application restarts.

pub fn get_ticks_msec(&self) -> i64

Returns the amount of time passed in milliseconds since the engine started.

pub fn get_ticks_usec(&self) -> i64

Returns the amount of time passed in microseconds since the engine started.

pub fn get_time(&self, utc: bool) -> Dictionary<Shared>

Returns current time as a dictionary of keys: hour, minute, second.

Default Arguments

utc - false

pub fn get_time_zone_info(&self) -> Dictionary<Shared>

Returns the current time zone as a dictionary with the keys: bias and name.

pub fn get_unique_id(&self) -> GodotString

Returns a string that is unique to the device. Note: Returns an empty string on HTML5 and UWP, as this method isn’t implemented on those platforms yet.

pub fn get_unix_time(&self) -> i64

Returns the current UNIX epoch timestamp in seconds. Important: This is the system clock that the user can manully set. Never use this method for precise time calculation since its results are also subject to automatic adjustments by the operating system. Always use get_ticks_usec or get_ticks_msec for precise time calculation instead, since they are guaranteed to be monotonic (i.e. never decrease).

pub fn get_unix_time_from_datetime(&self, datetime: Dictionary<Shared>) -> i64

Gets an epoch time value from a dictionary of time values. datetime must be populated with the following keys: year, month, day, hour, minute, second. If the dictionary is empty 0 is returned. You can pass the output from get_datetime_from_unix_time directly into this function. Daylight Savings Time (dst), if present, is ignored.

pub fn get_user_data_dir(&self) -> GodotString

Returns the absolute directory path where user data is written (user://). On Linux, this is ~/.local/share/godot/app_userdata/[project_name][project_name], or ~/.local/share/[custom_name][custom_name] if use_custom_user_dir is set. On macOS, this is ~/Library/Application Support/Godot/app_userdata/[project_name][project_name], or ~/Library/Application Support/[custom_name][custom_name] if use_custom_user_dir is set. On Windows, this is %APPDATA%\Godot\app_userdata\[project_name][project_name], or %APPDATA%\[custom_name][custom_name] if use_custom_user_dir is set. %APPDATA% expands to %USERPROFILE%\AppData\Roaming. If the project name is empty, user:// falls back to res://.

pub fn get_video_driver_count(&self) -> i64

Returns the number of video drivers supported on the current platform.

pub fn get_video_driver_name(&self, driver: i64) -> GodotString

Returns the name of the video driver matching the given driver index. This index is a value from VideoDriver, and you can use get_current_video_driver to get the current backend’s index.

pub fn get_virtual_keyboard_height(&self) -> i64

Returns the on-screen keyboard’s height in pixels. Returns 0 if there is no keyboard or if it is currently hidden.

pub fn window_per_pixel_transparency_enabled(&self) -> bool

If true, the window background is transparent and window frame is removed. Use get_tree().get_root().set_transparent_background(true) to disable main viewport background rendering. Note: This property has no effect if Project > Project Settings > Display > Window > Per-pixel transparency > Allowed setting is disabled. Note: This property is implemented on HTML5, Linux, macOS and Windows.

pub fn window_position(&self) -> Vector2

The window position relative to the screen, the origin is the top left corner, +Y axis goes to the bottom and +X axis goes to the right.

pub fn get_window_safe_area(&self) -> Rect2

Returns unobscured area of the window where interactive controls should be rendered.

pub fn window_size(&self) -> Vector2

The size of the window (without counting window manager decorations).

pub fn global_menu_add_item(
&self,
menu: impl Into<GodotString>,
label: impl Into<GodotString>,
id: impl OwnedToVariant,
meta: impl OwnedToVariant
)

Add a new item with text “label” to global menu. Use “_dock” menu to add item to the macOS dock icon menu. Note: This method is implemented on macOS.

pub fn global_menu_add_separator(&self, menu: impl Into<GodotString>)

Add a separator between items. Separators also occupy an index. Note: This method is implemented on macOS.

pub fn global_menu_clear(&self, menu: impl Into<GodotString>)

Clear the global menu, in effect removing all items. Note: This method is implemented on macOS.

pub fn global_menu_remove_item(&self, menu: impl Into<GodotString>, idx: i64)

Removes the item at index “idx” from the global menu. Note that the indexes of items after the removed item are going to be shifted by one. Note: This method is implemented on macOS.

pub fn has_environment(&self, variable: impl Into<GodotString>) -> bool

Returns true if the environment variable with the name variable exists. Note: Double-check the casing of variable. Environment variable names are case-sensitive on all platforms except Windows.

pub fn has_feature(&self, tag_name: impl Into<GodotString>) -> bool

Returns true if the feature for the given feature tag is supported in the currently running instance, depending on platform, build etc. Can be used to check whether you’re currently running a debug build, on a certain platform or arch, etc. Refer to the Feature Tags documentation for more details. Note: Tag names are case-sensitive.

pub fn has_touchscreen_ui_hint(&self) -> bool

Returns true if the device has a touchscreen or emulates one.

pub fn has_virtual_keyboard(&self) -> bool

Returns true if the platform has a virtual keyboard, false otherwise.

pub fn hide_virtual_keyboard(&self)

Hides the virtual keyboard if it is shown, does nothing otherwise.

pub fn is_debug_build(&self) -> bool

Returns true if the Godot binary used to run the project is a debug export template, or when running in the editor. Returns false if the Godot binary used to run the project is a release export template. To check whether the Godot binary used to run the project is an export template (debug or release), use OS.has_feature("standalone") instead.

pub fn is_in_low_processor_usage_mode(&self) -> bool

If true, the engine optimizes for low processor usage by only refreshing the screen if needed. Can improve battery consumption on mobile.

pub fn is_keep_screen_on(&self) -> bool

If true, the engine tries to keep the screen on while the game is running. Useful on mobile.

pub fn is_ok_left_and_cancel_right(&self) -> bool

Returns true if the OK button should appear on the left and Cancel on the right.

pub fn is_scancode_unicode(&self, code: i64) -> bool

Returns true if the input scancode corresponds to a Unicode character.

pub fn is_stdout_verbose(&self) -> bool

Returns true if the engine was executed with -v (verbose stdout).

pub fn is_userfs_persistent(&self) -> bool

If true, the user:// file system is persistent, so that its state is the same after a player quits and starts the game again. Relevant to the HTML5 platform, where this persistence may be unavailable.

pub fn is_vsync_enabled(&self) -> bool

If true, vertical synchronization (Vsync) is enabled.

pub fn is_vsync_via_compositor_enabled(&self) -> bool

If true and vsync_enabled is true, the operating system’s window compositor will be used for vsync when the compositor is enabled and the game is in windowed mode. Note: This option is experimental and meant to alleviate stutter experienced by some users. However, some users have experienced a Vsync framerate halving (e.g. from 60 FPS to 30 FPS) when using it. Note: This property is only implemented on Windows.

pub fn is_window_always_on_top(&self) -> bool

Returns true if the window should always be on top of other windows.

pub fn is_window_focused(&self) -> bool

Returns true if the window is currently focused. Note: Only implemented on desktop platforms. On other platforms, it will always return true.

pub fn is_window_fullscreen(&self) -> bool

If true, the window is fullscreen.

pub fn is_window_maximized(&self) -> bool

If true, the window is maximized.

pub fn is_window_minimized(&self) -> bool

If true, the window is minimized.

pub fn is_window_resizable(&self) -> bool

If true, the window is resizable by the user.

pub fn keyboard_get_current_layout(&self) -> i64

Returns active keyboard layout index. Note: This method is implemented on Linux, macOS and Windows.

pub fn keyboard_get_layout_count(&self) -> i64

Returns the number of keyboard layouts. Note: This method is implemented on Linux, macOS and Windows.

pub fn keyboard_get_layout_language(&self, index: i64) -> GodotString

Returns the ISO-639/BCP-47 language code of the keyboard layout at position index. Note: This method is implemented on Linux, macOS and Windows.

pub fn keyboard_get_layout_name(&self, index: i64) -> GodotString

Returns the localized name of the keyboard layout at position index. Note: This method is implemented on Linux, macOS and Windows.

pub fn keyboard_set_current_layout(&self, index: i64)

Sets active keyboard layout. Note: This method is implemented on Linux, macOS and Windows.

pub fn kill(&self, pid: i64) -> Result<(), GodotError>

Kill (terminate) the process identified by the given process ID (pid), e.g. the one returned by execute in non-blocking mode. Note: This method can also be used to kill processes that were not spawned by the game. Note: This method is implemented on Android, iOS, Linux, macOS and Windows.

pub fn move_window_to_foreground(&self)

Moves the window to the front. Note: This method is implemented on Linux, macOS and Windows.

pub fn native_video_is_playing(&self) -> bool

Returns true if native video is playing. Note: This method is implemented on Android and iOS.

pub fn native_video_pause(&self)

Pauses native video playback. Note: This method is implemented on Android and iOS.

pub fn native_video_play(
&self,
path: impl Into<GodotString>,
volume: f64,
audio_track: impl Into<GodotString>,
subtitle_track: impl Into<GodotString>
) -> Result<(), GodotError>

Plays native video from the specified path, at the given volume and with audio and subtitle tracks. Note: This method is implemented on Android and iOS, and the current Android implementation does not support the volume, audio_track and subtitle_track options.

pub fn native_video_stop(&self)

Stops native video playback. Note: This method is implemented on Android and iOS.

pub fn native_video_unpause(&self)

Resumes native video playback. Note: This method is implemented on Android and iOS.

pub fn open_midi_inputs(&self)

Initialises the singleton for the system MIDI driver. Note: This method is implemented on Linux, macOS and Windows.

pub fn print_all_resources(&self, tofile: impl Into<GodotString>)

Shows all resources in the game. Optionally, the list can be written to a file by specifying a file path in tofile.

Default Arguments

tofile - ""

pub fn print_all_textures_by_size(&self)

Shows the list of loaded textures sorted by size in memory.

pub fn print_resources_by_type(&self, types: TypedArray<GodotString>)

Shows the number of resources loaded by the game of the given types.

pub fn print_resources_in_use(&self, short: bool)

Shows all resources currently used by the game.

Default Arguments

short - false

pub fn request_attention(&self)

Request the user attention to the window. It’ll flash the taskbar button on Windows or bounce the dock icon on OSX. Note: This method is implemented on Linux, macOS and Windows.

pub fn request_permission(&self, name: impl Into<GodotString>) -> bool

At the moment this function is only used by AudioDriverOpenSL to request permission for RECORD_AUDIO on Android.

pub fn request_permissions(&self) -> bool

With this function you can request dangerous permissions since normal permissions are automatically granted at install time in Android application. Note: This method is implemented on Android.

pub fn set_borderless_window(&self, borderless: bool)

If true, removes the window frame. Note: Setting window_borderless to false disables per-pixel transparency.

pub fn set_clipboard(&self, clipboard: impl Into<GodotString>)

The clipboard from the host OS. Might be unavailable on some platforms.

pub fn set_current_screen(&self, screen: i64)

The current screen index (starting from 0).

pub fn set_current_tablet_driver(&self, name: impl Into<GodotString>)

The current tablet driver in use.

pub fn set_environment(
&self,
variable: impl Into<GodotString>,
value: impl Into<GodotString>
) -> bool

Sets the value of the environment variable variable to value. The environment variable will be set for the Godot process and any process executed with execute after running set_environment. The environment variable will not persist to processes run after the Godot process was terminated. Note: Double-check the casing of variable. Environment variable names are case-sensitive on all platforms except Windows.

pub fn set_exit_code(&self, code: i64)

The exit code passed to the OS when the main loop exits. By convention, an exit code of 0 indicates success whereas a non-zero exit code indicates an error. For portability reasons, the exit code should be set between 0 and 125 (inclusive). Note: This value will be ignored if using SceneTree.quit with an exit_code argument passed.

pub fn set_icon(&self, icon: impl AsArg<Image>)

Sets the game’s icon using an Image resource. The same image is used for window caption, taskbar/dock and window selection dialog. Image is scaled as needed. Note: This method is implemented on HTML5, Linux, macOS and Windows.

pub fn set_ime_active(&self, active: bool)

Sets whether IME input mode should be enabled. If active IME handles key events before the application and creates an composition string and suggestion list. Application can retrieve the composition status by using get_ime_selection and get_ime_text functions. Completed composition string is committed when input is finished. Note: This method is implemented on Linux, macOS and Windows.

pub fn set_ime_position(&self, position: Vector2)

Sets position of IME suggestion list popup (in window coordinates). Note: This method is implemented on Linux, macOS and Windows.

pub fn set_keep_screen_on(&self, enabled: bool)

If true, the engine tries to keep the screen on while the game is running. Useful on mobile.

pub fn set_low_processor_usage_mode(&self, enable: bool)

If true, the engine optimizes for low processor usage by only refreshing the screen if needed. Can improve battery consumption on mobile.

pub fn set_low_processor_usage_mode_sleep_usec(&self, usec: i64)

The amount of sleeping between frames when the low-processor usage mode is enabled (in microseconds). Higher values will result in lower CPU usage.

pub fn set_max_window_size(&self, size: Vector2)

The maximum size of the window (without counting window manager decorations). Does not affect fullscreen mode. Set to (0, 0) to reset to the system default value.

pub fn set_min_window_size(&self, size: Vector2)

The minimum size of the window (without counting window manager decorations). Does not affect fullscreen mode. Set to (0, 0) to reset to the system default value.

pub fn set_native_icon(&self, filename: impl Into<GodotString>)

Sets the game’s icon using a multi-size platform-specific icon file (*.ico on Windows and *.icns on macOS). Appropriate size sub-icons are used for window caption, taskbar/dock and window selection dialog. Note: This method is implemented on macOS and Windows.

pub fn set_screen_orientation(&self, orientation: i64)

The current screen orientation.

pub fn set_thread_name(
&self,
name: impl Into<GodotString>
) -> Result<(), GodotError>

Sets the name of the current thread.

pub fn set_use_file_access_save_and_swap(&self, enabled: bool)

Enables backup saves if enabled is true.

pub fn set_use_vsync(&self, enable: bool)

If true, vertical synchronization (Vsync) is enabled.

pub fn set_vsync_via_compositor(&self, enable: bool)

If true and vsync_enabled is true, the operating system’s window compositor will be used for vsync when the compositor is enabled and the game is in windowed mode. Note: This option is experimental and meant to alleviate stutter experienced by some users. However, some users have experienced a Vsync framerate halving (e.g. from 60 FPS to 30 FPS) when using it. Note: This property is only implemented on Windows.

pub fn set_window_always_on_top(&self, enabled: bool)

Sets whether the window should always be on top. Note: This method is implemented on Linux, macOS and Windows.

pub fn set_window_fullscreen(&self, enabled: bool)

If true, the window is fullscreen.

pub fn set_window_maximized(&self, enabled: bool)

If true, the window is maximized.

pub fn set_window_minimized(&self, enabled: bool)

If true, the window is minimized.

pub fn set_window_mouse_passthrough(&self, region: TypedArray<Vector2>)

Sample code is GDScript unless otherwise noted.

Sets a polygonal region of the window which accepts mouse events. Mouse events outside the region will be passed through. Passing an empty array will disable passthrough support (all mouse events will be intercepted by the window, which is the default behavior).

OS.set_window_mouse_passthrough($Path2D.curve.get_baked_points())

OS.set_window_mouse_passthrough($Polygon2D.polygon)

OS.set_window_mouse_passthrough([])

Note: On Windows, the portion of a window that lies outside the region is not drawn, while on Linux and macOS it is. Note: This method is implemented on Linux, macOS and Windows.

pub fn set_window_per_pixel_transparency_enabled(&self, enabled: bool)

If true, the window background is transparent and window frame is removed. Use get_tree().get_root().set_transparent_background(true) to disable main viewport background rendering. Note: This property has no effect if Project > Project Settings > Display > Window > Per-pixel transparency > Allowed setting is disabled. Note: This property is implemented on HTML5, Linux, macOS and Windows.

pub fn set_window_position(&self, position: Vector2)

The window position relative to the screen, the origin is the top left corner, +Y axis goes to the bottom and +X axis goes to the right.

pub fn set_window_resizable(&self, enabled: bool)

If true, the window is resizable by the user.

pub fn set_window_size(&self, size: Vector2)

The size of the window (without counting window manager decorations).

pub fn set_window_title(&self, title: impl Into<GodotString>)

Sets the window title to the specified string. Note: This should be used sporadically. Don’t set this every frame, as that will negatively affect performance on some window managers. Note: This method is implemented on HTML5, Linux, macOS and Windows.

pub fn shell_open(&self, uri: impl Into<GodotString>) -> Result<(), GodotError>

Requests the OS to open a resource with the most appropriate program. For example:

OS.shell_open("C:\\Users\name\Downloads") on Windows opens the file explorer at the user’s Downloads folder.

OS.shell_open("https://godotengine.org") opens the default web browser on the official Godot website.

OS.shell_open("mailto:example@example.com") opens the default email client with the “To” field set to example@example.com. See Customizing mailto: Links for a list of fields that can be added. Use ProjectSettings.globalize_path to convert a res:// or user:// path into a system path for use with this method. Note: This method is implemented on Android, iOS, HTML5, Linux, macOS and Windows.

pub fn show_virtual_keyboard(
&self,
existing_text: impl Into<GodotString>,
multiline: bool
)

Shows the virtual keyboard if the platform has one. The existing_text parameter is useful for implementing your own LineEdit or TextEdit, as it tells the virtual keyboard what text has already been typed (the virtual keyboard uses it for auto-correct and predictions). The multiline parameter needs to be set to true to be able to enter multiple lines of text, as in TextEdit. Note: This method is implemented on Android, iOS and UWP.

Default Arguments

existing_text - ""

multiline - false