Struct gdnative::api::Particles2D

pub struct Particles2D { /* private fields */ }

core class Particles2D inherits Node2D (manually managed).

This class has related types in the particles_2d module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

Particles2D is a reference-only type. Persistent references can only exist in the unsafe Ref<Particles2D> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

Particles2D inherits methods from:

Node2D

CanvasItem

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Particles2D

Constants

pub const DRAW_ORDER_INDEX: i64 = 0i64

pub const DRAW_ORDER_LIFETIME: i64 = 1i64

impl Particles2D

pub fn new() -> Ref<Particles2D, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn capture_rect(&self) -> Rect2

Returns a rectangle containing the positions of all existing particles.

pub fn amount(&self) -> i64

The number of particles emitted in one emission cycle (corresponding to the lifetime). Note: Changing amount will reset the particle emission, therefore removing all particles that were already emitted before changing amount.

pub fn draw_order(&self) -> DrawOrder

Particle draw order. Uses DrawOrder values.

pub fn explosiveness_ratio(&self) -> f64

How rapidly particles in an emission cycle are emitted. If greater than 0, there will be a gap in emissions before the next cycle begins.

pub fn fixed_fps(&self) -> i64

The particle system’s frame rate is fixed to a value. For instance, changing the value to 2 will make the particles render at 2 frames per second. Note this does not slow down the simulation of the particle system itself.

pub fn fractional_delta(&self) -> bool

If true, results in fractional delta calculation which has a smoother particles display effect.

pub fn lifetime(&self) -> f64

The amount of time each particle will exist (in seconds).

pub fn normal_map(&self) -> Option<Ref<Texture, Shared>>

Normal map to be used for the texture property. Note: Godot expects the normal map to use X+, Y-, and Z+ coordinates. See this page for a comparison of normal map coordinates expected by popular engines.

pub fn one_shot(&self) -> bool

If true, only one emission cycle occurs. If set true during a cycle, emission will stop at the cycle’s end.

pub fn pre_process_time(&self) -> f64

Particle system starts as if it had already run for this many seconds.

pub fn process_material(&self) -> Option<Ref<Material, Shared>>

Material for processing particles. Can be a ParticlesMaterial or a ShaderMaterial.

pub fn randomness_ratio(&self) -> f64

Emission lifetime randomness ratio.

pub fn speed_scale(&self) -> f64

Particle system’s running speed scaling ratio. A value of 0 can be used to pause the particles.

pub fn texture(&self) -> Option<Ref<Texture, Shared>>

Particle texture. If null, particles will be squares.

pub fn use_local_coordinates(&self) -> bool

If true, particles use the parent node’s coordinate space. If false, they use global coordinates.

pub fn visibility_rect(&self) -> Rect2

The Rect2 that determines the node’s region which needs to be visible on screen for the particle system to be active. Grow the rect if particles suddenly appear/disappear when the node enters/exits the screen. The Rect2 can be grown via code or with the Particles → Generate Visibility Rect editor tool.

pub fn is_emitting(&self) -> bool

If true, particles are being emitted.

pub fn restart(&self)

Restarts all the existing particles.

pub fn set_amount(&self, amount: i64)

The number of particles emitted in one emission cycle (corresponding to the lifetime). Note: Changing amount will reset the particle emission, therefore removing all particles that were already emitted before changing amount.

pub fn set_draw_order(&self, order: i64)

Particle draw order. Uses DrawOrder values.

pub fn set_emitting(&self, emitting: bool)

If true, particles are being emitted.

pub fn set_explosiveness_ratio(&self, ratio: f64)

How rapidly particles in an emission cycle are emitted. If greater than 0, there will be a gap in emissions before the next cycle begins.

pub fn set_fixed_fps(&self, fps: i64)

The particle system’s frame rate is fixed to a value. For instance, changing the value to 2 will make the particles render at 2 frames per second. Note this does not slow down the simulation of the particle system itself.

pub fn set_fractional_delta(&self, enable: bool)

If true, results in fractional delta calculation which has a smoother particles display effect.

pub fn set_lifetime(&self, secs: f64)

The amount of time each particle will exist (in seconds).

pub fn set_normal_map(&self, texture: impl AsArg<Texture>)

Normal map to be used for the texture property. Note: Godot expects the normal map to use X+, Y-, and Z+ coordinates. See this page for a comparison of normal map coordinates expected by popular engines.

pub fn set_one_shot(&self, secs: bool)

If true, only one emission cycle occurs. If set true during a cycle, emission will stop at the cycle’s end.

pub fn set_pre_process_time(&self, secs: f64)

Particle system starts as if it had already run for this many seconds.

pub fn set_process_material(&self, material: impl AsArg<Material>)

Material for processing particles. Can be a ParticlesMaterial or a ShaderMaterial.

pub fn set_randomness_ratio(&self, ratio: f64)

Emission lifetime randomness ratio.

pub fn set_speed_scale(&self, scale: f64)

Particle system’s running speed scaling ratio. A value of 0 can be used to pause the particles.

pub fn set_texture(&self, texture: impl AsArg<Texture>)

Particle texture. If null, particles will be squares.

pub fn set_use_local_coordinates(&self, enable: bool)

If true, particles use the parent node’s coordinate space. If false, they use global coordinates.

pub fn set_visibility_rect(&self, visibility_rect: Rect2)

The Rect2 that determines the node’s region which needs to be visible on screen for the particle system to be active. Grow the rect if particles suddenly appear/disappear when the node enters/exits the screen. The Rect2 can be grown via code or with the Particles → Generate Visibility Rect editor tool.