Struct gdnative::core_types::Variant

pub struct Variant(/* private fields */);

A Variant can represent all Godot values (core types or Object class instances).

The underlying data is either stored inline or reference-counted on the heap, depending on the size of the type and whether the it is trivially copyable.

If you compile godot-rust with the serde feature enabled, you will have access to serialization/deserialization support: the traits Serialize and Deserialize will be automatically implemented on VariantDispatch as well as most of the types in core_types.

Implementations

impl Variant

pub fn new<T>(from: T) -> Variant
where T: OwnedToVariant,

Creates a Variant from a value that implements ToVariant.

pub fn nil() -> Variant

Creates an empty Variant.

pub fn to<T>(&self) -> Option<T>
where T: FromVariant,

Performs a strongly-typed, structure-aware conversion to T from this variant, if it is a valid representation of said type. This is the same as T::from_variant(self).ok().

This is the same conversion used to parse arguments of exported methods. See FromVariant for more details.

pub fn try_to<T>(&self) -> Result<T, FromVariantError>
where T: FromVariant,

Performs a strongly-typed, structure-aware conversion to T from this variant, if it is a valid representation of said type. This is the same as T::from_variant(self).

This is the same conversion used to parse arguments of exported methods. See FromVariant for more details.

pub fn coerce_to<T>(&self) -> T
where T: CoerceFromVariant,

Coerce a value of type T out of this variant, through what Godot presents as a “best-effort” conversion, possibly returning a default value.

See CoerceFromVariant for more details.

See also Variant::to and Variant::try_to for strongly-typed, structure-aware conversions into Rust types.

pub fn to_object<T>(&self) -> Option<Ref<T>>
where T: GodotObject,

Convenience method to extract a Ref<T, Shared> from this variant, if the type matches. This is the same as Ref::<T, Shared>::from_variant(self).ok().

This is the same conversion used to parse arguments of exported methods. See FromVariant for more details.

pub fn try_to_object<T>(&self) -> Result<Ref<T>, FromVariantError>
where T: GodotObject,

Convenience method to extract a Ref<T, Shared> from this variant, if the type matches. This is the same as Ref::<T, Shared>::from_variant(self).

This is the same conversion used to parse arguments of exported methods. See FromVariant for more details.

pub fn get_type(&self) -> VariantType

Returns this variant’s type.

pub fn dispatch(&self) -> VariantDispatch

Converts this variant to a primitive value depending on its type.

Examples

let variant = 42.to_variant();
let number_as_float = match variant.dispatch() {
    VariantDispatch::I64(i) => i as f64,
    VariantDispatch::F64(f) => f,
    _ => panic!("not a number"),
};
approx::assert_relative_eq!(42.0, number_as_float);

pub fn is_nil(&self) -> bool

Returns true if this is an empty variant.

pub fn has_method(&self, method: impl Into<GodotString>) -> bool

pub unsafe fn call( &mut self, method: impl Into<GodotString>, args: &[Variant], ) -> Result<Variant, CallError>

Invokes a method on the held object.

Safety

This method may invoke [Object::call()] internally, which is unsafe, as it allows execution of arbitrary code (including user-defined code in GDScript or unsafe Rust).

pub fn evaluate( &self, op: VariantOperator, rhs: &Variant, ) -> Result<Variant, InvalidOp>

Evaluates a variant operator on self and rhs and returns the result on success.

Errors

Returns Err(InvalidOp) if the result is not valid.