Trait gdnative::core_types::FromVariant

pub trait FromVariant: Sized {
    // Required method
    fn from_variant(variant: &Variant) -> Result<Self, FromVariantError>;
}

Types that can be converted from a Variant. Conversions are performed in Rust, and can be implemented for custom types.

FromVariant generally avoids inexact conversions, favoring returning an error instead, with the noted exception of integer and float types that are not supported by Godot natively. This is stricter than what GDScript performs by default. For weakly-typed coercions of GDScript built-in typed, see CoerceFromVariant instead.

This trait is used for argument types of exported methods.

Option<T> and MaybeNot<T>

Option<T> requires the Variant to be T or Nil, in that order. For looser semantics, use MaybeNot<T>, which will catch all variant values that are not T as well.

Vec<T>

The FromVariant implementation for Vec<T> only allow homogeneous arrays. If you want to manually handle potentially heterogeneous values e.g. for error reporting, use VariantArray directly or compose with an appropriate wrapper: Vec<Option<T>> or Vec<MaybeNot<T>>.

Deriving FromVariant

The derive macro provides implementation consistent with derived ToVariant. See ToVariant for detailed documentation.

Required Methods

fn from_variant(variant: &Variant) -> Result<Self, FromVariantError>

Object Safety

This trait is not object safe.