Trait gdnative::core_types::PoolElement

pub trait PoolElement: Sealed { }

Trait for element types that can be contained in PoolArray. This trait is sealed and has no public interface.

Object Safety

This trait is not object safe.