Crate gdnative::derive

Macros

godot_wrap_method

Convenience macro to wrap an object’s method into a Method implementor that can be passed to the engine when registering a class.

Attribute Macros

methods

Collects method signatures of all functions in a NativeClass that have the #[method] attribute and registers them with Godot.

monomorphize

Wires up necessary internals for a concrete monomorphization of a generic NativeClass, represented as a type alias, so it can be registered.

profiled

Makes a function profiled in Godot’s built-in profiler. This macro automatically creates a tag using the name of the current module and the function by default.

Derive Macros

FromVarargs

Enable struct types to be parsed as argument lists.

FromVariant

NativeClass

Makes it possible to use a type as a NativeScript. Automatically registers the type if the inventory feature is enabled on supported platforms.

OwnedToVariant

ToVariant