Derive Macro gdnative::derive::NativeClass

#[derive(NativeClass)]
{
    // Attributes available to this derive:
    #[inherit]
    #[export]
    #[opt]
    #[user_data]
    #[property]
    #[register_with]
    #[no_constructor]
}

Makes it possible to use a type as a NativeScript.

Type attributes

The behavior of the derive macro can be customized using attributes on the type deriving NativeClass. All type attributes are optional.

#[inherit(gdnative::api::BaseClass)]

Sets gdnative::api::BaseClass as the base class for the script. This must be a type from the generated Godot API (that implements GodotObject). All owner arguments of exported methods must be references (TRef, Ref, or &) to this type.

Inheritance from other scripts, either in Rust or other languages, is not supported.

If no #[inherit(...)] is provided, gdnative::api::Reference is used as a base class. This behavior is consistent with GDScript: omitting the extends keyword will inherit Reference.

#[user_data(gdnative::user_data::SomeWrapper<Self>)]

Use the given type as the user-data wrapper. See the module-level docs on gdnative::user_data for more information.

#[register_with(path::to::function)]

Use a custom function to register signals, properties or methods, in addition to the one generated by #[methods]:

use gdnative::prelude::*;
use gdnative::nativescript::export::property::{RangeHint, FloatHint};

#[derive(NativeClass)]
#[inherit(Reference)]
#[register_with(Self::my_register_function)]
struct Foo;

#[methods]
impl Foo {
    fn new(_: &Reference) -> Self {
        Self {}
    }
    fn my_register_function(builder: &ClassBuilder<Foo>) {
        builder.add_signal(Signal { name: "foo", args: &[] });
        builder.add_property::<f32>("bar")
            .with_getter(|_, _| 42.0)
            .with_hint(FloatHint::Range(RangeHint::new(0.0, 100.0)))
            .done();
    }
}

#[no_constructor]

Indicates that this type has no zero-argument constructor. Instances of such scripts can only be created from Rust using Instance::emplace. Instance::new or ScriptName.new from GDScript will result in panics at runtime.

See documentation on Instance::emplace for an example on how this can be used.

Field attributes

All field attributes are optional.

#[property]

Convenience attribute to register a field as a property. Possible arguments for the attribute are:

path = "my_category/my_property_name"

Puts the property under the my_category category and renames it to my_property_name in the inspector and for GDScript.

default = 42.0

Sets the default value in the inspector for this property. The setter is not guaranteed to be called by the engine with the value.

before_get / after_get / before_set / after_set = "Self::hook_method"

Call hook methods with self and owner before and/or after the generated property accessors.

no_editor

Hides the property from the editor. Does not prevent it from being sent over network or saved in storage.

#[methods]

Adds the necessary information to a an impl block to register the properties and methods with Godot.

Important: This needs to be added to one and only one impl block for a given NativeClass.

For additional details about how #[methods] expands, please refer to gdnative::methods

#[export]

Registers the attributed function signature to be used by Godot. A valid function signature must have:

&self or &mut self as its first parameter

&T or TRef<T> where T refers to the type declared in #[inherit(T)] attribute as it’s second parameter; this is typically called the owner.

Optionally, any number of additional parameters, which must have the type Variant or must implement the FromVariant trait. FromVariant is implemented for most common types.

Return values must implement the OwnedToVariant trait (automatically implemented by ToVariant) or be a Variant type.

#[export]
fn foo(&self, owner: &Reference);

Note: Marking a function with #[export] does not have any effect unless inside an impl block that has the #[methods] attribute.

Possible arguments for this attribute are

name = “overridden_function_name”

Overrides the function name as the method name to be registered in Godot.

rpc = “selected_rpc”

“selected_rpc” must be one of the following values, which refer to the associated Multiplayer API RPC Mode

“disabled” - RPCMode::RPC_MODE_DISABLED

“remote” - RPCMode::RPC_MODE_REMOTE

“remote_sync” - RPCMode::RPC_MODE_REMOTE_SYMC

“master” - RPCMode::RPC_MODE_MASTER

“puppet” - RPCMode::RPC_MODE_PUPPET

“master_sync” - RPCMode::RPC_MODE_MASTERSYNC

“puppet_sync” - RPCMode::RPC_MODE_PUPPETSYNC

This enables you to set the Multiplayer API RPC Mode for the function.

Refer to Godot’s Remote Procedure documentation for more details.

Node virtual functions

This is a list of common Godot virtual functions that are automatically called via notifications.

fn _ready(&self, owner: &Node);

Called when both the node and its children have entered the scene tree.
See Godot docs for more information.

fn _enter_tree(&self, owner: &Node);

Called when the node enters the scene tree.
See Godot docs for more information.

fn _exit_tree(&self, owner: &Node);

Called when the node is removed from the scene tree.
See Godot docs for more information.

fn _get_configuration_warning(&self, owner: &Node) -> GodotString;

The string returned from this method is displayed as a warning in the Scene Dock if the script that overrides it is a tool script.
See Godot docs for more information.

fn _process(&mut self, owner: &Node, delta: f64);

Called during processing step of the main loop.
See Godot docs for more information.

fn _physics_process(&self, owner: &Node, delta: f64);

Called during physics update, with a fixed timestamp.
See Godot docs for more information.

fn _input(&self, owner: &Node, event: Ref<InputEvent>);

Called when there is an input event.
See Godot docs for more information.

fn _unhandled_input(&self, owner: &Node, event: Ref<InputEvent>);

Called when an InputEvent hasn’t been consumed by _input() or any GUI.
See Godot docs for more information.

fn _unhandled_key_input (&self, owner: &Node, event: Ref<InputKeyEvent>);

Called when an InputEventKey hasn’t been consumed by _input() or any GUI.
See Godot docs for more information.

Control virtual functions

This is a list of common Godot virtual functions that are automatically called via notifications.

fn _clips_input(&self, owner: &Control) -> bool;

Returns whether _gui_input() should not be called for children controls outside this control’s rectangle.
See Godot docs for more information.

fn _get_minimum_size(&self, owner: &Control) -> Vector2;

Returns the minimum size for this control.
See Godot docs for more information.

fn _gui_input(&self, owner: &Control, event: Ref<InputEvent>);

Use this method to process and accept inputs on UI elements.
See Godot docs for more information.

fn _make_custom_tooltip(&self, owner: &Control, for_text: String) -> Ref<Control>;

Returns a Control node that should be used as a tooltip instead of the default one.
See Godot docs for more information.