Module gdnative::init

Global initialization and termination of the library.

This module provides all the plumbing required for global initialization and shutdown of godot-rust.

Modules

diagnostics

Run-time tracing functions to help debug the init process.

Macros

godot_gdnative_initDeprecated

This macro now does nothing. It is provided only for limited backwards compatibility.

godot_gdnative_terminateDeprecated

This macro now does nothing. It is provided only for limited backwards compatibility.

godot_initDeprecated

Legacy macro that declares all the API endpoints necessary to initialize a NativeScript library.

godot_nativescript_initDeprecated

Legacy macro that declares all the API endpoints necessary to initialize a NativeScript library.

Structs

InitHandle

A handle that can register new classes to the engine during initialization.

InitializeInfo

Context for the godot_gdnative_init callback.

TerminateHandle

A handle passed from the engine during NativeScript termination. It’s purpose is currently unclear.

TerminateInfo

Context for the godot_gdnative_terminate callback.

Traits

GDNativeCallbacks

Trait for declaring library-level GDNative callbacks. See module-level docs for examples.

Attribute Macros

callbacks

Declares the library-level GDNative callbacks. See [gdnative::init::GDNativeCallbacks].