Module gdnative::init::diagnostics

Run-time tracing functions to help debug the init process.

The provided functions are designed to convey any issues found through human-readable error output, while programmatically providing only an overall indication of whether any problems were found. This is so that they can be freely improved without compatibility concerns.

Functions

godot_version_mismatch

Checks the version number of the host Godot instance to see if it matches the generated API. Returns true if the test isn’t applicable, or if no mismatch was found.

missing_manual_registration

Checks for any NativeClass types that are registered automatically, but not manually. Returns true if the test isn’t applicable, or if no such types were found.

missing_suggested_diagnostics

Checks if all suggested diagnostics have been ran depending on the current platform, at the point of invocation. This is automatically ran as part of the init macro, and do not usually need to be manually invoked.