Function gdnative::init::diagnostics::missing_suggested_diagnostics

pub fn missing_suggested_diagnostics() -> bool

Checks if all suggested diagnostics have been ran depending on the current platform, at the point of invocation. This is automatically ran as part of the init macro, and do not usually need to be manually invoked.

Returns true in a release build, or if no such diagnostics were found.