Module gdnative::log

Functions for using the engine’s logging system in the editor.

Macros

godot_dbg

Prints and returns the value of a given expression for quick and dirty debugging, using the engine’s logging system (visible in the editor).

godot_error

Print an error using the engine’s logging system (visible in the editor).

godot_print

Print a message using the engine’s logging system (visible in the editor).

godot_site

Creates a Site value from the current position in code, optionally with a function path for identification.

godot_warn

Print a warning using the engine’s logging system (visible in the editor).

Structs

Site

Value representing a call site for errors and warnings. Can be constructed using the godot_site macro, or manually.

Functions

error

Print an error to the Godot console.

print

Print a message to the Godot console.

warn

Print a warning to the Godot console.