Macro gdnative::godot_site

macro_rules! godot_site {
    () => { ... };
    ($($path:tt)+) => { ... };
}

Creates a Site value from the current position in code, optionally with a function path for identification.

Examples

use gdnative::log;

// WARN: <unset>: foo At: path/to/file.rs:123
log::warn(log::godot_site!(), "foo");

// WARN: Foo::my_func: bar At: path/to/file.rs:123
log::error(log::godot_site!(Foo::my_func), "bar");