Module gdnative::object::bounds

Various traits to verify memory policy, ownership policy or lifetime bounds

The symbols defined in this module are internal and used to enhance type safety. You typically will not need to work with them.

Structs

Forget

Simply releases the held object without deallocating it.

RefImplBound

Struct to be used for various Ref trait bounds.

UnRef

Decrements the reference count on the held object, deallocating it if it’s the last ref.

Traits

LifetimeConstraint

Trait for constraining assume_safe lifetimes to the one of &self when T is reference-counted. This is an internal interface.

MemorySpec

Specialization trait depending on Memory. This is an internal interface.

PtrWrapper

Specialization trait for Drop behavior.

SafeAsRaw

Trait for persistent Refs that point to valid objects. This is an internal interface.

SafeDeref

Trait for combinations of Memory and Ownership that can be dereferenced safely. This is an internal interface.