Trait gdnative::object::ownership::Ownership

pub trait Ownership: Sealed { }

Trait to parametrize over the ownership markers Unique, Shared and ThreadLocal.

This trait is sealed and has no public members.

It specifies the ownership policy of godot-rust smart pointers such as Ref. Ownership specifies how references own an object, i.e. how they point to it and who is responsible for its destruction (in case of RefCounted). Furthermore, it defines from where the object can be accessed, and if sharing the object across threads is possible.