Struct gdnative::object::RawObject

#[repr(C)]
pub struct RawObject<T> { /* fields omitted */ }

An opaque struct representing Godot objects. This should never be created on the stack.

This is an internal interface. Users are expected to use references to named generated types instead.

Implementations

impl<T> RawObject<T> where
T: GodotObject,

pub unsafe fn from_sys_ref_unchecked<'a>(
obj: NonNull<c_void>
) -> &'a RawObject<T>

Creates a typed reference from a pointer, without checking the type of the pointer.

Safety

The obj pointer must be pointing to a valid Godot object of type T during the entirety of 'a.

pub unsafe fn try_from_sys_ref<'a>(
obj: NonNull<c_void>
) -> Option<&'a RawObject<T>>

Creates a typed reference from a pointer if the pointer is pointing to an object of the correct type. Returns None otherwise.

Safety

The obj pointer must be pointing to a valid Godot object during the entirety of 'a.

pub fn sys(&self) -> NonNull<c_void>

Casts a reference to this opaque object to *const sys::godot_object.

pub fn is_class<U>(&self) -> bool where
U: GodotObject,

Checks whether the object is of a certain Godot class.

pub fn is_class_by_name(&self, class_name: &str) -> bool

Checks whether the object is of a certain Godot class by name.

pub fn class_name(&self) -> String

Returns the class name of this object dynamically using Object::get_class.

pub fn cast<U>(&self) -> Option<&RawObject<U>> where
U: GodotObject,

Attempt to cast a Godot object to a different class type.

pub unsafe fn cast_unchecked<U>(&self) -> &RawObject<U> where
U: GodotObject,

Attempt to cast a Godot object to a different class type without checking the type at runtime.

Safety

The types must be compatible.

pub unsafe fn free(&self)

Free the underlying object.

Safety

Further operations must not be performed on the same reference.

impl<T> RawObject<T> where
T: GodotObject<Memory = RefCounted>,

pub fn add_ref(&self)

Increase the reference count of the object.

pub unsafe fn unref(&self) -> bool

Decrease the reference count of the object. Returns true if this is the last reference.

Safety

Further operations must not be performed on the same reference if this is the last reference.

pub unsafe fn unref_and_free_if_last(&self) -> bool

Decrease the reference count of the object. Frees the object and returns true if this is the last reference.

Safety

Further operations must not be performed on the same reference if this is the last reference.

pub unsafe fn init_ref_count(&self)

Initialize the reference count of the object.

Safety

This function assumes that no other references are held at the time.