Trait gdnative::object::AsVariant

pub trait AsVariant: AsArg<Self::Target> {
    type Target;
}

Trait for safe conversion from Godot object references into Variant. This is a sealed trait with no public interface.

Used for Variant methods and implementations as a trait bound to improve type inference.

Required Associated Types

type Target