Trait gdnative::object::Receiver

pub trait Receiver<C>: Sized
where
    C: NativeClass,
{ }

Trait for types that can be used as the self or #[self] argument (receiver) for methods exported through the #[methods] attribute macro. This trait has no public interface, and is not intended to be implemented by users.

Notably, this is implemented for Instance and TInstance along with the usual self reference types. For types using ArcData as the wrapper specifically, Arc<C> is also allowed.

The trait is unsealed for technical coherence issues, but is not intended to be implemented by users. Changes to the definition of this trait are not considered breaking changes under semver.

Object Safety

This trait is not object safe.