Attribute Macro gdnative::prelude::methods

#[methods]

Collects method signatures of all functions in a NativeClass that have the #[export] attribute and registers them with Godot.

For example, in the following class

use gdnative::prelude::*;

#[derive(NativeClass)]
#[inherit(Reference)]
#[no_constructor]
struct Foo {}

#[methods]
impl Foo {
    #[export]
    fn foo(&self, _owner: &Reference, bar: i64) -> i64 {
        bar
    }
}

Will expand to

use gdnative::prelude::*;
struct Foo {}
impl NativeClass for Foo {
    type Base = gdnative::api::Reference;
    type UserData = gdnative::export::user_data::LocalCellData<Self>;
}
impl gdnative::export::StaticallyNamed for Foo {
    const CLASS_NAME: &'static str = "Foo";
}
impl gdnative::export::NativeClassMethods for Foo {
    fn register(builder: &ClassBuilder<Self>) {
        use gdnative::export::*;
        builder.method("foo", gdnative::export::godot_wrap_method!(Foo, false, fn foo(&self, #[base] _owner: &Reference, bar: i64) -> i64))
            .with_rpc_mode(RpcMode::Disabled)
            .done_stateless();
    }
}
impl Foo {
    fn foo(&self, _owner: &Reference, bar: i64) -> i64 {
        bar
    }
}

Important: Only one impl block per struct may be attributed with #[methods].

For more context, please refer to gdnative::derive::NativeClass.