Struct gdnative::prelude::Viewport

pub struct Viewport { /* private fields */ }

core class Viewport inherits Node (manually managed).

This class has related types in the viewport module.

Official documentation

See the documentation of this class in the Godot engine’s official documentation. The method descriptions are generated from it and typically contain code samples in GDScript, not Rust.

Memory management

Non-reference-counted objects, such as the ones of this type, are usually owned by the engine.

Viewport is a reference-only type. Persistent references can only exist in the unsafe Ref<Viewport> form.

In the cases where Rust code owns an object of this type, for example if the object was just created on the Rust side and not passed to the engine yet, ownership should be either given to the engine or the object must be manually destroyed using Ref::free, or Ref::queue_free if it is a Node.

Class hierarchy

Viewport inherits methods from:

Node

Object

Safety

All types in the Godot API have interior mutability in Rust parlance. To enforce that the official thread-safety guidelines are followed, the typestate pattern is used in the Ref and TRef smart pointers, and the Instance API. The typestate Ownership in these types tracks whether ownership is unique, shared, or exclusive to the current thread. For more information, see the type-level documentation on Ref.

Implementations

impl Viewport

Constants

pub const CLEAR_MODE_ALWAYS: i64 = 0i64

pub const DEBUG_DRAW_DISABLED: i64 = 0i64

pub const MSAA_DISABLED: i64 = 0i64

pub const RENDER_INFO_OBJECTS_IN_FRAME: i64 = 0i64

pub const SHADOW_ATLAS_QUADRANT_SUBDIV_DISABLED: i64 = 0i64

pub const UPDATE_DISABLED: i64 = 0i64

pub const USAGE_2D: i64 = 0i64

pub const CLEAR_MODE_NEVER: i64 = 1i64

pub const DEBUG_DRAW_UNSHADED: i64 = 1i64

pub const MSAA_2X: i64 = 1i64

pub const RENDER_INFO_VERTICES_IN_FRAME: i64 = 1i64

pub const SHADOW_ATLAS_QUADRANT_SUBDIV_1: i64 = 1i64

pub const UPDATE_ONCE: i64 = 1i64

pub const USAGE_2D_NO_SAMPLING: i64 = 1i64

pub const CLEAR_MODE_ONLY_NEXT_FRAME: i64 = 2i64

pub const DEBUG_DRAW_OVERDRAW: i64 = 2i64

pub const MSAA_4X: i64 = 2i64

pub const RENDER_INFO_MATERIAL_CHANGES_IN_FRAME: i64 = 2i64

pub const SHADOW_ATLAS_QUADRANT_SUBDIV_4: i64 = 2i64

pub const UPDATE_WHEN_VISIBLE: i64 = 2i64

pub const USAGE_3D: i64 = 2i64

pub const DEBUG_DRAW_WIREFRAME: i64 = 3i64

pub const MSAA_8X: i64 = 3i64

pub const RENDER_INFO_SHADER_CHANGES_IN_FRAME: i64 = 3i64

pub const SHADOW_ATLAS_QUADRANT_SUBDIV_16: i64 = 3i64

pub const UPDATE_ALWAYS: i64 = 3i64

pub const USAGE_3D_NO_EFFECTS: i64 = 3i64

pub const MSAA_16X: i64 = 4i64

pub const RENDER_INFO_SURFACE_CHANGES_IN_FRAME: i64 = 4i64

pub const SHADOW_ATLAS_QUADRANT_SUBDIV_64: i64 = 4i64

pub const RENDER_INFO_DRAW_CALLS_IN_FRAME: i64 = 5i64

pub const SHADOW_ATLAS_QUADRANT_SUBDIV_256: i64 = 5i64

pub const RENDER_INFO_2D_ITEMS_IN_FRAME: i64 = 6i64

pub const SHADOW_ATLAS_QUADRANT_SUBDIV_1024: i64 = 6i64

pub const RENDER_INFO_2D_DRAW_CALLS_IN_FRAME: i64 = 7i64

pub const SHADOW_ATLAS_QUADRANT_SUBDIV_MAX: i64 = 7i64

pub const RENDER_INFO_MAX: i64 = 8i64

impl Viewport

pub fn new() -> Ref<Viewport, Unique>

Creates a new instance of this object.

Because this type is not reference counted, the lifetime of the returned object is not automatically managed.

Immediately after creation, the object is owned by the caller, and can be passed to the engine (in which case the engine will be responsible for destroying the object) or destroyed manually using Ref::free, or preferably Ref::queue_free if it is a Node.

pub fn find_world(&self) -> Option<Ref<World, Shared>>

Returns the 3D world of the viewport, or if none the world of the parent viewport.

pub fn find_world_2d(&self) -> Option<Ref<World2D, Shared>>

Returns the 2D world of the viewport.

pub fn get_camera(&self) -> Option<Ref<Camera, Shared>>

Returns the active 3D camera.

pub fn canvas_transform(&self) -> Transform2D

The canvas transform of the viewport, useful for changing the on-screen positions of all child CanvasItems. This is relative to the global canvas transform of the viewport.

pub fn clear_mode(&self) -> ClearMode

The clear mode when viewport used as a render target. Note: This property is intended for 2D usage.

pub fn debug_draw(&self) -> DebugDraw

The overlay mode for test rendered geometry in debug purposes.

pub fn get_final_transform(&self) -> Transform2D

Returns the total transform of the viewport.

pub fn global_canvas_transform(&self) -> Transform2D

The global canvas transform of the viewport. The canvas transform is relative to this.

pub fn hdr(&self) -> bool

If true, the viewport rendering will receive benefits from High Dynamic Range algorithm. High Dynamic Range allows the viewport to receive values that are outside the 0-1 range. In Godot HDR uses 16 bits, meaning it does not store the full range of a floating point number. Note: Requires usage to be set to USAGE_3D or USAGE_3D_NO_EFFECTS, since HDR is not supported for 2D.

pub fn keep_3d_linear(&self) -> bool

If true, the result after 3D rendering will not have a linear to sRGB color conversion applied. This is important when the viewport is used as a render target where the result is used as a texture on a 3D object rendered in another viewport. It is also important if the viewport is used to create data that is not color based (noise, heightmaps, pickmaps, etc.). Do not enable this when the viewport is used as a texture on a 2D object or if the viewport is your final output. For the GLES2 driver this will convert the sRGB output to linear, this should only be used for VR plugins that require input in linear color space!

pub fn get_modal_stack_top(&self) -> Option<Ref<Control, Shared>>

Returns the topmost modal in the stack.

pub fn get_mouse_position(&self) -> Vector2

Returns the mouse position relative to the viewport.

pub fn msaa(&self) -> Msaa

The multisample anti-aliasing mode. A higher number results in smoother edges at the cost of significantly worse performance. A value of 4 is best unless targeting very high-end systems.

pub fn physics_object_picking(&self) -> bool

If true, the objects rendered by viewport become subjects of mouse picking process.

pub fn get_render_info(&self, info: i64) -> i64

Returns information about the viewport from the rendering pipeline.

pub fn shadow_atlas_quadrant_subdiv(
&self,
quadrant: i64
) -> ShadowAtlasQuadrantSubdiv

The subdivision amount of the fourth quadrant on the shadow atlas.

pub fn shadow_atlas_size(&self) -> i64

The shadow atlas’ resolution (used for omni and spot lights). The value will be rounded up to the nearest power of 2. Note: If this is set to 0, shadows won’t be visible. Since user-created viewports default to a value of 0, this value must be set above 0 manually.

pub fn sharpen_intensity(&self) -> f64

If set to a value greater than 0.0, contrast-adaptive sharpening will be applied to the 3D viewport. This has a low performance cost and can be used to recover some of the sharpness lost from using FXAA. Values around 0.5 generally give the best results. See also [fxaa][Self::fxaa].

pub fn size(&self) -> Vector2

The width and height of viewport. Must be set to a value greater than or equal to 2 pixels on both dimensions. Otherwise, nothing will be displayed.

pub fn get_size_override(&self) -> Vector2

Returns the size override set with set_size_override.

pub fn get_texture(&self) -> Option<Ref<ViewportTexture, Shared>>

Sample code is GDScript unless otherwise noted.

Returns the viewport’s texture. Note: Due to the way OpenGL works, the resulting ViewportTexture is flipped vertically. You can use Image.flip_y on the result of Texture.get_data to flip it back, for example:

var img = get_viewport().get_texture().get_data()
img.flip_y()

pub fn update_mode(&self) -> UpdateMode

The update mode when viewport used as a render target.

pub fn usage(&self) -> Usage

The rendering mode of viewport.

pub fn use_debanding(&self) -> bool

If true, uses a fast post-processing filter to make banding significantly less visible. In some cases, debanding may introduce a slightly noticeable dithering pattern. It’s recommended to enable debanding only when actually needed since the dithering pattern will make lossless-compressed screenshots larger. Note: Only available on the GLES3 backend. hdr must also be true for debanding to be effective.

pub fn use_fxaa(&self) -> bool

Enables fast approximate antialiasing. FXAA is a popular screen-space antialiasing method, which is fast but will make the image look blurry, especially at lower resolutions. It can still work relatively well at large resolutions such as 1440p and 4K. Some of the lost sharpness can be recovered by enabling contrast-adaptive sharpening (see sharpen_intensity).

pub fn vflip(&self) -> bool

If true, the result of rendering will be flipped vertically.

pub fn get_viewport_rid(&self) -> Rid

Returns the viewport’s RID from the VisualServer.

pub fn get_visible_rect(&self) -> Rect2

Returns the visible rectangle in global screen coordinates.

pub fn world(&self) -> Option<Ref<World, Shared>>

The custom World which can be used as 3D environment source.

pub fn world_2d(&self) -> Option<Ref<World2D, Shared>>

The custom World2D which can be used as 2D environment source.

pub fn gui_get_drag_data(&self) -> Variant

Returns the drag data from the GUI, that was previously returned by [Control.get_drag_data][Control::get_drag_data].

pub fn gui_has_modal_stack(&self) -> bool

Returns true if there are visible modals on-screen.

pub fn gui_is_dragging(&self) -> bool

Returns true if the viewport is currently performing a drag operation.

pub fn has_transparent_background(&self) -> bool

If true, the viewport should render its background as transparent.

pub fn input(&self, local_event: impl AsArg<InputEvent>)

pub fn is_3d_disabled(&self) -> bool

If true, the viewport will disable 3D rendering. For actual disabling use usage.

pub fn is_audio_listener(&self) -> bool

If true, the viewport will process 3D audio streams.

pub fn is_audio_listener_2d(&self) -> bool

If true, the viewport will process 2D audio streams.

pub fn is_handling_input_locally(&self) -> bool

pub fn is_input_disabled(&self) -> bool

If true, the viewport will not receive input events.

pub fn is_input_handled(&self) -> bool

pub fn is_size_override_enabled(&self) -> bool

Returns true if the size override is enabled. See set_size_override.

pub fn is_size_override_stretch_enabled(&self) -> bool

If true, the size override affects stretch as well.

pub fn is_snap_controls_to_pixels_enabled(&self) -> bool

If true, the GUI controls on the viewport will lay pixel perfectly.

pub fn is_using_own_world(&self) -> bool

If true, the viewport will use World defined in world property.

pub fn is_using_render_direct_to_screen(&self) -> bool

If true, renders the Viewport directly to the screen instead of to the root viewport. Only available in GLES2. This is a low-level optimization and should not be used in most cases. If used, reading from the Viewport or from SCREEN_TEXTURE becomes unavailable. For more information see VisualServer.viewport_set_render_direct_to_screen.

pub fn set_as_audio_listener(&self, enable: bool)

If true, the viewport will process 3D audio streams.

pub fn set_as_audio_listener_2d(&self, enable: bool)

If true, the viewport will process 2D audio streams.

pub fn set_attach_to_screen_rect(&self, rect: Rect2)

Attaches this Viewport to the root Viewport with the specified rectangle. This bypasses the need for another node to display this Viewport but makes you responsible for updating the position of this Viewport manually.

pub fn set_canvas_transform(&self, xform: Transform2D)

The canvas transform of the viewport, useful for changing the on-screen positions of all child CanvasItems. This is relative to the global canvas transform of the viewport.

pub fn set_clear_mode(&self, mode: i64)

The clear mode when viewport used as a render target. Note: This property is intended for 2D usage.

pub fn set_debug_draw(&self, debug_draw: i64)

The overlay mode for test rendered geometry in debug purposes.

pub fn set_disable_3d(&self, disable: bool)

If true, the viewport will disable 3D rendering. For actual disabling use usage.

pub fn set_disable_input(&self, disable: bool)

If true, the viewport will not receive input events.

pub fn set_global_canvas_transform(&self, xform: Transform2D)

The global canvas transform of the viewport. The canvas transform is relative to this.

pub fn set_handle_input_locally(&self, enable: bool)

pub fn set_hdr(&self, enable: bool)

If true, the viewport rendering will receive benefits from High Dynamic Range algorithm. High Dynamic Range allows the viewport to receive values that are outside the 0-1 range. In Godot HDR uses 16 bits, meaning it does not store the full range of a floating point number. Note: Requires usage to be set to USAGE_3D or USAGE_3D_NO_EFFECTS, since HDR is not supported for 2D.

pub fn set_input_as_handled(&self)

Stops the input from propagating further down the SceneTree.

pub fn set_keep_3d_linear(&self, keep_3d_linear: bool)

If true, the result after 3D rendering will not have a linear to sRGB color conversion applied. This is important when the viewport is used as a render target where the result is used as a texture on a 3D object rendered in another viewport. It is also important if the viewport is used to create data that is not color based (noise, heightmaps, pickmaps, etc.). Do not enable this when the viewport is used as a texture on a 2D object or if the viewport is your final output. For the GLES2 driver this will convert the sRGB output to linear, this should only be used for VR plugins that require input in linear color space!

pub fn set_msaa(&self, msaa: i64)

The multisample anti-aliasing mode. A higher number results in smoother edges at the cost of significantly worse performance. A value of 4 is best unless targeting very high-end systems.

pub fn set_physics_object_picking(&self, enable: bool)

If true, the objects rendered by viewport become subjects of mouse picking process.

pub fn set_shadow_atlas_quadrant_subdiv(&self, quadrant: i64, subdiv: i64)

The subdivision amount of the fourth quadrant on the shadow atlas.

pub fn set_shadow_atlas_size(&self, size: i64)

The shadow atlas’ resolution (used for omni and spot lights). The value will be rounded up to the nearest power of 2. Note: If this is set to 0, shadows won’t be visible. Since user-created viewports default to a value of 0, this value must be set above 0 manually.

pub fn set_sharpen_intensity(&self, intensity: f64)

If set to a value greater than 0.0, contrast-adaptive sharpening will be applied to the 3D viewport. This has a low performance cost and can be used to recover some of the sharpness lost from using FXAA. Values around 0.5 generally give the best results. See also [fxaa][Self::fxaa].

pub fn set_size(&self, size: Vector2)

The width and height of viewport. Must be set to a value greater than or equal to 2 pixels on both dimensions. Otherwise, nothing will be displayed.

pub fn set_size_override(&self, enable: bool, size: Vector2, margin: Vector2)

Sets the size override of the viewport. If the enable parameter is true the override is used, otherwise it uses the default size. If the size parameter is (-1, -1), it won’t update the size.

Default Arguments

size - Vector2( -1, -1 )

margin - Vector2( 0, 0 )

pub fn set_size_override_stretch(&self, enabled: bool)

If true, the size override affects stretch as well.

pub fn set_snap_controls_to_pixels(&self, enabled: bool)

If true, the GUI controls on the viewport will lay pixel perfectly.

pub fn set_transparent_background(&self, enable: bool)

If true, the viewport should render its background as transparent.

pub fn set_update_mode(&self, mode: i64)

The update mode when viewport used as a render target.

pub fn set_usage(&self, usage: i64)

The rendering mode of viewport.

pub fn set_use_arvr(&self, use_: bool)

If true, the viewport will be used in AR/VR process.

pub fn set_use_debanding(&self, enable: bool)

If true, uses a fast post-processing filter to make banding significantly less visible. In some cases, debanding may introduce a slightly noticeable dithering pattern. It’s recommended to enable debanding only when actually needed since the dithering pattern will make lossless-compressed screenshots larger. Note: Only available on the GLES3 backend. hdr must also be true for debanding to be effective.

pub fn set_use_fxaa(&self, enable: bool)

Enables fast approximate antialiasing. FXAA is a popular screen-space antialiasing method, which is fast but will make the image look blurry, especially at lower resolutions. It can still work relatively well at large resolutions such as 1440p and 4K. Some of the lost sharpness can be recovered by enabling contrast-adaptive sharpening (see sharpen_intensity).

pub fn set_use_own_world(&self, enable: bool)

If true, the viewport will use World defined in world property.

pub fn set_use_render_direct_to_screen(&self, enable: bool)

If true, renders the Viewport directly to the screen instead of to the root viewport. Only available in GLES2. This is a low-level optimization and should not be used in most cases. If used, reading from the Viewport or from SCREEN_TEXTURE becomes unavailable. For more information see VisualServer.viewport_set_render_direct_to_screen.

pub fn set_vflip(&self, enable: bool)

If true, the result of rendering will be flipped vertically.

pub fn set_world(&self, world: impl AsArg<World>)

The custom World which can be used as 3D environment source.

pub fn set_world_2d(&self, world_2d: impl AsArg<World2D>)

The custom World2D which can be used as 2D environment source.

pub fn unhandled_input(&self, local_event: impl AsArg<InputEvent>)

pub fn update_worlds(&self)

Forces update of the 2D and 3D worlds.

pub fn use_arvr(&self) -> bool

If true, the viewport will be used in AR/VR process.

pub fn warp_mouse(&self, to_position: Vector2)

Warps the mouse to a position relative to the viewport.