Module gdnative::prelude::user_data

User-data attributes from export::user_data module.

Structs

Aether

Special user-data wrapper intended for zero-sized types, that does not perform any allocation or synchronization at runtime. Does not implement MapMut.

ArcData

User-data wrapper encapsulating a Arc<T>. Does not implement MapMut.

LocalCellData

User-data wrapper analogous to a Arc<RefCell<T>>, that is restricted to the thread where it was originally created. The destructor of T is not guaranteed to be run if this is actually shared across multiple threads.

MutexData

User-data wrapper encapsulating a Arc<Mutex<T>>.

RwLockData

User-data wrapper encapsulating a Arc<RwLock<T>>.